### 2024-08-13 #7100 @editorialbot
[PRE REVIEW]: kifmm-rs: A Kernel Independent Fast Multipole Framework in Rust
**Submitting author:** @skailasa (<a href="http://orcid.org/0000-0001-9734-8318">Srinath Kailasa</a>)
**Repository:** https://github.com/bempp/kifmm/
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
