### 2024-08-13 #7100 @danielskatz
👋 @skailasa - thanks for your submission.

I have a few comments initially:
1. The paper is about 2000 words, while JOSS papers are typically closer to 1000 maximum. You should probably move some of the papers content to the repo/documentation, and point at it there.
2. I see 4 contributors to the repo, but you are the only author. Can you say why this is?
3. In the README, you will need to add some information on contributing to the software; this is in the JOSS review criteria
4. I don't know much about Rust packaging, so I am unsure if you are following best practices in this regard

While you consider these comments, and work on shortening the paper, I'll try to find an editor
