### 2024-08-14 #7100 @skailasa
Hi Danial

> 👋 @skailasa - thanks for your submission.
>
> I have a few comments initially:
>
> 1. The paper is about 2000 words, while JOSS papers are typically closer to 1000 maximum. You should probably move some of the papers content to the repo/documentation, and point at it there.

Sure I can do this, thank you

> 2. I see 4 contributors to the repo, but you are the only author. Can you say why this is?

This code has been extracted out of a larger mono-repo in order to expose functionality in a simpler way to related projects, the contributors to the mono repo are a part of the commit history which we did not want to delete, whereas I am the sole contributor to the code as presented in this repo, as well as the code present in the mono-repo which was extracted into this one.

> 3. In the README, you will need to add some information on contributing to the software; this is in the JOSS review criteria

Sure I will update this

> 4. I don't know much about Rust packaging, so I am unsure if you are following best practices in this regard
>

Ok no worries, Rust projects can be built from source by pointing to a github link, and is the easiest way to install this software. I can look into publishing the release on crates.io which is a repository of built projects, however this is likely to be done at a later stage as part of a software release of other tools also being developed in our group

> While you consider these comments, and work on shortening the paper, I'll try to find an editor
