+++ paper/paper.md
@@ -13,19 +13,18 @@ affiliations:
    index: 1
 date: 5 April 2024
 bibliography: paper.bib
-
 ---
 
 # Summary
 
-We present `kifmm-rs` a Rust based implementation of the kernel independent Fast Multipole Method (kiFMM), with C bindings, that serves as a implementation framework for kiFMMs [@Ying2004; @Greengard1987]. The FMM is a key algorithm for scientific computing, commonly cited as one of the top algorithmic advances of the twentieth century [@cipra2000best] due to its acceleration of the computation of $N$-body potential evaluation problems of the form,
+We present `kifmm-rs` a Rust based implementation of the kernel independent Fast Multipole Method (kiFMM) with C bindings, that serves as a implementation framework for kiFMMs [@Ying2004; @Greengard1987]. The FMM is a key algorithm for scientific computing due to its acceleration of the computation of $N$-body potential evaluation problems of the form,
 
 \begin{equation}
     \phi(x_i) = \sum_{j=1}^N K(x_i, y_j) q(y_j)
     \label{eq:sec:summary:potential}
 \end{equation}
 
-from $O(N^2)$ to $O(N)$ or $O(N \log(N))$, where the potential $\phi$ is evaluated at a set of target points, $\{x_i\}_{i=1}^M$, due to a set of densities, $\{ q_j \}_{j=1}^N$ and $K(.,.)$ is the interaction kernel. Compatible kernels commonly arise from second-order elliptic partial differential equations, such as the Laplace kernel which models the electrostatic or gravitational potentials corresponding to a set of source points on a set of target points,
+from $O(N^2)$ to $O(N)$ or $O(N \log(N))$, where the potential $\phi$ is evaluated at a set of target points, $\{x_i\}_{i=1}^M$, due to a set of densities, $\{ q_j \}_{j=1}^N$ and $K(.,.)$ is the interaction kernel. Compatible kernels commonly arise in science and engineering, such as the Laplace kernel which models the electrostatic or gravitational potentials corresponding to a set of source points on a set of target points,
 
 \begin{equation}
     K(x, y) = \begin{cases}
@@ -35,128 +34,41 @@ from $O(N^2)$ to $O(N)$ or $O(N \log(N))$, where the potential $\phi$ is evaluat
     \label{eq:sec:summary:laplace_kernel}
 \end{equation}
 
-The FMM also finds a major application in the acceleration of Boundary Element Methods (BEM) for elliptic boundary value problems [@steinbach2007numerical], which can be used to model a wide range of natural phenomena. Kernel independent variants of the FMM (kiFMMs) replace the analytical series approximations used to compress far-field interactions between clusters of source and target points, with approximation schemes that are based on kernel evaluations and extensions of the algorithm can also handle oscillatory problems specified by the Helmholtz kernel [@Ying2004; @engquist2010fast], with a common underlying algorithmic and software machinery that can be optimised in a kernel-independent manner.
-
-The FMM splits (\ref{eq:sec:summary:potential}) for a given target cluster into \textit{near} and \textit{far} components, the latter of which are taken to be _admissable_, i.e. amenable to approximation via an expansion scheme or alternative interpolation method such as [@Ying2004; @Fong2009],
+FMMs split (\ref{eq:sec:summary:potential}) for a given target cluster into \textit{near} and \textit{far} components, the latter of which are taken to be amenable to approximation,
 
 \begin{equation}
     \phi(x_i) = \sum_{y_j \in \text{Near}(x_i)} K(x_i, y_j) q_j + \sum_{y_j \in \text{Far}(x_i)} K(x_i, y_j) q_j
     \label{eq:sec:summary:near_far_split}
 \end{equation}
 
[…]
