+++ README.md
@@ -2,10 +2,19 @@
 
 ## Install
 
-Currently only supported on Unix.
+We currently only support Unix distributions.
+
+The current head can be built from source by adding the following to your `Cargo.toml` file.
+
 
 ```toml
 kifmm = { git = "https://github.com/bempp/kifmm" }
+```
+
+The current release version can be installed using the deployment on `crates.io`, and can be added to your `Cargo.toml` file with
+
+```bash
+cargo add kifmm
 ```
 
 ## Documentation
@@ -23,10 +32,27 @@ The functionality of the library can be tested by running:
 cargo test
 ```
 
-## Getting help
+## Contributing
+
+We welcome contributions to this project, and are grateful for your interest. Contributions can take many forms, including bug reports, feature requests code contributions and documentation.
+
+### Reporting Bugs and Issues
+
 Errors in the library should be added to the [GitHub issue tracker](https://github.com/bempp/kifmm/issues).
 
+### Feature Requests
+
+We're always open to suggestions for new features, please check the [GitHub issue tracker](https://github.com/bempp/kifmm/issues) to see if your concern has been discussed. If not, please open a new issue with the 'enhancement' label.
+
+
+### Submitting Code or Documentation Changes
+
+Your pull request should include a description of changes, any relevant issue numbers, and tests to cover the proposed changes.
+
+### Questions and Discussion
+
 Questions about the library and its use can be asked on the [Bempp Discourse](https://bempp.discourse.group).
+
 
 ## Licence
 KiFMM is licensed under a BSD 3-Clause licence. Full text of the licence can be found [here](LICENSE.md).

[…]
