### 2024-08-14 c924b50
CLN: Update README.md (#94)
* Update README.md

* Remove ds store
