+++ kifmm/Cargo.toml
@@ -26,35 +26,19 @@ num = "0.4"
 approx = "0.5"
 rayon = "1.9"
 rand = "0.8.5"
-<<<<<<< HEAD
-kifmm-fftw-sys = { version = "0.1" }
-green-kernels = { version = "0.1" }
-=======
-kifmm-fftw-sys = { path = "../fftw-sys" }
+kifmm-fftw-sys = { version = "0.2.0"}
 green-kernels = { version = "0.2.0"}
->>>>>>> main
 lazy_static = "1.4.0"
 num-complex = "0.4.5"
 mpi = { version = "0.8.*", optional = true}
 memoffset = {version = "0.9.1", optional = true}
 superslice = {version = "1.0.0", optional = true}
-<<<<<<< HEAD
-pyo3 = { version =  "0.21.2", features = ["extension-module", "multiple-pymethods"] , optional = true }
-numpy = { version = "0.21", optional = true }
-
-[target.aarch64-apple-darwin.dependencies]
-rlst = { version = "0.1", features=["metal"]}
-
-[target.x86_64-unknown-linux-gnu.dependencies]
-rlst = { version = "0.1", features=["blis"]}
-=======
 pulp = "0.18.12"
 bytemuck = "1.16.0"
 rand_distr = "0.4.3"
 bitflags = "2.6.0"
 rlst = {version = "0.2.0"}
 coe-rs = "0.1.2"
->>>>>>> main
 
 [dev-dependencies]
 criterion = { version = "0.5.*", features = ["html_reports"]}
