### 2024-08-14 #7100 @danielskatz
The submission is connected to the repo, so yes on that. If you've made paper changes, you can rebuild the paper, check the length, and check the references with editorialbot commands.  Each command should be the first thing in a new comment. Use `@editorialbot commands` to see what the commands are
