### 2024-08-15 2f9e2a9
Start working on directional compression scheme
