+++ kifmm/src/tree/morton.rs
@@ -538,6 +538,7 @@ where
         result
     }
 
+
     /// The anchor corresponding to this key.
     pub fn anchor(&self) -> &[u64; 3] {
         &self.anchor
