### 2024-08-16 bb06988
Add deny(unused_crate_dependencies) to "strict" feature, and remove unused dependencies
