+++ kifmm/Cargo.toml
@@ -31,9 +31,9 @@ lazy_static = "1.4.0"
 num-complex = "0.4.5"
 mpi = { version = "0.8.*", optional = true}
 memoffset = {version = "0.9.1", optional = true}
-superslice = {version = "1.0.0", optional = true}
 pulp = "0.18.12"
 bytemuck = "1.16.0"
+superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
 rlst = {version = "0.2.0"}
 
