+++ kifmm/src/fmm/field_translation/source_to_target/fft.rs
@@ -22,7 +22,7 @@ use crate::{
         fftw::Dft,
         fmm::{FmmOperatorData, HomogenousKernel, SourceToTargetTranslation},
         general::{AsComplex, Hadamard8x8},
-        tree::{FmmTree, Tree},
+        tree::{FmmTree, SingleNodeTree},
         types::FmmError,
     },
     tree::{
