+++ kifmm/src/tree/morton.rs
@@ -398,29 +398,30 @@ where
     T: RlstScalar + Float,
 {
     /// Constructor for Morton key
-    pub fn new(anchor: &[u64; 3], morton: u64) -> Self {
+    pub fn new(anchor: &[u64; 3], morton: u64, rank: i32) -> Self {
         Self {
             anchor: *anchor,
             morton,
+            rank,
             scalar: PhantomData::<T>,
         }
     }
 
     /// The Morton key corresponding to an octree root node
-    pub fn root() -> Self {
-        Self::new(&[0, 0, 0], 0)
+    pub fn root(rank: i32) -> Self {
+        Self::new(&[0, 0, 0], 0, rank)
     }
 
     /// Construct a `MortonKey` type from a Morton index
-    pub fn from_morton(morton: u64) -> Self {
+    pub fn from_morton(morton: u64, rank: i32) -> Self {
         let anchor = decode_key(morton);
-        Self::new(&anchor, morton)
+        Self::new(&anchor, morton, rank)
     }
 
     /// Construct a `MortonKey` type from the anchor at a given level
-    pub fn from_anchor(anchor: &[u64; 3], level: u64) -> Self {
+    pub fn from_anchor(anchor: &[u64; 3], level: u64, rank: i32) -> Self {
         let morton = encode_anchor(anchor, level);
-        Self::new(anchor, morton)
+        Self::new(anchor, morton, rank)
     }
 
     /// Construct a `MortonKey` associated with the box that encloses the point on the deepest level.
@@ -429,9 +430,9 @@ where
     /// * `point` - Cartesian coordinate for a given point.
     /// * `domain` - Domain associated with a given tree encoding.
     /// * `level` - level of octree on which to find the encoding.
-    pub fn from_point(point: &[T; 3], domain: &Domain<T>, level: u64) -> Self {
+    pub fn from_point(point: &[T; 3], domain: &Domain<T>, level: u64, rank: i32) -> Self {
         let anchor = point_to_anchor(point, level, domain).unwrap();
-        MortonKey::from_anchor(&anchor, level)
+        MortonKey::from_anchor(&anchor, level, rank)
     }
 
     /// Find the transfer vector between two Morton keys in component form.
@@ -566,17 +567,17 @@ where
 
         let parent_morton = (parent_morton_without_level << LEVEL_DISPLACEMENT) | parent_level;
 
-        MortonKey::from_morton(parent_morton)
+        MortonKey::from_morton(parent_morton, self.rank)
     }
 
     /// Return the first child of a Morton Key.
     pub fn first_child(&self) -> Self {
-        Self::new(&self.anchor, 1 + self.morton)
+        Self::new(&self.anchor, 1 + self.morton, self.rank)
     }
 
     /// Return the first child of a Morton Key on the deepest level.
     pub fn finest_first_child(&self) -> Self {
-        Self::new(&self.anchor, DEEPEST_LEVEL - self.level() + self.morton)
+        Self::new(&self.anchor, DEEPEST_LEVEL - self.level() + self.morton, self.rank)
     }
 
     /// Return the last child of a Morton Key on the deepest level.
@@ -614,7 +615,7 @@ where
         }
 
         for &child_morton in children_morton.iter() {
-            children.push(MortonKey::from_morton(child_morton))
+            children.push(MortonKey::from_morton(child_morton, self.rank))
         }
 
         children.sort();
@@ -646,6 +647,22 @@ where
         ancestors.insert(current);
 
         while current.level() > 0 {
+            current = current.parent();
+            ancestors.insert(current);
[…]
