### 2024-08-18 917967b
Prototype of new multinode trees
