+++ kifmm/src/fmm/builder.rs
@@ -116,8 +116,10 @@ where
 
             let depth = source_depth.max(target_depth); // refine source and target trees to same depth
 
-            let source_tree = SingleNodeTree::new(sources, depth, prune_empty, self.domain)?;
-            let target_tree = SingleNodeTree::new(targets, depth, prune_empty, self.domain)?;
+            let source_tree =
+                SingleNodeTree::new(sources, depth, prune_empty, self.domain, None, None, None)?;
+            let target_tree =
+                SingleNodeTree::new(targets, depth, prune_empty, self.domain, None, None, None)?;
 
             let fmm_tree = SingleNodeFmmTree {
                 source_tree,
