### 2024-08-19 697d4c3
Remove unused dependencies (#98)
* Add deny(unused_crate_dependencies) to "strict" feature, and remove unused dependencies

* unused deps

* nope, those deps were needed

* Fix #99
