+++ kifmm/src/fmm/field_translation/source_to_target/hadamard.rs
@@ -1,5 +1,6 @@
 //! Implementations of 8x8 operation during Hadamard product in FFT based M2L operations.
 use crate::{fmm::types::Isa, traits::general::Hadamard8x8};
+use bytemuck as _;
 use rlst::{c32, c64, RlstScalar};
 
 /// The 8x8 Hadamard operation, always inlined.
