+++ kifmm/Cargo.toml
@@ -23,7 +23,6 @@ crate-type = ["cdylib", "lib"]
 [dependencies]
 itertools = "0.13.*"
 num = "0.4"
-approx = "0.5"
 rayon = "1.9"
 rand = "0.8.5"
 kifmm-fftw-sys = { path = "../fftw-sys" }
@@ -32,18 +31,15 @@ lazy_static = "1.4.0"
 num-complex = "0.4.5"
 mpi = { version = "0.8.*", optional = true}
 memoffset = {version = "0.9.1", optional = true}
-superslice = {version = "1.0.0", optional = true}
 pulp = "0.18.12"
 bytemuck = "1.16.0"
+superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
-bitflags = "2.6.0"
 rlst = {version = "0.2.0"}
-coe-rs = "0.1.2"
 
 [dev-dependencies]
+approx = "0.5"
 criterion = { version = "0.5.*", features = ["html_reports"]}
-csv = "1.3.0"
-num_cpus = "1.16.0"
 
 [build-dependencies]
 cbindgen = "0.26.0"
