### 2024-08-19 75dc28c
Layout distributed algorithm
