### 2024-08-20 #7100 @diehlpk
> > Hi @rfj82982 do you have time to review this paper?
>
> Hi @diehlpk I can review the paper, but only after the end of September. If the review can wait for beginning of October I'll be happy to do it.

Yes, that is fine.
