### 2024-08-27 2e1162f
Add most of multinode builder
