+++ kifmm/examples/mpi_test_multi_node_fmm.rs
@@ -1,0 +1,59 @@
+//? mpirun -n {{NPROCESSES}} --features "mpi"
+
+#[cfg(feature = "mpi")]
+fn main() {
+    // Setup an MPI environment
+
+    use green_kernels::laplace_3d::Laplace3dKernel;
+    use kifmm::{
+        fmm::types::MultiNodeBuilder,
+        tree::{helpers::points_fixture, types::MultiNodeTreeNew},
+        FftFieldTranslation,
+    };
+    use mpi::{environment::Universe, traits::Communicator};
+    use rlst::RawAccess;
+
+    let universe: Universe = mpi::initialize().unwrap();
+    let world = universe.world();
+    let comm = world.duplicate();
+
+    // Setup tree parameters
+    let prune_empty = false;
+    let depth = 5;
+    let n_points = 10000;
+    let local_depth = 3;
+    let global_depth = 1;
+
+    let expansion_order = 6;
+
+    // Generate some random test data local to each process
+    let points = points_fixture::<f32>(n_points, None, None, None);
+
+    let charges = vec![1f32; n_points];
+
+    // Create a uniform tree
+    let fmm = MultiNodeBuilder::new()
+        .tree(
+            points.data(),
+            points.data(),
+            local_depth,
+            global_depth,
+            prune_empty,
+            &world,
+        )
+        .unwrap()
+        .parameters(
+            expansion_order,
+            Laplace3dKernel::<f32>::new(),
+            FftFieldTranslation::<f32>::new(None),
+        )
+        .unwrap()
+        .build()
+        .unwrap();
+
+    // Charge setting has to be after construction in multi-node
+    // fmm.set_charges();
+
+    // Evaluate distributed FMM
+    // fmm.evaluate();
+}
