+++ kifmm/src/fmm/multi_node.rs
@@ -1,12 +1,19 @@
 //! Multi Node FMM
 //! Single Node FMM
-use std::time::Instant;
+use std::{collections::HashMap, time::Instant};
 
 use green_kernels::traits::Kernel as KernelTrait;
 
-use mpi::traits::{Communicator, Equivalence};
+use itertools::Itertools;
+use mpi::{
+    datatype::PartitionMut,
+    topology::SimpleCommunicator,
+    traits::{Communicator, Equivalence},
+};
 use num::Float;
 use rlst::RlstScalar;
+
+use mpi::collective::CommunicatorCollectives;
 
 use crate::{
     fmm::types::{FmmEvalType, KiFmm},
@@ -21,6 +28,7 @@ use crate::{
         },
         types::{FmmError, FmmOperatorTime, FmmOperatorType},
     },
+    tree::types::MortonKey,
     Fmm, MultiNodeFmmTree,
 };
 
@@ -29,20 +37,18 @@ use super::{
     types::KiFmmMultiNode,
 };
 
-impl<Scalar, Kernel, SourceToTargetData, C> MultiNodeFmm
-    for KiFmmMultiNode<Scalar, Kernel, SourceToTargetData, C>
+impl<Scalar, Kernel, SourceToTargetData> MultiNodeFmm
+    for KiFmmMultiNode<Scalar, Kernel, SourceToTargetData, SimpleCommunicator>
 where
     Scalar: RlstScalar + Default + Equivalence + Float,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
     SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     KiFmm<Scalar, Kernel, SourceToTargetData>: SourceToTargetTranslation + FmmOperatorData,
-    C: Communicator + Default,
-    MultiNodeFmmTree<Scalar, C>: Default,
 {
     type Scalar = Scalar;
     type Kernel = Kernel;
-    type Tree = MultiNodeFmmTree<Scalar, C>;
+    type Tree = MultiNodeFmmTree<Scalar, SimpleCommunicator>;
 
     fn dim(&self) -> usize {
         3
@@ -50,28 +56,12 @@ where
 
     fn evaluate(&mut self, timed: bool) -> Result<(), FmmError> {
         // Run upward pass on local trees
-        {
-            for fmm in self.fmms.iter() {
-                fmm.p2m()?;
-            }
-
-            for level in (self.tree.source_tree.global_depth
-                ..(self.tree.source_tree.global_depth + self.tree.source_tree.local_depth))
-                .rev()
-            {
-                for fmm in self.fmms.iter() {
-                    fmm.m2m(level)?;
-                }
-            }
-        }
+        {}
 
         // At this point the exchange needs to happen of multipole data
         {
-            // 1. Gather ranges on other processes
-
-            // 2. Form packets
-
             // 3. Exchange packets (point to point)
+            // self.exchange_multipoles();
 
             // 4. Pass all root multipole data to root node so that final part of upward pass can occur on root node
         }
@@ -80,22 +70,11 @@ where
         {
             if self.communicator.rank() == 0 {
                 // Global upward pass
-                for level in (1..self.tree.source_tree.global_depth).rev() {
-                    for fmm in self.fmms.iter() {
-                        fmm.m2m(level)?;
-                    }
-                }
+                for level in (1..self.tree.source_tree.global_depth).rev() {}
 
                 // Global downward pass
                 for level in 2..=self.tree.target_tree.global_depth {
-                    if level > 2 {
-                        for fmm in self.fmms.iter() {
-                            fmm.l2l(level)?;
-                        }
[…]
