+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -8,7 +8,7 @@ use mpi::{
 };
 use num::{Float, Zero};
 use pulp::Scalar;
-use rayon::iter::IntoParallelIterator;
+use rayon::iter::{IntoParallelIterator, ParallelIterator};
 use rlst::{
     empty_array, rlst_array_from_slice2, rlst_dynamic_array2, rlst_dynamic_array3, Array,
     BaseArray, MatrixSvd, MultIntoResize, RawAccess, RawAccessMut, RlstScalar, VectorContainer,
@@ -17,13 +17,13 @@ use rlst::{
 use crate::{
     fmm::{
         helpers::{flip3, homogenous_kernel_scale},
-        types::FftMetadata,
+        types::{FftFieldTranslationMultiNode, FftMetadata},
     },
     linalg::pinv::pinv,
     traits::{
         fftw::{Dft, DftType},
         field::SourceToTargetTranslationMetadata,
-        fmm::MultiNodeFmm,
+        fmm::{FmmMetadata, MultiNodeFmm},
         general::AsComplex,
     },
     tree::{
@@ -297,7 +297,7 @@ impl<Scalar> SourceToTargetTranslationMetadata
     for KiFmmMultiNode<
         Scalar,
         Laplace3dKernel<Scalar>,
-        FftFieldTranslation<Scalar>,
+        FftFieldTranslationMultiNode<Scalar>,
         SimpleCommunicator,
     >
 where
@@ -309,59 +309,65 @@ where
         + Dft<InputType = Scalar, OutputType = <Scalar as AsComplex>::ComplexType>,
     <Scalar as RlstScalar>::Real: RlstScalar + Default + Equivalence + Float,
 {
+    // Must be run AFTER multipole exchange.
     fn displacements(&mut self) {
         let mut displacements = Vec::new();
         let depth = self.tree.source_tree.local_depth + self.tree.target_tree.global_depth;
 
-        // for level in 2..=depth {
-        //     let targets = self.tree.target_tree.keys(level).unwrap();
-        //     let targets_parents: HashSet<MortonKey<_>> =
-        //         targets.iter().map(|target| target.parent()).collect();
-        //     let mut targets_parents = targets_parents.into_iter().collect_vec();
-        //     targets_parents.sort();
-        //     let ntargets_parents = targets_parents.len();
+        for fmm_index in 0..self.nfmms {
+            let mut tmp = Vec::new();
 
-        //     let sources = self.tree.source_tree().keys(level).unwrap();
+            for level in 2..=depth {
+                let targets = self.tree.target_tree.trees[fmm_index].keys(level).unwrap();
+                let targets_parents: HashSet<MortonKey<_>> =
+                    targets.iter().map(|target| target.parent()).collect();
+                let mut targets_parents = targets_parents.into_iter().collect_vec();
+                targets_parents.sort();
+                let ntargets_parents = targets_parents.len();
 
-        //     let sources_parents: HashSet<MortonKey<_>> =
-        //         sources.iter().map(|source| source.parent()).collect();
-        //     let mut sources_parents = sources_parents.into_iter().collect_vec();
-        //     sources_parents.sort();
-        //     let nsources_parents = sources_parents.len();
+                let sources = self.tree.source_tree.trees[fmm_index].keys(level).unwrap();
 
-        //     let result = vec![Vec::new(); NHALO];
-        //     let result = result.into_iter().map(RwLock::new).collect_vec();
+                let sources_parents: HashSet<MortonKey<_>> =
+                    sources.iter().map(|source| source.parent()).collect();
+                let mut sources_parents = sources_parents.into_iter().collect_vec();
+                sources_parents.sort();
+                let nsources_parents = sources_parents.len();
 
-        //     let targets_parents_neighbors = targets_parents
-        //         .iter()
[…]
