+++ kifmm/examples/mpi_test_multi_node_fmm.rs
@@ -19,7 +19,6 @@ fn main() {
 
     // Setup tree parameters
     let prune_empty = false;
-    let depth = 5;
     let n_points = 10000;
     let local_depth = 3;
     let global_depth = 1;
@@ -50,6 +49,13 @@ fn main() {
         .unwrap()
         .build()
         .unwrap();
+
+    println!("RANK {:?} {:?}", fmm.communicator.rank(), fmm.tree.source_tree.trees.len());
+    // if fmm.communicator.rank() == 0 {
+        // for source_tree in fmm.tree.source_tree.trees.iter() {
+        //     println!("RANK {:?} {:?} {:?} {:?}", fmm.communicator.rank(), source_tree.leaves.iter().min().unwrap(), source_tree.leaves.iter().max().unwrap(), source_tree.leaves.len());
+        // }
+    // }
 
     // Charge setting has to be after construction in multi-node
     // fmm.set_charges();
