+++ kifmm/src/tree/multi_node.rs
@@ -1,9 +1,12 @@
 //! Implementation of constructors for MPI distributed multi node trees, from distributed point data.
 use crate::{
-    samplesort::samplesort, simplesort, traits::tree::{MultiNodeTreeTrait, SingleNodeTreeTrait}, tree::{
+    samplesort::samplesort,
+    simplesort,
+    traits::tree::{MultiNodeTreeTrait, SingleNodeTreeTrait},
+    tree::{
         constants::DEEPEST_LEVEL,
         types::{Domain, MortonKey, MortonKeys, MultiNodeTree, Point, Points, SingleNodeTree},
-    }
+    },
 };
 
 use crate::hyksort::hyksort;
@@ -18,16 +21,18 @@ use mpi::{
 use num::{zero, Float, Zero};
 use pulp::Scalar;
 use rlst::RlstScalar;
-use std::{collections::{HashMap, HashSet}, fmt::Debug};
+use std::{
+    collections::{HashMap, HashSet},
+    fmt::Debug,
+};
 
 use super::types::MultiNodeTreeNew;
 
-pub enum SortKind <T: Equivalence + Ord  + Default + Clone + Debug>{
-    Hyksort{k: usize},
-    Samplesort{k: usize},
-    Simplesort{splitters: Vec<T>}
+pub enum SortKind<T: Equivalence + Ord + Default + Clone + Debug> {
+    Hyksort { k: usize },
+    Samplesort { k: usize },
+    Simplesort { splitters: Vec<T> },
 }
-
 
 impl<T, C: Communicator> MultiNodeTreeNew<T, C>
 where
@@ -70,19 +75,24 @@ where
         // samplesort(&mut points, &comm, 500).unwrap();
 
         let splitters = MortonKey::root(None).descendants(local_depth)?;
-        let mut splitters = splitters.into_iter().map(|m|
-            Point {coordinate: [T::zero(); 3], global_index: 0, encoded_key: m, base_key: m}).collect_vec();
+        let mut splitters = splitters
+            .into_iter()
+            .map(|m| Point {
+                coordinate: [T::zero(); 3],
+                global_index: 0,
+                encoded_key: m,
+                base_key: m,
+            })
+            .collect_vec();
         splitters.sort();
         let splitters = &splitters[1..];
         simplesort(&mut points, &comm, splitters)?;
-
 
         // Find unique leaves specified by points on each processor
         let leaves: HashSet<MortonKey<_>> = points.iter().map(|p| p.encoded_key).collect();
         let mut leaves = MortonKeys::from(leaves);
         leaves.sort();
         // leaves.complete();
-
 
         // These define all the single node trees to be constructed
         let trees = SingleNodeTree::from_roots(
@@ -96,9 +106,12 @@ where
         );
 
         for tree in trees.iter() {
-            println!("HERE RANK {:?} ROOTS {:?}", comm.rank(), tree.n_coordinates_tot());
+            println!(
+                "HERE RANK {:?} ROOTS {:?}",
+                comm.rank(),
+                tree.n_coordinates_tot()
+            );
         }
-
 
         Ok(MultiNodeTreeNew {
             world: world.duplicate(),
