+++ kifmm/src/fmm/field_translation/source.rs
@@ -3,6 +3,11 @@ use std::collections::HashSet;
 
 use green_kernels::{traits::Kernel as KernelTrait, types::EvalType};
 use itertools::Itertools;
+use mpi::{
+    topology::SimpleCommunicator,
+    traits::{Communicator, Equivalence},
+};
+use num::Float;
 use rayon::prelude::*;
 
 use rlst::{
@@ -13,8 +18,8 @@ use rlst::{
 use crate::{
     fmm::{
         constants::{M2M_MAX_BLOCK_SIZE, P2M_MAX_BLOCK_SIZE},
-        helpers::chunk_size,
-        types::{FmmEvalType, KiFmm},
+        helpers::{chunk_size, homogenous_kernel_scale},
+        types::{FmmEvalType, KiFmm, KiFmmMultiNode},
     },
     traits::{
         field::SourceToTargetData as SourceToTargetDataTrait,
@@ -391,5 +396,145 @@ where
                 Ok(())
             }
         }
+    }
+}
+
+impl<Scalar, Kernel, SourceToTargetData> SourceTranslation
+    for KiFmmMultiNode<Scalar, Kernel, SourceToTargetData>
+where
+    Scalar: RlstScalar + Equivalence + Float,
+    Kernel: KernelTrait<T = Scalar> + HomogenousKernel,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default + Equivalence,
+    Self: FmmOperatorData,
+{
+    fn p2m(&self) -> Result<(), FmmError> {
+        let dim = 3;
+
+        println!(
+            "Rank {:?} running P2M nfmm: {:?}",
+            self.communicator.rank(),
+            self.tree.source_tree.trees.len()
+        );
+        // Run P2M for each local root at this processor
+        for (fmm_idx, tree) in self.tree.source_tree.trees.iter().enumerate() {
+            // let &ncoeffs
+            let Some(_leaves) = tree.all_leaves() else {
+                return Err(FmmError::Failed(
+                    "P2M failed, no leaves found in source tree".to_string(),
+                ));
+            };
+
+            let ncoeffs_equivalent_surface = self.ncoeffs_equivalent_surface;
+            let ncoeffs_check_surface = self.ncoeffs_check_surface;
+            let n_leaves = tree.n_leaves().unwrap();
+            let check_surface_size = ncoeffs_check_surface * dim;
+
+            let coordinates = tree.all_coordinates().unwrap();
+            let ncoordinates = coordinates.len() / dim;
+            let operator_index = 0;
+            let depth = self.tree.source_tree.local_depth + self.tree.source_tree.global_depth;
+
+            let mut check_potentials =
+                rlst_dynamic_array2!(Scalar, [n_leaves * ncoeffs_check_surface, 1]);
+
+            let charges = &self.charges[fmm_idx];
+            let kernel = &self.kernel;
+
+            check_potentials
+                .data_mut()
+                .par_chunks_exact_mut(ncoeffs_check_surface)
+                .zip(
+                    self.leaf_upward_check_surfaces_sources[fmm_idx]
+                        .par_chunks_exact(check_surface_size),
+                )
+                .zip(&self.charge_index_pointers_sources[fmm_idx])
+                .for_each(
+                    |((check_potential, upward_check_surface), charge_index_pointer)| {
+                        let charges = &charges[charge_index_pointer.0..charge_index_pointer.1];
+
+                        let coordinates_row_major = &coordinates
+                            [charge_index_pointer.0 * dim..charge_index_pointer.1 * dim];
+                        let nsources = coordinates_row_major.len() / dim;
+
+                        if nsources > 0 {
+                            kernel.evaluate_st(
+                                EvalType::Value,
+                                coordinates_row_major,
[…]
