+++ TODO.md
@@ -1,0 +1,5 @@
+- check the depths on local/global tree creation regarding metadata as well as meshing of morton keys.
+- multipole data exchange implementation
+- remainder of upward pass on elected node.
+- local data exchange to each local root
+- independent downward pass

[…]
