+++ kifmm/examples/mpi_test_sample_sort.rs
@@ -1,4 +1,0 @@
-// fn main () {
-
-//     let
-// }
