### 2024-09-01 ee6288d
Add subcommunicagtor splitting for multipole exchange
