+++ kifmm/src/fmm/multi_node.rs
@@ -1,12 +1,12 @@
 //! Multi Node FMM
 //! Single Node FMM
-use std::{collections::HashMap, time::Instant};
+use std::{collections::{HashMap, HashSet}, hash::Hash, time::Instant};
 
 use green_kernels::traits::Kernel as KernelTrait;
 
 use itertools::Itertools;
 use mpi::{
-    collective::SystemOperation, ffi::RSMPI_SUM, request::WaitGuard, topology::SimpleCommunicator, traits::{Communicator, Destination, Equivalence, Source}
+    collective::SystemOperation, ffi::RSMPI_SUM, raw::AsRaw, request::WaitGuard, topology::{Color, SimpleCommunicator}, traits::{Communicator, Destination, Equivalence, Group, Root, Source}
 };
 use num::Float;
 use rlst::RlstScalar;
@@ -105,7 +105,7 @@ where
     }
 
     fn exchange_multipoles(&mut self) {
-        let rank = self.rank;
+        let world_rank = self.rank;
         let size = self.communicator.size();
 
         // 1. Gather ranges_min on all processes (should be defined also by interaction lists)
@@ -155,6 +155,8 @@ where
 
         // Using query packet figure out where to send each query physically, using the computed ranges_min.
         {
+
+            // This can be multithreaded
             let mut packets = vec![Vec::new(); size as usize];
 
             for &query in self.query_packet.iter() {
@@ -165,7 +167,8 @@ where
                     }
                 }
             }
-            let messages = packets.iter().map(|p| if !p.is_empty() { 1} else { 0}).collect_vec();
+
+            let messages = packets.iter().map(|p| if !p.is_empty() { 1 } else { 0 }).collect_vec();
 
             let packet_destinations = packets
                 .iter()
@@ -192,10 +195,12 @@ where
             let packet_sizes = packets.iter().map(|p| p.len() as i32 ).collect_vec();
 
             let mut total_messages = vec![0i32; size as usize];
+
             // All to all to figure out with whom to communicate
             self.communicator.all_reduce_into(&messages, &mut total_messages, SystemOperation::sum());
 
-            let recv_count = total_messages[rank as usize];
+            // Break into subcommunicators so multipoles can be broadcast
+            let recv_count = total_messages[world_rank as usize];
             let send_count = packets.len() as i32;
             let nreqs = recv_count + send_count;
 
@@ -203,10 +208,12 @@ where
             let mut received_packet_sizes = vec![0 as i32; recv_count as usize];
             let mut received_packet_sources = vec![0 as i32; recv_count as usize];
 
+            let color = self.communicator.rank();
+            // println!("ABOUT TO START {:?} {:?} {:?}", rank, send_count, recv_count);
             mpi::request::multiple_scope(nreqs as usize, |scope, coll| {
-                for (i, &rank) in packet_destinations.iter().enumerate() {
-                    let tag = rank;
-                    let sreq = self.communicator.process_at_rank(rank).immediate_send_with_tag(
+                for (i, &destination_rank) in packet_destinations.iter().enumerate() {
+                    let tag = destination_rank;
+                    let sreq = self.communicator.process_at_rank(destination_rank).immediate_send_with_tag(
                         scope,
                         &packet_sizes[i],
                         tag,
@@ -219,7 +226,7 @@ where
                         // Spin for message availability. There is no guarantee that
                         // immediate sends, even to the same process, will be immediately
                         // visible to an immediate probe.
-                        let preq = self.communicator.any_process().immediate_matched_probe_with_tag(rank);
+                        let preq = self.communicator.any_process().immediate_matched_probe_with_tag(world_rank);
                         if let Some(p) = preq {
                             break p;
[…]
