+++ kifmm/src/fmm/multi_node.rs
@@ -1,14 +1,25 @@
 //! Multi Node FMM
 //! Single Node FMM
-use std::{collections::{HashMap, HashSet}, hash::Hash, time::Instant};
+use std::{
+    collections::{HashMap, HashSet},
+    hash::Hash,
+    time::Instant,
+};
 
 use green_kernels::traits::Kernel as KernelTrait;
 
 use itertools::Itertools;
 use mpi::{
-    collective::SystemOperation, ffi::RSMPI_SUM, raw::AsRaw, request::WaitGuard, topology::{Color, SimpleCommunicator}, traits::{Communicator, Destination, Equivalence, Group, Root, Source}
+    collective::SystemOperation,
+    ffi::RSMPI_SUM,
+    raw::AsRaw,
+    request::WaitGuard,
+    topology::{Color, SimpleCommunicator},
+    traits::{Communicator, Destination, Equivalence, Group, Root, Source},
 };
 use num::Float;
+use rayon::iter::IntoParallelIterator;
+use rayon::iter::ParallelIterator;
 use rlst::RlstScalar;
 
 use mpi::collective::CommunicatorCollectives;
@@ -18,8 +29,8 @@ use crate::{
     traits::{
         field::SourceToTargetData as SourceToTargetDataTrait,
         fmm::{
-            FmmOperatorData, HomogenousKernel, MultiNodeFmm, SourceToTargetTranslation,
-            SourceTranslation, TargetTranslation,
+            ExchangeGhostData, FmmOperatorData, HomogenousKernel, MultiNodeFmm,
+            SourceToTargetTranslation, SourceTranslation, TargetTranslation,
         },
         tree::SingleNodeTreeTrait,
         types::{FmmError, FmmOperatorTime, FmmOperatorType},
@@ -37,7 +48,7 @@ where
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
     SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
-    Self: SourceTranslation,
+    Self: SourceTranslation + ExchangeGhostData,
 {
     type Scalar = Scalar;
     type Kernel = Kernel;
@@ -70,7 +81,7 @@ where
         // At this point the exchange needs to happen of multipole data
         {
             //     // 3. Exchange packets (point to point)
-            self.exchange_multipoles();
+            self.v_list();
 
             //     // 4. Pass all root multipole data to root node so that final part of upward pass can occur on root node
         }
@@ -104,314 +115,8 @@ where
         &self.kernel
     }
 
-    fn exchange_multipoles(&mut self) {
-        let world_rank = self.rank;
-        let size = self.communicator.size();
-
-        // 1. Gather ranges_min on all processes (should be defined also by interaction lists)
-        let mut ranges_min = Vec::new();
-        let mut ranges_max= Vec::new();
-        for fmm_idx in 0..self.nsource_trees {
-            ranges_min.push(self.tree.source_tree.trees[fmm_idx].owned_range().finest_first_child());
-            ranges_max.push(self.tree.source_tree.trees[fmm_idx].owned_range().finest_last_child());
-        }
-
-
-        let range = [
-            *ranges_min.iter().min().unwrap(),
-            *ranges_max.iter().max().unwrap(),
-        ];
-
-        // let nranges = ranges_min.len() as i32;
-        // let mut all_nranges = vec![0i32; size as usize];
-
-        // self.communicator.all_gather_into(&nranges, &mut all_nranges);
-
-        // println!("RANK {:?} {:?}", rank, all_nranges);
-
-        // let recv_displacements = all_nranges
-        //     .iter()
-        //     .scan(0, |acc, &x| {
[…]
