### 2024-09-02 f51c211
Add communication mode
