+++ kifmm/src/fmm/multi_node.rs
@@ -39,7 +39,7 @@ use crate::{
     Fmm, MultiNodeFmmTree,
 };
 
-use super::types::KiFmmMultiNode;
+use super::types::{CommunicationMode, KiFmmMultiNode};
 
 impl<Scalar, Kernel, SourceToTargetData> MultiNodeFmm
     for KiFmmMultiNode<Scalar, Kernel, SourceToTargetData>
@@ -80,33 +80,31 @@ where
 
         // At this point the exchange needs to happen of multipole data
         {
-            //     // 3. Exchange packets (point to point)
-            self.v_list();
+            // 3. Exchange packets (point to point)
+
+            match self.communication_mode {
+                CommunicationMode::P2P => self.v_list_p2p(),
+                CommunicationMode::Subcomm => self.v_list_subcomm(),
+            }
+
+            // Update metadata
+            self.update_v_list_metadata();
 
             //     // 4. Pass all root multipole data to root node so that final part of upward pass can occur on root node
         }
 
+        // Gather root multipoles at nominated node
+        self.gather_root_multipoles();
+
         // Now can proceed with remainder of the upward pass on chosen node, and some of the downward pass
-        // {
-        //     if self.communicator.rank() == 0 {
-        //         // Global upward pass
-        //         for level in (1..self.tree.source_tree.global_depth).rev() {}
+        {
 
-        //         // Global downward pass
-        //         for level in 2..=self.tree.target_tree.global_depth {
-        //             if level > 2 {}
-        //         }
-        //     }
+        }
 
-        //     // Exchange root multipole data back to required MPI processes
-        // }
+        // Scatter root locals back to local trees
+        self.scatter_root_locals();
 
         // Now remainder of downward pass can happen in parallel on each process
-        // {
-        //     // local leaf level operations
-        //     // fmm.p2p()?;
-        //     // fmm.l2p()?;
-        // }
 
         Ok(())
     }
@@ -182,6 +180,14 @@ where
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     Self: SourceTranslation,
 {
+    fn gather_root_multipoles(&mut self) {
+
+    }
+
+    fn scatter_root_locals(&mut self) {
+
+    }
+
     fn gather_ranges(&mut self) {
         let size = self.communicator.size();
 
@@ -212,7 +218,7 @@ where
         self.all_ranges = all_ranges;
     }
 
-    fn u_list(&mut self) {
+    fn u_list_p2p(&mut self) {
         let world_rank = self.rank;
         let size = self.communicator.size();
 
@@ -343,7 +349,6 @@ where
             let mut tmp = Vec::new();
 
             for key in recv_buffer.iter() {
-
                 let mut index = 0;
                 for (fmm_idx, tree) in self.tree.source_tree.trees.iter().enumerate() {
                     if tree.keys_set.contains(key) {
@@ -352,10 +357,9 @@ where
                     }
                 }
 
-                if let Some(_coords) = self.tree.source_tree.trees[index].coordinates(key)  {
+                if let Some(_coords) = self.tree.source_tree.trees[index].coordinates(key) {
                     tmp.push(*key)
                 }
-
             }
 
             available_received_packets_sizes.push(tmp.len() as i32);
@@ -367,7 +371,6 @@ where
         let mut send_buffers_index_pointers = Vec::new();
 
         for packet in available_received_packets.iter() {
-
             let mut tmp = Vec::new(); // index pointer
 
             let mut packet_size = 0;
@@ -382,11 +385,12 @@ where
                     }
                 }
 
-                let ncoords = if let Some(coords) = self.tree.source_tree.trees[index].coordinates(key) {
-                    coords.len()
-                } else {
-                    0
-                };
+                let ncoords =
[…]
