### 2024-09-02 bffd4f3
Add placeholder for ghost trees
