### 2024-09-02 9143311
Clean up ghost comm trait
