+++ kifmm/src/fmm/multi_node.rs
@@ -83,10 +83,7 @@ where
 
         // At this point the exchange needs to happen of multipole data
         {
-            match self.communication_mode {
-                CommunicationMode::P2P => self.v_list_p2p(),
-                CommunicationMode::Subcomm => self.v_list_subcomm(),
-            }
+            self.v_list_exchange()
         }
 
         // Build up local data structures containing ghost data
@@ -108,10 +105,6 @@ where
 
     fn kernel(&self) -> &Self::Kernel {
         &self.kernel
-    }
-
-    fn reset_metadata(&mut self) {
-        // have to update
     }
 
     fn check_surface_order(&self, level: u64) -> usize {
@@ -211,783 +204,796 @@ where
         self.all_ranges = all_ranges;
     }
 
-    fn u_list_p2p(&mut self) {
+    fn u_list_exchange(&mut self) {
         let world_rank = self.rank;
         let size = self.communicator.size();
 
-        // 1. Use query packet to figure out where to send each query physically
-        let mut packets = vec![Vec::new(); size as usize];
+        match self.communication_mode {
+            CommunicationMode::P2P => {
+                // 1. Use query packet to figure out where to send each query physically
+                let mut packets = vec![Vec::new(); size as usize];
 
-        for &query in self.particle_query_packet.iter() {
-            for (destination_rank, range) in self.all_ranges.chunks(2).enumerate() {
-                if range[0] <= query.finest_first_child() && query.finest_first_child() <= range[1]
-                {
-                    packets[destination_rank].push(query)
-                }
-            }
-        }
-
-        let messages = packets
-            .iter()
-            .map(|p| if !p.is_empty() { 1 } else { 0 })
-            .collect_vec();
-
-        let packet_destinations = packets
-            .iter()
-            .enumerate()
-            .filter_map(|(rank, packet)| {
-                if packet.len() > 0 {
-                    Some(rank as i32)
-                } else {
-                    None
-                }
-            })
-            .collect_vec();
-
-        let packets = packets
-            .into_iter()
-            .filter_map(|p| if p.len() > 0 { Some(p) } else { None })
-            .collect_vec();
-
-        let packet_sizes = packets.iter().map(|p| p.len() as i32).collect_vec();
-
-        let mut total_messages = vec![0i32; size as usize];
-
-        // 2. All to all to figure out with whom to communicate
-
-        self.communicator
-            .all_reduce_into(&messages, &mut total_messages, SystemOperation::sum());
-
-        // 2. Need to now find packet sizes and sources for queries for particle data
-        let recv_count = total_messages[world_rank as usize];
-        let send_count = packets.len() as i32;
-        let nreqs = recv_count + send_count;
-
-        let mut received_packet_sizes = vec![0 as i32; recv_count as usize];
-        let mut received_packet_sources = vec![0 as i32; recv_count as usize];
-
-        mpi::request::multiple_scope(nreqs as usize, |scope, coll| {
-            for (i, &destination_rank) in packet_destinations.iter().enumerate() {
-                let tag = destination_rank;
-                let sreq = self
-                    .communicator
-                    .process_at_rank(destination_rank)
-                    .immediate_send_with_tag(scope, &packet_sizes[i], tag);
-                coll.add(sreq);
-            }
-
-            for (i, size) in received_packet_sizes.iter_mut().enumerate() {
-                let (msg, status) = loop {
-                    // Spin for message availability. There is no guarantee that
-                    // immediate sends, even to the same process, will be immediately
-                    // visible to an immediate probe.
-                    let preq = self
-                        .communicator
-                        .any_process()
-                        .immediate_matched_probe_with_tag(world_rank);
-                    if let Some(p) = preq {
-                        break p;
-                    }
-                };
-
-                let rreq = msg.immediate_matched_receive_into(scope, size);
[…]
