+++ kifmm/src/python.rs
@@ -1,858 +1,0 @@
-//! Python bindings for constructors and basic methods
-use std::collections::HashMap;
-use std::time::Duration;
-
-use crate::fmm::types::FmmSvdMode;
-use crate::fmm::KiFmm;
-use crate::traits::fmm::Fmm;
-use crate::traits::tree::Tree;
-use crate::tree::types::MortonKey;
-use crate::{
-    BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder,
-};
-use green_kernels::helmholtz_3d::Helmholtz3dKernel;
-use green_kernels::traits::Kernel;
-use green_kernels::{laplace_3d::Laplace3dKernel, types::EvalType};
-use numpy::{
-    ndarray::Dim, npyffi::NPY_ORDER, PyArray, PyArray2, PyArrayMethods, PyReadonlyArrayDyn,
-    PyUntypedArrayMethods, ToPyArray,
-};
-use pyo3::exceptions::PyTypeError;
-use pyo3::prelude::*;
-
-use rlst::{
-    c32, c64, rlst_array_from_slice2, rlst_dynamic_array2, RawAccess, RawAccessMut, RlstScalar,
-};
-
-use pyo3::{pymodule, types::PyModule, Bound, PyResult};
-
-macro_rules! define_pyclass {
-    ($name: ident, $type: ident, $kernel: ident, $field_translation: ident) => {
-        /// Python interface
-        #[pyclass]
-        pub struct $name {
-            fmm: KiFmm<$type, $kernel<$type>, $field_translation<$type>>,
-            source_keys: Vec<u64>,
-            target_keys: Vec<u64>,
-            source_leaves: Vec<u64>,
-            target_leaves: Vec<u64>,
-            source_key_map: HashMap<u64, MortonKey<<$type as RlstScalar>::Real>>,
-            target_key_map: HashMap<u64, MortonKey<<$type as RlstScalar>::Real>>,
-        }
-    };
-}
-
-define_pyclass!(LaplaceFft64, f64, Laplace3dKernel, FftFieldTranslation);
-define_pyclass!(LaplaceFft32, f32, Laplace3dKernel, FftFieldTranslation);
-define_pyclass!(
-    LaplaceBlas64,
-    f64,
-    Laplace3dKernel,
-    BlasFieldTranslationSaRcmp
-);
-define_pyclass!(
-    LaplaceBlas32,
-    f32,
-    Laplace3dKernel,
-    BlasFieldTranslationSaRcmp
-);
-
-define_pyclass!(HelmholtzFft64, c64, Helmholtz3dKernel, FftFieldTranslation);
-define_pyclass!(HelmholtzFft32, c32, Helmholtz3dKernel, FftFieldTranslation);
-define_pyclass!(
-    HelmholtzBlas64,
-    c64,
-    Helmholtz3dKernel,
-    BlasFieldTranslationIa
-);
-define_pyclass!(
-    HelmholtzBlas32,
-    c32,
-    Helmholtz3dKernel,
-    BlasFieldTranslationIa
-);
-
-macro_rules! laplace_fft_constructors {
-    ($name: ident, $type: ident) => {
-        #[pymethods]
-        impl $name {
-            /// Constructor
-            #[new]
-            #[allow(clippy::too_many_arguments)]
-            pub fn new<'py>(
-                expansion_order: Vec<usize>,
-                sources: PyReadonlyArrayDyn<'py, <$type as RlstScalar>::Real>,
-                targets: PyReadonlyArrayDyn<'py, <$type as RlstScalar>::Real>,
-                charges: PyReadonlyArrayDyn<'py, $type>,
-                prune_empty: bool,
[…]
