### 2024-09-03 6b4d5ee
Add ghost v tree creation
