+++ kifmm/src/tree/types.rs
@@ -4,6 +4,7 @@
 use mpi::traits::{Communicator, Equivalence};
 
 use num::Float;
+use pulp::Scalar;
 use rlst::RlstScalar;
 
 use std::{
@@ -348,6 +349,7 @@ where
     pub global_rank: i32,
 
     pub local_rank: i32,
+
     /// Depth of a tree. taken from underlying tree
     pub depth: u64,
 
@@ -386,20 +388,14 @@ where
 
     pub local_rank: i32,
 
-    /// Associated root node
-    pub root: MortonKey<T>,
-
     /// Depth of a tree.
     pub depth: u64,
 
     /// Domain spanned by the points.
-    pub domain: Domain<T>,
+    pub domain: Domain<T::Real>,
 
     /// All global indices
     pub global_indices: Vec<usize>,
-
-    /// The leaves that span the tree, and associated Point data.
-    pub leaves: MortonKeys<T::Real>,
 
     /// All nodes in tree, and associated Node data.
     pub keys: MortonKeys<T::Real>,
@@ -415,4 +411,7 @@ where
 
     /// All keys, returned as a set.
     pub keys_set: HashSet<MortonKey<T::Real>>,
+
+    /// multipole data
+    pub multipoles: Vec<T>,
 }
