### 2024-09-03 754ab7b
Add ghost tree construction for U list
