+++ kifmm/src/fmm/multi_node.rs
@@ -47,7 +47,7 @@ use crate::{
 
 use super::{
     helpers::{expected_queries, sparse_point_to_point_v},
-    types::{CommunicationMode, KiFmmMultiNode},
+    types::{CommunicationMode, IndexPointer, KiFmmMultiNode},
 };
 
 impl<Scalar, Kernel, SourceToTargetData> MultiNodeFmm
@@ -216,210 +216,153 @@ where
 
         match self.communication_mode {
             CommunicationMode::P2P => {
-                // 1. Use query packet to figure out where to send each query physically
-                let mut packets = vec![Vec::new(); size as usize];
+                // Use query packet to figure out where to send each query physically
+                let mut queries = vec![Vec::new(); size as usize];
 
                 for &query in self.particle_query_packet.iter() {
                     for (destination_rank, range) in self.all_ranges.chunks(2).enumerate() {
                         if range[0] <= query.finest_first_child()
                             && query.finest_first_child() <= range[1]
                         {
-                            packets[destination_rank].push(query)
+                            queries[destination_rank].push(query.morton)
                         }
                     }
                 }
 
-                let messages = packets
-                    .iter()
-                    .map(|p| if !p.is_empty() { 1 } else { 0 })
-                    .collect_vec();
+                let mut queries_to_send = Vec::new();
+                let mut nqueries_to_send = Vec::new();
+                let mut queries_to_send_sizes = Vec::new();
+                let mut queries_to_send_destination_ranks = Vec::new();
+                for (destination_rank, query) in queries.into_iter().enumerate() {
+                    if !query.is_empty() {
+                        queries_to_send_destination_ranks.push(destination_rank as i32);
+                        queries_to_send_sizes.push(query.len() as i32);
+                        queries_to_send.push(query);
+                        nqueries_to_send.push(1)
+                    } else {
+                        nqueries_to_send.push(0)
+                    }
+                }
 
-                let packet_destinations = packets
-                    .iter()
-                    .enumerate()
-                    .filter_map(|(rank, packet)| {
-                        if packet.len() > 0 {
-                            Some(rank as i32)
-                        } else {
-                            None
-                        }
-                    })
-                    .collect_vec();
-
-                let packets = packets
-                    .into_iter()
-                    .filter_map(|p| if p.len() > 0 { Some(p) } else { None })
-                    .collect_vec();
-
-                let packet_sizes = packets.iter().map(|p| p.len() as i32).collect_vec();
-
-                let mut total_messages = vec![0i32; size as usize];
-
-                // 2. All to all to figure out with whom to communicate
-
+                // Mark ranks we expect to receive from with the number of expected messages
+                let mut nqueries_to_receive = vec![0i32; size as usize];
                 self.communicator.all_reduce_into(
-                    &messages,
-                    &mut total_messages,
+                    &nqueries_to_send,
+                    &mut nqueries_to_receive,
                     SystemOperation::sum(),
                 );
 
-                // 2. Need to now find packet sizes and sources for queries for particle data
-                let recv_count = total_messages[world_rank as usize];
-                let send_count = packets.len() as i32;
+                // Number of expected queries at this rank
+                let recv_count = nqueries_to_receive[world_rank as usize];
+
+                // Number of queries being send to other ranks
+                let send_count = queries_to_send.len() as i32;
+
+                // total number of requests being handled at this rank
                 let nreqs = recv_count + send_count;
 
-                let mut received_packet_sizes = vec![0 as i32; recv_count as usize];
-                let mut received_packet_sources = vec![0 as i32; recv_count as usize];
+                // Communicate query sizes, and with whom to communicate
+                // This is the size and origin of queries received from source ranks, and expected at this rank
[…]
