+++ kifmm/benches/laplace_f32.rs
@@ -9,6 +9,7 @@ use kifmm::traits::tree::{FmmTree, Tree};
 use kifmm::tree::helpers::points_fixture;
 use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut};
 
+
 fn laplace_potentials_f32(c: &mut Criterion) {
     let mut group = c.benchmark_group("F32 Potentials");
 
