+++ kifmm/Cargo.toml
@@ -52,7 +52,6 @@ harness = false
 name = "laplace_f64"
 harness = false
 
-
 [package.metadata.docs.rs]
 cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
 rustdoc-args = [ "--html-in-header", "./src/docs-header.html" ]
