### 2024-09-04 7cebfd0
CLN: Update AMD benches for web (#106)
