+++ kifmm/Cargo.toml
@@ -45,22 +45,6 @@ criterion = { version = "0.5.*", features = ["html_reports"]}
 cbindgen = "0.26.0"
 
 [[bench]]
-name = "direct_f32"
-harness = false
-
-[[bench]]
-name = "direct_f64"
-harness = false
-
-[[bench]]
-name = "laplace_mat_f64"
-harness = false
-
-[[bench]]
-name = "laplace_mat_f32"
-harness = false
-
-[[bench]]
 name = "laplace_f32"
 harness = false
 
