### 2024-09-06 13fda6a
Add subcomm ghost tree creation
