+++ kifmm/src/fmm/multi_node.rs
@@ -278,31 +278,25 @@ where
     }
 
     fn u_list_exchange(&mut self) {
-        let world_rank = self.rank;
-        let size = self.communicator.size();
-
-
-
-    }
-
-    fn v_list_exchange(&mut self) {
-        // Two sets of communications, first for the multipoles that exist
-        // secondly for the associated coefficient data
-
-        // Begin by calculating receive counts, this requires an all to all over the neighbourhood
-        // communicator only
-        let neighbourhood_size = self.neighbourhood_communicator_v.size();
+        // Similar to V list, begin by calculating receive counts
+        let neighbourhood_size = self.neighbourhood_communicator_u.size();
 
         let mut neighbourhood_send_counts = vec![0i32; neighbourhood_size as usize];
 
-        for (global_rank, &send_count) in self.v_list_send_counts.iter().enumerate() {
-            if let Some(local_rank) = self.neighbourhood_communicator_v.global_to_local_rank(global_rank as i32) {
+        for (global_rank, &send_count) in self.u_list_send_counts.iter().enumerate() {
+            if let Some(local_rank) = self
+                .neighbourhood_communicator_u
+                .global_to_local_rank(global_rank as i32)
+            {
                 neighbourhood_send_counts[local_rank as usize] = send_count
             }
         }
 
         let mut neighbourhood_receive_counts = vec![0i32; neighbourhood_size as usize];
-        self.neighbourhood_communicator_v.all_to_all_into(&neighbourhood_send_counts, &mut neighbourhood_receive_counts);
+        self.neighbourhood_communicator_u.all_to_all_into(
+            &neighbourhood_send_counts,
+            &mut neighbourhood_receive_counts,
+        );
 
         // Now can create displacements
         let mut neighbourhood_send_displacements = Vec::new();
@@ -311,10 +305,13 @@ where
         let mut send_counter = 0;
         let mut receive_counter = 0;
 
-        for (send_count, receive_count) in neighbourhood_send_counts.iter().zip(neighbourhood_receive_counts.iter()) {
+        for (send_count, receive_count) in neighbourhood_send_counts
+            .iter()
+            .zip(neighbourhood_receive_counts.iter())
+        {
             neighbourhood_send_displacements.push(send_counter);
             neighbourhood_receive_displacements.push(receive_counter);
-            send_counter +=  send_count;
+            send_counter += send_count;
             receive_counter += receive_count;
         }
 
@@ -324,27 +321,238 @@ where
         let mut receive_counter = 0;
         let mut received_queries_source_ranks = vec![0i32; total_receive_count as usize];
         for (i, &receive_count) in neighbourhood_receive_counts.iter().enumerate() {
-
             let curr_receive_counter = receive_counter;
-            receive_counter +=  receive_count as usize;
+            receive_counter += receive_count as usize;
+
+            let rank = self.neighbourhood_communicator_u.neighbours[i];
+            let tmp = vec![rank as i32; receive_count as usize];
+            received_queries_source_ranks[curr_receive_counter..receive_counter]
+                .copy_from_slice(tmp.as_slice());
+        }
+
+        // Communicate queries for coordinate data
+        let mut received_queries = vec![0u64; total_receive_count as usize];
+        {
+            let partition_send = Partition::new(
+                &self.u_list_queries,
+                neighbourhood_send_counts,
+                neighbourhood_send_displacements,
+            );
+            let mut partition_receive = PartitionMut::new(
+                &mut received_queries,
+                neighbourhood_receive_counts,
+                &neighbourhood_receive_displacements[..],
+            );
+            self.neighbourhood_communicator_u
+                .all_to_all_varcount_into(&partition_send, &mut partition_receive);
+        }
+
+        // Now have to check for locally contained keys from received queries
[…]
