### 2024-09-06 70b8a71
Start working on global fmm object
