+++ kifmm/src/fmm/multi_node.rs
@@ -21,6 +21,7 @@ use mpi::{
 };
 
 use num::Float;
+use pulp::Scalar;
 use rayon::iter::IntoParallelIterator;
 use rayon::iter::ParallelIterator;
 use rlst::RlstScalar;
@@ -35,7 +36,7 @@ use crate::{
     traits::{
         field::SourceToTargetData as SourceToTargetDataTrait,
         fmm::{
-            FmmOperatorData, GhostExchange, HomogenousKernel, MultiNodeFmm,
+            FmmMetadata, FmmOperatorData, GhostExchange, HomogenousKernel, MultiNodeFmm,
             SourceToTargetTranslation, SourceTranslation, TargetTranslation,
         },
         tree::SingleNodeTreeTrait,
@@ -45,12 +46,12 @@ use crate::{
         helpers::all_to_allv_sparse,
         types::{GhostTreeU, GhostTreeV, MortonKey},
     },
-    Fmm, MultiNodeFmmTree,
+    Fmm, MultiNodeFmmTree, SingleNodeFmmTree,
 };
 
 use super::{
     helpers::{expected_queries, sparse_point_to_point_v},
-    types::{IndexPointer, KiFmmMultiNode, Layout},
+    types::{IndexPointer, KiFmmMultiNode, KiFmmMultiNodeGlobal, Layout},
 };
 
 impl<Scalar, Kernel, SourceToTargetData> MultiNodeFmm
@@ -78,11 +79,10 @@ where
             self.times
                 .push(FmmOperatorTime::from_instant(FmmOperatorType::P2M, s));
 
-            let local_depth = self.tree.source_tree.local_depth;
-            let global_depth = self.tree.source_tree.global_depth;
-            for level in ((global_depth + 1)..=(local_depth + global_depth)).rev() {
+            let depth = self.tree.source_tree.local_depth + self.tree.source_tree.global_depth;
+            for level in ((self.tree.source_tree.global_depth + 1)..=depth).rev() {
                 let s = Instant::now();
-                self.m2m(level)?;
+                self.m2m(level).unwrap();
                 self.times.push(FmmOperatorTime::from_instant(
                     FmmOperatorType::M2M(level),
                     s,
@@ -91,20 +91,17 @@ where
         }
 
         // At this point the exchange needs to happen of multipole data
-        {
-            self.v_list_exchange()
-        }
-
-        // Build up local data structures containing ghost data
+        self.v_list_exchange();
 
         // Gather root multipoles at nominated node
-        self.gather_root_multipoles();
+        self.gather_global_tree_at_root();
+
+        // Set metadata for ghost data, and root multipoles
 
         // Now can proceed with remainder of the upward pass on chosen node, and some of the downward pass
-        {}
 
         // Scatter root locals back to local trees
-        self.scatter_root_locals();
+        self.scatter_global_tree_from_root();
 
         // Now remainder of downward pass can happen in parallel on each process, similar to how I've written the local upward passes
         // new kernels have to reflect ghost data, and potentially multiple local source trees
@@ -175,33 +172,235 @@ impl<Scalar, Kernel, SourceToTargetData> GhostExchange
 where
     Scalar: RlstScalar + Default + Equivalence + Float,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
-    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync + Default,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     Self: SourceTranslation,
 {
-    fn gather_root_multipoles(&mut self) {
+    fn gather_global_tree_at_root(&mut self) {
[…]
