+++ kifmm/src/fmm/field_translation/metadata/mod.rs
@@ -1,2 +1,3 @@
+pub mod ghost_tree;
 pub mod multi_node;
 pub mod single_node;
