+++ kifmm/src/fmm/field_translation/metadata/single_node.rs
@@ -39,7 +39,10 @@ use crate::{
             SourceAndTargetTranslationMetadata, SourceToTargetData as SourceToTargetDataTrait,
             SourceToTargetTranslationMetadata,
         },
-        fmm::{FmmMetadata, FmmOperatorData, HomogenousKernel, SourceToTargetTranslation},
+        fmm::{
+            FmmGlobalFmmMetadata, FmmMetadata, FmmOperatorData, HomogenousKernel,
+            SourceToTargetTranslation,
+        },
         general::{AsComplex, Epsilon},
         tree::{Domain as DomainTrait, FmmTreeNode, SingleNodeFmmTreeTrait, SingleNodeTreeTrait},
     },
@@ -48,7 +51,8 @@ use crate::{
             ALPHA_INNER, ALPHA_OUTER, NHALO, NSIBLINGS, NSIBLINGS_SQUARED, NTRANSFER_VECTORS_KIFMM,
         },
         helpers::find_corners,
-        types::MortonKey,
+        types::{Domain, MortonKey},
+        SingleNodeTree,
     },
     Fmm,
 };
@@ -607,6 +611,13 @@ where
     Scalar: RlstScalar<Complex = Scalar> + Default + MatrixSvd,
     <Scalar as RlstScalar>::Real: Default,
 {
+    fn displacements_explicit<T: RlstScalar + Float>(
+        &mut self,
+        source_trees: &[SingleNodeTree<T>],
+        target_trees: &[SingleNodeTree<T>],
+    ) {
+    }
+
     fn displacements(&mut self) {
         let mut displacements = Vec::new();
 
@@ -817,6 +828,13 @@ where
         + Dft<InputType = Scalar, OutputType = <Scalar as AsComplex>::ComplexType>,
     <Scalar as RlstScalar>::Real: RlstScalar + Default,
 {
+    fn displacements_explicit<T: RlstScalar + Float>(
+        &mut self,
+        source_trees: &[SingleNodeTree<T>],
+        target_trees: &[SingleNodeTree<T>],
+    ) {
+    }
+
     fn displacements(&mut self) {
         let mut displacements = Vec::new();
 
@@ -1375,6 +1393,13 @@ where
     Scalar: RlstScalar + Default + MatrixRsvd,
     <Scalar as RlstScalar>::Real: Default,
 {
+    fn displacements_explicit<T: RlstScalar + Float>(
+        &mut self,
+        source_trees: &[SingleNodeTree<T>],
+        target_trees: &[SingleNodeTree<T>],
+    ) {
+    }
+
     fn displacements(&mut self) {
         let mut displacements = Vec::new();
 
@@ -1874,6 +1899,13 @@ where
         + Dft<InputType = Scalar, OutputType = <Scalar as AsComplex>::ComplexType>,
     <Scalar as RlstScalar>::Real: RlstScalar + Default,
 {
+    fn displacements_explicit<T: RlstScalar + Float>(
+        &mut self,
+        source_trees: &[SingleNodeTree<T>],
+        target_trees: &[SingleNodeTree<T>],
+    ) {
+    }
+
     fn displacements(&mut self) {
         let mut displacements = Vec::new();
 
@@ -2281,6 +2313,172 @@ where
 
     fn displacement_index(&self, level: u64) -> usize {
         (level - 2) as usize
+    }
+}
+
+impl<Scalar, Kernel, SourceToTargetData> FmmGlobalFmmMetadata
+    for KiFmm<Scalar, Kernel, SourceToTargetData>
+where
+    Scalar: RlstScalar + Default + Float,
[…]

[…]
