### 2024-09-16 0497028
Remove a lot of dead code
