+++ kifmm/examples/mpi_test_ghost_exchange.rs
@@ -8,7 +8,7 @@ fn main() {
     use kifmm::{
         fmm::types::{FftFieldTranslationMultiNode, MultiNodeBuilder},
         traits::fmm::{GhostExchange, MultiNodeFmm},
-        tree::{helpers::points_fixture, types::MultiNodeTreeNew},
+        tree::{helpers::points_fixture, types::MultiNodeTree},
         FftFieldTranslation,
     };
     use mpi::{environment::Universe, traits::Communicator};

[…]
