+++ kifmm/src/fmm/field_translation/metadata/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "mpi")]
 pub mod ghost_tree;
+#[cfg(feature = "mpi")]
 pub mod multi_node;
 pub mod single_node;
