+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -1,4 +1,7 @@
-use std::{collections::HashSet, sync::RwLock};
+use std::{
+    collections::{HashMap, HashSet},
+    sync::RwLock,
+};
 
 use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait, types::EvalType};
 use itertools::Itertools;
@@ -13,7 +16,9 @@ use rlst::{
 use crate::{
     fmm::{
         constants::DEFAULT_M2L_FFT_BLOCK_SIZE,
-        field_translation::source_to_target::transfer_vector::compute_transfer_vectors_at_level,
+        field_translation::{
+            source_to_target::transfer_vector::compute_transfer_vectors_at_level, target,
+        },
         helpers::{
             coordinate_index_pointer_multinode, flip3, homogenous_kernel_scale,
             leaf_expansion_pointers_multinode, leaf_surfaces, level_expansion_pointers_multinode,
@@ -35,7 +40,7 @@ use crate::{
             SourceToTargetTranslation,
         },
         general::{AsComplex, Epsilon},
-        tree::{FmmTreeNode, SingleNodeTreeTrait},
+        tree::{FmmTreeNode, MultiNodeFmmTreeTrait, MultiNodeTreeTrait, SingleNodeTreeTrait},
     },
     tree::{
         constants::{ALPHA_INNER, ALPHA_OUTER, NHALO, NSIBLINGS, NSIBLINGS_SQUARED},
@@ -734,6 +739,7 @@ where
     fn metadata<'a>(&mut self, eval_type: EvalType, _charges: &'a [Self::Charges]) {
         // In a multinode setting this method sets the required metdata for the local upward passes, before ghost exchange.
 
+        let dim = 3;
         let alpha_outer = Scalar::real(ALPHA_OUTER);
         let alpha_inner = Scalar::real(ALPHA_INNER);
 
@@ -749,313 +755,470 @@ where
             EvalType::ValueDeriv => 4,
         };
 
-        let nsource_trees = self.tree.source_tree.n_trees;
-        let ntarget_trees = self.tree.target_tree.n_trees;
-
-        // Allocate buffers to store multipole and local data
-        let mut multipoles = Vec::new();
-        let mut locals = Vec::new();
-
-        // Allocate buffer to store potential data at target points
-        let mut potentials = Vec::new();
-
-        for fmm_idx in 0..nsource_trees {
-            let nsource_keys = self.tree.source_tree.trees[fmm_idx].n_keys_tot().unwrap();
-            multipoles.push(vec![
-                Scalar::default();
-                nsource_keys * self.ncoeffs_equivalent_surface
-            ]);
-        }
-
-        for fmm_idx in 0..ntarget_trees {
-            let ntarget_keys = self.tree.target_tree.trees[fmm_idx].n_keys_tot().unwrap();
-            let ntarget_points = self.tree.target_tree.trees[fmm_idx]
-                .n_coordinates_tot()
-                .unwrap();
-            potentials.push(vec![Scalar::default(); ntarget_points * eval_size]);
-            locals.push(vec![
-                Scalar::default();
-                ntarget_keys * self.ncoeffs_equivalent_surface
-            ]);
-        }
-
-        // Index pointer of multipole and local data, indexed by fmm index, then by level
-        let level_index_pointer_multipoles = level_index_pointer_multinode(
-            &self.tree.source_tree.trees,
-            self.tree.source_tree.local_depth,
-            self.tree.source_tree.global_depth,
-        );
-        let level_index_pointer_locals = level_index_pointer_multinode(
-            &self.tree.target_tree.trees,
-            self.tree.target_tree.local_depth,
[…]
