+++ kifmm/src/tree/multi_node.rs
@@ -1,54 +1,39 @@
 //! Implementation of constructors for MPI distributed multi node trees, from distributed point data.
 use crate::{
-    traits::tree::Tree,
+    sorting::{hyksort, samplesort, simplesort},
+    traits::tree::{MultiTree, SingleTree},
     tree::{
         constants::DEEPEST_LEVEL,
-        types::{Domain, MortonKey, MortonKeys, MultiNodeTree, Point, Points, SingleNodeTree},
+        types::{Domain, MortonKey, MortonKeys, Point, Points, SingleNodeTree},
     },
 };
 
-use crate::hyksort::hyksort;
 use itertools::Itertools;
 use mpi::topology::SimpleCommunicator;
-use mpi::{
-    traits::{Communicator, CommunicatorCollectives, Destination, Equivalence, Source},
-    Rank,
-};
+use mpi::traits::{Communicator, CommunicatorCollectives, Equivalence};
 use num::Float;
 use rlst::RlstScalar;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
+
+use super::types::{MultiNodeTree, SortKind};
 
 impl<T, C: Communicator> MultiNodeTree<T, C>
 where
     T: RlstScalar + Float + Equivalence + Default,
 {
-    /// Constructor for uniform trees, distributed with MPI, node refined to a user defined depth.
-    ///
-    /// The input point data is also assumed to be distributed across each node.
-    ///
-    /// # Arguments
-    /// * `coordinates_row_major` - A slice of point coordinates, expected in row major order.
-    /// [x_1, y_1, z_1,...x_N, y_N, z_N]
-    /// * `domain` - The (global) physical domain with which Morton Keys are being constructed with respect to.
-    /// * `depth` - The maximum depth of the tree, defines the level of recursion.
-    /// * `global_indices` - A slice of indices to uniquely identify the points.
-    /// * `world` - a global communicator for the tree.
-    /// * `hyksort_subcomm_size` - size of sub-communicator used by
-    /// [hyksort](https://dl.acm.org/doi/abs/10.1145/2464996.2465442?casa_token=vfaxtoyb_xsaaaaa:dqq1hfnp_gokaatn_d0svex37v_xooiqevdrong-4lyn_pmsuphmr3cp-0qvbisxtbwvuucaua).
-    /// must be a power of 2.
+    /// Construct uniform tree, pruned by default
     pub fn uniform_tree(
         coordinates_row_major: &[T],
         &domain: &Domain<T>,
-        depth: u64,
+        local_depth: u64,
+        global_depth: u64,
         global_indices: &[usize],
         world: &C,
-        hyksort_subcomm_size: i32,
+        sort_kind: SortKind,
+        prune_empty: bool,
     ) -> Result<MultiNodeTree<T, SimpleCommunicator>, std::io::Error> {
-        // Size of global communicator, and processor rank
-        let size = world.size();
         let rank = world.rank();
 
-        // Convert column major coordinate into `Point`, containing Morton encoding
         let dim = 3;
         let n_coords = coordinates_row_major.len() / dim;
 
@@ -58,7 +43,7 @@ where
                 .try_into()
                 .unwrap();
             let base_key = MortonKey::from_point(coord, &domain, DEEPEST_LEVEL);
-            let encoded_key = MortonKey::from_point(coord, &domain, depth);
+            let encoded_key = MortonKey::from_point(coord, &domain, global_depth);
 
             points.push(Point {
                 coordinate: *coord,
@@ -70,429 +55,96 @@ where
 
         // Perform parallel Morton sort over encoded points
         let comm = world.duplicate();
-        hyksort(&mut points, hyksort_subcomm_size, comm)?;
[…]
