+++ kifmm/src/tree/single_node.rs
@@ -8,7 +8,7 @@ use itertools::Itertools;
 use rlst::RlstScalar;
 
 use crate::{
-    traits::tree::Tree,
+    traits::tree::SingleTree,
     tree::{
         constants::{DEEPEST_LEVEL, LEVEL_SIZE},
         morton::encode_anchor,
@@ -36,6 +36,7 @@ where
     ) -> Result<SingleNodeTree<T>, std::io::Error> {
         let dim = 3;
         let n_coords = coordinates_row_major.len() / dim;
+        let root = MortonKey::root();
 
         // Convert column major coordinate into `Point`, containing Morton encoding
         let mut points: Vec<Point<T>> = Points::default();
@@ -168,6 +169,7 @@ where
         }
 
         Ok(SingleNodeTree {
+            root,
             depth,
             coordinates,
             global_indices,
@@ -199,6 +201,7 @@ where
         depth: u64,
         global_indices: &[usize],
     ) -> Result<SingleNodeTree<T>, std::io::Error> {
+        let root = MortonKey::root();
         let dim = 3;
         let n_coords = coordinates_row_major.len() / dim;
 
@@ -326,6 +329,7 @@ where
         }
 
         Ok(SingleNodeTree {
+            root,
             depth,
             coordinates,
             global_indices,
@@ -340,6 +344,219 @@ where
             keys_set,
             levels_to_keys,
         })
+    }
+
+    /// Constructor for uniform trees on a single node refined to a user defined depth, however excludes
+    /// empty nodes which don't contain particles and their ancestors, specify roots.
+    ///
+    /// # Arguments
+    /// * `coordinates_row_major` - A slice of point coordinates, expected in row major order.
+    /// [x_1, y_1, z_1,...x_N, y_N, z_N]
+    /// * `domain` - The physical domain with which Morton Keys are being constructed with respect to.
+    /// * `depth` - The maximum depth of the tree, defines the level of recursion.
+    /// * `global_indices` - A slice of indices to uniquely identify the points.
+    fn uniform_tree_pruned_with_root(
+        coordinates_row_major: &[T],
+        &domain: &Domain<T>,
+        depth: u64,
+        global_indices: &[usize],
+        root: MortonKey<T>,
+    ) -> Result<SingleNodeTree<T>, std::io::Error> {
+        let dim = 3;
+        let n_coords = coordinates_row_major.len() / dim;
+
+        // Convert column major coordinate into `Point`, containing Morton encoding
+        let mut points = Points::default();
+        for i in 0..n_coords {
+            let coord: &[T; 3] = &coordinates_row_major[i * dim..(i + 1) * dim]
+                .try_into()
+                .unwrap();
+
+            let base_key = MortonKey::from_point(coord, &domain, DEEPEST_LEVEL);
+            let encoded_key = MortonKey::from_point(coord, &domain, depth);
+            points.push(Point {
+                coordinate: *coord,
+                base_key,
+                encoded_key,
+                global_index: global_indices[i],
+            })
+        }
+
+        // Morton sort over points
+        points.sort();
+
+        // Group coordinates by leaves
+        let mut leaves_to_coordinates = HashMap::new();
+        let mut curr = points[0];
+        let mut curr_idx = 0;
+
+        for (i, point) in points.iter().enumerate() {
+            if point.encoded_key != curr.encoded_key {
+                leaves_to_coordinates.insert(curr.encoded_key, (curr_idx, i));
+                curr_idx = i;
+                curr = *point;
+            }
+        }
+        leaves_to_coordinates.insert(curr.encoded_key, (curr_idx, points.len()));
+
+        // Ensure that final leaf set contains siblings of all encoded keys
[…]
