+++ kifmm/src/fmm/field_translation/target.rs
@@ -13,7 +13,7 @@ use crate::{
     traits::{
         field::SourceToTargetData as SourceToTargetDataTrait,
         fmm::{FmmOperatorData, HomogenousKernel, TargetTranslation},
-        tree::{FmmTree, Tree},
+        tree::{SingleFmmTree, SingleTree},
         types::FmmError,
     },
     tree::{constants::NSIBLINGS, types::MortonKey},
