### 2024-09-17 69dc833
Lint as I go
