+++ kifmm/src/tree/single_node.rs
@@ -559,15 +559,19 @@ where
         result
     }
 
+    /// Construct a single node tree from U list ghost octants
     pub fn from_ghost_octants_u(
-        depth: u64,
-        coordinates_row_major: &[T],
-        global_indices: &[usize],
-        leaves: &MortonKeys<T>,
+        _depth: u64,
+        _coordinates_row_major: &[T],
+        _global_indices: &[usize],
+        _leaves: &MortonKeys<T>,
     ) {
     }
 
-    pub fn from_ghost_octants_v(depth: u64, keys: &MortonKeys<T>) {}
+    /// Construct a single node tree from V list ghost octants
+    pub fn from_ghost_octants_v(
+        _depth: u64,
+        _keys: &MortonKeys<T>) {}
 
     /// Calculates the minimum depth of a tree required to ensure that each leaf box contains at most `n_crit` points,
     /// assuming a uniform distribution of points.
