+++ kifmm/src/traits/tree.rs
@@ -128,8 +128,6 @@ pub trait MultiTree {
 
 /// Interface for trees required by the FMM, which requires separate trees for the source and target particle data
 pub trait SingleFmmTree
-where
-    Self::Tree: SingleTree,
 {
     /// Tree associated with this FMM tree
     type Tree: SingleTree;
@@ -150,10 +148,11 @@ where
     ) -> Option<Vec<<Self::Tree as SingleTree>::Node>>;
 }
 
+
 /// Interface for trees required by multinde FMM, which require separate trees for source and target data.
 /// Now however, each MPI rank can contain multiple single node source and target trees.
 #[cfg(feature = "mpi")]
-pub trait MultiNodeFmmTreeTrait {
+pub trait MultiFmmTree {
     /// Tree associated with FMM tree
     type Tree: MultiTree;
 
