+++ kifmm/src/fmm/types.rs
@@ -1,5 +1,5 @@
 //! Data structures for kernel independent FMM
-use std::{collections::HashMap, sync::RwLock};
+use std::{collections::{HashMap, HashSet}, sync::RwLock};
 
 use green_kernels::{traits::Kernel as KernelTrait, types::EvalType};
 use num::traits::Float;
@@ -8,14 +8,16 @@ use rlst::{rlst_dynamic_array2, Array, BaseArray, RlstScalar, SliceContainer, Ve
 use crate::{
     linalg::rsvd::Normaliser,
     traits::{
-        fftw::Dft, field::SourceToTargetData as SourceToTargetDataTrait, fmm::HomogenousKernel,
-        general::AsComplex, types::FmmOperatorTime,
+        fftw::Dft, field::SourceToTargetData as SourceToTargetDataTrait, fmm::HomogenousKernel, general::AsComplex, types::FmmOperatorTime
     },
     tree::types::{Domain, MortonKey, SingleNodeTree},
 };
 
 #[cfg(feature = "mpi")]
 use crate::tree::types::MultiNodeTree;
+
+#[cfg(feature = "mpi")]
+use mpi::topology::SimpleCommunicator;
 
 #[cfg(feature = "mpi")]
 use mpi::traits::{Communicator, Equivalence};
@@ -910,4 +912,220 @@ pub enum Isa {
     /// Default is no vectorisation
     #[default]
     Default,
+}
+
+
+/// Data structure holding data for multinode FMM
+#[cfg(feature = "mpi")]
+#[allow(clippy::type_complexity)]
+pub struct KiFmmMulti<Scalar, Kernel, SourceToTargetData>
+where
+    Scalar: RlstScalar + Equivalence + Float,
+    <Scalar as RlstScalar>::Real: RlstScalar + Equivalence + Float,
+    Kernel: KernelTrait<T = Scalar> + HomogenousKernel,
+    SourceToTargetData: SourceToTargetDataTrait,
+{
+
+    /// Operator runtimes
+    pub times: Vec<FmmOperatorTime>,
+
+    /// Instruction set architecture
+    pub isa: Isa,
+
+    /// Associated MPI communicator
+    pub communicator: SimpleCommunicator,
+
+    /// Neighbourhood communicator for V list communication
+    pub neighbourhood_communicator_v: NeighbourhoodCommunicator,
+
+    /// Neighbourhood communicator for U list communication
+    pub neighbourhood_communicator_u: NeighbourhoodCommunicator,
+
+    /// Associated MPI rank
+    pub rank: i32,
+
+    /// The associated kernel function
+    pub kernel: Kernel,
+
+    /// A multi node tree
+    pub tree: MultiNodeFmmTree<<Scalar as RlstScalar>::Real, SimpleCommunicator>,
+
+    /// Charges associated with each source tree
+    pub charges: Vec<Vec<Scalar>>,
+
+    /// The expansion order used to construct check surfaces
+    pub check_surface_order: usize,
+
+    /// The expansion order of the FMM, used to construct equivalent surfaces.
+    pub equivalent_surface_order: usize,
+
+    /// The number of coefficients, corresponding to points discretising the equivalent surface
+    pub ncoeffs_equivalent_surface: usize,
+
+    /// The number of coefficients, corresponding to points discretising the check surface
+    pub ncoeffs_check_surface: usize,
+
+    /// Set by the kernel evaluation type, either 1 or 4 corresponding to evaluating potentials or potentials and derivatives
+    pub kernel_eval_type: EvalType,
+
+    /// The FMM evaluation type, either for a vector or matrix of input charges.
+    pub fmm_eval_type: FmmEvalType,
+
[…]
