### 2024-09-17 5f57d6d
Compiling with just traits
