+++ kifmm/src/fmm/builder/multi_node.rs
@@ -1,0 +1,245 @@
+use mpi::{
+    topology::{Communicator, SimpleCommunicator},
+    traits::Equivalence,
+};
+use num::Float;
+use rlst::{MatrixSvd, RlstScalar};
+
+use crate::{
+    fmm::{
+        helpers::ncoeffs_kifmm,
+        types::{
+            FmmEvalType, Isa, KiFmmMulti, Layout, MultiNodeBuilder, NeighbourhoodCommunicator,
+        },
+    },
+    traits::{
+        field::{
+            SourceAndTargetTranslationMetadata, SourceToTargetData as SourceToTargetDataTrait,
+            SourceToTargetTranslationMetadata,
+        },
+        fmm::{FmmMetadata, HomogenousKernel},
+        general::Epsilon,
+    },
+    tree::{
+        types::{Domain, SortKind},
+        MultiNodeTree,
+    },
+    MultiNodeFmmTree,
+};
+use green_kernels::{traits::Kernel as KernelTrait, types::EvalType};
+
+impl<Scalar, Kernel, SourceToTargetData> MultiNodeBuilder<Scalar, Kernel, SourceToTargetData>
+where
+    Scalar: RlstScalar + Default + Epsilon + MatrixSvd + Equivalence + Float,
+    <Scalar as RlstScalar>::Real: Default + Epsilon + Equivalence + Float,
+    Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Clone + Default,
+    SourceToTargetData: SourceToTargetDataTrait + Default,
+    KiFmmMulti<Scalar, Kernel, SourceToTargetData>: SourceToTargetTranslationMetadata
+        + SourceAndTargetTranslationMetadata
+        + FmmMetadata<Scalar = Scalar>,
+{
+    /// Init
+    pub fn new() -> Self {
+        Self {
+            communicator: None,
+            isa: None,
+            tree: None,
+            kernel: None,
+            charges: None,
+            source_to_target: None,
+            domain: None,
+            equivalent_surface_order: None,
+            check_surface_order: None,
+            ncoeffs_equivalent_surface: None,
+            ncoeffs_check_surface: None,
+            kernel_eval_type: None,
+            fmm_eval_type: None,
+        }
+    }
+
+    /// Tree
+    pub fn tree(
+        mut self,
+        comm: &SimpleCommunicator,
+        sources: &[Scalar::Real],
+        targets: &[Scalar::Real],
+        local_depth: u64,
+        global_depth: u64,
+        prune_empty: bool,
+        sort_kind: SortKind,
+    ) -> Result<Self, std::io::Error> {
+        let dim = 3;
+        let nsources = sources.len() / dim;
+        let ntargets = targets.len() / dim;
+
+        let dims = sources.len() % dim;
+        let dimt = targets.len() % dim;
+
+        if dims != 0 || dimt != 0 {
+            Err(std::io::Error::new(
+                std::io::ErrorKind::InvalidData,
+                "Only 3D FMM supported",
+            ))
+        } else if nsources == 0 || ntargets == 0 {
+            Err(std::io::Error::new(
+                std::io::ErrorKind::InvalidData,
+                "Must have a positive number of source or target particles",
+            ))
+        } else {
+            // Source and target trees calcualted over the same domain
+            let source_domain = Domain::from_global_points(sources, comm);
+            let target_domain = Domain::from_global_points(targets, comm);
+
+            // Calculate union of domains for source and target points, needed to define operators
+            let domain = source_domain.union(&target_domain);
+
+            let source_tree = MultiNodeTree::new(
+                comm,
+                sources,
+                local_depth,
+                global_depth,
+                Some(domain),
+                sort_kind.clone(),
+                prune_empty,
+            )?;
+
[…]

[…]
