### 2024-09-17 c27be90
Finally in position to add back metadata functions
