+++ kifmm/src/fmm/field_translation/source_to_target/fft/single_node.rs
@@ -29,7 +29,7 @@ use crate::{
         constants::{NHALO, NSIBLINGS, NSIBLINGS_SQUARED},
         types::MortonKey,
     },
-    FftFieldTranslation, Fmm,
+    FftFieldTranslation, SingleFmm,
 };
 
 impl<Scalar, Kernel> SourceToTargetTranslation
