+++ kifmm/src/traits.rs
@@ -5,3 +5,6 @@ pub mod fmm;
 pub mod general;
 pub mod tree;
 pub mod types;
+
+#[cfg(feature = "mpi")]
+pub mod parallel;
