### 2024-09-18 36afe94
Linear multinode trees
