+++ kifmm/src/fmm/helpers/mod.rs
@@ -1,0 +1,4 @@
+//! Helper routines
+#[cfg(feature = "mpi")]
+pub mod multi_node;
+pub mod single_node;
