+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -8,6 +8,7 @@ use rlst::{
 };
 use std::collections::HashMap;
 
+use crate::fmm::helpers::level_index_pointer_multi_node;
 use crate::traits::fftw::Dft;
 use crate::traits::fmm::{FmmMetadata, MultiFmm};
 use crate::traits::general::AsComplex;
@@ -80,7 +81,6 @@ where
     type Scalar = Scalar;
 
     fn metadata(&mut self, eval_type: EvalType, charges: &[Self::Scalar]) {
-        let dim = self.dim();
         let alpha_outer = Scalar::real(ALPHA_OUTER);
         let alpha_inner = Scalar::real(ALPHA_INNER);
 
@@ -96,63 +96,34 @@ where
             EvalType::ValueDeriv => 4,
         };
 
+        let n_target_points = self.tree.target_tree.n_coordinates_tot().unwrap();
+        let n_source_points = self.tree.source_tree.n_coordinates_tot().unwrap();
         let source_trees = self.tree.source_tree().trees();
         let target_trees = self.tree.target_tree().trees();
+        let n_source_keys = self.tree.source_tree().n_keys_tot().unwrap();
+        let n_target_keys = self.tree.target_tree().n_keys_tot().unwrap();
         let global_depth = self.tree.source_tree().global_depth();
         let local_depth = self.tree.source_tree().local_depth();
         let total_depth = self.tree.source_tree().local_depth();
 
-        // Allocate buffers to store locally available multipole, local and potential data
-        // multipole and local data is arranged by level
+        // Allocate multipole and local buffers for all locally owned source/target octants
+        let mut multipoles =
+            vec![Scalar::default(); n_source_keys * self.ncoeffs_equivalent_surface];
+        let mut locals = vec![Scalar::default(); n_target_keys * self.ncoeffs_equivalent_surface];
 
-        let mut source_counts = Vec::new(); // arranged by level, then by tree
-        let mut source_displacement = Vec::new(); // arranged by level, then by tree
-        let mut n_sources = 0; // total number of source boxes
-        let mut level_displacement = 0;
+        // Index pointers of multipole and local data, indexed by level
+        let level_index_pointer_multipoles = level_index_pointer_multi_node(&self.tree.source_tree);
+        let level_index_pointer_locals = level_index_pointer_multi_node(&self.tree.target_tree);
 
-        let mut source_to_index = HashMap::new();
+        // Allocate buffers for local potential data
+        let potentials = vec![Scalar::default(); n_target_points * eval_size];
 
-        for level in global_depth..=total_depth {
-            let mut tree_displacement = Vec::new();
-            let mut tree_counts = Vec::new();
-            let mut tree_displacement_ = level_displacement;
-
-            for tree in source_trees.iter() {
-                let keys = tree.keys(level).unwrap();
-                let n_keys = keys.len();
-                tree_displacement.push(tree_displacement_);
-                tree_counts.push(n_keys);
-
-                for (key_idx, k) in keys.iter().enumerate() {
-                    source_to_index.insert(*k, key_idx + tree_displacement_);
-                }
-
-                tree_displacement_ += n_keys;
-                n_sources += n_keys;
-            }
-
-            level_displacement += tree_displacement_;
-            source_counts.push(tree_counts);
-            source_displacement.push(tree_displacement);
-        }
-
-        let mut tree_displacement = 0;
-        let mut source_to_leaf_index = HashMap::new();
-        for tree in source_trees.iter() {
-            for (leaf_idx, leaf) in tree.all_leaves().unwrap().iter().enumerate() {
-                source_to_leaf_index.insert(*leaf, leaf_idx + tree_displacement);
-            }
-
[…]
