+++ kifmm/src/fmm/field_translation/source_to_target/blas/single_node.rs
@@ -13,7 +13,7 @@ use green_kernels::traits::Kernel as KernelTrait;
 
 use crate::{
     fmm::{
-        helpers::{homogenous_kernel_scale, m2l_scale},
+        helpers::single_node::{homogenous_kernel_scale, m2l_scale},
         types::{BlasFieldTranslationIa, FmmEvalType, SendPtrMut},
         KiFmm,
     },
