### 2024-09-18 a3b6dea
Run cargo check for path changes
