+++ kifmm/src/fmm/helpers/multi_node.rs
@@ -4,7 +4,10 @@ use mpi::traits::{Communicator, Equivalence};
 use num::Float;
 use rlst::RlstScalar;
 
-use crate::tree::{types::MortonKey, MultiNodeTree};
+use crate::{
+    traits::tree::MultiTree,
+    tree::{types::MortonKey, MultiNodeTree},
+};
 
 /// Create index pointers for each key at each level of an octree
 pub fn level_index_pointer_multi_node<T, C>(
@@ -14,9 +17,9 @@ where
     T: RlstScalar + Float + Equivalence,
     C: Communicator,
 {
-    let mut result = vec![HashMap::new(); (tree.depth() + 1).try_into().unwrap()];
+    let mut result = vec![HashMap::new(); (tree.total_depth() + 1).try_into().unwrap()];
 
-    for level in 0..=tree.total_depth {
+    for level in 0..=tree.total_depth() {
         if let Some(keys) = tree.keys(level) {
             for (level_idx, key) in keys.iter().enumerate() {
                 result[level as usize].insert(*key, level_idx);
