+++ kifmm/src/fmm/field_translation/source_to_target/fft/multi_node.rs
@@ -15,7 +15,7 @@ use green_kernels::traits::Kernel as KernelTrait;
 use crate::{
     fftw::array::{AlignedAllocable, AlignedVec},
     fmm::{
-        helpers::{chunk_size, homogenous_kernel_scale, m2l_scale},
+        helpers::single_node::{chunk_size, homogenous_kernel_scale, m2l_scale},
         types::{FmmEvalType, KiFmmMulti, SendPtrMut},
         KiFmm,
     },
