+++ kifmm/src/fmm/helpers/single_node.rs
@@ -76,27 +76,20 @@ pub fn m2l_scale<T: RlstScalar>(level: u64) -> Result<T, std::io::Error> {
 ///
 /// # Arguments
 /// * `tree`- Single node tree
-/// * `ncoeffs`- Number of interpolation points on leaf box
-pub fn leaf_scales<T>(
-    tree: &SingleNodeTree<T::Real>,
-    homogenous: bool,
-    ncoeffs_leaf: usize,
-) -> Vec<T>
+/// * `n_coeffs`- Number of interpolation points on leaf box
+pub fn leaf_scales_single_node<T>(tree: &SingleNodeTree<T::Real>, n_coeffs_leaf: usize) -> Vec<T>
 where
     T: RlstScalar + Default,
 {
-    let mut result = vec![T::default(); tree.n_leaves().unwrap() * ncoeffs_leaf];
+    let mut result = vec![T::default(); tree.n_leaves().unwrap() * n_coeffs_leaf];
 
-    if homogenous {
-        for (i, leaf) in tree.all_leaves().unwrap().iter().enumerate() {
-            // Assign scales
-            let l = i * ncoeffs_leaf;
-            let r = l + ncoeffs_leaf;
+    for (i, leaf) in tree.all_leaves().unwrap().iter().enumerate() {
+        // Assign scales
+        let l = i * n_coeffs_leaf;
+        let r = l + n_coeffs_leaf;
 
-            result[l..r].copy_from_slice(
-                vec![homogenous_kernel_scale(leaf.level()); ncoeffs_leaf].as_slice(),
-            );
-        }
+        result[l..r]
+            .copy_from_slice(vec![homogenous_kernel_scale(leaf.level()); n_coeffs_leaf].as_slice());
     }
     result
 }
@@ -105,12 +98,12 @@ where
 ///
 /// # Arguments
 /// * `tree`- Single node tree
-/// * `ncoeffs`- Number of interpolation points on leaf box
+/// * `n_coeffs`- Number of interpolation points on leaf box
 /// * `alpha` - The multiplier being used to modify the diameter of the surface grid uniformly along each coordinate axis.
 /// * `expansion_order` - Expansion order of the FMM
-pub fn leaf_surfaces<T>(
+pub fn leaf_surfaces_single_node<T>(
     tree: &SingleNodeTree<T>,
-    ncoeffs_leaf: usize,
+    n_coeffs_leaf: usize,
     alpha: T,
     expansion_order_leaf: usize,
 ) -> Vec<T>
@@ -119,12 +112,12 @@ where
 {
     let dim = 3;
     let n_keys = tree.n_leaves().unwrap();
-    let mut result = vec![T::default(); ncoeffs_leaf * dim * n_keys];
+    let mut result = vec![T::default(); n_coeffs_leaf * dim * n_keys];
 
-    for (i, key) in tree.all_leaves().unwrap().iter().enumerate() {
-        let l = i * ncoeffs_leaf * dim;
-        let r = l + ncoeffs_leaf * dim;
-        let surface = key.surface_grid(expansion_order_leaf, tree.domain(), alpha);
+    for (i, leaf) in tree.all_leaves().unwrap().iter().enumerate() {
+        let l = i * n_coeffs_leaf * dim;
+        let r = l + n_coeffs_leaf * dim;
+        let surface = leaf.surface_grid(expansion_order_leaf, tree.domain(), alpha);
 
         result[l..r].copy_from_slice(&surface);
     }
@@ -134,7 +127,7 @@ where
 
 /// Create an index pointer for the coordinates in a source and a target tree
 /// between the local indices for each leaf and their associated charges
-pub fn coordinate_index_pointer<T>(tree: &SingleNodeTree<T>) -> Vec<(usize, usize)>
[…]
