+++ kifmm/src/fmm/helpers/multi_node.rs
@@ -1,21 +1,29 @@
-use std::collections::HashMap;
+//! Helper functions for MPI setting
+use std::collections::{HashMap, HashSet};
 
-use mpi::traits::{Communicator, Equivalence};
+use itertools::Itertools;
+use mpi::{
+    topology::SimpleCommunicator,
+    traits::{Communicator, Equivalence},
+    Count, Rank,
+};
 use num::Float;
 use rlst::RlstScalar;
 
 use crate::{
-    traits::tree::MultiTree,
+    fmm::types::{Layout, SendPtrMut},
+    traits::tree::{FmmTreeNode, MultiTree},
     tree::{types::MortonKey, MultiNodeTree},
 };
 
+use super::single_node::homogenous_kernel_scale;
+
 /// Create index pointers for each key at each level of an octree
-pub fn level_index_pointer_multi_node<T, C>(
-    tree: &MultiNodeTree<T, C>,
+pub fn level_index_pointer_multi_node<T>(
+    tree: &MultiNodeTree<T, SimpleCommunicator>,
 ) -> Vec<HashMap<MortonKey<T>, usize>>
 where
     T: RlstScalar + Float + Equivalence,
-    C: Communicator,
 {
     let mut result = vec![HashMap::new(); (tree.total_depth() + 1).try_into().unwrap()];
 
@@ -28,3 +36,278 @@ where
     }
     result
 }
+
+/// Compute the scaling for each leaf box in a tree
+///
+/// # Arguments
+/// * `tree`- Multi node tree
+/// * `ncoeffs`- Number of interpolation points on leaf box
+pub fn leaf_scales_multi_node<T>(
+    tree: &MultiNodeTree<T::Real, SimpleCommunicator>,
+    n_coeffs_leaf: usize,
+) -> Vec<T>
+where
+    T: RlstScalar + Float + Equivalence,
+    <T as RlstScalar>::Real: Float + Equivalence,
+{
+    let mut result = vec![T::default(); tree.n_leaves().unwrap() * n_coeffs_leaf];
+
+    for (i, leaf) in tree.all_leaves().unwrap().iter().enumerate() {
+        // Assign scales
+        let l = i * n_coeffs_leaf;
+        let r = l + n_coeffs_leaf;
+
+        result[l..r]
+            .copy_from_slice(vec![homogenous_kernel_scale(leaf.level()); n_coeffs_leaf].as_slice());
+    }
+    result
+}
+
+/// Compute surfaces for each leaf box
+pub fn leaf_surfaces_multi_node<T>(
+    tree: &MultiNodeTree<T, SimpleCommunicator>,
+    n_coeffs_leaf: usize,
+    alpha: T,
+    expansion_order_leaf: usize,
+) -> Vec<T>
+where
+    T: RlstScalar + Equivalence + Float + Default,
+{
+    let dim = 3;
+    let n_keys = tree.n_leaves().unwrap();
+    let mut result = vec![T::default(); n_coeffs_leaf * dim * n_keys];
+
+    for (i, leaf) in tree.all_leaves().unwrap().iter().enumerate() {
+        let l = i * n_coeffs_leaf * dim;
+        let r = l + n_coeffs_leaf * dim;
+        let surface = leaf.surface_grid(expansion_order_leaf, tree.domain(), alpha);
+
+        result[l..r].copy_from_slice(&surface);
+    }
+
+    result
+}
+
+/// Create mutable pointers corresponding to each multipole expansion at each level of an octree
+pub fn level_expansion_pointers_multi_node<T>(
+    tree: &MultiNodeTree<T::Real, SimpleCommunicator>,
+    n_coeffs: usize,
+    expansions: &[T],
[…]
