+++ kifmm/src/fmm/field_translation/source/multi_node.rs
@@ -13,21 +13,29 @@ use rlst::{
 
 use green_kernels::{traits::Kernel as KernelTrait, types::EvalType};
 
-use crate::{fmm::{constants::P2M_MAX_BLOCK_SIZE, helpers::single_node::chunk_size, types::{FmmEvalType, KiFmmMulti}}, traits::{field::SourceToTargetData as SourceToTargetDataTrait, fmm::{HomogenousKernel, SourceTranslation}, tree::{MultiFmmTree, MultiTree}}};
+use crate::{
+    fmm::{
+        constants::P2M_MAX_BLOCK_SIZE,
+        helpers::single_node::chunk_size,
+        types::{FmmEvalType, KiFmmMulti},
+    },
+    traits::{
+        field::SourceToTargetData as SourceToTargetDataTrait,
+        fmm::{HomogenousKernel, SourceTranslation},
+        tree::{MultiFmmTree, MultiTree},
+    },
+};
 
-
-impl <Scalar, Kernel, SourceToTargetData> SourceTranslation for KiFmmMulti<Scalar, Kernel, SourceToTargetData>
+impl<Scalar, Kernel, SourceToTargetData> SourceTranslation
+    for KiFmmMulti<Scalar, Kernel, SourceToTargetData>
 where
     Scalar: RlstScalar + Default + Equivalence + Float,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     SourceToTargetData: SourceToTargetDataTrait,
-    Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync
-
+    Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
 {
     fn p2m(&self) -> Result<(), crate::traits::types::FmmError> {
-
         if let Some(leaves) = self.tree.source_tree().all_leaves() {
-
             let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface;
             let n_coeffs_check_surface = self.n_coeffs_check_surface;
             let n_leaves = self.tree.source_tree().n_leaves().unwrap();
@@ -42,7 +50,8 @@ where
 
             match self.fmm_eval_type {
                 FmmEvalType::Vector => {
-                    let mut check_potentials = rlst_dynamic_array2!(Scalar, [n_leaves * n_coeffs_check_surface, 1]);
+                    let mut check_potentials =
+                        rlst_dynamic_array2!(Scalar, [n_leaves * n_coeffs_check_surface, 1]);
 
                     // Compute check potential for each box
                     check_potentials
@@ -58,9 +67,8 @@ where
                                 let charges =
                                     &all_charges[charge_index_pointer.0..charge_index_pointer.1];
 
-                                let coordinates_row_major = &coordinates[charge_index_pointer.0
-                                    * dim
-                                    ..charge_index_pointer.1 * dim];
+                                let coordinates_row_major = &coordinates
+                                    [charge_index_pointer.0 * dim..charge_index_pointer.1 * dim];
                                 let nsources = coordinates_row_major.len() / dim;
 
                                 if nsources > 0 {
@@ -76,28 +84,18 @@ where
                         );
 
                     let chunk_size = chunk_size(n_leaves, P2M_MAX_BLOCK_SIZE);
-                    
-
-
-
                 }
 
                 FmmEvalType::Matrix(_n) => {
                     panic!("unimplemented for matrix input")
                 }
             }
-
-
         }
 
         Ok(())
     }
 
-
     fn m2m(&self, level: u64) -> Result<(), crate::traits::types::FmmError> {
-
         Ok(())
     }
-
-
 }
