+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -5,7 +5,8 @@ use mpi::topology::SimpleCommunicator;
 use mpi::traits::Equivalence;
 use num::Float;
 use rlst::{
-    empty_array, rlst_dynamic_array2, rlst_dynamic_array3, Array, BaseArray, MatrixSvd, MultIntoResize, RawAccess, RawAccessMut, RlstScalar, VectorContainer
+    empty_array, rlst_dynamic_array2, rlst_dynamic_array3, Array, BaseArray, MatrixSvd,
+    MultIntoResize, RawAccess, RawAccessMut, RlstScalar, VectorContainer,
 };
 use std::collections::HashMap;
 
@@ -46,7 +47,6 @@ where
     SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
 {
     fn source(&mut self) {
-
         let root = MortonKey::<Scalar::Real>::root();
         let equivalent_surface_order = self.equivalent_surface_order;
         let check_surface_order = self.check_surface_order;
@@ -58,31 +58,30 @@ where
         let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
         let domain = self.tree.domain();
 
-        let mut m2m= rlst_dynamic_array2!(Scalar, [n_coeffs_equivalent_surface, 8*n_coeffs_equivalent_surface]);
-        let mut m2m_vec= Vec::new();
+        let mut m2m = rlst_dynamic_array2!(
+            Scalar,
+            [n_coeffs_equivalent_surface, 8 * n_coeffs_equivalent_surface]
+        );
+        let mut m2m_vec = Vec::new();
 
         // Compute required surfaces
-        let upward_equivalent_surface = root.surface_grid(
-            equivalent_surface_order,
-            domain,
-            alpha_inner
-        );
+        let upward_equivalent_surface =
+            root.surface_grid(equivalent_surface_order, domain, alpha_inner);
 
-        let upward_check_surface = root.surface_grid(
-            check_surface_order,
-            domain,
-            alpha_outer
-        );
+        let upward_check_surface = root.surface_grid(check_surface_order, domain, alpha_outer);
 
         // Assemble matrix of kernel evaluations between upward check to equivalent, and downward check to equivalent matrices
         // As well as estimating their inverses using SVD
-        let mut uc2e = rlst_dynamic_array2!(Scalar, [n_coeffs_check_surface, n_coeffs_equivalent_surface]);
+        let mut uc2e = rlst_dynamic_array2!(
+            Scalar,
+            [n_coeffs_check_surface, n_coeffs_equivalent_surface]
+        );
 
         self.kernel.assemble_st(
             EvalType::Value,
             &upward_check_surface,
             &upward_equivalent_surface,
-            uc2e.data_mut()
+            uc2e.data_mut(),
         );
 
         let (s, ut, v) = pinv(&uc2e, None, None).unwrap();
@@ -92,38 +91,35 @@ where
             mat_s[[i, i]] = Scalar::from_real(s[i]);
         }
 
-        let uc2e_inv_1 = vec![
-            empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view())
-        ];
+        let uc2e_inv_1 =
+            vec![empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view())];
 
-        let uc2e_inv_2 = vec![
-            ut
-        ];
+        let uc2e_inv_2 = vec![ut];
 
-        let parent_upward_check_surface = root.surface_grid(check_surface_order, domain, alpha_outer);
+        let parent_upward_check_surface =
+            root.surface_grid(check_surface_order, domain, alpha_outer);
         let children = root.children();
-
 
[…]
