+++ kifmm/src/fmm/ghost_exchange.rs
@@ -7,18 +7,29 @@ use mpi::datatype::Partition;
 use mpi::datatype::PartitionMut;
 use mpi::datatype::Partitioned;
 use mpi::topology::Communicator;
+use mpi::topology::SimpleCommunicator;
 use mpi::traits::Equivalence;
 use mpi::traits::Root;
+use mpi::Count;
 use num::Float;
+use pulp::Scalar;
 use rlst::RlstScalar;
 
 use crate::traits::field::SourceToTargetData as SourceToTargetDataTrait;
 use crate::traits::fmm::HomogenousKernel;
+use crate::traits::fmm::MultiFmm;
+use crate::traits::fmm::SourceToTargetTranslation;
 use crate::traits::parallel::GhostExchange;
+use crate::traits::tree::MultiFmmTree;
+use crate::traits::tree::MultiTree;
 use crate::traits::tree::SingleTree;
 use crate::tree::types::MortonKey;
+use crate::tree::MultiNodeTree;
+use crate::tree::SingleNodeTree;
+use crate::MultiNodeFmmTree;
 
 use super::types::{KiFmmMulti, Layout};
+use super::KiFmm;
 
 impl<Scalar, Kernel, SourceToTargetData> GhostExchange
     for KiFmmMulti<Scalar, Kernel, SourceToTargetData>
@@ -26,9 +37,98 @@ where
     Scalar: RlstScalar + Default + Equivalence + Float,
     <Scalar as RlstScalar>::Real: RlstScalar + Default + Equivalence + Float,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
-    SourceToTargetData: SourceToTargetDataTrait,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    Self: SourceToTargetTranslation,
 {
-    fn gather_global_fmm_at_root(&mut self) {}
+    fn gather_global_fmm_at_root(&mut self) {
+        let size = self.communicator.size();
+        let rank = self.communicator.rank();
+
+        // Nominated rank chosen to run the global upward pass
+        let root_rank = 0;
+        let root_process = self.communicator.process_at_rank(root_rank);
+
+        if rank == root_rank {
+            // 1. Gather multipole data from root processes on all ranks
+            let n_root_multipoles = self.tree.source_tree().n_trees();
+            let mut global_multipoles_counts = vec![0 as Count; size as usize];
+            root_process.gather_into_root(&n_root_multipoles, &mut global_multipoles_counts);
+
+            // 1.1 Calculate displacements and counts for associated morton keys
+            let mut global_multipoles_displacements = Vec::new();
+            let mut displacement = 0;
+            for &count in global_multipoles_counts.iter() {
+                global_multipoles_displacements.push(displacement);
+                displacement += count;
+            }
+
+            // 1.1 Allocate emory for locally contained data
+            let mut multipole_roots = Vec::new();
+            let mut multipoles =
+                vec![Scalar::default(); n_root_multipoles * self.n_coeffs_equivalent_surface];
+
+            for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
+                let tmp = self.multipole(&tree.root()).unwrap();
+                multipoles[i * self.n_coeffs_equivalent_surface
+                    ..(i + 1) * self.n_coeffs_equivalent_surface]
+                    .copy_from_slice(tmp);
+                multipole_roots.push(tree.root())
+            }
+
+            // 1.1 Calculate displacements and counts for multipole data
+            let global_multipoles_bufs_counts = global_multipoles_counts
+                .iter()
+                .map(|c| c * self.n_coeffs_equivalent_surface as i32)
+                .collect_vec();
[…]
