### 2024-09-19 8b4c154
Add all local metadata
