+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -1,29 +1,34 @@
 use green_kernels::laplace_3d::Laplace3dKernel;
 use green_kernels::traits::Kernel as KernelTrait;
 use green_kernels::types::EvalType;
-use mpi::topology::SimpleCommunicator;
+use itertools::Itertools;
 use mpi::traits::Equivalence;
-use num::Float;
+use num::{Float, Zero};
+use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 use rlst::{
-    empty_array, rlst_dynamic_array2, rlst_dynamic_array3, Array, BaseArray, MatrixSvd,
-    MultIntoResize, RawAccess, RawAccessMut, RlstScalar, VectorContainer,
+    empty_array, rlst_array_from_slice2, rlst_dynamic_array2, rlst_dynamic_array3, Array,
+    BaseArray, MatrixSvd, MultIntoResize, RawAccess, RawAccessMut, RlstScalar, Shape, SvdMode,
+    UnsafeRandomAccessByRef, UnsafeRandomAccessMut, VectorContainer,
 };
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
+use std::sync::{Mutex, RwLock};
 
 use crate::fmm::helpers::multi_node::{
     coordinate_index_pointer_multi_node, leaf_expansion_pointers_multi_node,
     leaf_scales_multi_node, leaf_surfaces_multi_node, level_expansion_pointers_multi_node,
     level_index_pointer_multi_node, potential_pointers_multi_node,
 };
-use crate::fmm::helpers::single_node::homogenous_kernel_scale;
+use crate::fmm::helpers::single_node::{flip3, homogenous_kernel_scale};
+use crate::fmm::types::{BlasMetadataSaRcmp, FftMetadata};
 use crate::linalg::pinv::pinv;
-use crate::traits::fftw::Dft;
+use crate::traits::fftw::{Dft, DftType};
 use crate::traits::fmm::{FmmMetadata, MultiFmm};
 use crate::traits::general::AsComplex;
 use crate::traits::parallel::GhostExchange;
-use crate::traits::tree::{FmmTreeNode, MultiFmmTree, MultiTree};
+use crate::traits::tree::{Domain, FmmTreeNode, MultiFmmTree, MultiTree};
+use crate::tree::constants::{NHALO, NSIBLINGS, NSIBLINGS_SQUARED, NTRANSFER_VECTORS_KIFMM};
+use crate::tree::helpers::find_corners;
 use crate::tree::types::MortonKey;
-use crate::FftFieldTranslation;
 use crate::{
     fmm::types::{KiFmmMulti, NeighbourhoodCommunicator},
     linalg::rsvd::MatrixRsvd,
@@ -32,12 +37,15 @@ use crate::{
             SourceAndTargetTranslationMetadata, SourceToTargetData as SourceToTargetDataTrait,
             SourceToTargetTranslationMetadata,
         },
-        fmm::{HomogenousKernel, SourceToTargetTranslation},
+        fmm::HomogenousKernel,
         general::Epsilon,
     },
     tree::constants::{ALPHA_INNER, ALPHA_OUTER},
     BlasFieldTranslationSaRcmp,
 };
+use crate::{FftFieldTranslation, FmmSvdMode};
+
+use super::single_node::find_cutoff_rank;
 
 impl<Scalar, SourceToTargetData> SourceAndTargetTranslationMetadata
     for KiFmmMulti<Scalar, Laplace3dKernel<Scalar>, SourceToTargetData>
@@ -221,9 +229,302 @@ where
     Scalar: RlstScalar + Default + MatrixRsvd + Equivalence + Float,
[…]
