### 2024-09-19 644fddf
Add global tree construction
