### 2024-09-19 78f83fd
Add global FMM scatter
