+++ kifmm/src/fmm/ghost_exchange.rs
@@ -25,7 +25,7 @@ use crate::traits::tree::MultiTree;
 use crate::traits::tree::SingleFmmTree;
 use crate::traits::tree::SingleTree;
 use crate::tree::types::Domain;
-use crate::tree::types::MortonKey;
+use crate::tree::types::{MortonKey, MortonKeys};
 use crate::tree::SingleNodeTree;
 use crate::SingleFmm;
 use crate::SingleNodeFmmTree;
@@ -349,7 +349,256 @@ where
 
     fn u_list_exchange(&mut self) {}
 
-    fn v_list_exchange(&mut self) {}
+    fn v_list_exchange(&mut self) {
+        // Begin by calculating the receive counts, this requires an all to all over the neighbourhood communicators
+        let size = self.neighbourhood_communicator_v.size();
+
+        let mut send_counts = vec![0 as Count; size as usize];
+
+        for (global_rank, &send_count) in self.tree.v_list_query.send_counts.iter().enumerate() {
+            if let Some(local_rank) = self
+                .neighbourhood_communicator_v
+                .global_to_local_rank(global_rank as Rank)
+            {
+                send_counts[local_rank as usize] = send_count;
+            }
+        }
+
+        let mut receive_counts = vec![0 as Rank; size as usize];
+        self.neighbourhood_communicator_v
+            .all_to_all_into(&send_counts, &mut receive_counts);
+
+        // Now can calculate displacements
+        let mut send_displacements = Vec::new();
+        let mut receive_displacements = Vec::new();
+        let mut send_counter = 0;
+        let mut receive_counter = 0;
+
+        for (send_count, receive_count) in send_counts.iter().zip(receive_counts.iter()) {
+            send_displacements.push(send_counter);
+            receive_displacements.push(receive_counter);
+            send_counter += send_count;
+            receive_counter += receive_count;
+        }
+
+        let total_receive_count = receive_counter;
+
+        // Assign origin ranks to all received queries
+        let mut receive_counter = 0;
+        let mut received_queries_source_ranks = vec![0i32; total_receive_count as usize];
+        for (i, &receive_count) in receive_counts.iter().enumerate() {
+            let curr_receive_counter = receive_counter;
+            receive_counter += receive_count as usize;
+
+            let rank = self.neighbourhood_communicator_v.neighbours[i];
+            let tmp = vec![rank as i32; receive_count as usize];
+            received_queries_source_ranks[curr_receive_counter..receive_counter]
+                .copy_from_slice(tmp.as_slice());
+        }
+
+        // Communicate queries for multipole data
+        let mut received_queries = vec![0u64; total_receive_count as usize];
+        {
+            let partition_send = Partition::new(
+                &self.tree.v_list_query.queries,
+                send_counts,
+                send_displacements,
+            );
+            let mut partition_receive = PartitionMut::new(
+                &mut received_queries,
+                receive_counts,
+                &receive_displacements[..],
+            );
+
+            self.neighbourhood_communicator_v
+                .all_to_all_varcount_into(&partition_send, &mut partition_receive);
+        }
+
+        // Now have to check for locally contained keys from received queries
+        let mut available_queries = Vec::new();
+        let mut available_queries_source_ranks = Vec::new();
+        for (&query, &source_rank) in received_queries
+            .iter()
+            .zip(received_queries_source_ranks.iter())
+        {
+            let key = MortonKey::from_morton(query);
+            if self.tree.source_tree.keys_set.contains(&key) {
+                available_queries.push(key.morton);
+                available_queries_source_ranks.push(source_rank);
+            }
+        }
+
+        // Calculate send counts of available queries to send back
[…]

[…]
