+++ kifmm/src/fmm/builder/multi_node.rs
@@ -226,8 +226,10 @@ where
                 kernel_eval_size: 1,
                 source: tmp_arr,
                 charges: Vec::default(),
+                ghost_charges: Vec::default(),
                 charge_index_pointer_sources: Vec::default(),
                 charge_index_pointer_targets: Vec::default(),
+                charge_index_pointer_ghost_sources: Vec::default(),
                 leaf_upward_check_surfaces_sources: Vec::default(),
                 leaf_downward_equivalent_surfaces_targets: Vec::default(),
                 leaf_upward_equivalent_surfaces_sources: Vec::default(),
