+++ kifmm/src/traits/general/mod.rs
@@ -1,3 +1,4 @@
+//! General traits associated with FMM
 #[cfg(feature = "mpi")]
 pub mod multi_node;
 pub mod single_node;

[…]
