+++ kifmm/src/fmm/field_translation/source/multi_node.rs
@@ -1,6 +1,6 @@
 //! Multipole to Multipole field translation
 
-use std::{collections::HashSet, sync::Mutex};
+use std::collections::HashSet;
 
 use itertools::Itertools;
 use mpi::traits::Equivalence;
