+++ kifmm/src/fmm/field_translation/source_to_target/blas/multi_node.rs
@@ -1,31 +1,18 @@
 //! Multipole to local field translation trait implementation using BLAS.
 
-use std::sync::Mutex;
-
-use itertools::Itertools;
 use mpi::traits::Equivalence;
 use num::Float;
-use rayon::prelude::*;
-use rlst::{
-    empty_array, rlst_array_from_slice2, rlst_dynamic_array2, MultIntoResize, RawAccess,
-    RawAccessMut, RlstScalar,
-};
+use rlst::RlstScalar;
 
 use green_kernels::traits::Kernel as KernelTrait;
 
 use crate::{
-    fmm::{
-        helpers::single_node::{homogenous_kernel_scale, m2l_scale},
-        types::{BlasFieldTranslationIa, FmmEvalType, KiFmmMulti, SendPtrMut},
-        KiFmm,
-    },
+    fmm::types::KiFmmMulti,
     traits::{
-        fmm::{FmmOperatorData, HomogenousKernel, SourceToTargetTranslation},
-        tree::{SingleFmmTree, SingleTree},
+        fmm::{HomogenousKernel, SourceToTargetTranslation},
         types::FmmError,
     },
-    tree::constants::NTRANSFER_VECTORS_KIFMM,
-    BlasFieldTranslationSaRcmp, SingleFmm,
+    BlasFieldTranslationSaRcmp,
 };
 
 impl<Scalar, Kernel> SourceToTargetTranslation
@@ -35,11 +22,11 @@ where
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
 {
-    fn m2l(&self, level: u64) -> Result<(), FmmError> {
+    fn m2l(&self, _level: u64) -> Result<(), FmmError> {
         Ok(())
     }
 
-    fn p2l(&self, level: u64) -> Result<(), FmmError> {
+    fn p2l(&self, _level: u64) -> Result<(), FmmError> {
         Ok(())
     }
 }
