### 2024-09-20 2bb9a09
Separate out evaluate routine
