+++ kifmm/src/fmm/eval/single_node.rs
@@ -186,80 +186,94 @@ where
         Some(&self.potentials)
     }
 
-    fn evaluate(&mut self, timed: bool) -> Result<(), FmmError> {
-        // Upward pass
+    #[inline(always)]
+    fn evaluate_leaf_sources(&mut self, timed: bool) -> Result<(), FmmError> {
         if timed {
-            {
-                let s = Instant::now();
-                self.p2m()?;
-                self.times
-                    .push(FmmOperatorTime::from_instant(FmmOperatorType::P2M, s));
-
-                for level in (1..=self.tree().source_tree().depth()).rev() {
-                    let s = Instant::now();
-                    self.m2m(level)?;
-                    self.times.push(FmmOperatorTime::from_instant(
-                        FmmOperatorType::M2M(level),
-                        s,
-                    ));
-                }
-            }
-
-            // Downward pass
-            {
-                for level in 2..=self.tree().target_tree().depth() {
-                    if level > 2 {
-                        let s = Instant::now();
-                        self.l2l(level)?;
-                        self.times.push(FmmOperatorTime::from_instant(
-                            FmmOperatorType::L2L(level),
-                            s,
-                        ));
-                    }
-                    let s = Instant::now();
-                    self.m2l(level)?;
-                    self.times.push(FmmOperatorTime::from_instant(
-                        FmmOperatorType::M2L(level),
-                        s,
-                    ));
-                }
-
-                // Leaf level computation
-                let s = Instant::now();
-                self.p2p()?;
-                self.times
-                    .push(FmmOperatorTime::from_instant(FmmOperatorType::P2P, s));
-                let s = Instant::now();
-                self.l2p()?;
-                self.times
-                    .push(FmmOperatorTime::from_instant(FmmOperatorType::L2P, s));
-            }
-            Ok(())
+            let s = Instant::now();
+            self.p2m()?;
+            self.times
+                .push(FmmOperatorTime::from_instant(FmmOperatorType::P2M, s));
         } else {
-            // Upward pass
-            {
-                self.p2m()?;
-
-                for level in (1..=self.tree().source_tree().depth()).rev() {
-                    self.m2m(level)?;
-                }
-            }
-
-            // // Downward pass
-            {
-                for level in 2..=self.tree().target_tree().depth() {
-                    if level > 2 {
-                        self.l2l(level)?;
-                    }
-                    self.m2l(level)?;
-                }
-
-                // Leaf level computation
-                self.p2p()?;
-                self.l2p()?;
-            }
-            Ok(())
+            self.p2m()?;
         }
+
+        Ok(())
+    }
+
+    #[inline(always)]
+    fn evaluate_upward_pass(&mut self, timed: bool) -> Result<(), FmmError> {
+        if timed {
+            for level in (1..=self.tree().source_tree().depth()).rev() {
+                let s = Instant::now();
+                self.m2m(level)?;
+                self.times.push(FmmOperatorTime::from_instant(
+                    FmmOperatorType::M2M(level),
+                    s,
+                ));
+            }
+        } else {
+            for level in (1..=self.tree().source_tree().depth()).rev() {
+                self.m2m(level)?;
+            }
+        }
+
+        Ok(())
+    }
+
+    #[inline(always)]
+    fn evaluate_downward_pass(&mut self, timed: bool) -> Result<(), FmmError> {
+        if timed {
+            for level in 2..=self.tree().target_tree().depth() {
+                if level > 2 {
+                    let s = Instant::now();
+                    self.l2l(level)?;
+                    self.times.push(FmmOperatorTime::from_instant(
[…]
