+++ kifmm/src/fmm/eval/multi_node.rs
@@ -1,5 +1,6 @@
 use std::time::Instant;
 
+use crate::fmm::KiFmm;
 use crate::traits::fmm::{HomogenousKernel, MultiFmm, SourceTranslation, TargetTranslation};
 use crate::traits::general::multi_node::GhostExchange;
 use crate::traits::tree::{MultiFmmTree, MultiTree};
@@ -14,7 +15,7 @@ use num::Float;
 use rlst::RlstScalar;
 
 use crate::fmm::types::{FmmEvalType, KiFmmMulti};
-use crate::MultiNodeFmmTree;
+use crate::{MultiNodeFmmTree, SingleFmm};
 
 impl<Scalar, Kernel, SourceToTargetData> MultiFmm for KiFmmMulti<Scalar, Kernel, SourceToTargetData>
 where
@@ -23,6 +24,7 @@ where
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
     SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
     Self: SourceToTargetTranslation + GhostExchange,
+    KiFmm<Scalar, Kernel, SourceToTargetData>: SingleFmm,
 {
     type Scalar = Scalar;
     type Kernel = Kernel;
@@ -71,113 +73,125 @@ where
         3
     }
 
-    fn evaluate(&mut self, timed: bool) -> Result<(), crate::traits::types::FmmError> {
+    #[inline(always)]
+    fn evaluate_leaf_sources(&mut self, timed: bool) -> Result<(), crate::traits::types::FmmError> {
+        if timed {
+            let s = Instant::now();
+            self.p2m()?;
+            self.times
+                .push(FmmOperatorTime::from_instant(FmmOperatorType::P2M, s));
+        } else {
+            self.p2m()?;
+        }
+
+        Ok(())
+    }
+
+    #[inline(always)]
+    fn evaluate_upward_pass(&mut self, timed: bool) -> Result<(), crate::traits::types::FmmError> {
         let global_depth = self.tree.source_tree().global_depth();
         let total_depth = self.tree.source_tree().total_depth();
 
         if timed {
-            // Upward pass
-            {
+            for level in ((global_depth + 1)..=total_depth).rev() {
                 let s = Instant::now();
-                self.p2m()?;
-                self.times
-                    .push(FmmOperatorTime::from_instant(FmmOperatorType::P2M, s));
-
-                for level in ((global_depth + 1)..=total_depth).rev() {
-                    let s = Instant::now();
-                    self.m2m(level)?;
-                    self.times.push(FmmOperatorTime::from_instant(
-                        FmmOperatorType::M2M(level),
-                        s,
-                    ));
-                }
-            }
-
-            // Exchange ghost multipole data
-            self.v_list_exchange();
-
-            // Gather root multipoles at nominated node(s)
-            self.gather_global_fmm_at_root();
-
-            // Calculate metadata for received ghosts
-
-            // Calculate metadata for global FMM
-
-            // Perform upward/downward pass on nominated node(s)
-
-            // Scatter root locals back to local tree
-            self.scatter_global_fmm_from_root();
-
-            // Downward pass
-            {
-                for level in global_depth..=total_depth {
-                    if level > 2 {
-                        let s = Instant::now();
-                        self.l2l(level)?;
-                        self.times.push(FmmOperatorTime::from_instant(
-                            FmmOperatorType::L2L(level),
-                            s,
-                        ));
-                    }
-
-                    let s = Instant::now();
-                    self.m2l(level)?;
-                    self.times.push(FmmOperatorTime::from_instant(
-                        FmmOperatorType::M2L(level),
-                        s,
-                    ));
-                }
-
[…]
