### 2024-09-23 95abe00
Clean up ghost exchange
