+++ kifmm/src/bindings.rs
@@ -271,7 +271,7 @@ pub unsafe extern "C" fn free_fmm_evaluator(fmm_p: *mut FmmEvaluator) {
 pub mod constructors {
     use std::ffi::c_void;
 
-    use green_kernels::{helmholtz_3d::Helmholtz3dKernel, types::EvalType};
+    use green_kernels::{helmholtz_3d::Helmholtz3dKernel, types::GreenKernelEvalType};
 
     use crate::{BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, FftFieldTranslation};
 
@@ -327,9 +327,9 @@ pub mod constructors {
         surface_diff: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
         let sources = unsafe { std::slice::from_raw_parts(sources as *const f32, n_sources) };
@@ -427,9 +427,9 @@ pub mod constructors {
         surface_diff: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
         let sources = unsafe { std::slice::from_raw_parts(sources as *const f64, n_sources) };
@@ -531,9 +531,9 @@ pub mod constructors {
         n_oversamples: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
         let sources = unsafe { std::slice::from_raw_parts(sources as *const f32, n_sources) };
@@ -647,9 +647,9 @@ pub mod constructors {
         n_oversamples: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
         let sources = unsafe { std::slice::from_raw_parts(sources as *const f64, n_sources) };
@@ -756,9 +756,9 @@ pub mod constructors {
         block_size: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
         let sources = unsafe { std::slice::from_raw_parts(sources as *const f32, n_sources) };
@@ -844,9 +844,9 @@ pub mod constructors {
         block_size: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
         let sources = unsafe { std::slice::from_raw_parts(sources as *const f64, n_sources) };
@@ -938,9 +938,9 @@ pub mod constructors {
         surface_diff: usize,
     ) -> *mut FmmEvaluator {
         let eval_type = if eval_type {
-            EvalType::Value
+            GreenKernelEvalType::Value
         } else {
-            EvalType::ValueDeriv
+            GreenKernelEvalType::ValueDeriv
         };
 
[…]

[…]
