### 2024-09-23 20bdcf5
Global upward pass converging
