+++ kifmm/src/fmm/ghost_exchange.rs
@@ -1,4 +1,3 @@
-use std::borrow::BorrowMut;
 use std::collections::HashMap;
 use std::collections::HashSet;
 
@@ -8,15 +7,11 @@ use itertools::Itertools;
 use mpi::datatype::Partition;
 use mpi::datatype::PartitionMut;
 use mpi::topology::Communicator;
-use mpi::traits::Collection;
-use mpi::traits::CommunicatorCollectives;
 use mpi::traits::Equivalence;
 use mpi::traits::Partitioned;
-use mpi::traits::Pointer;
 use mpi::traits::Root;
 use mpi::Count;
 use mpi::Rank;
-use mpi::traits::Group;
 use num::Float;
 use rlst::RlstScalar;
 
@@ -66,7 +61,7 @@ where
         // Gather multipole data
         if rank == root_rank {
             // Gather multipole data from root processes on all ranks
-            let n_root_multipoles = self.tree.source_tree().n_trees();
+            let n_root_multipoles = self.tree.source_tree().n_trees() as Count;
             let mut global_multipoles_counts = vec![0 as Count; size as usize];
             root_process.gather_into_root(&n_root_multipoles, &mut global_multipoles_counts);
 
@@ -80,8 +75,11 @@ where
 
             // Allocate memory for locally contained data
             let mut multipole_roots = Vec::new();
-            let mut multipoles =
-                vec![Scalar::default(); n_root_multipoles * self.n_coeffs_equivalent_surface];
+            let mut multipoles = vec![
+                Scalar::default();
+                (n_root_multipoles as usize)
+                    * self.n_coeffs_equivalent_surface
+            ];
 
             for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
                 let tmp = self.multipole(&tree.root()).unwrap();
@@ -141,7 +139,13 @@ where
                 let ancestors_siblings = leaf
                     .ancestors()
                     .iter()
-                    .flat_map(|a| if a.level() != 0 { a.siblings() } else { vec![] })
+                    .flat_map(|a| {
+                        if a.level() != 0 {
+                            a.siblings()
+                        } else {
+                            vec![*a]
+                        }
+                    })
                     .collect_vec();
 
                 for &key in siblings.iter() {
@@ -186,13 +190,16 @@ where
             );
         } else {
             // 1. Send multipole roots, if they exist, and associated coeffient data
-            let n_root_multipoles = self.tree.source_tree().n_trees();
+            let n_root_multipoles = self.tree.source_tree().n_trees() as Count;
             root_process.gather_into(&n_root_multipoles);
 
             // 1.1 Create buffers of multipole data to be sent
             let mut multipole_roots = Vec::new();
-            let mut multipoles =
-                vec![Scalar::default(); n_root_multipoles * self.n_coeffs_equivalent_surface];
+            let mut multipoles = vec![
+                Scalar::default();
+                (n_root_multipoles as usize)
+                    * self.n_coeffs_equivalent_surface
+            ];
             for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
                 let tmp = self.multipole(&tree.root()).unwrap();
                 multipoles[i * self.n_coeffs_equivalent_surface
@@ -207,7 +214,7 @@ where
 
         // Gather local data
         if rank == root_rank {
-            let n_root_locals = self.tree.target_tree().n_trees();
+            let n_root_locals = self.tree.target_tree().n_trees() as i32;
             let mut global_locals_counts = vec![0 as Count; size as usize];
             root_process.gather_into_root(&n_root_locals, &mut global_locals_counts);
 
@@ -252,7 +259,13 @@ where
                 let ancestors_siblings = leaf
                     .ancestors()
[…]
