+++ kifmm/src/fmm/neighbour_comm.rs
@@ -1,23 +1,13 @@
 //! Neighbourhood communicator wrapper
-use std::mem::MaybeUninit;
-
 use itertools::Itertools;
 use mpi::{
-    collective::CommunicatorCollectives, raw::{AsRaw, FromRaw}, topology::{Communicator, SimpleCommunicator, UserGroup}, traits::{Buffer, BufferMut, PartitionedBuffer, PartitionedBufferMut}, Count, Rank
+    raw::{AsRaw, FromRaw},
+    topology::{Communicator, SimpleCommunicator},
+    traits::{Buffer, BufferMut, PartitionedBuffer, PartitionedBufferMut},
 };
 use mpi_sys;
 
 use super::types::NeighbourhoodCommunicator;
-
-unsafe fn with_uninitialized<F, U, R>(f: F) -> (R, U)
-where
-    F: FnOnce(*mut U) -> R,
-{
-    let mut uninitialized = MaybeUninit::uninit();
-    let res = f(uninitialized.as_mut_ptr());
-    (res, uninitialized.assume_init())
-}
-
 
 impl NeighbourhoodCommunicator {
     /// Number of associated ranks
@@ -67,37 +57,17 @@ impl NeighbourhoodCommunicator {
         }
     }
 
-    pub fn translate_ranks(&self, global_comm: &SimpleCommunicator, ranks: &[Rank]) -> Vec<Rank>{
-
-        let n_ranks = ranks.len();
-        let mut result = vec![0 as Rank; n_ranks];
-        unsafe {
-            let global_comm_group = global_comm.group().as_raw();
-            let neighbour_comm_group = with_uninitialized(|group| mpi::ffi::MPI_Comm_group(self.raw.as_raw(), group)).1;
-            mpi::ffi::MPI_Group_translate_ranks(global_comm_group, n_ranks as Count, ranks.as_ptr(), neighbour_comm_group, result.as_mut_ptr());
-        }
-
-        result
-    }
-
-    /// Map from the global rank to the local rank
-    pub fn global_to_local_rank(&self, global_rank: i32) -> Option<i32> {
-        if let Some(idx) = self.neighbours.iter().position(|&g| global_rank == g) {
-            Some(self.neighbours[idx])
-        } else {
-            None
-        }
-    }
-
     /// Barrier synchronisation for all processes in neighbourhoood
     pub fn barrier(&self) {
-        unsafe {
-            mpi::ffi::MPI_Barrier(self.raw.as_raw())
-        };
+        unsafe { mpi::ffi::MPI_Barrier(self.raw.as_raw()) };
     }
 
     /// Constructor from locations to send to
-    pub fn new(world_comm: &SimpleCommunicator, send_marker: &[i32], receive_marker: &[i32]) -> Self {
+    pub fn new(
+        world_comm: &SimpleCommunicator,
+        send_marker: &[i32],
+        receive_marker: &[i32],
+    ) -> Self {
         let size = world_comm.size();
 
         // Now create neighbours, with send and receive displacements
