+++ kifmm/src/fmm/field_translation/target/single_node.rs
@@ -1,0 +1,459 @@
+//! Local expansion translations
+
+use std::collections::HashSet;
+
+use itertools::Itertools;
+use rayon::prelude::*;
+use rlst::{empty_array, rlst_dynamic_array2, MultIntoResize, RawAccess, RawAccessMut, RlstScalar};
+
+use green_kernels::traits::Kernel as KernelTrait;
+
+use crate::{
+    fmm::{
+        constants::L2L_MAX_BLOCK_SIZE, helpers::single_node::chunk_size, types::FmmEvalType, KiFmm,
+    },
+    traits::{
+        field::SourceToTargetData as SourceToTargetDataTrait,
+        fmm::{FmmOperatorData, HomogenousKernel, TargetTranslation},
+        tree::{SingleFmmTree, SingleTree},
+        types::FmmError,
+    },
+    tree::{constants::NSIBLINGS, types::MortonKey},
+    SingleFmm,
+};
+
+impl<Scalar, Kernel, SourceToTargetData> TargetTranslation
+    for KiFmm<Scalar, Kernel, SourceToTargetData>
+where
+    Scalar: RlstScalar,
+    Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Send + Sync,
+    SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default,
+    Self: FmmOperatorData + SingleFmm<Scalar = Scalar>,
+{
+    fn l2l(&self, level: u64) -> Result<(), FmmError> {
+        let Some(child_targets) = self.tree.target_tree().keys(level) else {
+            return Err(FmmError::Failed(format!(
+                "L2L failed at level {:?}, no sources found",
+                level
+            )));
+        };
+
+        let parent_sources: HashSet<MortonKey<_>> =
+            child_targets.iter().map(|source| source.parent()).collect();
+        let mut parent_sources = parent_sources.into_iter().collect_vec();
+        parent_sources.sort();
+        let nparents = parent_sources.len();
+        let operator_index = self.l2l_operator_index(level);
+        let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
+        let ncoeffs_equivalent_surface_parent = self.n_coeffs_equivalent_surface(level - 1);
+
+        match self.fmm_eval_type {
+            FmmEvalType::Vector => {
+                let mut parent_locals = Vec::new();
+                for parent in parent_sources.iter() {
+                    let parent_index_pointer = *self.level_index_pointer_locals
+                        [(level - 1) as usize]
+                        .get(parent)
+                        .unwrap();
+                    let parent_local =
+                        &self.level_locals[(level - 1) as usize][parent_index_pointer][0];
+                    parent_locals.push(parent_local);
+                }
+
+                let mut max_chunk_size = nparents;
+                if max_chunk_size > L2L_MAX_BLOCK_SIZE {
+                    max_chunk_size = L2L_MAX_BLOCK_SIZE
+                }
+                let chunk_size = chunk_size(nparents, max_chunk_size);
+
+                let child_locals = &self.level_locals[level as usize];
+
+                parent_locals
+                    .par_chunks_exact(chunk_size)
+                    .zip(child_locals.par_chunks_exact(NSIBLINGS * chunk_size))
+                    .for_each(|(parent_local_pointer_chunk, child_local_pointers_chunk)| {
+                        let mut parent_locals = rlst_dynamic_array2!(
+                            Scalar,
+                            [ncoeffs_equivalent_surface_parent, chunk_size]
+                        );
+                        for (chunk_idx, parent_local_pointer) in parent_local_pointer_chunk
+                            .iter()
+                            .enumerate()
[…]
