+++ kifmm/src/fmm/eval/multi_node.rs
@@ -1,7 +1,11 @@
 use std::time::Instant;
 
 use crate::fmm::KiFmm;
-use crate::traits::fmm::{HomogenousKernel, MultiFmm, SourceTranslation, TargetTranslation};
+use crate::traits::field::SourceAndTargetTranslationMetadata;
+use crate::traits::fmm::{
+    FmmDataAccess, FmmDataAccessMulti, HomogenousKernel, MultiFmm, SourceTranslation,
+    TargetTranslation,
+};
 use crate::traits::general::multi_node::GhostExchange;
 use crate::traits::tree::{MultiFmmTree, MultiTree};
 use crate::traits::types::{FmmOperatorTime, FmmOperatorType};
@@ -9,13 +13,12 @@ use crate::traits::{
     field::SourceToTargetData as SourceToTargetDataTrait, fmm::SourceToTargetTranslation,
 };
 use green_kernels::traits::Kernel as KernelTrait;
-use mpi::topology::SimpleCommunicator;
 use mpi::traits::{Communicator, Equivalence};
 use num::Float;
 use rlst::RlstScalar;
 
-use crate::fmm::types::{FmmEvalType, KiFmmMulti};
-use crate::{MultiNodeFmmTree, SingleFmm};
+use crate::fmm::types::KiFmmMulti;
+use crate::SingleFmm;
 
 impl<Scalar, Kernel, SourceToTargetData> MultiFmm for KiFmmMulti<Scalar, Kernel, SourceToTargetData>
 where
@@ -23,56 +26,14 @@ where
     <Scalar as RlstScalar>::Real: Default + Float + Equivalence,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
     SourceToTargetData: SourceToTargetDataTrait + Send + Sync,
-    Self: SourceToTargetTranslation + GhostExchange,
-    KiFmm<Scalar, Kernel, SourceToTargetData>: SingleFmm,
+    Self: SourceToTargetTranslation
+        + SourceTranslation
+        + TargetTranslation
+        + FmmDataAccessMulti
+        + GhostExchange,
+    KiFmm<Scalar, Kernel, SourceToTargetData>:
+        FmmDataAccess + SourceAndTargetTranslationMetadata + SingleFmm,
 {
-    type Scalar = Scalar;
-    type Kernel = Kernel;
-    type Tree = MultiNodeFmmTree<Scalar::Real, SimpleCommunicator>;
-
-    fn multipoles(&self, level: u64) -> Option<&[Self::Scalar]> {
-        if let Some(n_sources) = self.tree().source_tree().n_keys(level) {
-            let multipole_ptr = &self.level_multipoles[level as usize][0];
-
-            unsafe {
-                match self.fmm_eval_type {
-                    FmmEvalType::Vector => Some(std::slice::from_raw_parts(
-                        multipole_ptr.raw,
-                        self.n_coeffs_equivalent_surface * n_sources,
-                    )),
-                    FmmEvalType::Matrix(_n) => None,
-                }
-            }
-        } else {
-            None
-        }
-    }
-
-    fn multipole(
-        &self,
-        key: &<<<Self::Tree as MultiFmmTree>::Tree as MultiTree>::SingleTree as crate::traits::tree::SingleTree>::Node,
-    ) -> Option<&[Self::Scalar]> {
-        if let Some(&key_idx) = self.tree().source_tree().level_index(key) {
-            let multipole_ptr = &self.level_multipoles[key.level() as usize][key_idx];
-
-            unsafe {
-                match self.fmm_eval_type {
-                    FmmEvalType::Vector => Some(std::slice::from_raw_parts(
-                        multipole_ptr.raw,
[…]
