+++ kifmm/benches/laplace_mat_f32.rs
@@ -4,7 +4,7 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
 use kifmm::fmm::types::FmmSvdMode;
 use kifmm::fmm::types::{BlasFieldTranslationSaRcmp, SingleNodeBuilder};
-use kifmm::traits::fmm::{SingleFmm, SourceToTargetTranslation, TargetTranslation};
+use kifmm::traits::fmm::{FmmDataAccess, SingleFmm, SourceToTargetTranslation, TargetTranslation};
 use kifmm::traits::tree::{SingleFmmTree, SingleTree};
 use kifmm::tree::helpers::points_fixture;
 use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut};
