+++ kifmm/src/fmm/field_translation/source_to_target/hadamard.rs
@@ -401,8 +401,8 @@ pub mod aarch64 {
 
 #[cfg(all(target_arch = "x86_64", target_feature = "avx2"))]
 pub mod x86 {
-    use bytemuck::{cast_slice, cast_slice_mut};
     use super::{c32, c64, Hadamard8x8Data};
+    use bytemuck::{cast_slice, cast_slice_mut};
     use pulp::x86::V3;
     use pulp::{f32x8, f64x4};
 
@@ -414,8 +414,7 @@ pub mod x86 {
                 {
                     let n =
                         std::mem::size_of::<$array_type>() / std::mem::size_of::<$scalar_type>();
-                    let out: &mut [$scalar_type] =
-                        cast_slice_mut(std::slice::from_mut(&mut out));
+                    let out: &mut [$scalar_type] = cast_slice_mut(std::slice::from_mut(&mut out));
                     let x: &[$scalar_type] = cast_slice(std::slice::from_ref(&value));
 
                     for i in 0..n {
