+++ kifmm/Cargo.toml
@@ -33,12 +33,10 @@ mpi = { version = "0.8.*", optional = true}
 mpi-sys = { version="0.2.2", optional=true }
 memoffset = {version = "0.9.1", optional = true}
 pulp = "0.18.12"
+bytemuck = "1.16.0"
 superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
 rlst = {version = "0.2.0"}
-
-[target.'cfg(target_arch = "x86_64")'.dependencies]
-bytemuck = "1.16.0"
 
 [dev-dependencies]
 approx = "0.5"
