### 2024-09-25 30d05db
global downward pass running, untested
