+++ kifmm/src/fmm/types.rs
@@ -850,7 +850,7 @@ where
 /// We also store these in a permuted 'frequency' order for fast application in `kernel_data_f`, based on the techniques presented in [[Malhotra et. al, 2015](https://www.cambridge.org/core/journals/communications-in-computational-physics/article/pvfmm-a-parallel-kernel-independent-fmm-for-particle-and-volume-potentials/365109A4C15B126CD2A184F767D4C957)]
 ///
 /// $$ [[K_1^1, K_2^1, ..., K_{64}^1],  [K_1^2, K_2^2, ..., K_{64}^2], ..., [K_1^{P}, K_2^P, ..., K_{64}^P ] $$
-#[derive(Default)]
+#[derive(Default, Clone)]
 pub struct FftMetadata<T>
 where
     T: RlstScalar,
