+++ kifmm/src/lib.rs
@@ -18,7 +18,7 @@
 //!
 //! ```rust
 //! use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
-//! use kifmm::{SingleFmm, BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
+//! use kifmm::{Evaluate, DataAccess, BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
 //! use kifmm::tree::helpers::points_fixture;
 //! use rlst::{rlst_dynamic_array2, RawAccessMut, RawAccess};
 //!
@@ -108,6 +108,10 @@ pub use fmm::types::MultiNodeFmmTree;
 
 #[doc(inline)]
 pub use traits::fmm::Evaluate;
+
+#[doc(inline)]
+pub use traits::fmm::DataAccess;
+
 #[cfg_attr(feature = "strict", deny(warnings))]
 #[warn(missing_docs)]
 pub mod bindings;
