+++ kifmm/src/fmm/field_translation/metadata.rs
@@ -6,7 +6,7 @@ use std::{
 
 use green_kernels::{
     helmholtz_3d::Helmholtz3dKernel, laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait,
-    types::EvalType,
+    types::GreenKernelEvalType,
 };
 use itertools::Itertools;
 use num::{Float, Zero};
@@ -108,7 +108,7 @@ where
             // As well as estimating their inverses using SVD
             let mut uc2e = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
             self.kernel.assemble_st(
-                EvalType::Value,
+                GreenKernelEvalType::Value,
                 &upward_check_surface[..],
                 &upward_equivalent_surface[..],
                 uc2e.data_mut(),
@@ -161,7 +161,7 @@ where
 
                 // Note, this way around due to calling convention of kernel, source/targets are 'swapped'
                 self.kernel.assemble_st(
-                    EvalType::Value,
+                    GreenKernelEvalType::Value,
                     &parent_upward_check_surface,
                     &child_upward_equivalent_surface,
                     ce2pc.data_mut(),
@@ -222,7 +222,7 @@ where
             // As well as estimating their inverses using SVD
             let mut dc2e = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
             self.kernel.assemble_st(
-                EvalType::Value,
+                GreenKernelEvalType::Value,
                 &downward_check_surface[..],
                 &downward_equivalent_surface[..],
                 dc2e.data_mut(),
@@ -270,7 +270,7 @@ where
                 let mut pe2cc =
                     rlst_dynamic_array2!(Scalar, [ncheck_surface_child, nequiv_surface_parent]);
                 self.kernel.assemble_st(
-                    EvalType::Value,
+                    GreenKernelEvalType::Value,
                     &child_downward_check_surface,
                     &parent_downward_equivalent_surface,
                     pe2cc.data_mut(),
@@ -352,7 +352,7 @@ where
             // As well as estimating their inverses using SVD
             let mut uc2e = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
             self.kernel.assemble_st(
-                EvalType::Value,
+                GreenKernelEvalType::Value,
                 &upward_check_surface[..],
                 &upward_equivalent_surface[..],
                 uc2e.data_mut(),
@@ -430,7 +430,7 @@ where
                     rlst_dynamic_array2!(Scalar, [ncheck_surface_parent, nequiv_surface_child]);
 
                 self.kernel.assemble_st(
-                    EvalType::Value,
+                    GreenKernelEvalType::Value,
                     &parent_upward_check_surface,
                     &child_upward_equivalent_surface,
                     ce2pc.data_mut(),
@@ -504,7 +504,7 @@ where
             // As well as estimating their inverses using SVD
             let mut dc2e = rlst_dynamic_array2!(Scalar, [ncheck_surface, nequiv_surface]);
             self.kernel.assemble_st(
-                EvalType::Value,
+                GreenKernelEvalType::Value,
                 &downward_check_surface[..],
                 &downward_equivalent_surface[..],
                 dc2e.data_mut(),
@@ -574,7 +574,7 @@ where
                 let mut pe2cc =
                     rlst_dynamic_array2!(Scalar, [ncheck_surface_child, nequiv_surface_parent]);
                 self.kernel.assemble_st(
-                    EvalType::Value,
+                    GreenKernelEvalType::Value,
                     &child_downward_check_surface,
[…]
