+++ kifmm/src/fmm/field_translation/source_to_target/blas/multi_node.rs
@@ -1,19 +1,31 @@
 //! Multipole to local field translation trait implementation using BLAS.
 
-use mpi::traits::Equivalence;
+use std::sync::Mutex;
+
+use itertools::Itertools;
+use mpi::{topology::SimpleCommunicator, traits::Equivalence};
 use num::Float;
-use rlst::RlstScalar;
+use rayon::prelude::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
+use rlst::{
+    empty_array, rlst_array_from_slice2, rlst_dynamic_array2, MultIntoResize, RawAccess,
+    RawAccessMut, RlstScalar,
+};
 
 use green_kernels::traits::Kernel as KernelTrait;
 
 use crate::{
-    fmm::types::KiFmmMulti,
+    fmm::{
+        helpers::single_node::{homogenous_kernel_scale, m2l_scale},
+        types::{FmmEvalType, KiFmmMulti, SendPtrMut},
+    },
     traits::{
         field::SourceToTargetTranslation,
-        fmm::{DataAccessMulti, HomogenousKernel},
+        fmm::{DataAccessMulti, HomogenousKernel, MetadataAccess},
+        tree::{MultiFmmTree, MultiTree, SingleTree},
         types::FmmError,
     },
-    BlasFieldTranslationSaRcmp,
+    tree::constants::NTRANSFER_VECTORS_KIFMM,
+    BlasFieldTranslationSaRcmp, MultiNodeFmmTree,
 };
 
 impl<Scalar, Kernel> SourceToTargetTranslation
@@ -22,11 +34,249 @@ where
     Scalar: RlstScalar + Default + Equivalence + Float,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
-    Self: DataAccessMulti<Scalar = Scalar, Kernel = Kernel>,
+    Self: MetadataAccess
+        + DataAccessMulti<
+            Scalar = Scalar,
+            Kernel = Kernel,
+            Tree = MultiNodeFmmTree<Scalar::Real, SimpleCommunicator>,
+        >,
 {
     fn m2l(&self, level: u64) -> Result<(), FmmError> {
-        let _multipoles = self.multipoles(level);
-        Ok(())
+        match self.fmm_eval_type {
+            FmmEvalType::Vector => {
+                if let Some(targets) = self.tree().target_tree().keys(level) {
+                    // Metadata
+                    let m2l_operator_index = self.m2l_operator_index(level);
+                    let c2e_operator_index = self.c2e_operator_index(level);
+                    let displacement_index = self.displacement_index(level);
+                    let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
+
+                    // Parameters
+                    let n_targets = targets.len();
+                    let mut n_translations = 0;
+                    let mut all_sentinels = Vec::new();
+                    let mut all_n_sources = Vec::new();
+                    let mut all_multipoles = Vec::new();
+                    let mut all_displacements = Vec::new();
+
+                    // Handle locally contained source boxes
+                    if let Some(sources) = self.tree().target_tree().keys(level) {
+                        n_translations += 1;
+                        let sentinel = sources.len();
+
+                        all_displacements
+                            .push(&self.source_to_target.displacements[displacement_index]);
+                        // Number of sources at this level
+                        let n_sources = sources.len();
+
+                        // Lookup multipole data from source tree
+                        let multipoles = self.multipoles(level).unwrap();
+
+                        all_n_sources.push(n_sources);
+                        all_multipoles.push(multipoles);
+                        all_sentinels.push(sentinel);
+                    }
+
[…]
