+++ kifmm/src/fmm/field_translation/source_to_target/fft/multi_node.rs
@@ -1,22 +1,39 @@
 //! Multipole to local field translation trait implementation using FFT.
-use mpi::traits::Equivalence;
-use num::Float;
+use std::collections::HashSet;
 
-use rlst::RlstScalar;
+use itertools::Itertools;
+use mpi::{topology::SimpleCommunicator, traits::Equivalence};
+use num::{Float, One, Zero};
+
+use rayon::prelude::{
+    IndexedParallelIterator, IntoParallelIterator, ParallelIterator, ParallelSlice,
+    ParallelSliceMut,
+};
+use rlst::{
+    empty_array, rlst_dynamic_array2, MultIntoResize, RandomAccessMut, RawAccess, RlstScalar,
+};
 
 use green_kernels::traits::Kernel as KernelTrait;
 
 use crate::{
-    fftw::array::AlignedAllocable,
-    fmm::types::KiFmmMulti,
+    fftw::array::{AlignedAllocable, AlignedVec},
+    fmm::{
+        helpers::single_node::{chunk_size, homogenous_kernel_scale, m2l_scale},
+        types::{FmmEvalType, KiFmmMulti, SendPtrMut},
+    },
     traits::{
         fftw::Dft,
         field::SourceToTargetTranslation,
-        fmm::{DataAccessMulti, HomogenousKernel},
+        fmm::{DataAccessMulti, HomogenousKernel, MetadataAccess},
         general::single_node::{AsComplex, Hadamard8x8},
+        tree::{MultiFmmTree, MultiTree, SingleTree},
         types::FmmError,
     },
-    FftFieldTranslation,
+    tree::{
+        constants::{NHALO, NSIBLINGS, NSIBLINGS_SQUARED},
+        types::MortonKey,
+    },
+    FftFieldTranslation, MultiNodeFmmTree,
 };
 
 impl<Scalar, Kernel> SourceToTargetTranslation
@@ -34,11 +51,465 @@ where
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     <Scalar as Dft>::Plan: Sync,
-    Self: DataAccessMulti<Scalar = Scalar>,
+    Self: MetadataAccess
+        + DataAccessMulti<
+            Scalar = Scalar,
+            Kernel = Kernel,
+            Tree = MultiNodeFmmTree<Scalar::Real, SimpleCommunicator>,
+        >,
 {
     fn m2l(&self, level: u64) -> Result<(), FmmError> {
-        let _multipoles = self.multipoles(level);
-        Ok(())
+        match self.fmm_eval_type {
+            FmmEvalType::Vector => {
+                if let Some(targets) = self.tree().target_tree().keys(level) {
+                    let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
+
+                    let n_targets = targets.len();
+                    // Find parents of targets
+                    let targets_parents: HashSet<MortonKey<_>> =
+                        targets.iter().map(|target| target.parent()).collect();
+
+                    let targets_parents = targets_parents.into_iter().collect_vec();
+                    let ntargets_parents = targets_parents.len();
+
+                    // Size of input FFT sequence
+                    let shape_in = <Scalar as Dft>::shape_in(self.equivalent_surface_order(level));
+                    let size_in: usize =
+                        <Scalar as Dft>::size_in(self.equivalent_surface_order(level));
+
+                    // Size of transformed FFT sequence
+                    let size_out = <Scalar as Dft>::size_out(self.equivalent_surface_order(level));
+
+                    // Pre processing chunk size, in terms of number of source parents
+                    let max_chunk_size;
+                    if level == 2 {
[…]
