+++ TODO.md
@@ -1,5 +1,6 @@
 - all that is left are downward pass kernels
 - add kernels
     - local downward pass
+- Ghost trees in displacaement calcualtions in implementation of Ghost exchange exchage v list data
 
 - Not sure about scaling in P2M, why does it need extra now?
