### 2024-09-26 011247a
Add most of the downward pass kernels too
