+++ kifmm/src/fmm/field_translation/target/multi_node.rs
@@ -1,14 +1,28 @@
+use itertools::Itertools;
+use std::collections::HashSet;
+
 use crate::{
-    fmm::types::KiFmmMulti,
+    fmm::{
+        constants::L2L_MAX_BLOCK_SIZE,
+        helpers::single_node::chunk_size,
+        types::{FmmEvalType, KiFmmMulti},
+    },
     traits::{
         field::{FieldTranslation as FieldTranslationTrait, TargetTranslation},
-        fmm::HomogenousKernel,
+        fmm::{DataAccessMulti, HomogenousKernel, MetadataAccess},
+        tree::{MultiFmmTree, MultiTree},
+        types::FmmError,
     },
+    tree::{constants::NSIBLINGS, types::MortonKey},
+    MultiNodeFmmTree,
 };
 use green_kernels::traits::Kernel as KernelTrait;
-use mpi::traits::Equivalence;
+use mpi::{topology::SimpleCommunicator, traits::Equivalence};
 use num::Float;
-use rlst::RlstScalar;
+use rayon::prelude::{
+    IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator, ParallelSlice,
+};
+use rlst::{empty_array, rlst_dynamic_array2, MultIntoResize, RawAccess, RawAccessMut, RlstScalar};
 
 impl<Scalar, Kernel, FieldTranslation> TargetTranslation
     for KiFmmMulti<Scalar, Kernel, FieldTranslation>
@@ -17,17 +31,182 @@ where
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     FieldTranslation: FieldTranslationTrait,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
+    Self: MetadataAccess
+        + DataAccessMulti<
+            Scalar = Scalar,
+            Kernel = Kernel,
+            Tree = MultiNodeFmmTree<Scalar::Real, SimpleCommunicator>,
+        >,
 {
-    fn l2l(&self, _level: u64) -> Result<(), crate::traits::types::FmmError> {
+    fn l2l(&self, level: u64) -> Result<(), crate::traits::types::FmmError> {
+        if let Some(child_targets) = self.tree.target_tree().keys(level) {
+            let parent_sources: HashSet<MortonKey<_>> =
+                child_targets.iter().map(|source| source.parent()).collect();
+
+            let mut parent_sources = parent_sources.into_iter().collect_vec();
+            parent_sources.sort();
+            let n_parents = parent_sources.len();
+            let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
+            let n_coeffs_equivalent_surface_parent = self.n_coeffs_equivalent_surface(level - 1);
+
+            match self.fmm_eval_type {
+                FmmEvalType::Vector => {
+                    let mut parent_locals = Vec::new();
+                    for parent in parent_sources.iter() {
+                        let parent_index_pointer = *self.level_index_pointer_locals
+                            [(level - 1) as usize]
+                            .get(parent)
+                            .unwrap();
+                        let parent_local =
+                            &self.level_locals[(level - 1) as usize][parent_index_pointer];
+                        parent_locals.push(parent_local);
+                    }
+
+                    let mut max_chunk_size = n_parents;
+                    if max_chunk_size > L2L_MAX_BLOCK_SIZE {
+                        max_chunk_size = L2L_MAX_BLOCK_SIZE
+                    }
+                    let chunk_size = chunk_size(n_parents, max_chunk_size);
+
+                    let child_locals = &self.level_locals[level as usize];
+
+                    let target_vec = &self.target_vec;
+
+                    parent_locals
[…]
