+++ kifmm/Cargo.toml
@@ -50,6 +50,10 @@ name = "laplace_f32"
 harness = false
 
 [[bench]]
+name = "laplace_f32_m1"
+harness = false
+
+[[bench]]
 name = "laplace_f64"
 harness = false
 
