### 2024-09-30 5c69a42
Clean up trait
