+++ kifmm/src/traits/field.rs
@@ -14,15 +14,6 @@ pub trait FieldTranslation {
     fn surface_diff(&self) -> usize;
 }
 
-// /// Set M2M and L2L metadata associated with a kernel
-// pub trait SourceAndTargetTranslationMetadata {
-//     /// Source field translations
-//     fn source(&mut self);
-
-//     /// Target field translations
-//     fn target(&mut self);
-// }
-
 /// Source tree translation metadata
 pub trait SourceTranslationMetadata {
     /// Constructor
@@ -33,9 +24,6 @@ pub trait SourceTranslationMetadata {
 pub trait TargetTranslationMetadata {
     /// Constructor
     fn target(&mut self);
-
-    /// Graph to process U list
-    fn displacements(&mut self);
 }
 
 /// Set M2L metadata associated with a kernel
