### 2024-10-01 68f39cb
Isolate all troublesome parts
