+++ kifmm/src/fmm/field_translation/target/multi_node.rs
@@ -1,4 +1,5 @@
 use itertools::Itertools;
+use rayon::prelude::*;
 use std::collections::HashSet;
 
 use crate::{
@@ -10,7 +11,7 @@ use crate::{
     traits::{
         field::{FieldTranslation as FieldTranslationTrait, TargetTranslation},
         fmm::{DataAccessMulti, HomogenousKernel, MetadataAccess},
-        tree::{MultiFmmTree, MultiTree},
+        tree::{MultiFmmTree, MultiTree, SingleTree},
         types::FmmError,
     },
     tree::{constants::NSIBLINGS, types::MortonKey},
@@ -19,9 +20,6 @@ use crate::{
 use green_kernels::traits::Kernel as KernelTrait;
 use mpi::{topology::SimpleCommunicator, traits::Equivalence};
 use num::Float;
-use rayon::prelude::{
-    IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator, ParallelSlice,
-};
 use rlst::{empty_array, rlst_dynamic_array2, MultIntoResize, RawAccess, RawAccessMut, RlstScalar};
 
 impl<Scalar, Kernel, FieldTranslation> TargetTranslation
@@ -161,33 +159,33 @@ where
                                     ),
                                     potential_send_ptr,
                                 )| {
-                                    let target_coordinates_row_major = &coordinates
-                                        [charge_index_pointer.0 * dim
-                                            ..charge_index_pointer.1 * dim];
-                                    let ntargets = target_coordinates_row_major.len() / dim;
+                                    // let target_coordinates_row_major = &coordinates
+                                    //     [charge_index_pointer.0 * dim
+                                    //         ..charge_index_pointer.1 * dim];
+                                    // let ntargets = target_coordinates_row_major.len() / dim;
 
-                                    // Compute direct
-                                    if ntargets > 0 {
-                                        let result = unsafe {
-                                            std::slice::from_raw_parts_mut(
-                                                potential_send_ptr.raw,
-                                                ntargets * kernel_eval_size,
-                                            )
-                                        };
+                                    // // Compute direct
+                                    // if ntargets > 0 {
+                                    //     let result = unsafe {
+                                    //         std::slice::from_raw_parts_mut(
+                                    //             potential_send_ptr.raw,
+                                    //             ntargets * kernel_eval_size,
+                                    //         )
+                                    //     };
 
-                                        kernel.evaluate_st(
-                                            kernel_eval_type,
-                                            leaf_downward_equivalent_surface,
-                                            target_coordinates_row_major,
-                                            unsafe {
-                                                std::slice::from_raw_parts_mut(
-                                                    leaf_locals.raw,
-                                                    n_coeffs_equivalent_surface,
-                                                )
-                                            },
-                                            result,
-                                        );
-                                    }
+                                    //     kernel.evaluate_st(
+                                    //         kernel_eval_type,
+                                    //         leaf_downward_equivalent_surface,
+                                    //         target_coordinates_row_major,
+                                    //         unsafe {
+                                    //             std::slice::from_raw_parts_mut(
+                                    //                 leaf_locals.raw,
+                                    //                 n_coeffs_equivalent_surface,
+                                    //             )
+                                    //         },
+                                    //         result,
+                                    //     );
+                                    // }
                                 },
                             );
                     }
@@ -210,6 +208,119 @@ where
     }
 
     fn p2p(&self) -> Result<(), crate::traits::types::FmmError> {
+        if let Some(leaves) = self.tree.target_tree().all_leaves() {
+            let all_target_coordinates = self.tree.target_tree().all_coordinates().unwrap();
[…]
