+++ kifmm/src/fmm/field_translation/source_to_target/fft/multi_node.rs
@@ -73,7 +73,7 @@ where
                         targets.iter().map(|target| target.parent()).collect();
 
                     let targets_parents = targets_parents.into_iter().collect_vec();
-                    let ntargets_parents = targets_parents.len();
+                    let n_targets_parents = targets_parents.len();
 
                     // Size of input FFT sequence
                     let shape_in = <Scalar as Dft>::shape_in(self.equivalent_surface_order(level));
@@ -126,12 +126,21 @@ where
                     let mut all_signals_hat_f = Vec::new();
                     let mut all_signals_hat_f_ptr = Vec::new();
                     let mut all_chunk_size_pre_proc = Vec::new();
-                    let mut all_chunk_size_kernel = Vec::new();
-                    // let mut all_kernel_data_ft = Vec::new();
-                    let mut all_check_potentials_hat_f = Vec::new();
-                    let mut all_check_potentials_hat_c = Vec::new();
-                    let mut all_check_potentials = Vec::new();
                     let mut n_translations = 0;
+
+                    // Allocate buffer to store the check potentials in frequency order
+                    let mut check_potentials_hat_f =
+                        AlignedVec::<<Scalar as AsComplex>::ComplexType>::new(
+                            size_out * n_targets,
+                        );
+
+                    // Allocate buffer to store the check potentials in box order
+                    let mut check_potential_hat_c =
+                        AlignedVec::<<Scalar as AsComplex>::ComplexType>::new(
+                            size_out * n_targets,
+                        );
+                    let mut check_potential = AlignedVec::<Scalar>::new(size_in * n_targets);
+                    let chunk_size_kernel = chunk_size(n_targets_parents, max_chunk_size);
 
                     // Handle locally contained source boxes
                     if let Some(sources) = self.tree().source_tree().keys(level) {
@@ -158,20 +167,6 @@ where
                         };
 
                         let chunk_size_pre_proc = chunk_size(n_sources_parents, max_chunk_size);
-                        let chunk_size_kernel = chunk_size(ntargets_parents, max_chunk_size);
-
-                        // Allocate buffer to store the check potentials in frequency order
-                        let check_potentials_hat_f =
-                            AlignedVec::<<Scalar as AsComplex>::ComplexType>::new(
-                                size_out * n_targets,
-                            );
-
-                        // Allocate buffer to store the check potentials in box order
-                        let check_potential_hat_c =
-                            AlignedVec::<<Scalar as AsComplex>::ComplexType>::new(
-                                size_out * n_targets,
-                            );
-                        let check_potential = AlignedVec::<Scalar>::new(size_in * n_targets);
 
                         all_sources.push(sources);
                         all_n_sources.push(n_sources);
@@ -183,16 +178,12 @@ where
                         all_signals_hat_f.push(signals_hat_f);
                         all_signals_hat_f_ptr.push(signals_hat_f_ptr);
                         all_chunk_size_pre_proc.push(chunk_size_pre_proc);
-                        all_chunk_size_kernel.push(chunk_size_kernel);
-                        // all_kernel_data_ft.push(kernel_data_ft);
-                        all_check_potentials_hat_f.push(check_potentials_hat_f);
-                        all_check_potentials_hat_c.push(check_potential_hat_c);
-                        all_check_potentials.push(check_potential);
                     }
 
                     // Handle ghost sources
                     if let Some(sources) = self.ghost_fmm_v.tree.source_tree.keys(level) {
-                        //     n_translations += 1;
+                        n_translations += 1;
+
                         let n_sources = sources.len();
                         let sources_parents: HashSet<MortonKey<_>> =
                             sources.iter().map(|source| source.parent()).collect();
@@ -202,56 +193,36 @@ where
                         let displacement_index = self.displacement_index(level);
 
[…]
