### 2024-10-01 bde8f91
Leaf level operators broken, something ot do with raw access
