+++ kifmm/src/fmm/eval/multi_node.rs
@@ -124,7 +124,7 @@ where
         } else {
             // Leaf level computations
             // self.p2p()?;
-            // self.l2p()?;
+            self.l2p()?;
         }
 
         Ok(())
@@ -152,7 +152,7 @@ where
 
         // Perform remainder of downward pass
         self.evaluate_downward_pass(timed)?;
-        self.evaluate_leaf_targets(timed)?;
+        // self.evaluate_leaf_targets(timed)?;
 
         Ok(())
     }
