### 2024-10-02 dbd8904
Bug during eval, update todo
