+++ kifmm/examples/mpi_test_fmm.rs
@@ -23,7 +23,7 @@ fn main() {
     let prune_empty = true;
     let n_points = 10000;
     let local_depth = 2;
-    let global_depth = 1;
+    let global_depth = 2;
     let sort_kind = SortKind::Samplesort { k: 100 };
 
     // Fmm Parameters
