+++ TODO.md
@@ -1,6 +1,6 @@
-- all that is left are downward pass kernels
-- add kernels
-    - local downward pass
-- Ghost trees in displacaement calcualtions in implementation of Ghost exchange exchage v list data
+- Need to test end-to-end
+- Need to test threading controls i.e. trade-off between threading choices on MPI and rayon threads
+- Need to figure out how to pin MPI processes to nodes on Kathleen
 
+- Need to test algorithm spec during downward pass eval, are the l2L and m2l being called in the right order, from the right levels?
 - Not sure about scaling in P2M, why does it need extra now?
