### 2024-10-09 fd50dd4
update for new green-kernels release (#108)
* update for new green-kernel enum name
