+++ kifmm/src/lib.rs
@@ -17,7 +17,7 @@
 //! Basic usage for evaluating an FMM between a set of source and target points
 //!
 //! ```rust
-//! use green_kernels::{laplace_3d::Laplace3dKernel, types::EvalType};
+//! use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
 //! use kifmm::{Fmm, BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
 //! use kifmm::tree::helpers::points_fixture;
 //! use rlst::{rlst_dynamic_array2, RawAccessMut, RawAccess};
@@ -49,7 +49,7 @@
 //!             charges.data(),
 //!             &expansion_order,
 //!             Laplace3dKernel::new(), // Set the kernel
-//!             EvalType::Value, // Set the type of evaluation, either just potentials or potentials + potential gradients
+//!             GreenKernelEvalType::Value, // Set the type of evaluation, either just potentials or potentials + potential gradients
 //!             FftFieldTranslation::new(None), // Choose a field translation method, could replace with BLAS field translation
 //!         )
 //!         .unwrap()
