+++ kifmm/src/fmm/types.rs
@@ -2,7 +2,7 @@
 use pulp as _;
 use std::{collections::HashMap, sync::RwLock};
 
-use green_kernels::{traits::Kernel as KernelTrait, types::EvalType};
+use green_kernels::{traits::Kernel as KernelTrait, types::GreenKernelEvalType};
 use num::traits::Float;
 use rlst::{rlst_dynamic_array2, Array, BaseArray, RlstScalar, SliceContainer, VectorContainer};
 
@@ -207,7 +207,7 @@ where
     pub ncoeffs_check_surface: Vec<usize>, // Index corresponds to level
 
     /// The kernel evaluation type, either for potentials or potentials and gradients
-    pub kernel_eval_type: EvalType,
+    pub kernel_eval_type: GreenKernelEvalType,
 
     /// The FMM evaluation type, either for a vector or matrix of input charges.
     pub fmm_eval_type: FmmEvalType,
@@ -309,7 +309,7 @@ where
             equivalent_surface_order: Vec::default(),
             check_surface_order: Vec::default(),
             fmm_eval_type: FmmEvalType::Vector,
-            kernel_eval_type: EvalType::Value,
+            kernel_eval_type: GreenKernelEvalType::Value,
             kernel_eval_size: 0,
             dim: 0,
             ncoeffs_equivalent_surface: Vec::default(),
@@ -405,7 +405,7 @@ pub enum FmmEvalType {
 /// use kifmm::traits::tree::FmmTree;
 /// use kifmm::tree::helpers::points_fixture;
 /// use rlst::{rlst_dynamic_array2, RawAccessMut, RawAccess};
-/// use green_kernels::{laplace_3d::Laplace3dKernel, types::EvalType};
+/// use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
 ///
 /// /// Particle data
 /// let nsources = 1000;
@@ -436,7 +436,7 @@ pub enum FmmEvalType {
 ///         charges.data(),
 ///         &expansion_order,
 ///         Laplace3dKernel::new(),
-///         EvalType::Value,
+///         GreenKernelEvalType::Value,
 ///         FftFieldTranslation::new(None),
 ///     )
 ///     .unwrap()
@@ -488,7 +488,7 @@ where
     pub ncoeffs_check_surface: Option<Vec<usize>>,
 
     /// Kernel eval type
-    pub kernel_eval_type: Option<EvalType>,
+    pub kernel_eval_type: Option<GreenKernelEvalType>,
 
     /// FMM eval type
     pub fmm_eval_type: Option<FmmEvalType>,
