+++ kifmm/src/tree/single_node.rs
@@ -364,6 +364,8 @@ where
         global_indices: &[usize],
         root: MortonKey<T>,
     ) -> Result<SingleNodeTree<T>, std::io::Error> {
+
+
         let dim = 3;
         let n_coords = coordinates_row_major.len() / dim;
 
@@ -374,14 +376,19 @@ where
                 .try_into()
                 .unwrap();
 
-            let base_key = MortonKey::from_point(coord, &domain, DEEPEST_LEVEL);
             let encoded_key = MortonKey::from_point(coord, &domain, depth);
-            points.push(Point {
-                coordinate: *coord,
-                base_key,
-                encoded_key,
-                global_index: global_indices[i],
-            })
+            let base_key = MortonKey::from_point(coord, &domain, DEEPEST_LEVEL);
+            let ancestors = base_key.ancestors();
+
+            if ancestors.contains(&root) {
+
+                points.push(Point {
+                    coordinate: *coord,
+                    base_key,
+                    encoded_key,
+                    global_index: global_indices[i],
+                })
+            }
         }
 
         // Morton sort over points
@@ -446,6 +453,7 @@ where
         })
     }
 
+    // TEST: Test logic of from roots construction
     /// From a set of specified roots, create a number of single node trees of matching
     // length
     #[allow(unused)]
@@ -775,6 +783,10 @@ where
     ///   used directly; otherwise, it is computed from the point data, ensuring the tree encompasses
     ///   all points.
     ///
+    /// - `root` - Optionally specify a root node for the tree, defaults to the global root node.
+    ///
+    /// - `global_indices` - Optionally specify a set of global indices to map the specified points to,
+    ///    defaults to assigning index based on ordering of input points
     pub fn new(
         coordinates_row_major: &[T],
         depth: u64,
@@ -1232,6 +1244,10 @@ where
         Some(&self.coordinates)
     }
 
+    fn all_points(&self) -> Option<&[Point<Self::Scalar>]> {
+        Some(&self.points)
+    }
+
     fn global_indices(&self, leaf: &Self::Node) -> Option<&[usize]> {
         if let Some(&(l, r)) = self.leaves_to_coordinates.get(leaf) {
             if r - l > 0 {
@@ -1322,6 +1338,33 @@ mod test {
         assert_eq!(unique_leaves.len(), expected);
     }
 
+    #[test]
+    fn test_uniform_tree_from_root() {
+
+        let n_points = 1000;
+        let depth = 2;
+
+        // Use first child of global root as root node
+        let root = MortonKey::root().children()[0];
+
+        // Test data distributed over domain
+        let points = points_fixture(n_points, Some(0.), Some(1.0), None);
+        let domain = Domain::new(&[0., 0., 0.], &[1., 1., 1.]);
+        let tree = SingleNodeTree::<f64>::new(points.data(), depth, false, Some(domain), Some(root), None).unwrap();
+
+        // Test that only points contained within specified root node are mapped to this tree.
+        assert!(tree.all_points().unwrap().len() < n_points);
+
+        // Test data contained in first child
+        let points = points_fixture(n_points, Some(0.), Some(0.5), None);
+        let domain = Domain::new(&[0., 0., 0.], &[1., 1., 1.]);
+        let tree = SingleNodeTree::<f64>::new(points.data(), depth, false, Some(domain), Some(root), None).unwrap();
+
+        // Test that only points contained within specified root node are mapped to this tree.
+        assert_eq!(tree.all_points().unwrap().len(), n_points);
+
+    }
+
     pub fn test_no_overlaps_helper<T>(nodes: &[MortonKey<T>])
     where
         T: RlstScalar + Float,
@@ -1383,6 +1426,42 @@ mod test {
                 assert!((r - l) == n_points);
             }
         }
+    }
+
+    #[test]
[…]
