+++ kifmm/src/fmm/ghost_exchange.rs
@@ -65,253 +65,150 @@ where
 
         // Gather multipole data
         if rank == root_rank {
-            // Gather multipole data from root processes on all ranks
-            let n_root_multipoles = self.tree.source_tree().n_trees() as Count;
-            let mut global_multipoles_counts = vec![0 as Count; size as usize];
-            root_process.gather_into_root(&n_root_multipoles, &mut global_multipoles_counts);
-
-            // Calculate displacements and counts for associated morton keys
-            let mut global_multipoles_displacements = Vec::new();
-            let mut displacement = 0;
-            for &count in global_multipoles_counts.iter() {
-                global_multipoles_displacements.push(displacement);
-                displacement += count;
-            }
-
-            // Allocate memory for locally contained data
-            let mut multipole_roots = Vec::new();
-            let mut multipoles = vec![
-                Scalar::default();
-                (n_root_multipoles as usize)
-                    * self.n_coeffs_equivalent_surface
-            ];
+            // Allocate memory for locally contained data that will be communicated
+            let n_roots = self.tree.source_tree().n_trees() as Count;
+            let mut roots = Vec::new();
+            let mut multipoles =
+                vec![Scalar::default(); (n_roots as usize) * self.n_coeffs_equivalent_surface];
 
             for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
                 let tmp = self.multipole(&tree.root()).unwrap();
                 multipoles[i * self.n_coeffs_equivalent_surface
                     ..(i + 1) * self.n_coeffs_equivalent_surface]
                     .copy_from_slice(tmp);
-                multipole_roots.push(tree.root())
+                roots.push(tree.root())
             }
 
-            // Calculate displacements and counts for multipole data
-            let global_multipoles_bufs_counts = global_multipoles_counts
+            // Gather multipole data from root processes on all ranks
+            // Begin by counting the number of global leaves/local roots there are for source trees
+            let mut counts = vec![0 as Count; size as usize];
+            root_process.gather_into_root(&n_roots, &mut counts);
+
+            // Calculate associated displacements from the counts
+            let mut displacements = Vec::new();
+            let mut displacement = 0;
+            for &count in counts.iter() {
+                displacements.push(displacement);
+                displacement += count;
+            }
+
+            // Calculate displacements and counts for multipole data from that for keys
+            let counts_buffers = counts
                 .iter()
                 .map(|c| c * self.n_coeffs_equivalent_surface as i32)
                 .collect_vec();
-            let mut global_multipoles_bufs_displacements = Vec::new();
+
+            let mut displacements_buffers = Vec::new();
             let mut displacement = 0;
-            for &count in global_multipoles_bufs_counts.iter() {
-                global_multipoles_bufs_displacements.push(displacement);
+            for &count in counts_buffers.iter() {
+                displacements_buffers.push(displacement);
                 displacement += count
             }
 
             // Allocate memory to store received multipoles
-            let n = global_multipoles_bufs_counts.iter().sum::<i32>() as usize;
-            let mut global_multipoles =
-                vec![Scalar::default(); n * self.n_coeffs_equivalent_surface];
+            let total_receive_buffers = counts_buffers.iter().sum::<i32>() as usize;
+            let mut global_multipoles = vec![Scalar::default(); total_receive_buffers];
 
+            // Gather all multipole data
             let mut partition = PartitionMut::new(
                 &mut global_multipoles,
-                &global_multipoles_bufs_counts[..],
-                &global_multipoles_bufs_displacements[..],
+                &counts_buffers[..],
+                &displacements_buffers[..],
             );
+
             root_process.gather_varcount_into_root(&multipoles, &mut partition);
 
[…]
