+++ kifmm/examples/mpi_test_p2m.rs
@@ -1,0 +1,245 @@
+#[cfg(feature = "mpi")]
+fn main() {
+    use std::collections::HashMap;
+
+    use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+    use itertools::{izip, Itertools};
+    use kifmm::{
+        fmm::types::MultiNodeBuilder,
+        traits::{
+            field::SourceTranslation,
+            fmm::{DataAccessMulti, EvaluateMulti},
+            general::multi_node::GhostExchange,
+            tree::{MultiFmmTree, MultiTree, SingleFmmTree, SingleTree},
+        },
+        tree::{
+            helpers::points_fixture,
+            types::{MortonKey, SortKind},
+        },
+        DataAccess, Evaluate, FftFieldTranslation, SingleNodeBuilder,
+    };
+
+    use rayon::ThreadPoolBuilder;
+
+    use mpi::{collective::SystemOperation, datatype::PartitionMut, traits::*};
+    use rlst::RawAccess;
+
+    let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
+    let world = universe.world();
+    let comm = world.duplicate();
+
+    // Tree parameters
+    let prune_empty = true;
+    let n_points = 10000;
+    let local_depth = 3;
+    let global_depth = 2;
+    let sort_kind = SortKind::Samplesort { k: 1000 };
+
+    // Fmm Parameters
+    let expansion_order = 4;
+    let kernel = Laplace3dKernel::<f32>::new();
+    let source_to_target = FftFieldTranslation::<f32>::new(None);
+
+    // Generate some random test data local to each process
+    let points = points_fixture::<f32>(n_points, None, None, None);
+
+    ThreadPoolBuilder::new()
+        .num_threads(1)
+        .build_global()
+        .unwrap();
+
+    // Queries are set as a part of the build
+    let multi_fmm = MultiNodeBuilder::new()
+        .tree(
+            &comm,
+            points.data(),
+            points.data(),
+            local_depth,
+            global_depth,
+            prune_empty,
+            sort_kind,
+        )
+        .unwrap()
+        .parameters(expansion_order, kernel, source_to_target)
+        .unwrap()
+        .build()
+        .unwrap();
+
+    multi_fmm.p2m().unwrap();
+
+    // Test that the interaction lists remain the same with respect to a single node FMM
+    // Gather all data for the test
+
+    let root_process = comm.process_at_rank(0);
+    let n_coords = multi_fmm.tree().source_tree().coordinates.len() as i32;
+    let mut all_coordinates = vec![0f32; 3 * n_points * world.size() as usize];
+
+    let mut global_leaves_counts = vec![0i32; world.size() as usize];
+
+    if world.rank() == 0 {
+        let mut coordinates_counts = vec![0i32; comm.size() as usize];
+        root_process.gather_into_root(&n_coords, &mut coordinates_counts);
+
+        let mut coordinates_displacements = Vec::new();
+        let mut counter = 0;
+        for &count in coordinates_counts.iter() {
+            coordinates_displacements.push(counter);
+            counter += count;
+        }
+
+        let local_coords = multi_fmm.tree().source_tree().all_coordinates().unwrap();
+
+        let mut partition = PartitionMut::new(
+            &mut all_coordinates,
+            coordinates_counts,
+            coordinates_displacements,
+        );
+
+        root_process.gather_varcount_into_root(local_coords, &mut partition);
+
+        // Gather leaf counts
+        let n_leaves = multi_fmm.tree.source_tree.leaves.len() as i32;
+        root_process.gather_into_root(&n_leaves, &mut global_leaves_counts);
[…]
