+++ kifmm/examples/mpi_test_ghost_exchange.rs
@@ -1,19 +1,24 @@
 #[cfg(feature = "mpi")]
 fn main() {
-    use green_kernels::laplace_3d::Laplace3dKernel;
+    use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+    use itertools::{izip, Itertools};
     use kifmm::{
         fmm::types::MultiNodeBuilder,
-        traits::tree::{MultiFmmTree, SingleFmmTree, SingleTree},
+        traits::{
+            fmm::{DataAccessMulti, EvaluateMulti},
+            general::multi_node::GhostExchange,
+            tree::{MultiFmmTree, MultiTree, SingleFmmTree, SingleTree},
+        },
         tree::{
             helpers::points_fixture,
             types::{MortonKey, SortKind},
         },
-        FftFieldTranslation,
+        DataAccess, Evaluate, FftFieldTranslation, SingleNodeBuilder,
     };
 
     use rayon::ThreadPoolBuilder;
 
-    use mpi::{collective::SystemOperation, traits::*};
+    use mpi::{collective::SystemOperation, datatype::PartitionMut, traits::*};
     use rlst::RawAccess;
 
     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
@@ -23,10 +28,10 @@ fn main() {
     let rank = comm.rank();
 
     // Tree parameters
-    let prune_empty = true;
+    let prune_empty = false;
     let n_points = 10000;
     let local_depth = 3;
-    let global_depth = 1;
+    let global_depth = 2;
     let sort_kind = SortKind::Samplesort { k: 1000 };
 
     // Fmm Parameters
@@ -43,7 +48,7 @@ fn main() {
         .unwrap();
 
     // Queries are set as a part of the build
-    let fmm = MultiNodeBuilder::new()
+    let mut fmm = MultiNodeBuilder::new()
         .tree(
             &comm,
             points.data(),
@@ -58,6 +63,9 @@ fn main() {
         .unwrap()
         .build()
         .unwrap();
+
+    // fmm.v_list_exchange();
+    fmm.evaluate(false).unwrap();
 
     // Test neighbourhood communicator setup
 
@@ -79,16 +87,134 @@ fn main() {
         fmm.tree.source_tree().comm.size() as usize
     );
 
-    // Test that the received leaves in the U ghost tree are a subset of the requested leaves, and that they
-    // are associated with charge/coordinate data
-    for leaf in fmm.ghost_fmm_u.tree.source_tree().leaves.iter() {
-        assert!(fmm.tree.u_list_query.queries.contains(&leaf.morton));
-        assert!(fmm
-            .ghost_fmm_u
-            .tree
-            .source_tree()
-            .coordinates(leaf)
-            .is_some());
+    // Test that the interaction lists remain the same with respect to a single node FMM
+    // Gather all coordinates for the test
+    let root_process = comm.process_at_rank(0);
+    let n_coords = fmm.tree().source_tree().coordinates.len() as i32;
+    let mut all_coordinates = vec![0f32; 3 * n_points * world.size() as usize];
+
+    if world.rank() == 0 {
+        let mut coordinates_counts = vec![0i32; comm.size() as usize];
+        root_process.gather_into_root(&n_coords, &mut coordinates_counts);
+
+        let mut coordinates_displacements = Vec::new();
+        let mut counter = 0;
+        for &count in coordinates_counts.iter() {
+            coordinates_displacements.push(counter);
+            counter += count;
+        }
+
+        let local_coords = fmm.tree().source_tree().all_coordinates().unwrap();
+
+        let mut partition = PartitionMut::new(
+            &mut all_coordinates,
[…]

[…]
