+++ kifmm/src/fmm/builder/multi_node.rs
@@ -154,15 +154,6 @@ where
             fmm_tree.set_queries(true);
             fmm_tree.set_queries(false);
 
-            println!(
-                "U List Query size {:?}",
-                fmm_tree.u_list_query.queries.len()
-            );
-            println!(
-                "V List Query size {:?}",
-                fmm_tree.v_list_query.queries.len()
-            );
-
             self.communicator = Some(comm.duplicate());
             self.tree = Some(fmm_tree);
             Ok(self)
@@ -318,7 +309,7 @@ where
 
             // pass dummy charges for now.
             result.metadata(self.kernel_eval_type.unwrap(), &[Scalar::zero(); 1]);
-            result.displacements(None);
+            result.displacements(None, None);
 
             Ok(result)
         }
