+++ kifmm/src/fmm/field_translation/source_to_target/blas/multi_node.rs
@@ -62,8 +62,6 @@ where
                     let mut all_displacements = Vec::new();
                     let displacement_index = self.displacement_index(level);
 
-                    // println!("RANK {:?} {:?} {:?}", self.rank, displacement_index, level);
-
                     // Handle locally contained source boxes
                     if let Some(sources) = self.tree().source_tree().keys(level) {
                         n_translations += 1;
@@ -72,8 +70,6 @@ where
 
                         all_displacements
                             .push(&self.source_to_target.displacements[displacement_index]);
-
-                        // println!("found some displacements {:?}", &self.source_to_target.displacements[displacement_index]);
 
                         // Number of sources at this level
                         let n_sources = sources.len();
@@ -94,8 +90,6 @@ where
                         all_displacements.push(
                             &self.ghost_fmm_v.source_to_target.displacements[displacement_index],
                         );
-
-                        // println!("found some ghosts {:?}", &self.ghost_fmm_v.source_to_target.displacements[displacement_index]);
 
                         // Number of sources at this level
                         let n_sources = sources.len();
@@ -167,10 +161,6 @@ where
                                     .collect_vec()
                             })
                             .collect_vec();
-
-                        // if i == 1 {
-                        //     println!("level {:?} index {:?} idxs {:?} \n {:?} ", level, i, &local_idxs, self.level_index_pointer_locals.len());
-                        // }
 
                         let n_sources = all_n_sources[i];
                         let multipoles = &all_multipoles[i];
@@ -261,8 +251,6 @@ where
                                     }
                                 });
                         }
-
-                        // println!("level {:?} index {:?} multipoles {:?}", level, i, &compressed_check_potentials.data()[0..5]);
                     }
 
                     // 3. Compute local expansions from compressed check potentials
