+++ kifmm/src/fmm/field_translation/source_to_target/blas/single_node.rs
@@ -67,7 +67,7 @@ where
                     .iter()
                     .enumerate()
                     .filter(|(_, &d)| d != sentinel)
-                    .map(|(i, _)| i as usize)
+                    .map(|(i, _)| i)
                     .collect_vec()
             })
             .collect_vec();
