+++ kifmm/examples/mpi_test_upward_pass.rs
@@ -9,9 +9,10 @@ fn main() {
             tree::{FmmTreeNode, MultiFmmTree, MultiTree, SingleFmmTree, SingleTree},
         },
         tree::{constants::ALPHA_INNER, helpers::points_fixture, types::SortKind},
-        Evaluate, FftFieldTranslation,
+        Evaluate,
+        // FftFieldTranslation,
     };
-    use kifmm::{BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, FmmSvdMode};
+    use kifmm::{BlasFieldTranslationSaRcmp, FmmSvdMode};
     use mpi::{datatype::PartitionMut, traits::*};
     use rlst::{RawAccess, RlstScalar};
 
@@ -29,7 +30,7 @@ fn main() {
     // Fmm Parameters
     let expansion_order = 4;
     let kernel = Laplace3dKernel::<f32>::new();
-    let source_to_target = FftFieldTranslation::<f32>::new(None);
+    // let source_to_target = FftFieldTranslation::<f32>::new(None);
     let source_to_target =
         BlasFieldTranslationSaRcmp::<f32>::new(None, None, FmmSvdMode::Deterministic);
 
