+++ kifmm/src/sorting/hyksort.rs
@@ -18,30 +18,39 @@ use mpi::Count;
 use superslice::Ext;
 
 /// Modulo function compatible with signed integers.
-pub fn modulo(a: i32, b: i32) -> i32 {
+fn modulo(a: i32, b: i32) -> i32 {
     ((a % b) + b) % b
 }
 
-/// Parallel selection algorithm to determine 'k' splitters from the global array currently being
+/// Check if `n` is a power of two
+fn power_of_two(n: Rank) -> bool {
+    n != 0 && (n & (n - 1)) == 0
+}
+
+/// Parallel selection algorithm to determine 'n_splitters' splitters from the global array currently being
 /// considered in the communicator.
-pub fn parallel_select<T, C: Communicator>(arr: &Vec<T>, &k: &Rank, comm: C) -> Vec<T>
+pub fn parallel_select<T, C: Communicator>(
+    array: &Vec<T>,
+    &n_splitters: &Rank,
+    communicator: C,
+) -> Vec<T>
 where
     T: Default + Clone + Copy + Equivalence + Ord + std::fmt::Debug,
 {
-    let p: Rank = comm.size();
+    let p: Rank = communicator.size();
 
     // Store problem size in u64 to handle very large arrays
     let mut problem_size: u64 = 0;
     let mut min_arr_size: u64 = 0;
-    let arr_len: u64 = arr.len().try_into().unwrap();
+    let arr_len: u64 = array.len().try_into().unwrap();
 
     // Communicate the total problem size to each process in communicator
-    comm.all_reduce_into(&arr.len(), &mut problem_size, SystemOperation::sum());
-    comm.all_reduce_into(&arr.len(), &mut min_arr_size, SystemOperation::min());
+    communicator.all_reduce_into(&array.len(), &mut problem_size, SystemOperation::sum());
+    communicator.all_reduce_into(&array.len(), &mut min_arr_size, SystemOperation::min());
 
     // Determine number of samples for splitters, beta=20 taken from paper
     let beta = 20;
-    let numerator: f64 = (beta * k * (arr_len as Count)) as f64;
+    let numerator: f64 = (beta * n_splitters * (arr_len as Count)) as f64;
     let denominator: f64 = problem_size as f64;
     let mut split_count: Count = (numerator / denominator).ceil() as Count;
     if split_count > arr_len as Count {
@@ -56,13 +65,13 @@ where
     for i in 0..split_count {
         let mut idx: u64 = rng.gen::<u64>();
         idx %= arr_len;
-        splitters[i as usize] = arr[idx as usize];
+        splitters[i as usize] = array[idx as usize];
     }
 
     // Gather sampled splitters from all processes at each process
     let mut global_split_counts: Vec<Count> = vec![0; p as usize];
 
-    comm.all_gather_into(&split_count, &mut global_split_counts[..]);
+    communicator.all_gather_into(&split_count, &mut global_split_counts[..]);
 
     let global_split_displacements: Vec<Count> = global_split_counts
         .iter()
@@ -84,7 +93,7 @@ where
             global_split_counts,
             &global_split_displacements[..],
         );
-        comm.all_gather_varcount_into(&splitters[..], &mut partition)
+        communicator.all_gather_varcount_into(&splitters[..], &mut partition)
     }
 
     // Sort the sampled splitters
@@ -94,36 +103,37 @@ where
     let mut disp: Vec<u64> = vec![0; global_split_count as usize];
 
     for i in 0..global_split_count {
-        disp[i as usize] = arr.lower_bound(&global_splitters[i as usize]) as u64;
+        disp[i as usize] = array.lower_bound(&global_splitters[i as usize]) as u64;
     }
 
     // The global rank is found via a simple sum
[…]
