+++ kifmm/examples/mpi_test_trees.rs
@@ -29,7 +29,7 @@ fn main() {
     let n_points = 10000;
     let local_depth = 3;
     let global_depth = 2;
-    let sort_kind = SortKind::Samplesort { k: 100 };
+    let sort_kind = SortKind::Samplesort { n_samples: 100 };
 
     // Fmm Parameters
     let expansion_order = 4;
