+++ kifmm/src/sorting/simplesort.rs
@@ -46,61 +46,63 @@ where
     // Perform local sort
     array.sort();
 
-    // Sort local data into buckets
-    let mut ranks = Vec::new();
-    for item in array.iter() {
-        let mut rank_index = -1i32;
-        for (i, splitter) in splitters.iter().enumerate() {
-            if item < splitter {
-                rank_index = i as i32;
-                break;
+    if size > 1 {
+        // Sort local data into buckets
+        let mut ranks = Vec::new();
+        for item in array.iter() {
+            let mut rank_index = -1i32;
+            for (i, splitter) in splitters.iter().enumerate() {
+                if item < splitter {
+                    rank_index = i as i32;
+                    break;
+                }
+                if rank_index == -1i32 {
+                    rank_index = splitters.len() as i32
+                }
             }
-            if rank_index == -1i32 {
-                rank_index = splitters.len() as i32
-            }
+
+            ranks.push(rank_index);
         }
 
-        ranks.push(rank_index);
+        let mut counts_snd = vec![0i32; size as usize];
+        for &rank in ranks.iter() {
+            counts_snd[rank as usize] += 1
+        }
+
+        let displs_snd = counts_snd
+            .iter()
+            .scan(0, |acc, &x| {
+                let tmp = *acc;
+                *acc += x;
+                Some(tmp)
+            })
+            .collect_vec();
+
+        let mut counts_recv = vec![0 as Count; size as usize];
+
+        communicator.all_to_all_into(&counts_snd, &mut counts_recv);
+
+        let displs_recv = counts_recv
+            .iter()
+            .scan(0, |acc, &x| {
+                let tmp = *acc;
+                *acc += x;
+                Some(tmp)
+            })
+            .collect_vec();
+
+        let total = counts_recv.iter().sum::<Count>();
+
+        let mut received = vec![T::default(); total as usize];
+        let mut partition_received: PartitionMut<[T], Vec<i32>, &[i32]> =
+            PartitionMut::new(&mut received[..], counts_recv, &displs_recv[..]);
+        let partition_snd = Partition::new(&array[..], counts_snd, &displs_snd[..]);
+
+        communicator.all_to_all_varcount_into(&partition_snd, &mut partition_received);
+        received.sort();
+
+        *array = received;
     }
-
-    let mut counts_snd = vec![0i32; size as usize];
-    for &rank in ranks.iter() {
-        counts_snd[rank as usize] += 1
-    }
-
-    let displs_snd = counts_snd
-        .iter()
-        .scan(0, |acc, &x| {
-            let tmp = *acc;
-            *acc += x;
-            Some(tmp)
-        })
-        .collect_vec();
-
-    let mut counts_recv = vec![0 as Count; size as usize];
-
-    communicator.all_to_all_into(&counts_snd, &mut counts_recv);
-
-    let displs_recv = counts_recv
-        .iter()
-        .scan(0, |acc, &x| {
-            let tmp = *acc;
-            *acc += x;
-            Some(tmp)
-        })
-        .collect_vec();
-
-    let total = counts_recv.iter().sum::<Count>();
-
-    let mut received = vec![T::default(); total as usize];
-    let mut partition_received: PartitionMut<[T], Vec<i32>, &[i32]> =
-        PartitionMut::new(&mut received[..], counts_recv, &displs_recv[..]);
-    let partition_snd = Partition::new(&array[..], counts_snd, &displs_snd[..]);
-
-    communicator.all_to_all_varcount_into(&partition_snd, &mut partition_received);
-    received.sort();
-
-    *array = received;
 
     Ok(())
 }
