+++ kifmm/examples/mpi_test_p2m.rs
@@ -32,7 +32,7 @@ fn main() {
     let n_points = 10000;
     let local_depth = 3;
     let global_depth = 2;
-    let sort_kind = SortKind::Samplesort { k: 1000 };
+    let sort_kind = SortKind::Samplesort { n_samples: 1000 };
 
     // Fmm Parameters
     let expansion_order = 4;
