+++ kifmm/examples/mpi_test_ghost_exchange.rs
@@ -1,4 +1,9 @@
 #[cfg(feature = "mpi")]
+fn approx_equal(a: f32, b: f32, epsilon: f32) -> bool {
+    (a - b).abs() < epsilon
+}
+
+#[cfg(feature = "mpi")]
 fn main() {
     use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
     use itertools::Itertools;
@@ -18,7 +23,6 @@ fn main() {
     use rlst::RawAccess;
 
     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
-    // let universe = mpi::initialize().unwrap();
     let world = universe.world();
     let comm = world.duplicate();
     let rank = comm.rank();
@@ -28,7 +32,7 @@ fn main() {
     let n_points = 10000;
     let local_depth = 3;
     let global_depth = 2;
-    let sort_kind = SortKind::Samplesort { k: 1000 };
+    let sort_kind = SortKind::Samplesort { n_samples: 1000 };
 
     // Fmm Parameters
     let expansion_order = 4;
@@ -44,7 +48,7 @@ fn main() {
         .unwrap();
 
     // Queries are set as a part of the build
-    let mut fmm = MultiNodeBuilder::new()
+    let mut multi_fmm = MultiNodeBuilder::new()
         .tree(
             &comm,
             points.data(),
@@ -60,32 +64,31 @@ fn main() {
         .build()
         .unwrap();
 
-    fmm.evaluate(false).unwrap();
+    multi_fmm.evaluate(false).unwrap();
 
     // Test neighbourhood communicator setup
-
     // Expect these to match the global communicator
     assert_eq!(
-        fmm.tree.u_list_query.send_counts.len(),
-        fmm.tree.source_tree().comm.size() as usize
+        multi_fmm.tree.u_list_query.send_counts.len(),
+        multi_fmm.tree.source_tree().comm.size() as usize
     );
     assert_eq!(
-        fmm.tree.u_list_query.receive_counts.len(),
-        fmm.tree.source_tree().comm.size() as usize
+        multi_fmm.tree.u_list_query.receive_counts.len(),
+        multi_fmm.tree.source_tree().comm.size() as usize
     );
     assert_eq!(
-        fmm.tree.v_list_query.send_counts.len(),
-        fmm.tree.source_tree().comm.size() as usize
+        multi_fmm.tree.v_list_query.send_counts.len(),
+        multi_fmm.tree.source_tree().comm.size() as usize
     );
     assert_eq!(
-        fmm.tree.v_list_query.receive_counts.len(),
-        fmm.tree.source_tree().comm.size() as usize
+        multi_fmm.tree.v_list_query.receive_counts.len(),
+        multi_fmm.tree.source_tree().comm.size() as usize
     );
 
     // Test that the interaction lists remain the same with respect to a single node FMM
     // Gather all coordinates for the test
     let root_process = comm.process_at_rank(0);
-    let n_coords = fmm.tree().source_tree().coordinates.len() as i32;
+    let n_coords = multi_fmm.tree().source_tree().coordinates.len() as i32;
     let mut all_coordinates = vec![0f32; 3 * n_points * world.size() as usize];
 
     if world.rank() == 0 {
@@ -99,7 +102,7 @@ fn main() {
             counter += count;
         }
 
-        let local_coords = fmm.tree().source_tree().all_coordinates().unwrap();
+        let local_coords = multi_fmm.tree().source_tree().all_coordinates().unwrap();
 
         let mut partition = PartitionMut::new(
             &mut all_coordinates,
@@ -111,7 +114,7 @@ fn main() {
     } else {
         root_process.gather_into(&n_coords);
 
[…]
