+++ kifmm/examples/mpi_test_layout.rs
@@ -8,13 +8,12 @@ fn main() {
             helpers::points_fixture,
             types::{MortonKey, SortKind},
         },
-        // BlasFieldTranslationSaRcmp,
         FftFieldTranslation,
     };
 
     use rayon::ThreadPoolBuilder;
 
-    use mpi::{collective::SystemOperation, traits::*};
+    use mpi::{collective::SystemOperation, datatype::PartitionMut, traits::*};
     use rlst::RawAccess;
 
     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
@@ -58,8 +57,37 @@ fn main() {
         .build()
         .unwrap();
 
-    // Test U list queries
+    // Gather layout manually for testing
+    let local_domain = &fmm.tree.source_tree.roots;
+    let n_local_domain = local_domain.len() as i32;
 
+    let mut n_global_domain = vec![0i32; comm.size() as usize];
+
+    comm.all_gather_into(&n_local_domain, &mut n_global_domain);
+
+    // Now communicate the actual roots
+    let n_global_roots = n_global_domain.iter().sum::<i32>();
+    let mut global_roots = vec![MortonKey::<f32>::default(); n_global_roots as usize];
+
+    let counts = n_global_domain;
+
+    let mut displacement = 0;
+    let mut displacements = Vec::new();
+    for count in counts.iter() {
+        displacements.push(displacement);
+        displacement += *count;
+    }
+
+    let mut partition = PartitionMut::new(&mut global_roots, &counts[..], &displacements[..]);
+
+    comm.all_gather_varcount_into(&local_domain[..], &mut partition);
+
+    let mut ranks = Vec::new();
+    for (i, &count) in counts.iter().enumerate() {
+        ranks.append(&mut vec![i as i32; count as usize])
+    }
+
+    // Test U list queries
     // Test that requests are really not contained globally, but in global tree.
     for &query in fmm.tree.u_list_query.queries.iter() {
         let key = MortonKey::from_morton(query);
@@ -78,6 +106,15 @@ fn main() {
             .rank_from_key(intersection[0])
             .unwrap();
         assert_ne!(rank, fmm.rank);
+
+        // Calculate rank from layout directly, and test that it is
+        // as expected
+        let rank_idx = global_roots
+            .iter()
+            .position(|&x| x.morton == intersection[0].morton)
+            .unwrap();
+        let expected_rank = ranks[rank_idx];
+        assert_eq!(expected_rank, rank);
     }
 
     // Test that send and receive counts are properly matched up
@@ -112,6 +149,10 @@ fn main() {
         min = item
     }
 
+    if comm.rank() == 0 {
+        println!("...test_layout_u_list_queries passed")
+    }
+
     // Test V list queries
 
     // Test that requests are really not contained globally, but in global tree.
@@ -132,6 +173,51 @@ fn main() {
             .rank_from_key(intersection[0])
             .unwrap();
         assert_ne!(rank, fmm.rank);
+
+        // Calculate rank from layout directly, and test that it is
+        // as expected
+        let rank_idx = global_roots
+            .iter()
+            .position(|&x| x.morton == intersection[0].morton)
+            .unwrap();
+        let expected_rank = ranks[rank_idx];
+        assert_eq!(expected_rank, rank);
+    }
+
+    // Test that send and receive counts are properly matched up
+    let total_send_count = fmm.tree.v_list_query.send_counts.iter().sum::<i32>();
+    let mut total_receive_count = 0;
+
+    for rank in 0..comm.size() {
+        fmm.communicator.all_reduce_into(
+            &fmm.tree.v_list_query.receive_counts[rank as usize],
+            &mut total_receive_count,
[…]
