+++ kifmm/python/kifmm_py/__init__.py
@@ -31,7 +31,7 @@ def read_stl_triangle_mesh_vertices(filepath, dtype=np.float32):
     return (result, faces)
 
 
-class Times:
+class OperatorTimes:
     """
     Helper class to destructure operator runtimes.
     """
@@ -60,6 +60,76 @@ class Times:
                     operator_name = "L2P"
                 elif tag == lib.FmmOperatorType_P2P:
                     operator_name = "P2P"
+
+                self._times[operator_name] = time
+
+    def __repr__(self):
+        return str(self._times)
+
+class MetadataTimes:
+    """
+    Helper class to destructure metadata runtimes.
+    """
+
+    def __init__(self, times_p):
+        self.times_p = times_p
+        self._times = dict()
+        if self.times_p.length > 0:
+            for i in range(0, self.times_p.length):
+                metadata_time = self.times_p.times[i]
+                tag = metadata_time.operator_
+                time = metadata_time.time
+
+                if tag == 0:
+                    operator_name = "source_to_target_data"
+                elif tag == 1:
+                    operator_name = "source_data"
+                elif tag == 2:
+                    operator_name = "target_data"
+                elif tag == 3:
+                    operator_name = "global_fmm"
+                elif tag == 4:
+                    operator_name = "ghost_fmm_v"
+                elif tag == 5:
+                    operator_name = "ghost_fmm_u"
+
+                self._times[operator_name] = time
+
+    def __repr__(self):
+        return str(self._times)
+
+class CommunicationTimes:
+    """
+    Helper class to destructure communication runtimes.
+    """
+
+    def __init__(self, times_p):
+        self.times_p = times_p
+        self._times = dict()
+        if self.times_p.length > 0:
+            for i in range(0, self.times_p.length):
+                communication_time = self.times_p.times[i]
+                tag = communication_time.operator_
+                time = communication_time.time
+
+                if tag == 0:
+                    operator_name = "source_tree"
+                elif tag == 1:
+                    operator_name = "target_tree"
+                elif tag == 2:
+                    operator_name = "target_domain"
+                elif tag == 3:
+                    operator_name = "source_domain"
+                elif tag == 4:
+                    operator_name = "layout"
+                elif tag == 5:
+                    operator_name = "ghost_exchange_v"
+                elif tag == 6:
+                    operator_name = "ghost_exchange_u"
+                elif tag == 7:
+                    operator_name = "gather_global_fmm"
+                elif tag == 8:
+                    operator_name = "scatter_global_fmm"
 
                 self._times[operator_name] = time
 
@@ -369,6 +439,7 @@ class KiFmm:
             if isinstance(self._field_translation.kernel, LaplaceKernel):
                 if self._field_translation.kernel.dtype == np.float32:
                     self._fmm = lib.laplace_fft_f32_alloc(
+                        self._timed,
                         self._expansion_order_c,
                         self._nexpansion_order,
                         self._field_translation.kernel.eval_type,
@@ -387,6 +458,7 @@ class KiFmm:
 
                 elif self._field_translation.kernel.dtype == np.float64:
                     self._fmm = lib.laplace_fft_f64_alloc(
+                        self._timed,
                         self._expansion_order_c,
                         self._nexpansion_order,
                         self._field_translation.kernel.eval_type,
@@ -409,6 +481,7 @@ class KiFmm:
             elif isinstance(self._field_translation.kernel, HelmholtzKernel):
                 if self._field_translation.kernel.dtype == np.float32:
                     self._fmm = lib.helmholtz_fft_f32_alloc(
+                        self._timed,
                         self._expansion_order_c,
                         self._nexpansion_order,
                         self._field_translation.kernel.eval_type,
@@ -428,6 +501,7 @@ class KiFmm:
 
[…]
