### 2024-10-23 23d5a26
Start work on weak scaling
