+++ kifmm/src/fmm/helpers/multi_node.rs
@@ -209,12 +209,10 @@ where
     result
 }
 
-// /// Test weak scaling
-// pub fn weak_scaling(
-//     expansion_order: usize,
-//     n_points_per_process: usize,
-//     threads_per_process: usize
-// ) {
-
-
-// }
+/// Test weak scaling
+pub fn weak_scaling(
+    expansion_order: usize,
+    n_points_per_process: usize,
+    threads_per_process: usize,
+) {
+}
