+++ scripts/src/bin/grid_search.rs
@@ -1,0 +1,4 @@
+//! Optimal parameter search
+fn main() {
+    println!("RUNNING GRID SEARCH");
+}
