+++ hpc/hpc.md
@@ -1,0 +1,131 @@
+# Running kifmm-rs on HPC
+
+Here we document how to run this software on HPC systems, tried and tested configurations and optimal parameter search and setting.
+
+## Archer 2
+
+### Compilation
+
+- PrgEnv-aocc
+- set RUSTFLAGS as well as LD lib path to find BLAS/LAPACK
+- BLAS/LAPACK symlinked with libblas.so and liblapack.so
+- Same for MPI?
+
+### Dependencies
+
+- Clang - because of bindgen
+- MPI
+- BLAS/LAPACK installation
+
+Other dependencies such as fftw are built and linked by kifmm-rs as they contain custom bindings.
+
+### Hardware
+
+Three types of node available to users
+- login
+- data analysis
+- compute
+
+Nodes are grouped into partition, specified with the `--partition` option in the slurm jobscript.
+
+The relevant nodes for kifmm-rs are:
+
+- standard - CPU nodes, 2x AMD EPYC 7742 64 Core processor 256/512 GB memory, 5860 nodes available.
+
+### Software
+
+- As of 2024 Rust should be installed by the user, the easiest way to avoid building from source is to use the [rustup script](https://www.rust-lang.org/tools/install).
+
+
+
+#### File Systems
+
+Home and work filesystems at
+
+- `/home/[project id]/[project id]/[user id]`
+- `/work/[project id]/[project id]/[user id]`
+
+Home and work are available on the login nodes and data analysis nodes, the compute nodes only have access to work filesystem
+
+All three node types have access to solid state NVMe filesystem.
+
+- Note Rust binaries cannot be built in the HPE CCE which includes a specialised version of Clang, one must use the AMD-aocc compiler environment.
+
+- One should use the HPE Cray MPICH library, for optimised usage of slingshot.
+
+- One should use the HPE Cray numerical libraries, LibSci.
+
+
+
+### Scheduling with Slurm
+
+There are example slurm scheduling scripts in the `/hpc` directory for the weak scaling and parameter search scripts in this crate.
+
+Archer job limits defined by Quality of Service, specified by `--qos` directive. There are numerous listed on the site.
+
+The relevant ones are `standard`, which allows for a maximum of 1024 nodes per job with a max walltime of 24 hours. `short` may also be useful, with a max of 32 nodes and 20 minutes walltime.
+
+Potentially, if results are going well, `largescale` may be useful, which allows you to use the whole system but requires a minimum of 1025 nodes in use, max wall time of 12 hours.
+
+- optionally set a max wall clock time to help scheduler, and provide a custom name via the `--time` and `--job-name` directives
+
+To prevent batch scripts from being dependent on the user environment in the login node you have to specify
+
+`--export=none`
+
+This also means that any environment variables, programming environments, or linking, must be specified again in the job submission script.
+
+For parallel jobs need to specify the number of nodes `--nodes`, the tasks per node i.e. the number of parallel processes per compute node `--ntasks-per-node`, for parallel jobs that also use threading need to control the number of CPUs per task `--cpus-per-task`. Note that this is extra to also specifying the number of threads used by parallel threading modules like Rayon on BLAS.
+
+For jobs using less than 128 cores per node the cpus per task setting is set to the stride between parallel processes.
+
+E.g. 64 (mpi) processes per node and leave an empty core between each process, you'd set cpus-per-task to 2 and ntasks-per-node to 64.
+
+
+`sbatch`
+- used to submit a jobscript, typically contains one or more `srun` commands.
+
+`srun`
[…]
