+++ scripts/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "scripts"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+kifmm ={ path = "../kifmm"}
