+++ scripts/src/bin/grid_search_mpi.rs
@@ -1,0 +1,238 @@
+use clap::Parser;
+use green_kernels::laplace_3d::Laplace3dKernel;
+use kifmm::{
+    traits::field::FieldTranslation,
+    traits::types::{
+        CommunicationTime, CommunicationType, FmmOperatorTime, FmmOperatorType, MetadataTime,
+        MetadataType,
+    },
+    tree::{helpers::points_fixture, types::SortKind},
+    EvaluateMulti, FftFieldTranslation, MultiNodeBuilder,
+};
+use mpi::traits::*;
+use rayon::ThreadPoolBuilder;
+use rlst::RawAccess;
+use std::{collections::HashMap, env, time::Instant};
+
+/// Struct for parsing command-line arguments
+#[derive(Parser)]
+struct Args {
+    #[arg(long, default_value_t = 3)]
+    expansion_order: usize,
+
+    /// Whether to prune empty nodes
+    #[arg(long, default_value_t = false)]
+    prune_empty: bool,
+
+    /// Number of points per MPI process
+    #[arg(long, default_value_t = 10000)]
+    n_points: usize,
+
+    /// Local depth
+    #[arg(long, default_value_t = 1)]
+    local_depth: u64,
+
+    /// Global depth
+    #[arg(long, default_value_t = 1)]
+    global_depth: u64,
+
+    /// FFT Hadamard Block Size
+    #[arg(long, default_value_t = 16)]
+    block_size: usize,
+
+    /// Number of threads per MPI process
+    #[arg(long, default_value_t = 1)]
+    n_threads: usize,
+
+    /// Number of samples in parallel sample sort, must be
+    /// less than the number of samples and greater than 0
+    #[arg(long, default_value_t = 10)]
+    n_samples: usize,
+}
+
+fn main() {
+    let (universe, _threading) =
+        mpi::initialize_with_threading(mpi::Threading::Serialized).unwrap();
+    let world = universe.world();
+    let comm = world.duplicate();
+
+    // Tree parameters
+    let args = Args::parse();
+    let expansion_order = args.expansion_order;
+    let prune_empty = args.prune_empty;
+    let n_points = args.n_points;
+    let local_depth = args.local_depth;
+    let global_depth = args.global_depth;
+    let block_size = args.block_size;
+    let n_threads = args.n_threads;
+    let n_samples = args.n_samples;
+
+    assert!(n_samples > 0 && n_samples < n_points);
+
+    let sort_kind = SortKind::Samplesort { n_samples };
+
+    // Fmm Parameters
+    let kernel = Laplace3dKernel::<f32>::new();
+
+    ThreadPoolBuilder::new()
+        .num_threads(n_threads)
+        .build_global()
+        .unwrap();
+
+    let source_to_target = FftFieldTranslation::<f32>::new(Some(block_size));
+
+    // Generate some random test data local to each process
+    let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
+
+    let mut multi_fmm = MultiNodeBuilder::new(true)
+        .tree(
+            &comm,
+            points.data(),
+            points.data(),
+            local_depth.clone(),
+            global_depth.clone(),
+            prune_empty,
+            sort_kind.clone(),
+        )
+        .unwrap()
+        .parameters(expansion_order.clone(), kernel.clone(), source_to_target)
+        .unwrap()
+        .build()
+        .unwrap();
+
+    let start = Instant::now();
+    multi_fmm.evaluate().unwrap();
+    let runtime = start.elapsed().as_millis();
+
+    // Destructure operator times
+    let mut operator_times = HashMap::new();
+
+    for time in multi_fmm.operator_times.iter() {
+        match &time.operator {
+            FmmOperatorType::P2M => {
[…]
