### 2024-10-24 968cac6
Update slurm script
