### 2024-10-24 c854a2d
Fix script
