### 2024-10-24 4c886dc
Make script more useful
