+++ scripts/src/bin/fmm_m2l_blas_mpi_f32.rs
@@ -1,0 +1,240 @@
+
+//! Run a parametrised  distributed FMM with BLAS based M2L in f32
+use clap::Parser;
+use green_kernels::laplace_3d::Laplace3dKernel;
+use kifmm::{
+    traits::{field::FieldTranslation, types::{
+        CommunicationTime, CommunicationType, FmmOperatorTime, FmmOperatorType, MetadataTime,
+        MetadataType,
+    }}, tree::{helpers::points_fixture, types::SortKind}, BlasFieldTranslationSaRcmp, EvaluateMulti, FftFieldTranslation, FmmSvdMode, MultiNodeBuilder
+};
+use mpi::traits::*;
+use rayon::ThreadPoolBuilder;
+use rlst::RawAccess;
+use std::{collections::HashMap, time::Instant};
+
+/// Struct for parsing command-line arguments
+#[derive(Parser)]
+struct Args {
+    #[arg(long, default_value_t = 3)]
+    expansion_order: usize,
+
+    /// Whether to prune empty nodes
+    #[arg(long, default_value_t = false)]
+    prune_empty: bool,
+
+    /// Number of points per MPI process
+    #[arg(long, default_value_t = 10000)]
+    n_points: usize,
+
+    /// Local depth
+    #[arg(long, default_value_t = 1)]
+    local_depth: u64,
+
+    /// Global depth
+    #[arg(long, default_value_t = 1)]
+    global_depth: u64,
+
+    /// Singular Value Cutoff
+    #[arg(long, default_value_t = 1e-7)]
+    threshold: f32,
+
+    /// Number of threads per MPI process
+    #[arg(long, default_value_t = 1)]
+    n_threads: usize,
+
+    /// Number of samples in parallel sample sort, must be
+    /// less than the number of samples and greater than 0
+    #[arg(long, default_value_t = 10)]
+    n_samples: usize,
+}
+
+fn main() {
+    let (universe, _threading) =
+        mpi::initialize_with_threading(mpi::Threading::Serialized).unwrap();
+    let world = universe.world();
+    let comm = world.duplicate();
+
+    // Tree parameters
+    let args = Args::parse();
+    let expansion_order = args.expansion_order;
+    let prune_empty = args.prune_empty;
+    let n_points = args.n_points;
+    let local_depth = args.local_depth;
+    let global_depth = args.global_depth;
+    let threshold = args.threshold;
+    let n_threads = args.n_threads;
+    let n_samples = args.n_samples;
+
+    assert!(n_samples > 0 && n_samples < n_points);
+
+    let sort_kind = SortKind::Samplesort { n_samples };
+
+    // Fmm Parameters
+    let kernel = Laplace3dKernel::<f32>::new();
+
+    ThreadPoolBuilder::new()
+        .num_threads(n_threads)
+        .build_global()
+        .unwrap();
+
+    let source_to_target = BlasFieldTranslationSaRcmp::<f32>::new(Some(threshold), None, FmmSvdMode::Deterministic);
+
+    // Generate some random test data local to each process
+    let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
+
+    let mut multi_fmm = MultiNodeBuilder::new(true)
+        .tree(
+            &comm,
+            points.data(),
+            points.data(),
+            local_depth.clone(),
+            global_depth.clone(),
+            prune_empty,
+            sort_kind.clone(),
+        )
+        .unwrap()
+        .parameters(expansion_order.clone(), kernel.clone(), source_to_target)
+        .unwrap()
+        .build()
+        .unwrap();
+
+    let start = Instant::now();
+    multi_fmm.evaluate().unwrap();
+    let runtime = start.elapsed().as_millis();
+
[…]
