+++ kifmm/src/fmm/helpers/multi_node.rs
@@ -208,11 +208,3 @@ where
 
     result
 }
-
-/// Test weak scaling
-pub fn weak_scaling(
-    expansion_order: usize,
-    n_points_per_process: usize,
-    threads_per_process: usize,
-) {
-}
