+++ scripts/src/bin/weak.rs
@@ -1,4 +1,0 @@
-//! Weak scaling script
-fn main() {
-    println!("running weak scaling")
-}
