+++ TODO.md
@@ -9,6 +9,12 @@
             - thread/process pinning.
             - numa aware pinning specialised for archer 2.
 
+- repeat for BLAS translation
+- check budget
 
+- with optimum parameters for single node, have to try and run the same thing on 2 or more nodes.
+
+- run weak scaling per MPI process, as it's not actually useful in my code to do any other way as there is no load balancing at the moment.
 
 - Not sure about scaling in P2M, why does it need extra now?
+
