### 2024-10-25 aeb95cd
Update scripts
