+++ scripts/src/bin/fmm_m2l_blas_mpi_f32.rs
@@ -207,9 +207,8 @@ fn main() {
     println!(
         "{:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, \
          {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, \
-         {:?}, {:?}, {:?}, {:?}, {:?}, {:?} \
-         {:?}, {:?}, {:?}, {:?}. {:?}. {:?}, {:?},
-         ",
+         {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, \
+         {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}",
         id,
         multi_fmm.rank,
         runtime,
