### 2024-10-25 814a20c
Separate out runtime and precomputation ghost metadata
