+++ kifmm/src/fmm/ghost_exchange.rs
@@ -12,7 +12,6 @@ use mpi::traits::Equivalence;
 use mpi::traits::Partitioned;
 use mpi::traits::Root;
 use mpi::Count;
-use mpi::Rank;
 use num::Float;
 use rlst::RlstScalar;
 
@@ -56,7 +55,6 @@ where
         + SourceToTargetTranslationMetadata,
 {
     fn gather_global_fmm_at_root(&mut self) {
-        let size = self.communicator.size();
         let rank = self.communicator.rank();
 
         // Nominated rank chosen to run the global upward pass
@@ -67,7 +65,6 @@ where
         if rank == root_rank {
             // Allocate memory for locally contained data that will be communicated
             let n_roots = self.tree.source_tree().n_trees() as Count;
-            let mut roots = Vec::new();
             let mut multipoles =
                 vec![Scalar::default(); (n_roots as usize) * self.n_coeffs_equivalent_surface];
 
@@ -76,24 +73,12 @@ where
                 multipoles[i * self.n_coeffs_equivalent_surface
                     ..(i + 1) * self.n_coeffs_equivalent_surface]
                     .copy_from_slice(tmp);
-                roots.push(tree.root())
             }
 
-            // Gather multipole data from root processes on all ranks
-            // Begin by counting the number of global leaves/local roots there are for source trees
-            let mut counts = vec![0 as Count; size as usize];
-            root_process.gather_into_root(&n_roots, &mut counts);
-
-            // Calculate associated displacements from the counts
-            let mut displacements = Vec::new();
-            let mut displacement = 0;
-            for &count in counts.iter() {
-                displacements.push(displacement);
-                displacement += count;
-            }
-
-            // Calculate displacements and counts for multipole data from that for keys
-            let counts_buffers = counts
+            let counts_buffers = self
+                .tree
+                .source_tree
+                .all_roots_counts
                 .iter()
                 .map(|c| c * self.n_coeffs_equivalent_surface as i32)
                 .collect_vec();
@@ -118,110 +103,23 @@ where
 
             root_process.gather_varcount_into_root(&multipoles, &mut partition);
 
-            // Gather all associated keys
-            let total_receive = counts.iter().sum::<i32>();
-            let mut global_roots =
-                vec![MortonKey::<Scalar::Real>::default(); total_receive as usize];
-            let mut partition =
-                PartitionMut::new(&mut global_roots, &counts[..], &displacements[..]);
-
-            root_process.gather_varcount_into_root(&roots, &mut partition);
-
-            // Need to set metadata for global FMM
             self.global_fmm.global_fmm_multipole_metadata(
-                self.tree.domain(),
-                self.tree.source_tree().global_depth(),
-                global_roots,
+                self.tree.source_tree.all_roots.clone(),
                 global_multipoles,
             );
         } else {
-            // Send number of roots of local source trees
-            let n_roots = self.tree.source_tree().n_trees() as Count;
-            root_process.gather_into(&n_roots);
-
             // Create buffers of multipole data to be sent
+            let n_roots = self.tree.source_tree().n_trees() as Count;
             let mut multipoles =
                 vec![Scalar::default(); (n_roots as usize) * self.n_coeffs_equivalent_surface];
-            let mut roots = Vec::new();
             for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
                 let tmp = self.multipole(&tree.root()).unwrap();
                 multipoles[i * self.n_coeffs_equivalent_surface
                     ..(i + 1) * self.n_coeffs_equivalent_surface]
                     .copy_from_slice(tmp);
-                roots.push(tree.root())
             }
 
             root_process.gather_varcount_into(&multipoles);
-            root_process.gather_varcount_into(&roots);
-        }
-
-        // Gather local data
-        if rank == root_rank {
[…]
