+++ scripts/src/bin/fmm_m2l_fft_mpi_f32.rs
@@ -207,10 +207,10 @@ fn main() {
     }
 
     println!(
-        "{:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, \
-         {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, \
-         {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, \
-         {:?}, {:?}, {:?}, {:?}, {:?}, {:?}, {:?}",
+        "{:?},{:?},{:?},{:?},{:?},{:?},{:?},{:?},\
+         {:?},{:?},{:?},{:?},{:?},{:?},{:?},{:?},{:?},{:?}, \
+         {:?},{:?},{:?},{:?},{:?},{:?},\
+         {:?},{:?},{:?},{:?},{:?},{:?},{:?}",
         id,
         multi_fmm.rank,
         runtime,
