+++ TODO.md
@@ -12,6 +12,13 @@
 - repeat for BLAS translation
 - check budget
 
+
+- Docs and tests
+
+- Linting
+
+- Merge
+
 - with optimum parameters for single node, have to try and run the same thing on 2 or more nodes.
 
 - run weak scaling per MPI process, as it's not actually useful in my code to do any other way as there is no load balancing at the moment.

[…]
