+++ kifmm/src/tree/types.rs
@@ -1,4 +1,8 @@
 //! Data structures to create distributed octrees with MPI.
+use std::{
+    collections::{HashMap, HashSet},
+    marker::PhantomData,
+};
 
 #[cfg(feature = "mpi")]
 use mpi::{
@@ -8,11 +12,6 @@ use mpi::{
 
 use num::Float;
 use rlst::RlstScalar;
-
-use std::{
-    collections::{HashMap, HashSet},
-    marker::PhantomData,
-};
 
 /// Represents a three-dimensional box characterized by its origin and side-length along the Cartesian axes.
 ///
@@ -81,76 +80,87 @@ where
 /// Represents an MPI distributed tree structure equipped with spatial indexing capabilities for 3D
 /// particle data.
 ///
-/// The struct is similar to `SingleNodeTree`, however the associated data is now local to a given
-/// processor. The major difference is that each processor owns a contiguous, non-overlapping, set of
-/// leaf nodes, as well as __all__ ancestors to their owned leaf nodes. This means that ancestor nodes
-/// may be duplicated across processors, but not leaf nodes.
+/// `MultiNodeTrees` are split into 'local' and 'global' portions, where the local trees correspond
+/// to trees that are defined by roots that form the leaves of the global tree, such that that leaves
+/// of the local trees do not intersect with each other. Each rank can contain multiple local trees, the
+/// total number of local trees across all ranks is controlled with the `global_depth` parameter, as the
+/// leaves correspond to the local trees' roots.
+///
+/// The global tree, which contains common ancestor nodes of all local trees distributed across nodes,
+/// is stored on a single nominated node at rank 0 in the global communicator. Data from the global tree
+/// is shared with the local trees via collective operations.
+///
+/// This split between local and global trees allows one to achieve optimal scaling in the distributed FMM
+/// for all the ghost data exchanges.
 ///
 /// # Fields
-/// - `world` - The global MPI communicator for this tree.
+/// - `roots` - The roots associated with the local trees at this rank
 ///
-/// - `range` - The range represented as the minimum and maximum leaf key associated owned by
-/// each processor, as well as the processor's MPI rank.
+/// - `all_roots` - All the roots associated with local trees in the distributed MultiNodeTree.
 ///
-/// - `depth` - The depth of the tree.
+/// - `all_roots_ranks` - The associated ranks of `all_roots`.
 ///
-/// - `domain` - The spatial domain covered by the tree's associated point data.
+/// - `all_roots_displacements` - The displacements of all `all_roots`, for collective communications
+/// with the global tree.
+///
+/// - `all_roots_counts` - The counts of all `all_roots`, for collective communications
+/// with the global tree.
+///
+/// - `global_depth` - The depth of the global tree.
+///
+/// - `local_depth` - The depth of the local trees.
+///
+/// - `total_depth` - The local depth + the global depth
+///
+/// - `domain` - The spatial domain covered by the globally distributed associated point data.
 ///
 /// - `coordinates` - A flat vector storing the coordinates of all points managed by the tree,
 ///   in row-major format (e.g., `[x1, y1, z1, ..., xn, yn, zn]`), for efficient retrieval.
 ///
+/// - `points` - Coordinates stored as `Points`.
+///
 /// - `global_indices` - A vector of unique global indices corresponding to each point, allowing
 ///   for efficient identification and lookup of points across different parts of the system.
 ///
[…]

[…]
