+++ kifmm/src/tree/multi_node.rs
@@ -2,11 +2,7 @@
 use std::collections::{HashMap, HashSet};
 
 use itertools::Itertools;
-use mpi::{
-    datatype::{PartitionMut, Partitioned},
-    topology::SimpleCommunicator,
-    traits::Root,
-};
+use mpi::{datatype::PartitionMut, topology::SimpleCommunicator, traits::Root};
 use mpi::{
     traits::{Communicator, CommunicatorCollectives, Equivalence},
     Count, Rank,
@@ -15,14 +11,12 @@ use num::Float;
 use rlst::RlstScalar;
 
 use crate::{
-    fmm::types::{Layout, Query},
     sorting::{hyksort, samplesort, simplesort},
-    traits::tree::{MultiFmmTree, MultiTree, SingleTree},
+    traits::tree::{MultiTree, SingleTree},
     tree::{
         constants::DEEPEST_LEVEL, Domain, MortonKey, MortonKeys, MultiNodeTree, Point, Points,
         SingleNodeTree, SortKind,
     },
-    MultiNodeFmmTree,
 };
 
 impl<T> MultiNodeTree<T, SimpleCommunicator>
@@ -502,230 +496,6 @@ where
 
     fn leaf_index(&self, leaf: &<Self::SingleTree as SingleTree>::Node) -> Option<&usize> {
         self.leaf_to_index.get(leaf)
-    }
-}
-
-impl<T, C> MultiFmmTree for MultiNodeFmmTree<T, C>
-where
-    T: RlstScalar + Default + Float + Equivalence,
-    C: Communicator,
-{
-    type Tree = MultiNodeTree<T, C>;
-
-    fn domain(&self) -> &<<Self::Tree as MultiTree>::SingleTree as SingleTree>::Domain {
-        &self.domain
-    }
-
-    fn n_source_trees(&self) -> usize {
-        self.source_tree().n_trees()
-    }
-
-    fn n_target_trees(&self) -> usize {
-        self.target_tree().n_trees()
-    }
-
-    fn source_tree(&self) -> &Self::Tree {
-        &self.source_tree
-    }
-
-    fn target_tree(&self) -> &Self::Tree {
-        &self.target_tree
-    }
-}
-
-impl<T, C> MultiNodeFmmTree<T, C>
-where
-    T: RlstScalar + Default + Float + Equivalence,
-    C: Communicator,
-{
-    /// Configure queries to send from this rank for either U or V list data
-    pub fn set_queries(&mut self, admissible: bool) {
-        let mut queries = HashSet::new();
-
-        if admissible {
-            // V list queries
-            for key in self.target_tree.all_keys().unwrap() {
-                // Compute interaction list
-                let interaction_list = key
-                    .parent()
-                    .neighbors()
-                    .iter()
-                    .flat_map(|pn| pn.children())
-                    .filter(|pnc| !key.is_adjacent(pnc))
-                    .collect_vec();
-
-                // Filter for those contained on foreign ranks
-                let interaction_list = interaction_list
-                    .into_iter()
-                    .filter(|key| {
-                        // Check if the rank is not equal to this rank
-                        if let Some(&rank) = self.source_layout.rank_from_key(key) {
-                            rank != self.source_tree.rank()
-                        } else {
-                            false
-                        }
-                        // !self.source_tree.keys_set.contains(key)
-                    })
-                    .collect_vec();
-
-                queries.extend(interaction_list);
-            }
-        } else {
-            // U list queries
-
-            for leaf in self.target_tree.all_leaves().unwrap() {
-                let interaction_list = leaf.neighbors();
-
-                // Filter for those contained on foreign ranks
-                let interaction_list = interaction_list
-                    .into_iter()
-                    .filter(|key| {
-                        // Check if the rank is not equal to this rank
[…]
