+++ kifmm/examples/mpi_test_ghost_exchange.rs
@@ -1,332 +1,335 @@
-#[cfg(feature = "mpi")]
-fn approx_equal(a: f32, b: f32, epsilon: f32) -> bool {
-    (a - b).abs() < epsilon
-}
+// #[cfg(feature = "mpi")]
+// fn approx_equal(a: f32, b: f32, epsilon: f32) -> bool {
+//     (a - b).abs() < epsilon
+// }
+
+// #[cfg(feature = "mpi")]
+// fn main() {
+//     use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+//     use itertools::Itertools;
+//     use kifmm::{
+//         fmm::types::MultiNodeBuilder,
+//         traits::{
+//             fmm::{DataAccess, DataAccessMulti, EvaluateMulti},
+//             tree::{MultiFmmTree, MultiTree, SingleFmmTree, SingleTree},
+//         },
+//         tree::{helpers::points_fixture, types::SortKind},
+//         Evaluate, FftFieldTranslation, SingleNodeBuilder,
+//     };
+
+//     use rayon::ThreadPoolBuilder;
+
+//     use mpi::{
+//         datatype::PartitionMut,
+//         traits::{Communicator, Root},
+//     };
+//     use rlst::RawAccess;
+
+//     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
+//     let world = universe.world();
+//     let comm = world.duplicate();
+//     let rank = comm.rank();
+
+//     // Tree parameters
+//     let prune_empty = false;
+//     let n_points = 10000;
+//     let local_depth = 3;
+//     let global_depth = 2;
+//     let sort_kind = SortKind::Samplesort { n_samples: 1000 };
+
+//     // Fmm Parameters
+//     let expansion_order = 4;
+//     let kernel = Laplace3dKernel::<f32>::new();
+//     let source_to_target = FftFieldTranslation::<f32>::new(None);
+
+//     // Generate some random test data local to each process
+//     let points = points_fixture::<f32>(n_points, None, None, Some(rank as u64));
+
+//     ThreadPoolBuilder::new()
+//         .num_threads(1)
+//         .build_global()
+//         .unwrap();
+
+//     // Queries are set as a part of the build
+//     let mut multi_fmm = MultiNodeBuilder::new(false)
+//         .tree(
+//             &comm,
+//             points.data(),
+//             points.data(),
+//             local_depth,
+//             global_depth,
+//             prune_empty,
+//             sort_kind,
+//         )
+//         .unwrap()
+//         .parameters(expansion_order, kernel, source_to_target)
+//         .unwrap()
+//         .build()
+//         .unwrap();
+
+//     multi_fmm.evaluate().unwrap();
+
+//     // Test neighbourhood communicator setup
+//     // Expect these to match the global communicator
+//     assert_eq!(
+//         multi_fmm.tree.u_list_query.send_counts.len(),
+//         multi_fmm.tree.source_tree().communicator.size() as usize
+//     );
+//     assert_eq!(
+//         multi_fmm.tree.u_list_query.receive_counts.len(),
+//         multi_fmm.tree.source_tree().communicator.size() as usize
+//     );
+//     assert_eq!(
+//         multi_fmm.tree.v_list_query.send_counts.len(),
+//         multi_fmm.tree.source_tree().communicator.size() as usize
+//     );
+//     assert_eq!(
+//         multi_fmm.tree.v_list_query.receive_counts.len(),
+//         multi_fmm.tree.source_tree().communicator.size() as usize
+//     );
+
+//     // Test that the interaction lists remain the same with respect to a single node FMM
+//     // Gather all coordinates for the test
[…]
