### 2024-10-30 65b0c04
ENH: Experimental Distributed FMM (#107)
Add an experimental distributed FMM and associated traits for the Laplace kernel.

What is achieved in this PR:

- Significantly clearer trait layout, allowing for maximal code re-use between single and multi-node
- Multiple parallel sorting algorithms: Bucket sort, Sample sort, hyksort
- Ghost data exchange based on neighbourhood communicators
- Builder for multi-node experiments
- Currently only works with unit charges (for benchmarking)
- Fully distributed FMM for Laplace kernel
- Tests, benchmark and scaling scripts for Archer2 Supercomputer
