+++ kifmm/src/fmm/field_translation/metadata/single_node.rs
@@ -1,0 +1,3232 @@
+//! Implementation of traits to compute metadata for field translation operations.
+use std::{
+    collections::{HashMap, HashSet},
+    sync::{Mutex, RwLock},
+};
+
+use green_kernels::{
+    helmholtz_3d::Helmholtz3dKernel, laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait,
+    types::GreenKernelEvalType,
+};
+use itertools::Itertools;
+use num::{Float, Zero};
+use rayon::prelude::*;
+use rlst::{
+    empty_array, rlst_array_from_slice2, rlst_dynamic_array2, rlst_dynamic_array3, Array,
+    BaseArray, Gemm, MatrixSvd, MultIntoResize, RawAccess, RawAccessMut, RlstScalar, Shape,
+    SvdMode, UnsafeRandomAccessByRef, UnsafeRandomAccessMut, VectorContainer,
+};
+
+use crate::{
+    fmm::{
+        constants::DEFAULT_M2L_FFT_BLOCK_SIZE,
+        field_translation::source_to_target::transfer_vector::compute_transfer_vectors_at_level,
+        helpers::single_node::{
+            coordinate_index_pointer_single_node, flip3, homogenous_kernel_scale,
+            leaf_expansion_pointers_single_node, leaf_scales_single_node,
+            leaf_surfaces_single_node, level_expansion_pointers_single_node,
+            level_index_pointer_single_node, ncoeffs_kifmm, potential_pointers_single_node,
+        },
+        types::{
+            BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, BlasMetadataIa, BlasMetadataSaRcmp,
+            FftFieldTranslation, FftMetadata, FmmSvdMode,
+        },
+    },
+    linalg::{pinv::pinv, rsvd::MatrixRsvd},
+    traits::{
+        fftw::{Dft, DftType},
+        field::{
+            FieldTranslation as FieldTranslationTrait, SourceToTargetTranslationMetadata,
+            SourceTranslationMetadata, TargetTranslationMetadata,
+        },
+        fmm::{DataAccess, HomogenousKernel, Metadata, MetadataAccess},
+        general::single_node::{AsComplex, Epsilon},
+        tree::{Domain as DomainTrait, FmmTreeNode, SingleFmmTree, SingleTree},
+    },
+    tree::{
+        constants::{
+            ALPHA_INNER, ALPHA_OUTER, NHALO, NSIBLINGS, NSIBLINGS_SQUARED, NTRANSFER_VECTORS_KIFMM,
+        },
+        helpers::find_corners,
+        types::MortonKey,
+    },
+    Evaluate, KiFmm,
+};
+
+/// Compute the cutoff rank for an SVD decomposition of a matrix from its singular values
+/// using a specified `threshold` as a tolerance parameter
+pub fn find_cutoff_rank<T: Float + RlstScalar + Gemm>(
+    singular_values: &[T],
+    threshold: T,
+    max_rank: usize,
+) -> usize {
+    let len = singular_values.len().min(max_rank);
+
+    for (i, &s) in singular_values.iter().take(len).enumerate() {
+        if s <= threshold {
+            return i;
+        }
+    }
+
+    len - 1
+}
+impl<Scalar, FieldTranslation> SourceTranslationMetadata
+    for KiFmm<Scalar, Laplace3dKernel<Scalar>, FieldTranslation>
+where
+    Scalar: RlstScalar + Default + Epsilon + MatrixSvd,
+    FieldTranslation: FieldTranslationTrait + Send + Sync,
+    <Scalar as RlstScalar>::Real: Default,
+    Self: DataAccess,
+{
[…]
