+++ kifmm/Cargo.toml
@@ -5,7 +5,7 @@ mpi = ["dep:mpi", "dep:superslice", "dep:memoffset", "dep:mpi-sys"]
 
 [package]
 name = "kifmm"
-version = "1.0.2-dev"
+version = "2.0.0"
 edition = "2021"
 authors = ["Srinath Kailasa <srinathkailasa@gmail.com>"]
 description = "Kernel-independent fast multipole method"
