+++ kifmm/Cargo.toml
@@ -25,7 +25,7 @@ itertools = "0.13.*"
 num = "0.4"
 rayon = "1.9"
 rand = "0.8.5"
-kifmm-fftw-sys = { path = "../fftw-sys" }
+kifmm-fftw-sys = { version = "0.2.0"}
 green-kernels = { version = "0.2.2", default-features = false}
 lazy_static = "1.4.0"
 num-complex = "0.4.5"
