### 2024-10-31 fca1e3c
REL: Bump version 2.0.0 (#118)
