### 2024-10-31 7555479
CLN: Single node eval trait (#119)
* Remove templates in single node eval
