### 2024-12-03 e7eb5ba
Add some uniformity, rsvd, and test blkrcmp for helmholtz
