+++ kifmm/src/fmm/field_translation/metadata/single_node.rs
@@ -622,7 +622,7 @@ where
 impl<Scalar> SourceToTargetTranslationMetadata
     for KiFmm<Scalar, Helmholtz3dKernel<Scalar>, BlasFieldTranslationIa<Scalar>>
 where
-    Scalar: RlstScalar<Complex = Scalar> + Default + MatrixSvd,
+    Scalar: RlstScalar<Complex = Scalar> + Default + MatrixRsvd,
     <Scalar as RlstScalar>::Real: Default,
 {
     fn displacements(&mut self, start_level: Option<u64>) {
@@ -785,14 +785,48 @@ where
                     let mut sigma = vec![Scalar::zero().re(); k];
                     let mut vt = rlst_dynamic_array2!(Scalar, [k, nvt]);
 
-                    tmp_gram
-                        .into_svd_alloc(
-                            u.view_mut(),
-                            vt.view_mut(),
-                            &mut sigma[..],
-                            SvdMode::Reduced,
-                        )
-                        .unwrap();
+                    let target_rank;
+
+                    match &self.source_to_target.svd_mode {
+                        &FmmSvdMode::Random {
+                            n_components,
+                            normaliser,
+                            n_oversamples,
+                            random_state,
+                        } => {
+                            // Estimate targget rank if unspecified by user
+                            if let Some(n_components) = n_components {
+                                target_rank = n_components
+                            } else {
+                                let max_equivalent_surface_ncoeffs =
+                                    self.n_coeffs_equivalent_surface.iter().max().unwrap();
+                                let max_check_surface_ncoeffs =
+                                    self.n_coeffs_check_surface.iter().max().unwrap();
+                                target_rank =
+                                    *max_equivalent_surface_ncoeffs.max(max_check_surface_ncoeffs);
+                            }
+
+                            (sigma, u, vt) = Scalar::rsvd_fixed_rank(
+                                &tmp_gram,
+                                target_rank,
+                                n_oversamples,
+                                normaliser,
+                                random_state,
+                            )
+                            .unwrap();
+                        }
+
+                        FmmSvdMode::Deterministic => {
+                            tmp_gram
+                                .into_svd_alloc(
+                                    u.view_mut(),
+                                    vt.view_mut(),
+                                    &mut sigma[..],
+                                    SvdMode::Reduced,
+                                )
+                                .unwrap();
+                        }
+                    }
 
                     let mut sigma_mat = rlst_dynamic_array2!(Scalar, [k, k]);
 
@@ -832,6 +866,389 @@ where
         }
 
         self.source_to_target = result;
+    }
+}
+
+impl<Scalar> SourceToTargetTranslationMetadata
+    for KiFmm<Scalar, Helmholtz3dKernel<Scalar>, BlasFieldTranslationSaRcmp<Scalar>>
+where
+    Scalar: RlstScalar<Complex = Scalar> + Default + MatrixRsvd,
+    <Scalar as RlstScalar>::Real: Default,
+{
+    fn displacements(&mut self, start_level: Option<u64>) {
+        let mut displacements = Vec::new();
+        let start_level = if let Some(start_level) = start_level {
+            if start_level >= 2 {
+                start_level
+            } else {
+                2
+            }
+        } else {
+            2
+        };
+
+        for level in start_level..=self.tree.source_tree().depth() {
+            let sources = self.tree.source_tree().keys(level).unwrap();
+            let n_sources = sources.len();
+            let sentinel = -1i32;
+
+            let result = vec![vec![sentinel; n_sources]; 316];
+            let result = result.into_iter().map(RwLock::new).collect_vec();
+
[…]
