### 2024-12-11 f193be9
ENH: Add rcmp strategy for testing to Helmholtz (#121)
* Add some uniformity, rsvd, and test blkrcmp for helmholtz
