+++ kifmm/src/fmm/field_translation/source_to_target/blas/single_node.rs
@@ -97,6 +97,10 @@ where
                 let multipoles =
                     rlst_array_from_slice2!(multipoles, [n_coeffs_equivalent_surface, n_sources]);
 
+                println!(
+                    "HERE {:?} {:?}",
+                    self.source_to_target.cutoff_rank, m2l_operator_index
+                );
                 // Allocate buffer to store compressed check potentials
                 let compressed_check_potentials = rlst_dynamic_array2!(
                     Scalar,
@@ -459,7 +463,7 @@ where
         let displacement_index = self.displacement_index(level);
         let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
         let n_coeffs_check_surface = self.n_coeffs_check_surface(level);
-        let sentinel = sources.len();
+        let sentinel = -1i32;
 
         // Compute the displacements
         let all_displacements = &self.source_to_target.displacements[displacement_index];
@@ -487,7 +491,7 @@ where
                     .iter()
                     .enumerate()
                     .filter(|(_, &d)| d != sentinel)
-                    .map(|(_, &j)| j)
+                    .map(|(_, &j)| j as usize)
                     .collect_vec()
             })
             .collect_vec();
