+++ kifmm/examples/mpi_test_fmm.rs
@@ -17,11 +17,12 @@ fn main() {
         datatype::PartitionMut,
         traits::{Communicator, Root},
     };
-    use rlst::RawAccess;
+    use rlst::{rlst_dynamic_array1, RawAccess};
 
     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
     let world = universe.world();
     let comm = world.duplicate();
+    let rank = comm.rank();
 
     // Tree parameters
     let prune_empty = true;
@@ -45,6 +46,8 @@ fn main() {
 
         // Generate some random test data local to each process
         let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
+        let mut charges = rlst_dynamic_array1!(f32, [n_points]);
+        charges.fill_from_seed_equally_distributed(rank as usize);
 
         let mut multi_fmm = MultiNodeBuilder::new(false)
             .tree(
@@ -57,7 +60,7 @@ fn main() {
                 sort_kind.clone(),
             )
             .unwrap()
-            .parameters(expansion_order, kernel.clone(), source_to_target)
+            .parameters(charges.data(), expansion_order, kernel.clone(), source_to_target)
             .unwrap()
             .build()
             .unwrap();
@@ -96,171 +99,178 @@ fn main() {
             root_process.gather_varcount_into(local_coords);
         }
 
-        if world.rank() == 0 {
-            let mut single_fmm = SingleNodeBuilder::new(false)
-                .tree(
-                    &all_coordinates,
-                    &all_coordinates,
-                    None,
-                    Some(local_depth + global_depth),
-                    prune_empty,
-                )
-                .unwrap()
-                .parameters(
-                    &vec![1f32; all_coordinates.len() / 3],
-                    &vec![expansion_order; (local_depth + global_depth + 1) as usize],
-                    Laplace3dKernel::new(),
-                    GreenKernelEvalType::Value,
-                    FftFieldTranslation::new(None),
-                )
-                .unwrap()
-                .build()
-                .unwrap();
-            single_fmm.evaluate().unwrap();
-            let mut expected = vec![0f32; &multi_fmm.tree.target_tree.coordinates.len() / 3];
+        println!("RANK {:?} N local sources {:?}", multi_fmm.tree.source_tree.communicator.rank(), multi_fmm.tree.source_tree.n_coordinates_tot());
 
-            multi_fmm.kernel().evaluate_st(
-                GreenKernelEvalType::Value,
-                &all_coordinates,
-                &multi_fmm.tree.target_tree.coordinates,
-                &vec![1f32; n_points * world.size() as usize],
-                &mut expected,
-            );
+        // if world.rank() == 0 {
+        //     // let mut single_fmm = SingleNodeBuilder::new(false)
+        //     //     .tree(
+        //     //         &all_coordinates,
+        //     //         &all_coordinates,
+        //     //         None,
+        //     //         Some(local_depth + global_depth),
+        //     //         prune_empty,
+        //     //     )
+        //     //     .unwrap()
+        //     //     .parameters(
+        //     //         &vec![1f32; all_coordinates.len() / 3],
+        //     //         &vec![expansion_order; (local_depth + global_depth + 1) as usize],
+        //     //         Laplace3dKernel::new(),
+        //     //         GreenKernelEvalType::Value,
+        //     //         FftFieldTranslation::new(None),
+        //     //     )
+        //     //     .unwrap()
+        //     //     .build()
+        //     //     .unwrap();
 
-            let distributed = multi_fmm.potentials().unwrap();
 
-            distributed
-                .iter()
-                .zip(expected.iter())
[…]
