### 2024-12-16 49a01b7
Updated Cargo with new versions
