+++ kifmm/src/fmm/field_translation/source_to_target/blas/single_node.rs
@@ -132,7 +132,7 @@ where
                 let mut compressed_multipoles;
                 {
                     compressed_multipoles = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                        self.source_to_target.metadata[m2l_operator_index].st.view(),
+                        self.source_to_target.metadata[m2l_operator_index].st.r(),
                         multipoles,
                     );
 
@@ -181,10 +181,10 @@ where
 
                             let compressed_check_potential = empty_array::<Scalar, 2>()
                                 .simple_mult_into_resize(
-                                    c_u_sub.view(),
+                                    c_u_sub.r(),
                                     empty_array::<Scalar, 2>().simple_mult_into_resize(
-                                        c_vt_sub.view(),
-                                        compressed_multipoles_subset.view(),
+                                        c_vt_sub.r(),
+                                        compressed_multipoles_subset.r(),
                                     ),
                                 );
 
@@ -213,11 +213,11 @@ where
                 // 3. Compute local expansions from compressed check potentials
                 {
                     let locals = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                        self.dc2e_inv_1[c2e_operator_index].view(),
+                        self.dc2e_inv_1[c2e_operator_index].r(),
                         empty_array::<Scalar, 2>().simple_mult_into_resize(
-                            self.dc2e_inv_2[c2e_operator_index].view(),
+                            self.dc2e_inv_2[c2e_operator_index].r(),
                             empty_array::<Scalar, 2>().simple_mult_into_resize(
-                                self.source_to_target.metadata[m2l_operator_index].u.view(),
+                                self.source_to_target.metadata[m2l_operator_index].u.r(),
                                 compressed_check_potentials,
                             ),
                         ),
@@ -281,7 +281,7 @@ where
                 let mut compressed_multipoles;
                 {
                     compressed_multipoles = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                        self.source_to_target.metadata[m2l_operator_index].st.view(),
+                        self.source_to_target.metadata[m2l_operator_index].st.r(),
                         multipoles,
                     );
 
@@ -347,10 +347,10 @@ where
 
                             let compressed_check_potential = empty_array::<Scalar, 2>()
                                 .simple_mult_into_resize(
-                                    c_u_sub.view(),
+                                    c_u_sub.r(),
                                     empty_array::<Scalar, 2>().simple_mult_into_resize(
-                                        c_vt_sub.view(),
-                                        compressed_multipoles_subset.view(),
+                                        c_vt_sub.r(),
+                                        compressed_multipoles_subset.r(),
                                     ),
                                 );
 
@@ -396,11 +396,11 @@ where
                 // 3. Compute local expansions from compressed check potentials
                 {
                     let locals = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                        self.dc2e_inv_1[c2e_operator_index].view(),
+                        self.dc2e_inv_1[c2e_operator_index].r(),
                         empty_array::<Scalar, 2>().simple_mult_into_resize(
-                            self.dc2e_inv_2[c2e_operator_index].view(),
+                            self.dc2e_inv_2[c2e_operator_index].r(),
                             empty_array::<Scalar, 2>().simple_mult_into_resize(
-                                self.source_to_target.metadata[m2l_operator_index].u.view(),
+                                self.source_to_target.metadata[m2l_operator_index].u.r(),
                                 compressed_check_potentials,
                             ),
                         ),
@@ -563,10 +563,10 @@ where
 
                             let check_potential = empty_array::<Scalar, 2>()
                                 .simple_mult_into_resize(
-                                    u.view(),
+                                    u.r(),
[…]
