+++ kifmm/Cargo.toml
@@ -39,6 +39,7 @@ superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
 # rlst = {version = "0.2.0", default-features = false}
 rlst = { git = "https://github.com/linalg-rs/rlst.git", features = ["mpi"]}
+bempp-distributed-tools = { git = "https://github.com/bempp/distributed_tools.git" }
 
 [dev-dependencies]
 approx = "0.5"
