+++ kifmm/Cargo.toml
@@ -1,7 +1,7 @@
 [features]
 # Treat warnings as a build error.
 strict = []
-mpi = ["dep:mpi", "dep:superslice", "dep:memoffset", "dep:mpi-sys"]
+mpi = ["dep:mpi", "dep:superslice", "dep:memoffset", "dep:mpi-sys", "rlst/mpi"]
 
 [package]
 name = "kifmm"
@@ -38,7 +38,7 @@ bytemuck = "1.16.0"
 superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
 # rlst = {version = "0.2.0", default-features = false}
-rlst = { git = "https://github.com/linalg-rs/rlst.git", features = ["mpi"]}
+rlst = { git = "https://github.com/linalg-rs/rlst.git", default-features = false }
 
 [dev-dependencies]
 approx = "0.5"
