### 2024-12-16 b42903b
Timo/new rlst (#122)
* Updated Cargo with new versions

* Fixed all warnings

* Added distributed tools

* Removed distributed tools

* Updated MPI feature of RLST

* Reformatted code
