+++ kifmm/src/fmm/field_translation/metadata/single_node.rs
@@ -121,8 +121,7 @@ where
                 mat_s[[i, i]] = Scalar::from_real(s[i]);
             }
 
-            uc2e_inv_1
-                .push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view()));
+            uc2e_inv_1.push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.r(), mat_s.r()));
             uc2e_inv_2.push(ut);
         }
 
@@ -168,10 +167,10 @@ where
                 );
 
                 let tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                    uc2e_inv_1[self.expansion_index(parent_level as u64)].view(),
+                    uc2e_inv_1[self.expansion_index(parent_level as u64)].r(),
                     empty_array::<Scalar, 2>().simple_mult_into_resize(
-                        uc2e_inv_2[self.expansion_index(parent_level as u64)].view(),
-                        ce2pc.view(),
+                        uc2e_inv_2[self.expansion_index(parent_level as u64)].r(),
+                        ce2pc.r(),
                     ),
                 );
 
@@ -244,8 +243,7 @@ where
                 mat_s[[i, i]] = Scalar::from_real(s[i]);
             }
 
-            dc2e_inv_1
-                .push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view()));
+            dc2e_inv_1.push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.r(), mat_s.r()));
             dc2e_inv_2.push(ut);
         }
 
@@ -286,10 +284,10 @@ where
                 );
 
                 let mut tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                    dc2e_inv_1[self.expansion_index(parent_level + 1)].view(),
+                    dc2e_inv_1[self.expansion_index(parent_level + 1)].r(),
                     empty_array::<Scalar, 2>().simple_mult_into_resize(
-                        dc2e_inv_2[self.expansion_index(parent_level + 1)].view(),
-                        pe2cc.view(),
+                        dc2e_inv_2[self.expansion_index(parent_level + 1)].r(),
+                        pe2cc.r(),
                     ),
                 );
 
@@ -374,8 +372,7 @@ where
                 mat_s[[i, i]] = Scalar::from_real(s[i]);
             }
 
-            uc2e_inv_1
-                .push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.view(), mat_s.view()));
+            uc2e_inv_1.push(empty_array::<Scalar, 2>().simple_mult_into_resize(v.r(), mat_s.r()));
             uc2e_inv_2.push(ut);
 
             curr = curr.first_child();
@@ -446,9 +443,9 @@ where
                 );
 
                 let tmp = empty_array::<Scalar, 2>().simple_mult_into_resize(
-                    uc2e_inv_1[level as usize].view(),
+                    uc2e_inv_1[level as usize].r(),
                     empty_array::<Scalar, 2>()
-                        .simple_mult_into_resize(uc2e_inv_2[level as usize].view(), ce2pc.view()),
+                        .simple_mult_into_resize(uc2e_inv_2[level as usize].r(), ce2pc.r()),
                 );
                 let l = i * nequiv_surface_child * nequiv_surface_parent;
                 let r = l + nequiv_surface_child * nequiv_surface_parent;
@@ -535,8 +532,7 @@ where
                 mat_s[[i, i]] = Scalar::from_real(s[i]);
             }
 
-            dc2e_inv_1
[…]
