### 2024-12-17 69a3279
Add better charge handler
