### 2024-12-17 cacfab9
Make sure python example is ok
