### 2024-12-17 d067ca8
Better error handling
