### 2024-12-17 da4a513
ENH: Add better charge handler (#123)
* Add better charge handler

* Make sure python example is ok

* Better error handling
