+++ kifmm/src/fmm/eval/single_node.rs
@@ -135,7 +135,7 @@ mod test {
     use crate::{
         fmm::types::BlasFieldTranslationIa,
         traits::{
-            fmm::DataAccess,
+            fmm::ChargeHandler,
             tree::{FmmTreeNode, SingleFmmTree, SingleTree},
         },
         tree::{constants::ALPHA_INNER, helpers::points_fixture, types::MortonKey},
@@ -340,7 +340,7 @@ mod test {
         direct.iter().zip(potential).for_each(|(&d, &p)| {
             let abs_error = RlstScalar::abs(d - p);
             let rel_error = abs_error / p;
-            // println!("err {:?} \nd {:?} \np {:?}", rel_error, d, &p);
+            println!("err {:?} \nd {:?} \np {:?}", rel_error, d, &p);
             assert!(rel_error <= threshold)
         });
     }
@@ -625,7 +625,8 @@ mod test {
         let mut rng = StdRng::seed_from_u64(1);
         charges.data_mut().iter_mut().for_each(|c| *c = rng.gen());
 
-        fmm.clear(charges.data());
+        let _ = fmm.clear();
+        let _ = fmm.attach_charges_unordered(charges.data());
         fmm.evaluate().unwrap();
 
         let fmm = Box::new(fmm);
