+++ kifmm/python/kifmm_py/__init__.py
@@ -361,7 +361,7 @@ class SingleNodeTree:
         self.prune_empty = prune_empty
         self.n_sources = len(sources)
         self.n_targets = len(targets)
-        self.ncharges = len(charges)
+        self.n_charges = len(charges)
         self.sources_c = ffi.cast("void* ", self.sources.ctypes.data)
         self.targets_c = ffi.cast("void* ", self.targets.ctypes.data)
         self.charges_c = ffi.cast("void* ", self.charges.ctypes.data)
@@ -373,7 +373,7 @@ class SingleNodeTree:
         except:
             TypeError("Incompatible number of new_charges for sources")
 
-        self.ncharges = len(new_charges)
+        self.n_charges = len(new_charges)
         self.charges = new_charges
         self.charges_c = ffi.cast("void* ", new_charges.ctypes.data)
 
@@ -405,7 +405,7 @@ class KiFmm:
         dim = 3
 
         try:
-            if self._tree.ncharges // (self._tree.n_sources // dim) > 1:
+            if self._tree.n_charges // (self._tree.n_sources // dim) > 1:
                 assert isinstance(self._field_translation, BlasFieldTranslation)
         except:
             raise TypeError(
@@ -450,7 +450,7 @@ class KiFmm:
                         self._tree.targets_c,
                         self._tree.n_targets,
                         self._tree.charges_c,
-                        self._tree.ncharges,
+                        self._tree.n_charges,
                         self._tree.prune_empty,
                         self._tree.n_crit,
                         self._tree.depth,
@@ -469,7 +469,7 @@ class KiFmm:
                         self._tree.targets_c,
                         self._tree.n_targets,
                         self._tree.charges_c,
-                        self._tree.ncharges,
+                        self._tree.n_charges,
                         self._tree.prune_empty,
                         self._tree.n_crit,
                         self._tree.depth,
@@ -493,7 +493,7 @@ class KiFmm:
                         self._tree.targets_c,
                         self._tree.n_targets,
                         self._tree.charges_c,
-                        self._tree.ncharges,
+                        self._tree.n_charges,
                         self._tree.prune_empty,
                         self._tree.n_crit,
                         self._tree.depth,
@@ -513,7 +513,7 @@ class KiFmm:
                         self._tree.targets_c,
                         self._tree.n_targets,
                         self._tree.charges_c,
-                        self._tree.ncharges,
+                        self._tree.n_charges,
                         self._tree.prune_empty,
                         self._tree.n_crit,
                         self._tree.depth,
@@ -543,7 +543,7 @@ class KiFmm:
                             self._tree.targets_c,
                             self._tree.n_targets,
                             self._tree.charges_c,
-                            self._tree.ncharges,
+                            self._tree.n_charges,
                             self._tree.prune_empty,
                             self._tree.n_crit,
                             self._tree.depth,
@@ -566,7 +566,7 @@ class KiFmm:
                             self._tree.targets_c,
                             self._tree.n_targets,
                             self._tree.charges_c,
-                            self._tree.ncharges,
+                            self._tree.n_charges,
                             self._tree.prune_empty,
                             self._tree.n_crit,
                             self._tree.depth,
@@ -593,7 +593,7 @@ class KiFmm:
                             self._tree.targets_c,
                             self._tree.n_targets,
                             self._tree.charges_c,
-                            self._tree.ncharges,
+                            self._tree.n_charges,
                             self._tree.prune_empty,
                             self._tree.n_crit,
                             self._tree.depth,
@@ -613,7 +613,7 @@ class KiFmm:
[…]
