+++ kifmm/src/bindings.rs
@@ -93,6 +93,17 @@ pub mod types {
         pub scalar: ScalarType,
     }
 
+    /// Expansion data
+    #[repr(C)]
+    pub struct Expansion {
+        /// Length of underlying buffer, of length n_eval_mode*n_evals*n_coeffs
+        pub len: usize,
+        /// Pointer to underlying buffer
+        pub data: *const c_void,
+        /// Associated scalar type
+        pub scalar: ScalarType,
+    }
+
     /// Container for multiple Potentials. Used when FMM run over multiple
     /// charge vectors.
     #[repr(C)]
@@ -1553,14 +1564,14 @@ pub mod api {
         fmm::types::FmmEvalType,
         traits::{
             fmm::{ChargeHandler, DataAccess},
-            tree::{SingleFmmTree, SingleTree, TreeNode},
+            tree::{FmmTreeNode, SingleFmmTree, SingleTree, TreeNode},
         },
         tree::types::MortonKey,
         BlasFieldTranslationIa, Evaluate, FftFieldTranslation, KiFmm,
     };
 
     use super::{
-        c32, c64, BlasFieldTranslationSaRcmp, CommunicationTimes, Coordinates, FmmCType,
+        c32, c64, BlasFieldTranslationSaRcmp, CommunicationTimes, Coordinates, Expansion, FmmCType,
         FmmEvaluator, FmmOperatorTimes, FmmTranslationCType, GlobalIndices, Laplace3dKernel,
         MortonKeys, Potential, Potentials, ScalarType,
     };
@@ -2872,6 +2883,528 @@ pub mod api {
         }
     }
 
+    /// Query for locals at a specific key.
+    ///
+    /// # Parameters
+    ///
+    /// - `fmm`: Pointer to an `FmmEvaluator` instance.
+    /// - `key`: The identifier of a node.
+    ///
+    /// # Safety
+    /// This function is intended to be called from C. The caller must ensure that:
+    /// - Input data corresponds to valid pointers
+    /// - That they remain valid for the duration of the function call
+    #[no_mangle]
+    pub unsafe extern "C" fn local(fmm: *mut FmmEvaluator, key: u64) -> *mut Expansion {
+        assert!(!fmm.is_null());
+        let ctype = unsafe { (*fmm).get_ctype() };
+        let ctranslation_type = unsafe { (*fmm).get_ctranslation_type() };
+        let pointer = unsafe { (*fmm).get_pointer() };
+
+        match ctype {
+            FmmCType::Laplace32 => match ctranslation_type {
+                FmmTranslationCType::Fft => {
+                    let fmm =
+                        pointer as *mut KiFmm<f32, Laplace3dKernel<f32>, FftFieldTranslation<f32>>;
+
+                    let key = MortonKey::from_morton(key);
+
+                    let len;
+                    let ptr;
+                    unsafe {
+                        if let Some(m) = (*fmm).local(&key) {
+                            len = m.len();
+                            ptr = m.as_ptr()
+                        } else {
+                            len = 0;
+                            ptr = std::ptr::null()
+                        }
+                    }
+
+                    let data = ptr as *const c_void;
+
+                    let m = Expansion {
+                        len,
+                        data,
+                        scalar: ScalarType::F32,
+                    };
+
+                    Box::into_raw(Box::new(m))
+                }
+
+                FmmTranslationCType::Blas => {
+                    let fmm = pointer
+                        as *mut KiFmm<f32, Laplace3dKernel<f32>, BlasFieldTranslationSaRcmp<f32>>;
+
+                    let key = MortonKey::from_morton(key);
+
+                    let len;
+                    let ptr;
+                    unsafe {
+                        if let Some(m) = (*fmm).local(&key) {
+                            len = m.len();
+                            ptr = m.as_ptr()
+                        } else {
[…]
