### 2024-12-28 0ab60ca
ENH: Add More to Python API (#124)
* Add mt kernel

* Add more to API

* Clippy

* Privates

* Force update cargo mpirun
