+++ kifmm/pyproject.toml
@@ -4,6 +4,7 @@ build-backend = "maturin"
 
 [project]
 name = "kifmm_py"
+version="2.0.0"
 requires-python = "==3.10.*"
 classifiers = [
     "Programming Language :: Rust",
