### 2025-01-01 b0ef34f
Add random charges
