+++ kifmm/examples/mpi_test_ghost_exchange.rs
@@ -62,7 +62,7 @@
 //             sort_kind,
 //         )
 //         .unwrap()
-//         .parameters(expansion_order, kernel, source_to_target)
+//         .parameters(charges, expansion_order, kernel, source_to_target)
 //         .unwrap()
 //         .build()
 //         .unwrap();
