+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -1,8 +1,13 @@
 use green_kernels::laplace_3d::Laplace3dKernel;
 use green_kernels::traits::Kernel as KernelTrait;
 use green_kernels::types::GreenKernelEvalType;
+use itertools::izip;
 use itertools::Itertools;
-use mpi::traits::Equivalence;
+use mpi::datatype::Partition;
+use mpi::datatype::PartitionMut;
+use mpi::datatype::Partitioned;
+use mpi::traits::{Communicator, CommunicatorCollectives, Equivalence};
+use mpi::{Count, Rank};
 use num::{Float, Zero};
 use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 use rlst::{
@@ -935,7 +940,7 @@ where
 {
     type Scalar = Scalar;
 
-    fn metadata(&mut self, eval_type: GreenKernelEvalType, _charges: &[Self::Scalar]) {
+    fn metadata(&mut self, eval_type: GreenKernelEvalType, charges: &[Self::Scalar]) {
         // Check if computing potentials, or potentials and derivatives
         match eval_type {
             GreenKernelEvalType::Value => {}
@@ -1025,8 +1030,253 @@ where
         let potential_send_pointers =
             potential_pointers_multi_node(&self.tree.target_tree, kernel_eval_size, &potentials);
 
+        // Setup neighbourhood communication for charges
+        let n_sources = charges.len() as u64;
+        let size = self.tree.source_tree.communicator.size();
+        let mut counts = vec![0u64; size as usize];
+        self.tree
+            .source_tree
+            .communicator
+            .all_gather_into(&n_sources, &mut counts[..]);
+
+        let mut displacements = Vec::new();
+
+        let mut curr = 0;
+        for count in counts.iter() {
+            displacements.push((curr, curr + count - 1));
+            curr = curr + count;
+        }
+
+        let rank = self.tree.source_tree.communicator.rank();
+        let mut local_displacement = 0;
+        for count in counts.iter().take(rank as usize) {
+            local_displacement += count;
+        }
+
+        let global_indices = &self.tree.source_tree.global_indices;
+
+        let mut ranks = Vec::new();
+        let mut send_counts = vec![0 as Count; self.tree.source_tree.communicator.size() as usize];
+        let mut send_marker = vec![0 as Rank; self.tree.source_tree.communicator.size() as usize];
+
+        for &global_index in global_indices.iter() {
+            let rank = displacements
+                .iter()
+                .position(|&(start, end)| {
+                    global_index >= (start as usize) && global_index <= (end as usize)
+                })
+                .unwrap();
+            ranks.push(rank as i32);
+            send_counts[rank] += 1;
+        }
+
+        for (rank, &send_count) in send_counts.iter().enumerate() {
+            if send_count > 0 {
+                send_marker[rank] = 1;
+            }
+        }
+
+        // Sort queries by destination rank
+        let queries = {
+            let mut indices = (0..global_indices.len()).collect_vec();
+            indices.sort_by_key(|&i| ranks[i]);
+
+            let mut sorted_queries_ = Vec::with_capacity(global_indices.len());
+            for i in indices {
+                sorted_queries_.push(global_indices[i])
+            }
+            sorted_queries_
+        };
+
+        // Sort ranks of queries into rank order
+        ranks.sort();
+
+        // Compute the receive counts, and mark again processes involved
+        let mut receive_counts = vec![0i32; self.tree.source_tree.communicator.size() as usize];
+        let mut receive_marker = vec![0i32; self.tree.source_tree.communicator.size() as usize];
+
+        self.tree
+            .source_tree
+            .communicator
[…]
