### 2025-01-03 57e8b5d
Add benchmarks
