+++ kifmm/benches/helmholtz_direct_f64.rs
@@ -2,9 +2,10 @@ use std::time::Duration;
 
 use criterion::{criterion_group, criterion_main, Criterion};
 use green_kernels::traits::Kernel;
-use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+use green_kernels::{helmholtz_3d::Helmholtz3dKernel, types::GreenKernelEvalType};
 use kifmm::tree::helpers::points_fixture;
-use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut};
+use num::{Complex, One};
+use rlst::{c64, rlst_dynamic_array2, RawAccess, RawAccessMut};
 
 
 fn multithreaded_direct_f64(c: &mut Criterion) {
@@ -22,13 +23,14 @@ fn multithreaded_direct_f64(c: &mut Criterion) {
 
         // FFT based M2L for a vector of charges
         let nvecs = 1;
-        let tmp = vec![1.0; nsources * nvecs];
-        let mut charges = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        let tmp = vec![Complex::one(); nsources * nvecs];
+        let mut charges = rlst_dynamic_array2!(c64, [nsources, nvecs]);
         charges.data_mut().copy_from_slice(&tmp);
+        let wavenumber = 1f64;
 
-        let kernel = Laplace3dKernel::new();
+        let kernel = Helmholtz3dKernel::new(wavenumber);
 
-        let mut result = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        let mut result = rlst_dynamic_array2!(c64, [nsources, nvecs]);
 
         group.bench_function(format!("N={nsources}"), |b| {
             b.iter(|| {
@@ -51,13 +53,14 @@ fn multithreaded_direct_f64(c: &mut Criterion) {
 
         // FFT based M2L for a vector of charges
         let nvecs = 1;
-        let tmp = vec![1.0; nsources * nvecs];
-        let mut charges = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        let tmp = vec![Complex::one(); nsources * nvecs];
+        let mut charges = rlst_dynamic_array2!(c64, [nsources, nvecs]);
         charges.data_mut().copy_from_slice(&tmp);
+        let wavenumber =1f64;
 
-        let kernel = Laplace3dKernel::new();
+        let kernel = Helmholtz3dKernel::new(wavenumber);
 
-        let mut result = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        let mut result = rlst_dynamic_array2!(c64, [nsources, nvecs]);
 
         group.bench_function(format!("N={nsources}"), |b| {
             b.iter(|| {
@@ -80,13 +83,14 @@ fn multithreaded_direct_f64(c: &mut Criterion) {
 
         // FFT based M2L for a vector of charges
         let nvecs = 1;
-        let tmp = vec![1.0; nsources * nvecs];
-        let mut charges = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        let tmp = vec![Complex::one(); nsources * nvecs];
+        let mut charges = rlst_dynamic_array2!(c64, [nsources, nvecs]);
         charges.data_mut().copy_from_slice(&tmp);
+        let wavenumber = 1f64;
 
-        let kernel = Laplace3dKernel::new();
+        let kernel = Helmholtz3dKernel::new(wavenumber);
 
-        let mut result = rlst_dynamic_array2!(f64, [nsources, nvecs]);
+        let mut result = rlst_dynamic_array2!(c64, [nsources, nvecs]);
 
         group.bench_function(format!("N={nsources}"), |b| {
             b.iter(|| {
[…]
