### 2025-01-03 78538af
Local deps to avoid sleef
