+++ kifmm/Cargo.toml
@@ -27,7 +27,7 @@ rayon = "1.9"
 rand = "0.8.5"
 kifmm-fftw-sys = { path = "../fftw-sys" }
 #green-kernels = { version = "0.2.2", default-features = false}
-green-kernels = { git = "https://github.com/bempp/green-kernels" }
+green-kernels = { git = "https://github.com/skailasa/green-kernels" }
 lazy_static = "1.4.0"
 num-complex = "0.4.5"
 mpi = { version = "0.8.*", optional = true}
@@ -38,7 +38,7 @@ bytemuck = "1.16.0"
 superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
 # rlst = {version = "0.2.0", default-features = false}
-rlst = { git = "https://github.com/linalg-rs/rlst.git", default-features = false }
+rlst = { git = "https://github.com/skailasa/rlst.git", default-features = false }
 
 [dev-dependencies]
 approx = "0.5"
