### 2025-01-06 e189a6d
Start rewriting grid search script
