+++ kifmm/src/tree/helpers.rs
@@ -45,36 +45,110 @@ pub fn points_fixture<T: Float + RlstScalar + rand::distributions::uniform::Samp
     points
 }
 
-/// Points fixture for testing, uniformly samples on surface of a sphere of diameter 1.
-///
-/// # Arguments
-/// * `n_points` - The number of points to sample.
-/// * `min` - The minimum coordinate value along each axis.
-/// * `max` - The maximum coordinate value along each axis.
+// /// Points fixture for testing, uniformly samples on surface of a sphere of diameter 1.
+// ///
+// /// # Arguments
+// /// * `n_points` - The number of points to sample.
+// /// * `min` - The minimum coordinate value along each axis.
+// /// * `max` - The maximum coordinate value along each axis.
+// pub fn points_fixture_sphere<T: RlstScalar + rand::distributions::uniform::SampleUniform>(
+//     n_points: usize,
+// ) -> PointsMat<T> {
+//     // Generate a set of randomly distributed points
+//     let mut range = StdRng::seed_from_u64(0);
+//     let pi = T::from(3.14159).unwrap();
+//     let two = T::from(2.0).unwrap();
+//     let half = T::from(0.5).unwrap();
+
+//     let between = rand::distributions::Uniform::from(T::zero()..T::one());
+
+//     let mut points = rlst_dynamic_array2!(T, [3, n_points]);
+//     let mut phi = rlst_dynamic_array2!(T, [n_points, 1]);
+//     let mut theta = rlst_dynamic_array2!(T, [n_points, 1]);
+
+//     for i in 0..n_points {
+//         phi[[i, 0]] = between.sample(&mut range) * two * pi;
+//         theta[[i, 0]] = ((between.sample(&mut range) - half) * two).acos();
+//     }
+
+//     for i in 0..n_points {
+//         points[[0, i]] = half * theta[[i, 0]].sin() * phi[[i, 0]].cos() + half;
+//         points[[1, i]] = half * theta[[i, 0]].sin() * phi[[i, 0]].sin() + half;
+//         points[[2, i]] = half * theta[[i, 0]].cos() + half;
+//     }
+
+//     points
+// }
+
 pub fn points_fixture_sphere<T: RlstScalar + rand::distributions::uniform::SampleUniform>(
     n_points: usize,
 ) -> PointsMat<T> {
-    // Generate a set of randomly distributed points
-    let mut range = StdRng::seed_from_u64(0);
-    let pi = T::from(3.134159).unwrap();
+    use rand::distributions::Uniform;
+    use rand::rngs::StdRng;
+    use rand::SeedableRng;
+
+    // Seeded random number generator for reproducibility
+    let mut rng = StdRng::seed_from_u64(0);
+    let pi = T::from(std::f64::consts::PI).unwrap();
     let two = T::from(2.0).unwrap();
-    let half = T::from(0.5).unwrap();
+    let one = T::one();
 
-    let between = rand::distributions::Uniform::from(T::zero()..T::one());
+    // Uniform distributions for phi and cos(theta)
+    let phi_dist = Uniform::from(T::zero()..(two * pi));
+    let u_dist = Uniform::from(T::zero()..one);
 
+    // Initialize points array
     let mut points = rlst_dynamic_array2!(T, [3, n_points]);
-    let mut phi = rlst_dynamic_array2!(T, [n_points, 1]);
-    let mut theta = rlst_dynamic_array2!(T, [n_points, 1]);
 
     for i in 0..n_points {
-        phi[[i, 0]] = between.sample(&mut range) * two * pi;
-        theta[[i, 0]] = ((between.sample(&mut range) - half) * two).acos();
+        // Generate random phi and theta
[…]
