+++ kifmm/src/fmm/eval/single_node.rs
@@ -138,7 +138,7 @@ mod test {
             fmm::ChargeHandler,
             tree::{FmmTreeNode, SingleFmmTree, SingleTree},
         },
-        tree::{constants::ALPHA_INNER, helpers::points_fixture, types::MortonKey},
+        tree::{constants::ALPHA_INNER, helpers::{points_fixture, points_fixture_sphere}, types::MortonKey},
         BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, FmmSvdMode, SingleNodeBuilder,
         SingleNodeFmmTree,
     };
@@ -320,7 +320,17 @@ mod test {
             GreenKernelEvalType::ValueDeriv => 4,
         };
 
-        let leaf_idx = 0;
+        let mut leaf_idx = 0;
+
+        for (i, leaf) in fmm.tree().target_tree().all_leaves().unwrap().iter().enumerate() {
+            if let Some(n_coords) = fmm.tree().target_tree().n_coordinates(&leaf) {
+                if n_coords > 0 {
+                    leaf_idx = i;
+                    break;
+                }
+            }
+        }
+
         let leaf = fmm.tree().target_tree().all_leaves().unwrap()[leaf_idx];
         let potential = fmm.potential(&leaf).unwrap()[0];
 
@@ -591,6 +601,9 @@ mod test {
         let max = None;
         let sources = points_fixture::<f64>(n_sources, min, max, Some(0));
         let targets = points_fixture::<f64>(n_targets, min, max, Some(1));
+
+        let sources = points_fixture_sphere(n_sources);
+        let targets = points_fixture_sphere(n_targets);
 
         // FMM parameters
         let n_crit = Some(100);
