+++ scripts/src/bin/grid_search_single_node.rs
@@ -14,7 +14,7 @@ use kifmm::{
         general::single_node::{AsComplex, Epsilon, Hadamard8x8},
         tree::{SingleFmmTree, SingleTree},
     },
-    tree::helpers::points_fixture,
+    tree::helpers::{points_fixture, points_fixture_oblate_spheroid, points_fixture_sphere},
     BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder,
 };
 use num::Float;
@@ -23,6 +23,7 @@ use rlst::{rlst_dynamic_array2, MatrixSvd, RawAccess, RawAccessMut, RlstScalar};
 
 fn grid_search_laplace_blas<T>(
     filename: String,
+    geometry: &str,
     n_points: usize,
     expansion_order_vec: &[usize],
     svd_threshold_vec: &[Option<T>],
@@ -65,9 +66,23 @@ fn grid_search_laplace_blas<T>(
     // Setup random sources and targets
     let n_sources = n_points;
     let n_targets = n_points;
-    let sources = points_fixture::<T::Real>(n_sources, None, None, Some(0));
-    let targets = points_fixture::<T::Real>(n_targets, None, None, Some(1));
+    let sources;
+    let targets;
+    if geometry == "uniform" {
+        sources = points_fixture::<T::Real>(n_sources, None, None, Some(0));
+        targets = points_fixture::<T::Real>(n_targets, None, None, Some(1));
+    } else if geometry == "sphere" {
+        sources = points_fixture_sphere::<T::Real>(n_sources);
+        targets = points_fixture_sphere::<T::Real>(n_targets);
+    } else if geometry == "spheroid" {
+        sources = points_fixture_oblate_spheroid::<T::Real>(n_sources, T::from(1.0).unwrap(), T::from(0.5).unwrap());
+        targets = points_fixture_oblate_spheroid::<T::Real>(n_targets, T::from(1.0).unwrap(), T::from(0.5).unwrap());
+    } else {
+        panic!("Unsupported geometry")
+    }
+
     let n_vecs = 1;
+
     let tmp = vec![T::one(); n_sources * n_vecs];
     let mut charges = rlst_dynamic_array2!(T, [n_sources, n_vecs]);
     charges.data_mut().copy_from_slice(&tmp);
@@ -136,13 +151,6 @@ fn grid_search_laplace_blas<T>(
         s.elapsed().as_secs()
     );
 
-    // Setup random sources and targets
-    let n_sources = 1000000;
-    let sources = points_fixture::<T::Real>(n_sources, None, None, Some(0));
-    let n_vecs = 1;
-    let tmp = vec![T::one(); n_sources * n_vecs];
-    let mut charges = rlst_dynamic_array2!(T, [n_sources, n_vecs]);
-    charges.data_mut().copy_from_slice(&tmp);
 
     let mut progress = 0;
     for (i, fmm, setup_time) in fmms.lock().unwrap().iter_mut() {
@@ -158,11 +166,10 @@ fn grid_search_laplace_blas<T>(
         println!("BLAS Evaluated {:?}/{:?}", progress, n_params);
 
         let mut leaf_idx = 0;
-
-        for leaf in fmm.tree().target.tree().target_tree().all_leaves().unwrap().iter() {
+        for (i, leaf) in fmm.tree().target_tree().all_leaves().unwrap().iter().enumerate() {
             if let Some(n_targets) = fmm.tree().target_tree().n_coordinates(&leaf) {
                 if n_targets > 0 {
-                    leaf_idx = 0;
+                    leaf_idx = i;
                     break;
                 }
             }
@@ -175,7 +182,7 @@ fn grid_search_laplace_blas<T>(
[…]
