### 2025-01-11 d8afe35
Add f32 amd bences
