+++ kifmm/Cargo.toml
@@ -48,7 +48,15 @@ criterion = { version = "0.5.*", features = ["html_reports"]}
 cbindgen = "0.26.0"
 
 [[bench]]
-name = "laplace_f32"
+name = "laplace_f32_amd_uniform"
+harness = false
+
+[[bench]]
+name = "laplace_f32_amd_sphere"
+harness = false
+
+[[bench]]
+name = "laplace_f32_amd_spheroid"
 harness = false
 
 [[bench]]

[…]
