### 2025-01-11 792e95d
Add basic flop counter for fft m2l
