+++ kifmm/src/fmm/types.rs
@@ -97,6 +97,9 @@ where
     FieldTranslation: FieldTranslationTrait,
     <Scalar as RlstScalar>::Real: Default,
 {
+    /// Number of flops in M2L
+    pub nflops: f64,
+
     /// Operator runtimes
     pub operator_times: Vec<FmmOperatorTime>,
 
@@ -236,6 +239,7 @@ where
 {
     fn default() -> Self {
         KiFmm {
+            nflops: 0.,
             timed: false,
             operator_times: Vec::default(),
             communication_times: Vec::default(),

[…]
