### 2025-01-11 7964a45
Add flop counter
