+++ kifmm/src/fmm/field_translation/source_to_target/blas/single_node.rs
@@ -1,12 +1,12 @@
 //! Multipole to local field translation trait implementation using BLAS.
 
-use std::sync::Mutex;
+use std::{ops::DerefMut, sync::Mutex};
 
 use itertools::Itertools;
 use rayon::prelude::*;
 use rlst::{
     empty_array, rlst_array_from_slice2, rlst_dynamic_array2, MultIntoResize, RawAccess,
-    RawAccessMut, RlstScalar,
+    RawAccessMut, RlstScalar, Shape
 };
 
 use green_kernels::traits::Kernel as KernelTrait;
@@ -50,6 +50,7 @@ where
         let c2e_operator_index = self.c2e_operator_index(level);
         let displacement_index = self.displacement_index(level);
         let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
+        let mut nflops = 0.;
 
         // let sentinel = sources.len();
         let sentinel = -1i32;
@@ -140,13 +141,25 @@ where
                     }
                 }
 
+                // Count flops for applying scale
+                nflops += (compressed_multipoles.data().len()) as f64;
+
+                // Count flops for compressed check potentials
+                let [m, n] = self.source_to_target.metadata[m2l_operator_index].st.shape();
+                nflops += (m * n_sources * (2 * n - 1)) as f64;
+
                 // 2. Apply BLAS operation
+                let tmp1 = Mutex::new(0f64);
+
                 {
                     (0..NTRANSFER_VECTORS_KIFMM)
                         .into_par_iter()
                         .zip(multipole_idxs)
                         .zip(local_idxs)
                         .for_each(|((c_idx, multipole_idxs), local_idxs)| {
+
+                            let mut tmp2 = 0f64;
+
                             let c_u_sub =
                                 &self.source_to_target.metadata[m2l_operator_index].c_u[c_idx];
                             let c_vt_sub =
@@ -175,6 +188,9 @@ where
                                     );
                             }
 
+                            // Count flops for copying subset of data
+                            tmp2 += (compressed_multipoles_subset.data().len()) as f64;
+
                             let compressed_check_potential = empty_array::<Scalar, 2>()
                                 .simple_mult_into_resize(
                                     c_u_sub.r(),
@@ -183,6 +199,12 @@ where
                                         compressed_multipoles_subset.r(),
                                     ),
                                 );
+
+                            // Count flops for convolution
+                            let [m, n] = c_vt_sub.shape();
+                            tmp2 += (m * multipole_idxs.len() * (2 * n - 1)) as f64;
+                            let [m, n] = c_u_sub.shape();
+                            tmp2 += (m * multipole_idxs.len() * (2 * n - 1)) as f64;
 
                             for (multipole_idx, &local_idx) in local_idxs.iter().enumerate() {
                                 let check_potential_lock =
@@ -203,8 +225,14 @@ where
                                     .zip(tmp)
                                     .for_each(|(l, r)| *l += *r);
                             }
+
+                            // Count flops for check potentials
+                            tmp2 += (local_idxs.len() * self.source_to_target.cutoff_rank[m2l_operator_index]) as f64;
+                            *tmp1.lock().unwrap().deref_mut() += tmp2;
                         });
                 }
+
+                nflops += *tmp1.lock().unwrap().deref_mut();
 
                 // 3. Compute local expansions from compressed check potentials
                 {
@@ -219,6 +247,17 @@ where
                         ),
                     );
 
+                    // Count flops for evaluating locals
+                    let [m, n] = self.dc2e_inv_2[c2e_operator_index].shape();
[…]
