+++ kifmm/Cargo.toml
@@ -48,23 +48,7 @@ criterion = { version = "0.5.*", features = ["html_reports"]}
 cbindgen = "0.26.0"
 
 [[bench]]
-name = "laplace_f32_amd_uniform"
-harness = false
-
-[[bench]]
-name = "laplace_f32_amd_sphere"
-harness = false
-
-[[bench]]
-name = "laplace_f32_amd_spheroid"
-harness = false
-
-[[bench]]
-name = "laplace_f32_m1"
-harness = false
-
-[[bench]]
-name = "laplace_f64"
+name = "laplace_f32_amd"
 harness = false
 
 [package.metadata.docs.rs]

[…]
