### 2025-01-11 b7945a0
Add amd f64 benchmark
