### 2025-01-13 fb0da08
Add storage costs
