+++ scripts/src/bin/grid_search_single_node.rs
@@ -497,104 +497,117 @@ fn main() {
     // FFT Grid Search
     {
         // for geometry in geometries.iter() {
-        //     let precision = "f32";
-        //     let expansion_order_vec: Vec<usize> = vec![3, 4, 5];
-        //     let depth_vec: Vec<u64> = vec![4, 5];
-        //     let max_m2l_fft_block_size_vec = vec![16, 32, 64, 128];
+        //     //     let precision = "f32";
+        //     //     let expansion_order_vec: Vec<usize> = vec![3, 4, 5];
+        //     //     let depth_vec: Vec<u64> = vec![4, 5];
+        //     //     let max_m2l_fft_block_size_vec = vec![16, 32, 64, 128];
 
-        //     grid_search_laplace_fft::<f32>(
-        //         format!("grid_search_laplace_fft_{}_{}_{}", precision, arch, geometry).to_string(),
-        //         &geometry,
-        //         n_points,
-        //         &expansion_order_vec,
-        //         &depth_vec,
-        //         &max_m2l_fft_block_size_vec,
-        //     );
-        // }
+        //     //     grid_search_laplace_fft::<f32>(
+        //     //         format!("grid_search_laplace_fft_{}_{}_{}", precision, arch, geometry).to_string(),
+        //     //         &geometry,
+        //     //         n_points,
+        //     //         &expansion_order_vec,
+        //     //         &depth_vec,
+        //     //         &max_m2l_fft_block_size_vec,
+        //     //     );
+        //     // }
 
-        // for geometry in geometries.iter() {
-        //     let precision = "f64";
-        //     // let expansion_order_vec: Vec<usize> = vec![6, 7, 8, 9, 10, 11];
-        //     let expansion_order_vec: Vec<usize> = vec![11];
-        //     let depth_vec: Vec<u64> = vec![4, 5];
-        //     let max_m2l_fft_block_size_vec = vec![16, 32, 64, 128];
+        for geometry in geometries.iter() {
+            let precision = "f64";
+            // let expansion_order_vec: Vec<usize> = vec![6, 7, 8, 9, 10, 11];
+            let expansion_order_vec: Vec<usize> = vec![12, 13];
+            let depth_vec: Vec<u64> = vec![4, 5];
+            let max_m2l_fft_block_size_vec = vec![16, 32, 64];
 
-        //     grid_search_laplace_fft::<f64>(
-        //         format!("grid_search_laplace_fft_max_{}_{}_{}", precision, arch, geometry).to_string(),
-        //         &geometry,
-        //         n_points,
-        //         &expansion_order_vec,
-        //         &depth_vec,
-        //         &max_m2l_fft_block_size_vec,
-        //     );
-        // }
+            grid_search_laplace_fft::<f64>(
+                format!(
+                    "grid_search_laplace_fft_max_two_{}_{}_{}",
+                    precision, arch, geometry
+                )
+                .to_string(),
+                &geometry,
+                n_points,
+                &expansion_order_vec,
+                &depth_vec,
+                &max_m2l_fft_block_size_vec,
+            );
+        }
     }
 
     // BLAS Grid Search
 
-    {
-        // for geometry in geometries.iter() {
-        //     let precision = "f32";
-        //     let expansion_order_vec: Vec<usize> = vec![2, 3, 4, 5];
-        //     let depth_vec: Vec<u64> = vec![5];
+    // {
+    //     // for geometry in geometries.iter() {
+    //     //     let precision = "f32";
+    //     //     let expansion_order_vec: Vec<usize> = vec![2, 3, 4, 5];
+    //     //     let depth_vec: Vec<u64> = vec![5];
 
-        //     let rsvd_settings_vec = [
[…]
