+++ kifmm/benches/laplace_f64_amd.rs
@@ -26,16 +26,16 @@ fn fft_f64(c: &mut Criterion) {
 
     // Expansion order
     let e_vec = vec![
-        [7, 7, 7],    // 6 digits, for each geometry
-        [9, 9, 9],    // 8 digits for each geometry
-        [11, 11, 11], // 10 digits for each geometry
+        [6, 7, 7],    // 6 digits, for each geometry
+        [8, 9, 9],    // 8 digits for each geometry
+        [10, 11, 11], // 10 digits for each geometry
     ];
 
     // Block size
     let b_vec = vec![
-        [64, 32, 128], // 6 digits, for each geometry
+        [32, 32, 128], // 6 digits, for each geometry
         [32, 16, 64],  // 8 digits for each geometry
-        [16, 64, 64],  // 10 digits for each geometry
+        [32, 64, 64],  // 10 digits for each geometry
     ];
 
     let experiments = [6, 8, 10]; // number of digits sought
@@ -120,40 +120,40 @@ fn blas_f64(c: &mut Criterion) {
 
     // Expansion order
     let e_vec = vec![
-        [7, 7, 7],    // 6 digits, for each geometry
-        [8, 8, 9],    // 8 digits for each geometry
-        [10, 10, 11], // 10 digits for each geometry
+        [6, 6, 6],    // 6 digits, for each geometry
+        [8, 8, 8],    // 8 digits for each geometry
+        [10, 10, 10], // 10 digits for each geometry
     ];
 
     // SVD threshold
     let svd_threshold_vec = vec![
-        [Some(1e-5), Some(1e-5), Some(1e-9)],  // 6 digits
-        [Some(1e-5), Some(1e-7), Some(1e-7)],  // 8 digits
-        [Some(1e-9), Some(1e-7), Some(1e-11)], // 10 digits
+        [Some(1e-5), Some(1e-5), Some(1e-5)],  // 6 digits
+        [Some(1e-5), Some(1e-7), Some(1e-5)],  // 8 digits
+        [Some(1e-7), Some(1e-9), Some(1e-7)], // 10 digits
     ];
 
     let svd_mode_vec = vec![
         [
-            FmmSvdMode::new(true, None, None, Some(5), None),
+            FmmSvdMode::new(true, None, None, Some(20), None),
             FmmSvdMode::new(true, None, None, Some(20), None),
             FmmSvdMode::new(true, None, None, Some(10), None),
         ],
         [
+            FmmSvdMode::new(true, None, None, Some(10), None),
+            FmmSvdMode::new(true, None, None, Some(20), None),
+            FmmSvdMode::new(true, None, None, Some(20), None),
+        ],
+        [
+            FmmSvdMode::new(true, None, None, Some(10), None),
+            FmmSvdMode::new(true, None, None, Some(20), None),
             FmmSvdMode::new(false, None, None, None, None),
-            FmmSvdMode::new(true, None, None, Some(20), None),
-            FmmSvdMode::new(true, None, None, Some(10), None),
-        ],
-        [
-            FmmSvdMode::new(true, None, None, Some(20), None),
-            FmmSvdMode::new(true, None, None, Some(20), None),
-            FmmSvdMode::new(true, None, None, Some(20), None),
         ],
     ];
 
     let surface_diff_vec = vec![
-        [None, Some(1), None],       // 6 digits
+        [Some(1), Some(1), Some(1)],       // 6 digits
         [Some(1), Some(1), Some(1)], // 8 digits
-        [Some(2), Some(2), Some(1)], // 10 digits
+        [Some(2), Some(2), Some(2)], // 10 digits
     ];
 
     let nvecs = vec![1, 5, 10];
