+++ kifmm/benches/laplace_f64_m1.rs
@@ -10,6 +10,11 @@ use kifmm::traits::tree::{SingleFmmTree, SingleTree};
 use kifmm::tree::helpers::{points_fixture, points_fixture_oblate_spheroid, points_fixture_sphere};
 use rlst::RawAccess;
 
+
+fn configure_criterion() -> Criterion {
+    Criterion::default().sample_size(10) // Set global sample size to 10
+}
+
 fn fft_f64(c: &mut Criterion) {
     let mut group = c.benchmark_group("F64 Potentials FFT-M2L");
 
@@ -237,5 +242,10 @@ fn blas_f64(c: &mut Criterion) {
     }
 }
 
-criterion_group!(laplace_p_f64, fft_f64, blas_f64);
+criterion_group! {
+    name = laplace_p_f64;
+    config = configure_criterion();
+    targets = fft_f64, blas_f64
+}
+
 criterion_main!(laplace_p_f64);
