+++ kifmm/benches/laplace_f32_m1.rs
@@ -11,6 +11,10 @@ use kifmm::traits::tree::{SingleFmmTree, SingleTree};
 use kifmm::tree::helpers::{points_fixture, points_fixture_oblate_spheroid, points_fixture_sphere};
 use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut};
 
+fn configure_criterion() -> Criterion {
+    Criterion::default().sample_size(10) // Set global sample size to 10
+}
+
 fn fft_f32(c: &mut Criterion) {
     let mut group = c.benchmark_group("F32 Potentials FFT-M2L");
 
@@ -226,5 +230,11 @@ fn blas_f32(c: &mut Criterion) {
     }
 }
 
-criterion_group!(laplace_p_f32, fft_f32, blas_f32);
+
+criterion_group! {
+    name = laplace_p_f32;
+    config = configure_criterion();
+    targets = fft_f32, blas_f32
+}
+
 criterion_main!(laplace_p_f32);
