### 2025-01-13 ad60bc4
Reduce number of f32 amd samples
