### 2025-01-13 d223f8d
Add flop counter
