+++ kifmm/src/tree/helpers.rs
@@ -45,36 +45,76 @@ pub fn points_fixture<T: Float + RlstScalar + rand::distributions::uniform::Samp
     points
 }
 
-/// Points fixture for testing, uniformly samples on surface of a sphere of diameter 1.
-///
-/// # Arguments
-/// * `n_points` - The number of points to sample.
-/// * `min` - The minimum coordinate value along each axis.
-/// * `max` - The maximum coordinate value along each axis.
 pub fn points_fixture_sphere<T: RlstScalar + rand::distributions::uniform::SampleUniform>(
     n_points: usize,
 ) -> PointsMat<T> {
-    // Generate a set of randomly distributed points
-    let mut range = StdRng::seed_from_u64(0);
-    let pi = T::from(3.134159).unwrap();
+    use rand::distributions::Uniform;
+    use rand::rngs::StdRng;
+    use rand::SeedableRng;
+
+    // Seeded random number generator for reproducibility
+    let mut rng = StdRng::seed_from_u64(0);
+    let pi = T::from(std::f64::consts::PI).unwrap();
     let two = T::from(2.0).unwrap();
-    let half = T::from(0.5).unwrap();
+    let one = T::one();
 
-    let between = rand::distributions::Uniform::from(T::zero()..T::one());
+    // Uniform distributions for phi and cos(theta)
+    let phi_dist = Uniform::from(T::zero()..(two * pi));
+    let u_dist = Uniform::from(T::zero()..one);
 
+    // Initialize points array
     let mut points = rlst_dynamic_array2!(T, [3, n_points]);
-    let mut phi = rlst_dynamic_array2!(T, [n_points, 1]);
-    let mut theta = rlst_dynamic_array2!(T, [n_points, 1]);
 
     for i in 0..n_points {
-        phi[[i, 0]] = between.sample(&mut range) * two * pi;
-        theta[[i, 0]] = ((between.sample(&mut range) - half) * two).acos();
+        // Generate random phi and theta
+        let phi = phi_dist.sample(&mut rng);
+        let u = u_dist.sample(&mut rng);
+        let theta = (one - two * u).acos();
+
+        // Compute Cartesian coordinates
+        let x = theta.sin() * phi.cos();
+        let y = theta.sin() * phi.sin();
+        let z = theta.cos();
+
+        // Assign to points array
+        points[[0, i]] = x;
+        points[[1, i]] = y;
+        points[[2, i]] = z;
     }
 
+    points
+}
+
+pub fn points_fixture_oblate_spheroid<
+    T: RlstScalar + rand::distributions::uniform::SampleUniform,
+>(
+    n_points: usize,
+    a: T, // Semi-axis length along x- and y-axes
+    c: T, // Semi-axis length along the z-axis
+) -> PointsMat<T> {
+    use rand::distributions::Uniform;
+    use rand::rngs::StdRng;
+    use rand::SeedableRng;
+
+    let mut rng = StdRng::seed_from_u64(0); // Use a fixed seed for reproducibility
+    let pi = T::from(std::f64::consts::PI).unwrap();
+    let two = T::from(2.0).unwrap();
+    let one = T::one();
+
+    let phi_dist = Uniform::from(T::zero()..(two * pi)); // Azimuthal angle
+    let cos_theta_dist = Uniform::from(-one..one); // Cosine of polar angle
+
+    let mut points = rlst_dynamic_array2!(T, [3, n_points]);
+
     for i in 0..n_points {
-        points[[0, i]] = half * theta[[i, 0]].sin() * phi[[i, 0]].cos() + half;
-        points[[1, i]] = half * theta[[i, 0]].sin() * phi[[i, 0]].sin() + half;
[…]

[…]
