### 2025-01-13 dcce613
Add laplace f32
