+++ scripts/src/bin/storage_costs_helmholtz.rs
@@ -2,6 +2,7 @@ use core::panic;
 
 use green_kernels::helmholtz_3d::Helmholtz3dKernel;
 use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+use itertools::{izip, Itertools};
 use kifmm::fmm::types::FmmSvdMode;
 use kifmm::fmm::types::{BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
 use kifmm::traits::fftw::Dft;
@@ -12,15 +13,32 @@ use kifmm::traits::tree::{SingleFmmTree, SingleTree};
 use kifmm::tree::helpers::{points_fixture, points_fixture_oblate_spheroid, points_fixture_sphere};
 use kifmm::{fmm, BlasFieldTranslationIa, KiFmm};
 use num::Float;
-use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut, RlstScalar, Shape};
+use rlst::{c32, c64, rlst_dynamic_array2, RawAccess, RawAccessMut, RlstScalar, Shape};
 
 use std::mem;
+
+
+fn p_vec(p_leaf: usize, p_scale: f64) -> Vec<usize> {
+
+    let mut res = Vec::new();
+
+    let mut curr = p_leaf;
+    for l in 0..5 {
+        res.push(curr as usize);
+        curr = ((curr as f64) * p_scale) as usize;
+    }
+
+    res = res.into_iter().rev().collect_vec();
+    res[0] = 0;
+    res[1] = 0;
+    res
+}
 
 fn calculate_fmm_storage_fft_m2l<Scalar>(
     fmm: &KiFmm<Scalar, Helmholtz3dKernel<Scalar>, FftFieldTranslation<Scalar>>,
 ) -> (f64, f64, f64)
 where
-    Scalar: RlstScalar<Complex = Scalar> + Float + AsComplex + Dft,
+    Scalar: RlstScalar<Complex = Scalar> + AsComplex + Dft,
 
 {
     let element_size = mem::size_of::<Scalar>(); // Assuming f32 is the element type
@@ -67,7 +85,7 @@ where
         size_m2l += vec.len();
     }
     size_m2l *= 2; // need it twice, freq re-ordered
-    storage_m2l = size_m2l * element_size * 2; // complex numbers
+    storage_m2l = size_m2l * element_size; // complex numbers
 
     // Convert storage to megabytes
     (
@@ -135,7 +153,7 @@ where
         size_m2l += size;
     }
 
-    storage_m2l = size_m2l * element_size * 2;
+    storage_m2l = size_m2l * element_size;
 
     // Convert storage to megabytes
     (
@@ -145,36 +163,51 @@ where
     )
 }
 
-fn fft_f32() {
-    let n_points = 1000000;
+fn fft_f32(arch: String) {
+    let n_points = 1000;
     let geometries = ["uniform", "sphere", "spheroid"];
 
-    // TODO: 4 digits for sphere and spheroid are dummies for now
 
-    // Tree depth
-    let depth_vec = vec![
-        [5, 5, 5], // 3 digits, for each geometry
-        [5, 5, 5], // 4 digits for each geometry
-    ];
+    let depth = 4;
 
-    // Expansion order
-    let e_vec = vec![
-        [4, 5, 5], // 3 digits, for each geometry
-        [5, 5, 5], // 4 digits for each geometry
-    ];
+    let p_leaf_vec;
+    let p_scale_vec;
+    if arch == "amd" {
+        p_leaf_vec = vec![
+            [4, 5, 6, 8, 11], // uniform
[…]
