+++ scripts/src/bin/storage_costs_laplace.rs
@@ -142,30 +142,31 @@ fn calculate_fmm_storage_blas_m2l<Scalar: RlstScalar + Float + AsComplex + Dft>(
 
 fn fft_f32() {
     let n_points = 1000000;
-    let geometries = ["uniform"];
-
-    // TODO: 4 digits for sphere and spheroid are dummies for now
+    let geometries = ["uniform", "sphere", "spheroid"];
 
     // Tree depth
     let depth_vec = vec![
-        [5], // 3 digits, for each geometry
+        [5, 5, 5], // 3 digits, for each geometry
+        [5, 5, 5], // 4 digits for each geometry
     ];
 
     // Expansion order
     let e_vec = vec![
-        [4], // 3 digits, for each geometry
+        [3, 3, 4], // 3 digits, for each geometry
+        [4, 5, 5], // 4 digits for each geometry
     ];
 
     // Block size
     let b_vec = vec![
-        [32], // 3 digits, for each geometry
+        [64, 64, 128], // 3 digits, for each geometry
+        [32, 128, 128],   // 4 digits for each geometry
     ];
 
-    let experiments = [3]; // number of digits sought
+    let experiments = [3, 4]; // number of digits sought
 
     let prune_empty = true;
 
-    for (i, _digits) in experiments.iter().enumerate() {
+    for (i, digits) in experiments.iter().enumerate() {
         for (j, &geometry) in geometries.iter().enumerate() {
             let sources;
             let targets;
@@ -190,7 +191,7 @@ fn fft_f32() {
             let expansion_order = vec![e; depth.unwrap() as usize + 1];
             let block_size = Some(b);
 
-            let fmm_fft = SingleNodeBuilder::new(false)
+            let fmm = SingleNodeBuilder::new(false)
                 .tree(sources.data(), targets.data(), None, depth, prune_empty)
                 .unwrap()
                 .parameters(
@@ -204,9 +205,10 @@ fn fft_f32() {
                 .build()
                 .unwrap();
 
-            let (m2m, l2l, m2l) = calculate_fmm_storage_fft_m2l(&fmm_fft);
+            let (m2m, l2l, m2l) = calculate_fmm_storage_fft_m2l(&fmm);
 
-            println!("STORAGE COSTS {:?} {:?} {:?}", m2m, l2l, m2l);
+            println!("FFT M2L geometry {:?}, digits {:?}", geometry, digits);
+            println!("M2M {:?} mb  L2L {:?} mb M2L {:?} mb", m2m, l2l, m2l);
         }
     }
 }
@@ -220,40 +222,40 @@ fn blas_f32() {
     // Tree depth
     let depth_vec = vec![
         [5, 5, 5], // 3 digits, for each geometry
-        [5, 4, 5], // 4 digits for each geometry
+        [5, 5, 5], // 4 digits for each geometry
     ];
 
     // Expansion order
     let e_vec = vec![
         [3, 3, 3], // 3 digits, for each geometry
-        [5, 5, 5], // 4 digits for each geometry
+        [3, 3, 5], // 4 digits for each geometry
     ];
 
     // SVD threshold
     let svd_threshold_vec = vec![
-        [Some(0.001), Some(0.001), None], // 3 digits
-        [Some(1e-5), Some(0.001), None],  // 4 digits
+        [Some(0.1), Some(0.00001), Some(0.00001)], // 3 digits
+        [Some(0.001), Some(0.001), None],  // 4 digits
     ];
 
     let svd_mode_vec = vec![
         [
-            FmmSvdMode::new(false, None, None, Some(10), None),
+            FmmSvdMode::new(true, None, None, Some(5), None),
             FmmSvdMode::new(false, None, None, None, None),
[…]
