+++ kifmm/Cargo.toml
@@ -52,6 +52,14 @@ name = "laplace_f32_amd"
 harness = false
 
 [[bench]]
+name = "laplace_f32_m1"
+harness = false
+
+[[bench]]
+name = "laplace_f64_m1"
+harness = false
+
+[[bench]]
 name = "laplace_f64_amd"
 harness = false
 

[…]
