+++ scripts/src/bin/count_flops_f32_m1.rs
@@ -1,0 +1,187 @@
+use core::panic;
+use std::time::Duration;
+
+use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+use kifmm::fmm::types::FmmSvdMode;
+use kifmm::fmm::types::{BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
+use kifmm::traits::field::{SourceToTargetTranslation, TargetTranslation};
+use kifmm::traits::fmm::{DataAccess, Evaluate};
+use kifmm::traits::tree::{SingleFmmTree, SingleTree};
+use kifmm::tree::helpers::{points_fixture, points_fixture_oblate_spheroid, points_fixture_sphere};
+use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut};
+
+fn fft_f32() {
+    let n_points = 1000000;
+    let geometries = ["uniform", "sphere", "spheroid"];
+
+    // Tree depth
+    let depth_vec = vec![
+        [5, 5, 5], // 3 digits, for each geometry
+        [5, 5, 5], // 4 digits for each geometry
+    ];
+
+    // Expansion order
+    let e_vec = vec![
+        [3, 3, 4], // 3 digits, for each geometry
+        [4, 5, 5], // 4 digits for each geometry
+    ];
+
+    // Block size
+    let b_vec = vec![
+        [64, 64, 128], // 3 digits, for each geometry
+        [32, 128, 128],   // 4 digits for each geometry
+    ];
+
+    let experiments = [3, 4]; // number of digits sought
+
+    let prune_empty = true;
+
+    for (i, digits) in experiments.iter().enumerate() {
+        for (j, &geometry) in geometries.iter().enumerate() {
+            let sources;
+            let targets;
+            if geometry == "uniform" {
+                sources = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
+                targets = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
+            } else if geometry == "sphere" {
+                sources = points_fixture_sphere(n_points);
+                targets = points_fixture_sphere(n_points);
+            } else if geometry == "spheroid" {
+                sources = points_fixture_oblate_spheroid(n_points, 1.0, 0.5);
+                targets = points_fixture_oblate_spheroid(n_points, 1.0, 0.5);
+            } else {
+                panic!("invalid geometry");
+            }
+
+            let charges = vec![1f32; n_points];
+
+            let depth = Some(depth_vec[i][j]);
+            let e = e_vec[i][j];
+            let b = b_vec[i][j];
+            let expansion_order = vec![e; depth.unwrap() as usize + 1];
+            let block_size = Some(b);
+
+            let mut fmm = SingleNodeBuilder::new(false)
+                .tree(sources.data(), targets.data(), None, depth, prune_empty)
+                .unwrap()
+                .parameters(
+                    &charges,
+                    &expansion_order,
+                    Laplace3dKernel::new(),
+                    GreenKernelEvalType::Value,
+                    FftFieldTranslation::new(block_size),
+                )
+                .unwrap()
+                .build()
+                .unwrap();
+
+            fmm.evaluate();
+
+            println!("FFT-M2L geometry {:?}, digits {:?}, flops {:?}", geometry, digits, fmm.nflops)
+
+        }
+    }
+}
+
+fn blas_f32() {
+
+    let n_points = 1000000;
+    let geometries = ["uniform", "sphere", "spheroid"];
+
+    // Tree depth
+    let depth_vec = vec![
+        [4, 4, 4], // 3 digits, for each geometry
[…]

[…]
