### 2025-01-17 d92cfb9
Check f64 bench
