### 2025-01-17 0b6a2ef
Add calculate compression script
