+++ scripts/src/bin/storage_costs_laplace.rs
@@ -221,39 +221,40 @@ fn blas_f32() {
 
     // Tree depth
     let depth_vec = vec![
-        [5, 5, 5], // 3 digits, for each geometry
-        [5, 5, 5], // 4 digits for each geometry
+        [4, 5, 5], // 3 digits, for each geometry
+        [4, 5, 5], // 4 digits for each geometry
     ];
 
     // Expansion order
     let e_vec = vec![
-        [3, 3, 3], // 3 digits, for each geometry
-        [3, 3, 5], // 4 digits for each geometry
+        [3, 2, 3], // 3 digits, for each geometry
+        [3, 3, 4], // 4 digits for each geometry
     ];
 
     // SVD threshold
     let svd_threshold_vec = vec![
-        [Some(0.1), Some(0.00001), Some(0.00001)], // 3 digits
-        [Some(0.001), Some(0.001), None],  // 4 digits
+        [Some(0.00001), Some(0.00001), Some(0.001)],              // 3 digits
+        [None, None, None], // 4 digits
     ];
 
     let svd_mode_vec = vec![
         [
-            FmmSvdMode::new(true, None, None, Some(5), None),
-            FmmSvdMode::new(false, None, None, None, None),
-            FmmSvdMode::new(true, None, None, Some(20), None),
-        ],
-        [
+            FmmSvdMode::new(true, None, None, Some(10), None),
             FmmSvdMode::new(true, None, None, Some(10), None),
             FmmSvdMode::new(false, None, None, None, None),
+        ],
+        [
             FmmSvdMode::new(false, None, None, None, None),
+            FmmSvdMode::new(true, None, None, Some(20), None),
+            FmmSvdMode::new(true, None, None, Some(5), None),
         ],
     ];
 
     let surface_diff_vec = vec![
-        [Some(1), None, None], // 3 digits
-        [Some(1), Some(1), None],    // 4 digits
+        [None, Some(2), Some(2)], // 3 digits
+        [Some(1), Some(2), Some(1)],       // 4 digits
     ];
+
 
     let nvecs = vec![1];
 
@@ -261,53 +262,56 @@ fn blas_f32() {
 
     let prune_empty = true;
 
-    for (i, digits) in experiments.iter().enumerate() {
+    for (i, &digits) in experiments.iter().enumerate() {
         for (j, &geometry) in geometries.iter().enumerate() {
-            for (k, nvec) in nvecs.iter().enumerate() {
-                let sources;
-                let targets;
-                if geometry == "uniform" {
-                    sources = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
-                    targets = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
-                } else if geometry == "sphere" {
-                    sources = points_fixture_sphere(n_points);
-                    targets = points_fixture_sphere(n_points);
-                } else if geometry == "spheroid" {
-                    sources = points_fixture_oblate_spheroid(n_points, 1.0, 0.5);
-                    targets = points_fixture_oblate_spheroid(n_points, 1.0, 0.5);
-                } else {
-                    panic!("invalid geometry");
+
+            if geometry == "spheroid" && digits == 3 {
+                for (k, nvec) in nvecs.iter().enumerate() {
+                    let sources;
+                    let targets;
+                    if geometry == "uniform" {
+                        sources = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
+                        targets = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
+                    } else if geometry == "sphere" {
+                        sources = points_fixture_sphere(n_points);
+                        targets = points_fixture_sphere(n_points);
+                    } else if geometry == "spheroid" {
[…]
