+++ scripts/src/bin/compression_f32.rs
@@ -1,0 +1,316 @@
+use core::panic;
+use std::time::Instant;
+
+use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+use kifmm::fmm::types::FmmSvdMode;
+use kifmm::fmm::types::{BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
+use kifmm::traits::fftw::Dft;
+use kifmm::traits::field::{SourceToTargetTranslation, TargetTranslation};
+use kifmm::traits::fmm::{DataAccess, Evaluate};
+use kifmm::traits::general::single_node::AsComplex;
+use kifmm::traits::tree::{SingleFmmTree, SingleTree};
+use kifmm::tree::helpers::{points_fixture, points_fixture_oblate_spheroid, points_fixture_sphere};
+use kifmm::{fmm, KiFmm};
+use num::Float;
+use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut, RlstScalar, Shape};
+
+use std::mem;
+
+
+fn calculate_compression<Scalar: RlstScalar + Float + AsComplex + Dft>(
+    fmm: &KiFmm<Scalar, Laplace3dKernel<Scalar>, BlasFieldTranslationSaRcmp<Scalar>>,
+    fmm_full: &KiFmm<Scalar, Laplace3dKernel<Scalar>, BlasFieldTranslationSaRcmp<Scalar>>,
+) -> f64 {
+    let element_size = mem::size_of::<Scalar>(); // Assuming f32 is the element type
+
+    // M2L
+    let data = &fmm.source_to_target.metadata;
+    let mut size_m2l_compressed = 0;
+
+    for metadata in data {
+        let u = &metadata.u;
+        let st = &metadata.st;
+        let c_u = &metadata.c_u;
+        let c_vt = &metadata.c_vt;
+
+        let mut size: usize = u.shape()[0] * u.shape()[1] + st.shape()[0] + st.shape()[1];
+        for (l, r) in c_u.iter().zip(c_vt.iter()) {
+            size += l.shape()[0] + l.shape()[1] + r.shape()[0] * r.shape()[1];
+        }
+        size_m2l_compressed += size;
+    }
+
+    let data = &fmm_full.source_to_target.metadata;
+    let mut size_m2l = 0;
+
+    for metadata in data {
+        let u = &metadata.u;
+        let st = &metadata.st;
+        let c_u = &metadata.c_u;
+        let c_vt = &metadata.c_vt;
+
+        let mut size: usize = u.shape()[0] * u.shape()[1] + st.shape()[0] + st.shape()[1];
+        for (l, r) in c_u.iter().zip(c_vt.iter()) {
+            size += l.shape()[0] + l.shape()[1] + r.shape()[0] * r.shape()[1];
+        }
+        size_m2l += size;
+    }
+
+    println!("HERE {:?} {:?}", size_m2l, size_m2l_compressed);
+
+    let compression = 100. - 100. * ((size_m2l_compressed as f64) / (size_m2l as f64));
+
+    compression
+}
+
+
+fn blas_f32() {
+    let n_points = 1000000;
+    let geometries = ["uniform", "sphere", "spheroid"];
+
+    // TODO: 4 digits for sphere and spheroid are dummies for now
+
+    // Tree depth
+    let depth_vec = vec![
+        [4, 5, 5], // 3 digits, for each geometry
+        [4, 5, 5], // 4 digits for each geometry
+    ];
+
+    // Expansion order
+    let e_vec = vec![
+        [3, 2, 3], // 3 digits, for each geometry
+        [3, 3, 4], // 4 digits for each geometry
+    ];
+
+    // SVD threshold
+    let svd_threshold_vec = vec![
[…]
