### 2025-01-17 bfa2687
Add temp setup script for fft
