+++ scripts/src/bin/compression_f32.rs
@@ -64,6 +64,252 @@ fn calculate_compression<Scalar: RlstScalar + Float + AsComplex + Dft>(
 }
 
 
+// fn blas_f32() {
+//     let n_points = 1000000;
+//     let geometries = ["uniform", "sphere", "spheroid"];
+
+//     // TODO: 4 digits for sphere and spheroid are dummies for now
+
+//     // Tree depth
+//     let depth_vec = vec![
+//         [4, 5, 5], // 3 digits, for each geometry
+//         [4, 5, 5], // 4 digits for each geometry
+//     ];
+
+//     // Expansion order
+//     let e_vec = vec![
+//         [3, 2, 3], // 3 digits, for each geometry
+//         [3, 3, 4], // 4 digits for each geometry
+//     ];
+
+//     // SVD threshold
+//     let svd_threshold_vec = vec![
+//         [Some(0.00001), Some(0.00001), Some(0.001)],              // 3 digits
+//         [None, None, None], // 4 digits
+//     ];
+
+//     let svd_mode_vec = vec![
+//         [
+//             FmmSvdMode::new(true, None, None, Some(10), None),
+//             FmmSvdMode::new(true, None, None, Some(10), None),
+//             FmmSvdMode::new(false, None, None, None, None),
+//         ],
+//         [
+//             FmmSvdMode::new(false, None, None, None, None),
+//             FmmSvdMode::new(true, None, None, Some(20), None),
+//             FmmSvdMode::new(true, None, None, Some(5), None),
+//         ],
+//     ];
+
+//     let surface_diff_vec = vec![
+//         [None, Some(2), Some(2)], // 3 digits
+//         [Some(1), Some(2), Some(1)],       // 4 digits
+//     ];
+
+//     let nvecs = vec![1];
+
+//     let experiments = [3, 4]; // number of digits sought
+
+//     let prune_empty = true;
+
+//     for (i, &digits) in experiments.iter().enumerate() {
+//         for (j, &geometry) in geometries.iter().enumerate() {
+
+//             // if geometry == "spheroid" && digits == 3 {
+//                 for (k, nvec) in nvecs.iter().enumerate() {
+//                     let sources;
+//                     let targets;
+//                     if geometry == "uniform" {
+//                         sources = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
+//                         targets = points_fixture::<f32>(n_points, Some(0.), Some(1.), None);
+//                     } else if geometry == "sphere" {
+//                         sources = points_fixture_sphere(n_points);
+//                         targets = points_fixture_sphere(n_points);
+//                     } else if geometry == "spheroid" {
+//                         sources = points_fixture_oblate_spheroid(n_points, 1.0, 0.5);
+//                         targets = points_fixture_oblate_spheroid(n_points, 1.0, 0.5);
+//                     } else {
+//                         panic!("invalid geometry");
+//                     }
+
+//                     let charges = vec![1f32; n_points * nvec];
+
+//                     let depth = Some(depth_vec[i][j]);
+//                     let e = e_vec[i][j];
+//                     let threshold = svd_threshold_vec[i][j];
+//                     let surface_diff = surface_diff_vec[i][j];
+//                     let svd_mode = svd_mode_vec[i][j];
+
+//                     let expansion_order = vec![e; depth.unwrap() as usize + 1];
+
+//                     let s = Instant::now();
+//                     let fmm = SingleNodeBuilder::new(false)
+//                         .tree(sources.data(), targets.data(), None, depth, prune_empty)
+//                         .unwrap()
+//                         .parameters(
+//                             &charges,
+//                             &expansion_order,
[…]
