### 2025-01-20 e02a3e2
Add test for attaching charges
