+++ kifmm/examples/mpi_test_fmm.rs
@@ -46,7 +46,6 @@ fn main() {
 
         // Generate some random test data local to each process
         let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
-        // let charges = vec![0.1f32; n_points];
         let mut rng = StdRng::seed_from_u64(comm.rank() as u64);
         let mut charges = rlst_dynamic_array1!(f32, [n_points]);
         charges.data_mut().iter_mut().for_each(|c| *c = rng.gen());
@@ -62,7 +61,12 @@ fn main() {
                 sort_kind.clone(),
             )
             .unwrap()
-            .parameters(charges.data(), expansion_order, kernel.clone(), source_to_target)
+            .parameters(
+                charges.data(),
+                expansion_order,
+                kernel.clone(),
+                source_to_target,
+            )
             .unwrap()
             .build()
             .unwrap();
@@ -97,7 +101,7 @@ fn main() {
             }
 
             let local_coords = multi_fmm.tree().source_tree().all_coordinates().unwrap();
-            let local_charges= multi_fmm.charges().unwrap();
+            let local_charges = multi_fmm.charges().unwrap();
 
             let mut partition = PartitionMut::new(
                 &mut all_coordinates,
@@ -107,11 +111,8 @@ fn main() {
 
             root_process.gather_varcount_into_root(local_coords, &mut partition);
 
-            let mut partition = PartitionMut::new(
-                &mut all_charges,
-                charges_counts,
-                charges_displacements,
-            );
+            let mut partition =
+                PartitionMut::new(&mut all_charges, charges_counts, charges_displacements);
 
             root_process.gather_varcount_into_root(local_charges, &mut partition);
         } else {
@@ -121,17 +122,11 @@ fn main() {
             let local_coords = multi_fmm.tree().source_tree().all_coordinates().unwrap();
             root_process.gather_varcount_into(local_coords);
 
-
-            let local_charges= multi_fmm.charges().unwrap();
+            let local_charges = multi_fmm.charges().unwrap();
             root_process.gather_varcount_into(local_charges);
         }
 
         if world.rank() == 0 {
-
-            // let mut rng = StdRng::seed_from_u64(comm.rank() as u64);
-            // let mut charges = rlst_dynamic_array1!(f32, [n_points]);
-            // charges.data_mut().iter_mut().for_each(|c| *c = rng.gen());
-
             let mut single_fmm = SingleNodeBuilder::new(false)
                 .tree(
                     &all_coordinates,
@@ -164,16 +159,12 @@ fn main() {
 
             let distributed = multi_fmm.potentials().unwrap();
 
-            // let leaf = &single_fmm.tree.target_tree.leaves[0];
-            // let found_single = single_fmm.potential(&leaf);
-            // let found = fmm.potential(&leaf);
-
-            println!(
-                "{:?} expected: {:?} \n found: {:?}",
-                world.rank(),
-                &distributed[0..5],
-                &expected[0..5]
-            );
+            // println!(
+            //     "{:?} expected: {:?} \n found: {:?}",
+            //     world.rank(),
+            //     &distributed[0..5],
+            //     &expected[0..5]
+            // );
 
             distributed
                 .iter()
@@ -181,122 +172,132 @@ fn main() {
                 .for_each(|(f, e)| assert!(((f - e).abs() / e.abs()) < 1e-3));
 
             println!("...test_fmm M2L=FFT passed");
-
         }
     }
 
-    // // Test BLAS field translation
-    // {
-    //     // let source_to_target = FftFieldTranslation::<f32>::new(None);
-    //     let source_to_target =
-    //         BlasFieldTranslationSaRcmp::<f32>::new(None, None, kifmm::FmmSvdMode::Deterministic);
+    // Test BLAS field translation
+    {
+        let source_to_target =
[…]
