+++ kifmm/src/fmm/charge_handler/multi_node.rs
@@ -1,5 +1,9 @@
 use green_kernels::traits::Kernel as KernelTrait;
-use mpi::traits::Equivalence;
+use itertools::izip;
+use mpi::{
+    datatype::{Partition, PartitionMut},
+    traits::Equivalence,
+};
 use num::Float;
 use rlst::RlstScalar;
 
@@ -8,6 +12,7 @@ use crate::{
     traits::{
         field::FieldTranslation as FieldTranslationTrait,
         fmm::{ChargeHandler, HomogenousKernel},
+        general::multi_node::GhostExchange,
         tree::{MultiFmmTree, MultiTree},
         types::FmmError,
     },
@@ -21,6 +26,7 @@ where
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
     FieldTranslation: FieldTranslationTrait + Send + Sync,
+    Self: GhostExchange,
 {
     type Scalar = Scalar;
 
@@ -41,6 +47,54 @@ where
             *c = Scalar::zero();
         }
 
+        for m in self.ghost_fmm_u.multipoles.iter_mut() {
+            *m = Scalar::zero();
+        }
+
+        for l in self.ghost_fmm_u.locals.iter_mut() {
+            *l = Scalar::zero();
+        }
+
+        for p in self.ghost_fmm_u.potentials.iter_mut() {
+            *p = Scalar::zero();
+        }
+
+        for c in self.ghost_fmm_u.charges.iter_mut() {
+            *c = Scalar::zero();
+        }
+
+        for m in self.ghost_fmm_v.multipoles.iter_mut() {
+            *m = Scalar::zero();
+        }
+
+        for l in self.ghost_fmm_v.locals.iter_mut() {
+            *l = Scalar::zero();
+        }
+
+        for p in self.ghost_fmm_v.potentials.iter_mut() {
+            *p = Scalar::zero();
+        }
+
+        for c in self.ghost_fmm_v.charges.iter_mut() {
+            *c = Scalar::zero();
+        }
+
+        for m in self.global_fmm.multipoles.iter_mut() {
+            *m = Scalar::zero();
+        }
+
+        for l in self.global_fmm.locals.iter_mut() {
+            *l = Scalar::zero();
+        }
+
+        for p in self.global_fmm.potentials.iter_mut() {
+            *p = Scalar::zero();
+        }
+
+        for c in self.global_fmm.charges.iter_mut() {
+            *c = Scalar::zero();
+        }
+
         Ok(())
     }
 
@@ -57,6 +111,7 @@ where
         let n_source_points_input = charges.len() / n_matvecs;
 
         if n_source_points == n_source_points_input {
+            self.clear().unwrap();
             self.charges = charges.to_vec();
             Ok(())
         } else {
@@ -72,10 +127,68 @@ where
 
     fn attach_charges_unordered(
         &mut self,
-        _charges: &[<Self as ChargeHandler>::Scalar],
+        charges: &[<Self as ChargeHandler>::Scalar],
     ) -> Result<(), FmmError> {
-        Err(FmmError::Unimplemented(
-            "Attaching unordered charges currently unimplemented.".to_string(),
-        ))
+        if (charges.len() as u64) == (self.local_count_charges) {
+            self.clear().unwrap();
+
+            let mut available_queries = Vec::new();
+
+            for (&count, &displacement) in izip!(
+                &self.ghost_received_queries_charge_counts,
+                &self.ghost_received_queries_charge_displacements
+            ) {
+                let l = displacement as usize;
+                let r = l + (count as usize);
+
+                // Received queries from this rank
+                let received_queries_rank = &self.ghost_received_queries_charge[l..r];
+
[…]
