### 2025-01-20 7f64556
Add random charges (#129)
* Add support for attaching random charge data and general charge handling to multi-node FMM
* Closes #115 and #114
