+++ kifmm/examples/mpi_test_fmm.rs
@@ -11,13 +11,14 @@ fn main() {
         BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, SingleNodeBuilder,
     };
 
+    use rand::{rngs::StdRng, Rng, SeedableRng};
     use rayon::ThreadPoolBuilder;
 
     use mpi::{
         datatype::PartitionMut,
         traits::{Communicator, Root},
     };
-    use rlst::RawAccess;
+    use rlst::{rlst_dynamic_array1, RawAccess, RawAccessMut};
 
     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
     let world = universe.world();
@@ -45,6 +46,9 @@ fn main() {
 
         // Generate some random test data local to each process
         let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
+        let mut rng = StdRng::seed_from_u64(comm.rank() as u64);
+        let mut charges = rlst_dynamic_array1!(f32, [n_points]);
+        charges.data_mut().iter_mut().for_each(|c| *c = rng.gen());
 
         let mut multi_fmm = MultiNodeBuilder::new(false)
             .tree(
@@ -57,21 +61,30 @@ fn main() {
                 sort_kind.clone(),
             )
             .unwrap()
-            .parameters(expansion_order, kernel.clone(), source_to_target)
+            .parameters(
+                charges.data(),
+                expansion_order,
+                kernel.clone(),
+                source_to_target,
+            )
             .unwrap()
             .build()
             .unwrap();
 
         multi_fmm.evaluate().unwrap();
 
-        // Gather all coordinates for the test
+        // Gather all coordinates and charges for the test
         let root_process = comm.process_at_rank(0);
         let n_coords = multi_fmm.tree().source_tree().coordinates.len() as i32;
+        let n_charges = multi_fmm.tree().source_tree().global_indices.len() as i32;
+        let mut all_charges = vec![0f32; n_points * world.size() as usize];
         let mut all_coordinates = vec![0f32; 3 * n_points * world.size() as usize];
 
         if world.rank() == 0 {
+            let mut charges_counts = vec![0i32; comm.size() as usize];
             let mut coordinates_counts = vec![0i32; comm.size() as usize];
             root_process.gather_into_root(&n_coords, &mut coordinates_counts);
+            root_process.gather_into_root(&n_charges, &mut charges_counts);
 
             let mut coordinates_displacements = Vec::new();
             let mut counter = 0;
@@ -80,7 +93,15 @@ fn main() {
                 counter += count;
             }
 
+            let mut charges_displacements = Vec::new();
+            let mut counter = 0;
+            for &count in charges_counts.iter() {
+                charges_displacements.push(counter);
+                counter += count;
+            }
+
             let local_coords = multi_fmm.tree().source_tree().all_coordinates().unwrap();
+            let local_charges = multi_fmm.charges().unwrap();
 
             let mut partition = PartitionMut::new(
                 &mut all_coordinates,
@@ -89,11 +110,20 @@ fn main() {
             );
 
             root_process.gather_varcount_into_root(local_coords, &mut partition);
+
+            let mut partition =
+                PartitionMut::new(&mut all_charges, charges_counts, charges_displacements);
+
+            root_process.gather_varcount_into_root(local_charges, &mut partition);
         } else {
             root_process.gather_into(&n_coords);
+            root_process.gather_into(&n_charges);
 
             let local_coords = multi_fmm.tree().source_tree().all_coordinates().unwrap();
             root_process.gather_varcount_into(local_coords);
+
+            let local_charges = multi_fmm.charges().unwrap();
+            root_process.gather_varcount_into(local_charges);
         }
 
[…]
