+++ paper/paper.md
@@ -11,80 +11,67 @@ authors:
 affiliations:
  - name: Department of Mathematics, University College London, UK
    index: 1
-date: 5 April 2024
+date: 22 January 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-We present `kifmm-rs` a Rust based implementation of the kernel independent Fast Multipole Method (kiFMM) with C bindings, that serves as a implementation framework for kiFMMs [@Ying2004; @Greengard1987]. The FMM is a key algorithm for scientific computing due to its acceleration of the computation of $N$-body potential evaluation problems of the form,
+`kifmm-rs` is an open-source framework for implementing Fast Multipole Methods (FMMs) in shared and distributed memory. FMMs accelerate the calculation of $N$-body potential evaluation problems that arise in computational science and engineering of the form,
 
 \begin{equation}
     \phi(x_i) = \sum_{j=1}^N K(x_i, y_j) q(y_j)
     \label{eq:sec:summary:potential}
 \end{equation}
 
-from $O(N^2)$ to $O(N)$ or $O(N \log(N))$, where the potential $\phi$ is evaluated at a set of target points, $\{x_i\}_{i=1}^M$, due to a set of densities, $\{ q_j \}_{j=1}^N$ and $K(.,.)$ is the interaction kernel. Compatible kernels commonly arise in science and engineering, such as the Laplace kernel which models the electrostatic or gravitational potentials corresponding to a cluster of source points on a cluster of target points,
+where the potential $\phi$ is evaluated at a set of target particles, $\{x_i\}_{i=1}^M$, due to a set of densities, $\{ q_j \}_{j=1}^N$ collocated at a set of source particles $\{y_j\}_{j=1}^N$. This evaluation can be interpreted as a matrix-vector multiplication,
+
+\begin{equation}
+    \symbfup{\phi} = \mathbf{K} \mathbf{q}
+    \label{eq:sec:summary:matvec}
+\end{equation}
+
+where $\symbfup{\phi}$ is an $M$ dimensional vector, $\mathbf{K}$ is a dense $M \times N$ matrix and $\mathbf{q}$ is an $N$ dimensional vector. As $\mathbf{K}$ is dense, the naive evaluation of \eqref{eq:sec:summary:matvec} is of $O(M \cdot N)$ complexity. FMMs provide a way of evaluating \eqref{eq:sec:summary:matvec} in $O(P(M+N))$ where $P$ is a user defined parameter $P \ll M, N$ called the _expansion order_, for a range of interaction kernels $K(\cdot)$ that commonly arise from elliptic partial differential equations (PDEs). The prototypical example for which the FMM was first presented is the Laplace kernel [@Greengard1987], which describes particles interacting electrostatically or gravitationally,
 
 \begin{equation}
     K(x, y) = \begin{cases}
-	\frac{1}{2 \pi} \log(\frac{1}{\|x-y\|}),  \> \> (2D) \\
-	\frac{1}{4 \pi \|x-y\|}, \> \> (3D)
+	\frac{1}{2 \pi} \log(\frac{1}{\|x-y\|}),  \> \> \text{, 2D} \\
+	\frac{1}{4 \pi \|x-y\|}, \> \> \text{, 3D}
     \end{cases}
-    \label{eq:sec:summary:laplace_kernel}
+å    \label{eq:sec:summary:laplace_kernel}
 \end{equation}
 
[…]
