### 2025-01-23 61ff929
Small paper edits
