+++ paper/paper.md
@@ -53,15 +53,15 @@ Since their initial introduction FMMs have been developed for a wide variety of 
 
 As well as the Stokes kernel describing fluid flow [@tornberg2008fast], and the Navier kernel used in elastostatics [@fu1998fast] among others. The major application of FMMs is found in the accelerated application of dense operator matrices that arise from the integral formulation of partial differential equations, and as a result FMM algorithms are a crucial component across simulations in many domains, from medical imaging to geophysics.
 
-FMMs accelerate the evaluation of \eqref{eq:sec:summary:matvec} by decomposing the problem domain using a hierarchical tree, a quadtree in 2D and an octree in 3D, the algorithm consisting of a series of operations collectively referred to as _field translations_. Evaluations of the off-diagonal blocks of the matrix $\mathbf{K}$ in \eqref{eq:sec:summary:matvec} correspond to the `multipole to local' or M2L field translation which summarise the long-range interactions of a local cluster of target points, and the evaluation of the diagonal blocks correspond to unavoidable direct evaluations of the kernel function for evaluating near-range interactions.
+FMMs accelerate the evaluation of \eqref{eq:sec:summary:matvec} by decomposing the problem domain using a hierarchical tree, a quadtree in 2D and an octree in 3D, the algorithm consisting of a series of operations collectively referred to as _field translations_. Evaluations of the off-diagonal blocks of the matrix $\mathbf{K}$ in \eqref{eq:sec:summary:matvec} correspond to the `multipole to local' or M2L field translation which summarise the long-range interactions of a local cluster of target points, and the evaluation of the diagonal blocks correspond to unavoidable direct evaluations of the kernel function for evaluating near-range interactions. M2L is the most challenging optimisation bottleneck in FMMs due to its memory-bound nature.
 
 Many parts of an implementation are common across FMM algorithms, such as the tree setup and the kernel function implementation. As a result, our framework is presented as a set of modular, re-usable, sub-components of each of the key algorithmic sub-components: (i) the tree in shared and distributed memory, (ii) the field translation operations and (iii) the kernel evaluation. We use our framework to develop an implementation of the so-called kernel independent Fast Multiple Method (kiFMM) [@Ying2004], compatible with a wide variety of elliptic PDE kernels with an implementation for the Laplace and Helmholtz kernels provided currently. The kiFMM has favourable implementation properties due to its formulation in terms of high operational-intensity BLAS operations.
 
 # Statement of need
 
-Other notable software efforts for the FMM include PVFMM [@Malhotra2015], ExaFMM [@wang2021exafmm], ScalFMM [@blanchard2015scalfmm] and TBFMM [@bramas2020tbfmm]. PVFMM, ScalFMM and TBFMM are fully distributed implementations of the black box FMM [@fong2009black] and kiFMM respectively. The former two are notable for being distributed using a task-based runtime, with the latter using a more traditional MPI based approach. ExaFMM offers a shared memory implementation, with Python interfaces and a simple template based design. A commonality of previous implementations is the coupling of algorithmic optimisation with implementation optimisation. For example, ExaFMM and PVFMM both offer field translations that are highly optimised for x64 architectures and lack ARM support, with ScalFMM and TBFMM being tailored to the runtime systems they are designed for.
[…]
