### 2025-01-23 dff8794
EDIT: JOSS review comments. (#132)
* Update paper
