### 2025-01-23 8a6445b
EDIT: Update JOSS Paper (#133)
* Update paper

* Small paper edits

* A word

* Fix merge conflicts
