### 2025-01-24 105ce9d
Make attach charges test better
