### 2025-01-24 fc88a81
Add test for ordered charge handling
