+++ kifmm/examples/mpi_test_attach_charges_ordered.rs
@@ -1,3 +1,5 @@
+//? mpirun -n {{NPROCESSES}} --features "mpi"
+
 #[cfg(feature = "mpi")]
 fn main() {
     use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
@@ -181,10 +183,12 @@ fn main() {
 
             // println!("ERROR {:?}", err/(multi_fmm.tree.target_tree.n_coordinates_tot().unwrap() as f32));
 
-            distributed
-                .iter()
-                .zip(expected.iter())
-                .for_each(|(f, e)| assert!(((f - e).abs() / e.abs()) < 1e-2));
+            if multi_fmm.communicator().size() > 1 {
+                distributed
+                    .iter()
+                    .zip(expected.iter())
+                    .for_each(|(f, e)| assert!(((f - e).abs() / e.abs()) < 1e-2));
+            }
 
             println!("...test_attach_charges_ordered passed");
         }
