+++ kifmm/examples/mpi_test_p2m.rs
@@ -1,5 +1,3 @@
-//? mpirun -n {{NPROCESSES}} --features "mpi"
-
 #[cfg(feature = "mpi")]
 fn main() {
     use std::collections::HashMap;
