### 2025-01-24 a449f85
FIX/TST: Make attach charges test more robust (#135)
* Add test for ordered charge handling
* Bug in ordered charge handling
