### 2025-01-29 bf7e5fc
Patched multi node constructor
