+++ scripts/src/bin/fmm_m2l_blas_mpi_f32.rs
@@ -98,7 +98,13 @@ fn main() {
             sort_kind.clone(),
         )
         .unwrap()
-        .parameters(&charges, expansion_order, kernel.clone(), source_to_target)
+        .parameters(
+            &charges,
+            expansion_order,
+            kernel.clone(),
+            green_kernels::types::GreenKernelEvalType::Value,
+            source_to_target,
+        )
         .unwrap()
         .build()
         .unwrap();
