+++ kifmm/examples/mpi_test_trees.rs
@@ -61,7 +61,13 @@ fn main() {
             sort_kind,
         )
         .unwrap()
-        .parameters(&charges, expansion_order, kernel, source_to_target)
+        .parameters(
+            &charges,
+            expansion_order,
+            kernel,
+            GreenKernelEvalType::Value,
+            source_to_target,
+        )
         .unwrap()
         .build()
         .unwrap();
