+++ kifmm/examples/mpi_test_fmm_gradients.rs
@@ -3,6 +3,7 @@
 #[cfg(feature = "mpi")]
 fn main() {
     use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
+    use itertools::izip;
     use kifmm::{
         fmm::types::MultiNodeBuilder,
         traits::{
@@ -10,7 +11,7 @@ fn main() {
             tree::{MultiFmmTree, MultiTree},
         },
         tree::{helpers::points_fixture, types::SortKind},
-        ChargeHandler, Evaluate, FftFieldTranslation, SingleNodeBuilder,
+        BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, SingleNodeBuilder,
     };
 
     use rand::{rngs::StdRng, Rng, SeedableRng};
@@ -67,18 +68,13 @@ fn main() {
                 charges.data(),
                 expansion_order,
                 kernel.clone(),
-                GreenKernelEvalType::Value,
+                GreenKernelEvalType::ValueDeriv,
                 source_to_target,
             )
             .unwrap()
             .build()
             .unwrap();
 
-        multi_fmm.evaluate().unwrap();
-
-        // Change the charges associated with the FMM object
-        let new_charges = vec![multi_fmm.rank() as f32; n_points];
-        multi_fmm.attach_charges_unordered(&new_charges).unwrap();
         multi_fmm.evaluate().unwrap();
 
         // Gather all coordinates and charges for the test
@@ -148,17 +144,18 @@ fn main() {
                     &all_charges,
                     &vec![expansion_order; (local_depth + global_depth + 1) as usize],
                     Laplace3dKernel::new(),
-                    GreenKernelEvalType::Value,
+                    GreenKernelEvalType::ValueDeriv,
                     FftFieldTranslation::new(None),
                 )
                 .unwrap()
                 .build()
                 .unwrap();
             single_fmm.evaluate().unwrap();
-            let mut expected = vec![0f32; &multi_fmm.tree.target_tree.coordinates.len() / 3];
+            let mut expected = vec![0f32; 4 * &multi_fmm.tree.target_tree.coordinates.len() / 3];
+            let n_expected = expected.len();
 
             multi_fmm.kernel().evaluate_st(
-                GreenKernelEvalType::Value,
+                GreenKernelEvalType::ValueDeriv,
                 &all_coordinates,
                 &multi_fmm.tree.target_tree.coordinates,
                 &all_charges,
@@ -167,29 +164,156 @@ fn main() {
 
             let distributed = multi_fmm.potentials().unwrap();
 
+            let mut err = 0.;
+            for (l, r) in izip!(expected, distributed) {
+                err += (l - r).abs() / l;
+            }
+
+            err /= n_expected as f32;
+
             // println!(
             //     "{:?} expected: {:?} \n found: {:?}",
             //     world.rank(),
-            //     &distributed[0..100],
-            //     &expected[0..100]
+            //     &distributed[0..25],
+            //     &expected[0..25]
             // );
 
-            // let mut err = 0f32;
-            // for (l, r) in izip!(distributed, expected) {
-            //     println!("l {:?}, r {:?}", l, r);
-            //     err += (l-r).abs()/r.abs();
-            // }
+            assert!(err.abs() < 1e-2);
 
-            // println!("ERROR {:?}", err/(multi_fmm.tree.target_tree.n_coordinates_tot().unwrap() as f32));
+            println!("...test_fmm_gradients M2L=FFT passed");
+        }
+    }
 
-            if multi_fmm.communicator().size() > 1 {
-                distributed
-                    .iter()
-                    .zip(expected.iter())
[…]
