### 2025-01-30 898cb45
ENH: Add gradient calculations to distributed FMM (#137)
* Add gradient evaluation to distributed FMM
