### 2025-02-07 8a1da0b
Fix readme install
