### 2025-02-18 8f10aab
Start working to make variable expansion order available on multi-node
