+++ kifmm/examples/mpi_test_fmm_gradients.rs
@@ -142,7 +142,8 @@ fn main() {
                 .unwrap()
                 .parameters(
                     &all_charges,
-                    &vec![expansion_order; (local_depth + global_depth + 1) as usize],
+                    expansion_order,
+                    None,
                     Laplace3dKernel::new(),
                     GreenKernelEvalType::ValueDeriv,
                     FftFieldTranslation::new(None),
@@ -284,7 +285,8 @@ fn main() {
                 .unwrap()
                 .parameters(
                     &all_charges,
-                    &vec![expansion_order; (local_depth + global_depth + 1) as usize],
+                    expansion_order,
+                    None,
                     Laplace3dKernel::new(),
                     GreenKernelEvalType::ValueDeriv,
                     FftFieldTranslation::new(None),
