+++ kifmm/examples/mpi_test_p2m.rs
@@ -194,7 +194,8 @@ fn main() {
             .unwrap()
             .parameters(
                 &vec![1f32; all_coordinates.len() / 3],
-                &vec![expansion_order; (local_depth + global_depth + 1) as usize],
+                expansion_order,
+                None,
                 Laplace3dKernel::new(),
                 GreenKernelEvalType::Value,
                 FftFieldTranslation::new(None),
