### 2025-02-18 7e91417
Getting started on making pre-computations work
