+++ scripts/src/bin/fmm_m2l_fft_mpi_f32.rs
@@ -100,6 +100,7 @@ fn main() {
         .parameters(
             &charges,
             expansion_order,
+            None,
             kernel.clone(),
             green_kernels::types::GreenKernelEvalType::Value,
             source_to_target,
