+++ kifmm/examples/mpi_test_upward_pass.rs
@@ -56,6 +56,7 @@ fn main() {
         .parameters(
             &charges,
             expansion_order,
+            None,
             kernel,
             GreenKernelEvalType::Value,
             source_to_target,
