+++ kifmm/examples/mpi_test_fmm.rs
@@ -2,311 +2,311 @@
 
 #[cfg(feature = "mpi")]
 fn main() {
-    use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
-    use kifmm::{
-        fmm::types::MultiNodeBuilder,
-        traits::{
-            fmm::{DataAccessMulti, EvaluateMulti},
-            tree::{MultiFmmTree, MultiTree},
-        },
-        tree::{helpers::points_fixture, types::SortKind},
-        BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, SingleNodeBuilder,
-    };
+    // use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
+    // use kifmm::{
+    //     fmm::types::MultiNodeBuilder,
+    //     traits::{
+    //         fmm::{DataAccessMulti, EvaluateMulti},
+    //         tree::{MultiFmmTree, MultiTree},
+    //     },
+    //     tree::{helpers::points_fixture, types::SortKind},
+    //     BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, SingleNodeBuilder,
+    // };
 
-    use rand::{rngs::StdRng, Rng, SeedableRng};
-    use rayon::ThreadPoolBuilder;
+    // use rand::{rngs::StdRng, Rng, SeedableRng};
+    // use rayon::ThreadPoolBuilder;
 
-    use mpi::{
-        datatype::PartitionMut,
-        traits::{Communicator, Root},
-    };
-    use rlst::{rlst_dynamic_array1, RawAccess, RawAccessMut};
+    // use mpi::{
+    //     datatype::PartitionMut,
+    //     traits::{Communicator, Root},
+    // };
+    // use rlst::{rlst_dynamic_array1, RawAccess, RawAccessMut};
 
-    let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
-    let world = universe.world();
-    let comm = world.duplicate();
+    // let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
+    // let world = universe.world();
+    // let comm = world.duplicate();
 
-    // Tree parameters
-    let prune_empty = true;
-    let n_points = 10000;
-    let local_depth = 1;
-    let global_depth = 2;
-    let sort_kind = SortKind::Samplesort { n_samples: 100 };
+    // // Tree parameters
+    // let prune_empty = true;
+    // let n_points = 10000;
+    // let local_depth = 1;
+    // let global_depth = 2;
+    // let sort_kind = SortKind::Samplesort { n_samples: 100 };
 
-    // Fmm Parameters
-    let expansion_order = 4;
-    let kernel = Laplace3dKernel::<f32>::new();
+    // // Fmm Parameters
+    // let expansion_order = 4;
+    // let kernel = Laplace3dKernel::<f32>::new();
 
-    ThreadPoolBuilder::new()
-        .num_threads(2)
-        .build_global()
-        .unwrap();
+    // ThreadPoolBuilder::new()
+    //     .num_threads(2)
+    //     .build_global()
+    //     .unwrap();
 
-    // Test FFT field translation
-    {
-        let source_to_target = FftFieldTranslation::<f32>::new(None);
+    // // Test FFT field translation
+    // {
+    //     let source_to_target = FftFieldTranslation::<f32>::new(None);
 
-        // Generate some random test data local to each process
-        let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
-        let mut rng = StdRng::seed_from_u64(comm.rank() as u64);
[…]
