+++ kifmm/src/fmm/field_translation/metadata/single_node.rs
@@ -621,15 +621,7 @@ where
 {
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
-        let start_level = if let Some(start_level) = start_level {
-            if start_level >= 2 {
-                start_level
-            } else {
-                2
-            }
-        } else {
-            2
-        };
+        let start_level = start_level.unwrap_or(2).max(2);
 
         for level in start_level..=self.tree.source_tree().depth() {
             let sources = self.tree.source_tree().keys(level).unwrap();
@@ -871,15 +863,7 @@ where
 {
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
-        let start_level = if let Some(start_level) = start_level {
-            if start_level >= 2 {
-                start_level
-            } else {
-                2
-            }
-        } else {
-            2
-        };
+        let start_level = start_level.unwrap_or(2).max(2);
 
         for level in start_level..=self.tree.source_tree().depth() {
             let sources = self.tree.source_tree().keys(level).unwrap();
@@ -1254,15 +1238,7 @@ where
 {
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
-        let start_level = if let Some(start_level) = start_level {
-            if start_level >= 2 {
-                start_level
-            } else {
-                2
-            }
-        } else {
-            2
-        };
+        let start_level = start_level.unwrap_or(2).max(2);
 
         for level in start_level..=self.tree.source_tree().depth() {
             let targets = self.tree.target_tree().keys(level).unwrap();
@@ -1818,23 +1794,14 @@ where
 {
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
-
-        let start_level = if let Some(start_level) = start_level {
-            if start_level >= 2 {
-                start_level
-            } else {
-                2
-            }
-        } else {
-            2
-        };
+        let start_level = start_level.unwrap_or(2).max(2);
 
         for level in start_level..=self.tree.source_tree().depth() {
             let mut result = Vec::default();
 
             if let Some(sources) = self.tree.source_tree().keys(level) {
                 let n_sources = sources.len();
-                let sentinel = -1_i32;
+                let sentinel = -1i32;
 
                 let tmp = vec![vec![sentinel; n_sources]; 316];
                 result = tmp.into_iter().map(RwLock::new).collect_vec();
@@ -2329,16 +2296,7 @@ where
 {
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
-
-        let start_level = if let Some(start_level) = start_level {
-            if start_level >= 2 {
-                start_level
-            } else {
-                2
-            }
-        } else {
-            2
-        };
+        let start_level = start_level.unwrap_or(2).max(2);
 
         for level in start_level..=self.tree.source_tree().depth() {
             let mut result = Vec::default();
@@ -3330,7 +3288,7 @@ mod test {
             .unwrap();
 
         let level = 3;
-        let coeff_idx = fmm.c2e_operator_index(level);
+        let coeff_idx = fmm.expansion_index(level);
 
         let mut multipole = rlst_dynamic_array2!(f64, [fmm.n_coeffs_equivalent_surface(level), 1]);
 
@@ -3499,7 +3457,7 @@ mod test {
             .unwrap();
 
         let level = 2;
-        let coeff_index = fmm.expansion_index(level);
+        let coeff_idx = fmm.expansion_index(level);
[…]
