### 2025-02-23 1d4b945
Remove redundant parameter
