+++ kifmm/src/fmm/builder/single_node.rs
@@ -50,7 +50,6 @@ where
             n_coeffs_check_surface: None,
             kernel_eval_type: None,
             fmm_eval_type: None,
-            depth_set: None,
         }
     }
 
@@ -119,7 +118,6 @@ where
             if depth.is_some() && n_crit.is_none() {
                 source_depth = depth.unwrap();
                 target_depth = depth.unwrap();
-                self.depth_set = Some(true);
             } else if depth.is_none() && n_crit.is_some() {
                 // Estimate depth based on a uniform distribution
                 source_depth = SingleNodeTree::<Scalar::Real>::minimum_depth(
@@ -130,7 +128,6 @@ where
                     n_targets as u64,
                     n_crit.unwrap(),
                 );
-                self.depth_set = Some(false);
             } else {
                 return Err(std::io::Error::new(
                     std::io::ErrorKind::InvalidData,
