### 2025-02-23 01f040d
Add data accessors
