+++ kifmm/src/fmm/ghost_exchange.rs
@@ -49,144 +49,150 @@ where
         + SourceToTargetTranslationMetadata,
 {
     fn gather_global_fmm_at_root(&mut self) {
-        let rank = self.communicator.rank();
 
-        // Nominated rank chosen to run the global upward pass
-        let root_rank = 0;
-        let root_process = self.communicator.process_at_rank(root_rank);
+        // TODO
 
-        // Gather multipole data
-        if rank == root_rank {
-            // Allocate memory for locally contained data that will be communicated
-            let n_roots = self.tree.source_tree().n_trees() as Count;
-            let mut multipoles =
-                vec![Scalar::default(); (n_roots as usize) * self.n_coeffs_equivalent_surface];
+        // let rank = self.communicator.rank();
 
-            for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
-                let tmp = self.multipole(&tree.root()).unwrap();
-                multipoles[i * self.n_coeffs_equivalent_surface
-                    ..(i + 1) * self.n_coeffs_equivalent_surface]
-                    .copy_from_slice(tmp);
-            }
+        // // Nominated rank chosen to run the global upward pass
+        // let root_rank = 0;
+        // let root_process = self.communicator.process_at_rank(root_rank);
 
-            let counts_buffers = self
-                .tree
-                .source_tree
-                .all_roots_counts
-                .iter()
-                .map(|c| c * self.n_coeffs_equivalent_surface as i32)
-                .collect_vec();
+        // // Gather multipole data
+        // if rank == root_rank {
+        //     // Allocate memory for locally contained data that will be communicated
+        //     let n_roots = self.tree.source_tree().n_trees() as Count;
+        //     let mut multipoles =
+        //         vec![Scalar::default(); (n_roots as usize) * self.n_coeffs_equivalent_surface];
 
-            let mut displacements_buffers = Vec::new();
-            let mut displacement = 0;
-            for &count in counts_buffers.iter() {
-                displacements_buffers.push(displacement);
-                displacement += count
-            }
+        //     for (i, tree) in self.tree.source_tree().trees().iter().enumerate() {
+        //         let tmp = self.multipole(&tree.root()).unwrap();
+        //         multipoles[i * self.n_coeffs_equivalent_surface
+        //             ..(i + 1) * self.n_coeffs_equivalent_surface]
+        //             .copy_from_slice(tmp);
+        //     }
 
-            // Allocate memory to store received multipoles
-            let total_receive_buffers = counts_buffers.iter().sum::<i32>() as usize;
-            let mut global_multipoles = vec![Scalar::default(); total_receive_buffers];
+        //     let counts_buffers = self
+        //         .tree
+        //         .source_tree
+        //         .all_roots_counts
+        //         .iter()
+        //         .map(|c| c * self.n_coeffs_equivalent_surface as i32)
+        //         .collect_vec();
 
-            // Gather all multipole data
-            let mut partition = PartitionMut::new(
-                &mut global_multipoles,
-                &counts_buffers[..],
-                &displacements_buffers[..],
-            );
+        //     let mut displacements_buffers = Vec::new();
+        //     let mut displacement = 0;
+        //     for &count in counts_buffers.iter() {
+        //         displacements_buffers.push(displacement);
+        //         displacement += count
+        //     }
 
-            root_process.gather_varcount_into_root(&multipoles, &mut partition);
+        //     // Allocate memory to store received multipoles
+        //     let total_receive_buffers = counts_buffers.iter().sum::<i32>() as usize;
+        //     let mut global_multipoles = vec![Scalar::default(); total_receive_buffers];
 
-            self.global_fmm.global_fmm_multipole_metadata(
-                self.tree.source_tree.all_roots.clone(),
[…]
