+++ kifmm/src/fmm/data_access/multi_node.rs
@@ -96,18 +96,15 @@ where
     fn multipoles(&self, level: u64) -> Option<&[Self::Scalar]> {
         if let Some(n_sources) = self.tree().source_tree().n_keys(level) {
             let multipole_ptr = &self.level_multipoles[level as usize][0];
-            // TODO
-            None
-
-            // unsafe {
-            //     match self.fmm_eval_type {
-            //         FmmEvalType::Vector => Some(std::slice::from_raw_parts(
-            //             multipole_ptr.raw,
-            //             self.n_coeffs_equivalent_surface * n_sources,
-            //         )),
-            //         FmmEvalType::Matrix(_n) => None,
-            //     }
-            // }
+            unsafe {
+                match self.fmm_eval_type {
+                    FmmEvalType::Vector => Some(std::slice::from_raw_parts(
+                        multipole_ptr.raw,
+                        self.n_coeffs_equivalent_surface(level) * n_sources,
+                    )),
+                    FmmEvalType::Matrix(_n) => None,
+                }
+            }
         } else {
             None
         }
@@ -117,17 +114,15 @@ where
         if let Some(n_targets) = self.tree().target_tree().n_keys(level) {
             let local_ptr = &self.level_locals[level as usize][0];
 
-            // TODO
-            None
-            // unsafe {
-            //     match self.fmm_eval_type {
-            //         FmmEvalType::Vector => Some(std::slice::from_raw_parts(
-            //             local_ptr.raw,
-            //             self.n_coeffs_equivalent_surface * n_targets,
-            //         )),
-            //         FmmEvalType::Matrix(_n) => None,
-            //     }
-            // }
+            unsafe {
+                match self.fmm_eval_type {
+                    FmmEvalType::Vector => Some(std::slice::from_raw_parts(
+                        local_ptr.raw,
+                        self.n_coeffs_equivalent_surface(level) * n_targets,
+                    )),
+                    FmmEvalType::Matrix(_n) => None,
+                }
+            }
         } else {
             None
         }
@@ -140,17 +135,15 @@ where
         if let Some(&key_idx) = self.tree().source_tree().level_index(key) {
             let multipole_ptr = &self.level_multipoles[key.level() as usize][key_idx];
 
-            // TODO
-            None
-            // unsafe {
-            //     match self.fmm_eval_type {
-            //         FmmEvalType::Vector => Some(std::slice::from_raw_parts(
-            //             multipole_ptr.raw,
-            //             self.n_coeffs_equivalent_surface,
-            //         )),
-            //         FmmEvalType::Matrix(_n) => None,
-            //     }
-            // }
+            unsafe {
+                match self.fmm_eval_type {
+                    FmmEvalType::Vector => Some(std::slice::from_raw_parts(
+                        multipole_ptr.raw,
+                        self.n_coeffs_equivalent_surface(key.level()),
+                    )),
+                    FmmEvalType::Matrix(_n) => None,
+                }
+            }
         } else {
             None
         }
@@ -162,17 +155,15 @@ where
     ) -> Option<&mut [Self::Scalar]> {
         if let Some(&key_idx) = self.tree().source_tree().level_index(key) {
             let multipole_ptr = &self.level_multipoles[key.level() as usize][key_idx];
-            // TODO
-            None
-            // unsafe {
-            //     match self.fmm_eval_type {
-            //         FmmEvalType::Vector => Some(std::slice::from_raw_parts_mut(
-            //             multipole_ptr.raw,
-            //             self.n_coeffs_equivalent_surface,
-            //         )),
-            //         FmmEvalType::Matrix(_n) => None,
-            //     }
-            // }
+            unsafe {
+                match self.fmm_eval_type {
[…]
