+++ kifmm/src/fmm/field_translation/source/multi_node.rs
@@ -21,7 +21,7 @@ use crate::{
     },
     traits::{
         field::{FieldTranslation as FieldTranslationTrait, SourceTranslation},
-        fmm::{DataAccessMulti, HomogenousKernel},
+        fmm::{DataAccessMulti, HomogenousKernel, MetadataAccess},
         tree::{MultiFmmTree, MultiTree},
         types::FmmError,
     },
@@ -35,237 +35,243 @@ where
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
     FieldTranslation: FieldTranslationTrait,
     Kernel: KernelTrait<T = Scalar> + HomogenousKernel + Default + Send + Sync,
-    Self: DataAccessMulti<Scalar = Scalar>,
+    Self: MetadataAccess + DataAccessMulti<Scalar = Scalar>,
 {
     fn p2m(&self) -> Result<(), crate::traits::types::FmmError> {
-        // TODO
+        if let Some(_leaves) = self.tree.source_tree().all_leaves() {
 
-        // if let Some(_leaves) = self.tree.source_tree().all_leaves() {
-        //     let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface;
-        //     let n_coeffs_check_surface = self.n_coeffs_check_surface;
-        //     let n_leaves = self.tree.source_tree().n_leaves().unwrap();
-        //     let check_surface_size = n_coeffs_check_surface * self.dim;
-        //     let coordinates = self.tree.source_tree().all_coordinates().unwrap();
+            let &n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface.last().unwrap();
+            let &n_coeffs_check_surface = self.n_coeffs_check_surface.last().unwrap();
+            let n_leaves = self.tree.source_tree().n_leaves().unwrap();
+            let check_surface_size = n_coeffs_check_surface * self.dim;
 
-        //     let all_charges = &self.charges;
-        //     let kernel = &self.kernel;
-        //     let dim = self.dim;
+            let coordinates = self.tree.source_tree().all_coordinates().unwrap();
+            let all_charges = &self.charges;
+            let kernel = &self.kernel;
+            let dim = self.dim;
+            let depth = self.tree.source_tree().total_depth();
+            let operator_index = self.c2e_operator_index(depth);
 
-        //     match self.fmm_eval_type {
-        //         FmmEvalType::Vector => {
-        //             let mut check_potentials =
-        //                 rlst_dynamic_array2!(Scalar, [n_leaves * n_coeffs_check_surface, 1]);
+            if self.rank == 0 {
+                println!("HERE {:?} {:?} {:?} {:?}", operator_index, self.source.len(), self.uc2e_inv_1.len(), n_coeffs_equivalent_surface)
+            }
 
-        //             // Compute check potential for each box
-        //             check_potentials
-        //                 .data_mut()
-        //                 .par_chunks_exact_mut(n_coeffs_check_surface)
-        //                 .zip(
-        //                     self.leaf_upward_check_surfaces_sources
-        //                         .par_chunks_exact(check_surface_size),
-        //                 )
-        //                 .zip(&self.charge_index_pointer_sources)
-        //                 .for_each(
-        //                     |((check_potential, upward_check_surface), charge_index_pointer)| {
-        //                         let charges =
-        //                             &all_charges[charge_index_pointer.0..charge_index_pointer.1];
+            match self.fmm_eval_type {
+                FmmEvalType::Vector => {
+                    let mut check_potentials =
+                        rlst_dynamic_array2!(Scalar, [n_leaves * n_coeffs_check_surface, 1]);
 
-        //                         let coordinates_row_major = &coordinates
[…]
