### 2025-02-24 1f87ce7
Start work on distributed helmholtz
