### 2025-02-25 02de91a
Debug upward pass variable expansion order, mostly
