+++ kifmm/examples/mpi_test_upward_pass.rs
@@ -1,6 +1,9 @@
 //? mpirun -n {{NPROCESSES}} --features "mpi"
 
 #[cfg(feature = "mpi")]
+fn test_upward_pass() {}
+
+#[cfg(feature = "mpi")]
 fn main() {
     use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
     use kifmm::traits::general::multi_node::GhostExchange;
@@ -21,173 +24,86 @@ fn main() {
     };
     use rlst::{RawAccess, RlstScalar};
 
-    let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
-    let world = universe.world();
-    let comm = world.duplicate();
+    fn test_upward_pass(name: String, expansion_order: &[usize], world: &mpi::topology::SimpleCommunicator) {
 
-    // Tree parameters
-    let prune_empty = true;
-    let n_points = 10000;
-    let local_depth = 3;
-    let global_depth = 2;
-    let sort_kind = SortKind::Samplesort { n_samples: 100 };
+        let comm = world.duplicate();
 
-    // Fmm Parameters
-    let expansion_order = [5, 4, 5, 4, 5, 4];
-    let kernel = Laplace3dKernel::<f32>::new();
-    let source_to_target =
-        BlasFieldTranslationSaRcmp::<f32>::new(None, None, FmmSvdMode::Deterministic);
+        // Tree parameters
+        let prune_empty = true;
+        let n_points = 10000;
+        let local_depth = 3;
+        let global_depth = 2;
+        let sort_kind = SortKind::Samplesort { n_samples: 100 };
 
-    // Generate some random test data local to each process
-    let points = points_fixture::<f32>(n_points, None, None, None);
-    let charges = vec![1f32; n_points];
+        // Fmm Parameters
 
-    let mut fmm = MultiNodeBuilder::new(false)
-        .tree(
-            &comm,
-            points.data(),
-            points.data(),
-            local_depth,
-            global_depth,
-            prune_empty,
-            sort_kind,
-        )
-        .unwrap()
-        .parameters(
-            &charges,
-            &expansion_order,
-            kernel,
-            GreenKernelEvalType::Value,
-            source_to_target,
-        )
-        .unwrap()
-        .build()
-        .unwrap();
+        let kernel = Laplace3dKernel::<f32>::new();
+        let source_to_target =
+            BlasFieldTranslationSaRcmp::<f32>::new(None, None, FmmSvdMode::Deterministic);
 
-    // Perform partial upward pass on each rank
-    fmm.evaluate_leaf_sources().unwrap();
-    fmm.evaluate_upward_pass().unwrap();
+        // Generate some random test data local to each process
+        let points = points_fixture::<f32>(n_points, None, None, None);
+        let charges = vec![1f32; n_points];
 
-    // Test at roots of local trees for result of partial upward passes
-    let roots = fmm.tree().source_tree().roots();
+        let mut fmm = MultiNodeBuilder::new(false)
+            .tree(
+                &comm,
+                points.data(),
+                points.data(),
+                local_depth,
+                global_depth,
+                prune_empty,
+                sort_kind,
+            )
+            .unwrap()
+            .parameters(
+                &charges,
+                &expansion_order,
+                kernel,
+                GreenKernelEvalType::Value,
+                source_to_target,
+            )
+            .unwrap()
+            .build()
+            .unwrap();
 
-    let distant_point = vec![100000., 0., 0.];
-    let mut expected = vec![0.];
-    let mut found = vec![0.];
+        // Perform partial upward pass on each rank
[…]
