### 2025-02-26 ad611ca
Add target metadata
