+++ kifmm/src/fmm/field_translation/source_to_target/blas/multi_node.rs
@@ -50,6 +50,7 @@ where
                     // Metadata
                     let m2l_operator_index = self.m2l_operator_index(level);
                     let c2e_operator_index = self.c2e_operator_index(level);
+                    let displacement_index = self.displacement_index(level);
                     let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface(level);
 
                     // Parameters
@@ -59,7 +60,6 @@ where
                     let mut all_n_sources = Vec::new();
                     let mut all_multipoles = Vec::new();
                     let mut all_displacements = Vec::new();
-                    let displacement_index = self.displacement_index(level);
 
                     // Handle locally contained source boxes
                     if let Some(sources) = self.tree().source_tree().keys(level) {
