+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -56,6 +56,42 @@ use crate::{FftFieldTranslation, FmmSvdMode, SingleNodeFmmTree};
 
 use super::single_node::find_cutoff_rank;
 
+/// Calculate load for precomputation based on a block distribution strategy
+pub(crate) fn calculate_precomputation_load(
+    n_precomputations: i32,
+    size: i32,
+) -> Option<(Vec<Count>, Vec<Count>)> {
+    if n_precomputations > 0 {
+        let mut counts;
+
+        if n_precomputations > 1 {
+            // Distributed pre-computation
+            let q = n_precomputations / size; // Base number of calculations per processor
+            let r = n_precomputations % size; // Extra calculations to distribute evenly among ranks
+
+            // Block distribution strategy
+            counts = (0..size)
+                .map(|i| if i < r { q + 1 } else { q })
+                .collect_vec();
+        } else {
+            // If only have one pre-computation, carry out on a single rank (root rank)
+            counts = vec![0; size as usize];
+            counts[0] = n_precomputations;
+        }
+
+        let mut curr = 0;
+        let mut displacements = Vec::new();
+        for &count in counts.iter() {
+            displacements.push(curr);
+            curr += count;
+        }
+
+        Some((counts, displacements))
+    } else {
+        None
+    }
+}
+
 impl<Scalar, FieldTranslation> SourceTranslationMetadata
     for KiFmmMulti<Scalar, Laplace3dKernel<Scalar>, FieldTranslation>
 where
@@ -86,34 +122,11 @@ where
             1
         };
 
-        // Vector characterising load of pre-computation at each rank (denoted by index)
-        let mut load;
-
-        if n_precomputations > 1 {
-            // Distributed pre-computation
-            let q = n_precomputations / size; // Base number of calculations per processor
-            let r = n_precomputations % size; // Extra calculations to distribute evenly among ranks
-
-            // Block distribution strategy
-            load = (0..size)
-                .map(|i| if i < r { q + 1 } else { q })
-                .collect_vec();
-        } else {
-            // If only have one pre-computation, carry out on a single rank (root rank)
-            let size = self.communicator.size(); // Number of processors available
-            load = vec![0; size as usize];
-            load[0] = n_precomputations;
-        }
-
-        let mut curr = 0;
-        let mut load_displacement = Vec::new();
-        for &count in load.iter() {
-            load_displacement.push(curr);
-            curr += count;
-        }
+        let (load_counts, load_displacement) =
+            calculate_precomputation_load(n_precomputations, size).unwrap();
 
         // Compute mandated local portion
-        let local_load_count = load[rank as usize];
+        let local_load_count = load_counts[rank as usize];
         let local_load_displacement = load_displacement[rank as usize];
         let equivalent_surface_order = &self.equivalent_surface_order[(local_load_displacement
             as usize)
@@ -163,7 +176,7 @@ where
         let mut shared_dim = vec![0usize; n_precomputations as usize];
         {
             let mut partition =
-                PartitionMut::new(&mut shared_dim, &load[..], &load_displacement[..]);
+                PartitionMut::new(&mut shared_dim, &load_counts[..], &load_displacement[..]);
             self.communicator
[…]
