+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -18,6 +18,7 @@ use rlst::{
 use std::collections::{HashMap, HashSet};
 use std::sync::{Mutex, RwLock};
 
+use crate::fmm::field_translation::source_to_target::transfer_vector::compute_transfer_vectors_at_level;
 use crate::fmm::helpers::multi_node::{
     coordinate_index_pointer_multi_node, leaf_expansion_pointers_multi_node,
     leaf_surfaces_multi_node, level_expansion_pointers_multi_node, level_index_pointer_multi_node,
@@ -102,7 +103,7 @@ where
 {
     fn source(&mut self) {
         let root = MortonKey::<Scalar::Real>::root();
-        let size = self.communicator.size(); // Number of processors available
+        let size = self.communicator.size();
         let rank = self.communicator.rank();
 
         // Cast surface parameters
@@ -827,6 +828,150 @@ where
     }
 
     fn source_to_target(&mut self) {
+        let size = self.communicator.size();
+        let rank = self.communicator.rank();
+        let total_depth = self.tree.source_tree().total_depth();
+        let alpha = Scalar::real(ALPHA_INNER);
+
+        // Distribute SVD by level
+        // Number of pre-computations
+        let n_precomputations = if self.variable_expansion_order {
+            self.equivalent_surface_order.len() as i32
+        } else {
+            1
+        };
+
+        let (load_counts, load_displacement) =
+            calculate_precomputation_load(n_precomputations, size).unwrap();
+
+        // Compute mandated local portion
+        let local_load_count = load_counts[rank as usize];
+        let local_load_displacement = load_displacement[rank as usize];
+        let equivalent_surface_order = &self.equivalent_surface_order[(local_load_displacement
+            as usize)
+            ..((local_load_displacement + local_load_count) as usize)];
+        let check_surface_order = &self.check_surface_order[(local_load_displacement as usize)
+            ..((local_load_displacement + local_load_count) as usize)];
+        // let mut local_shared_dim = Vec::new();
+
+        for (&equivalent_surface_order, &check_surface_order, level) in izip!(equivalent_surface_order, check_surface_order, 0..=total_depth)
+        {
+            let transfer_vectors =
+                compute_transfer_vectors_at_level::<Scalar::Real>(level).unwrap();
+
+            let n_rows = ncoeffs_kifmm(check_surface_order);
+            let n_cols = ncoeffs_kifmm(equivalent_surface_order);
+
+            let mut se2tc_fat =
+                rlst_dynamic_array2!(Scalar, [n_rows, n_cols * NTRANSFER_VECTORS_KIFMM]);
+            let mut se2tc_thin =
+                rlst_dynamic_array2!(Scalar, [n_rows * NTRANSFER_VECTORS_KIFMM, n_cols]);
+
+            transfer_vectors.iter().enumerate().for_each(|(i, t)| {
+                let source_equivalent_surface = t.source.surface_grid(
+                    equivalent_surface_order,
+                    self.tree.source_tree().domain(),
+                    alpha,
+                );
+                let target_check_surface = t.target.surface_grid(
+                    check_surface_order,
+                    self.tree.source_tree().domain(),
+                    alpha,
+                );
+
+                let mut tmp_gram = rlst_dynamic_array2!(Scalar, [n_rows, n_cols]);
+
+                self.kernel.assemble_st(
+                    GreenKernelEvalType::Value,
+                    &target_check_surface[..],
+                    &source_equivalent_surface[..],
+                    tmp_gram.data_mut(),
+                );
+
+                let mut block = se2tc_fat
+                    .r_mut()
+                    .into_subview([0, i * n_cols], [n_rows, n_cols]);
+                block.fill_from(tmp_gram.r());
+
+                let mut block_column = se2tc_thin
[…]
