### 2025-03-04 1bd17e4
Start blas source to target precomputaion, distributed
