### 2025-04-09 #7124 @rdfriese
> Hi [@rdfriese](https://github.com/rdfriese) would have time to review this paper?

I think I should be able to, what sort of timeline are we looking at?
