### 2025-04-09 #7124 @diehlpk
> > Hi [@rdfriese](https://github.com/rdfriese) would have time to review this paper?
>
> I think I should be able to, what sort of timeline are we looking at?

If you could finish it within 6 weeks that would be great.
