+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -762,6 +762,7 @@ where
     Scalar: RlstScalar + Default + MatrixRsvd + Equivalence + Float,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
 {
+    // TODO even this can be parallelized with MPI
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
 
@@ -835,11 +836,31 @@ where
 
         // Distribute SVD by level
         // Number of pre-computations
-        let n_precomputations = if self.variable_expansion_order {
-            self.equivalent_surface_order.len() as i32
+        // Distribute the SVD by level
+        let iterator = if self.variable_expansion_order {
+            (2..=total_depth)
+                .zip(self.equivalent_surface_order.iter().skip(2).cloned())
+                .zip(self.check_surface_order.iter().skip(2).cloned())
+                .collect_vec()
         } else {
-            1
+            (2..=total_depth)
+                .zip(vec![
+                    *self.equivalent_surface_order.last().unwrap();
+                    (total_depth - 1) as usize
+                ])
+                .zip(vec![
+                    *self.check_surface_order.last().unwrap();
+                    (total_depth - 1) as usize
+                ])
+                .collect_vec()
         };
+
+        let n_precomputations;
+        if self.variable_expansion_order {
+            n_precomputations = iterator.len() as i32;
+        } else {
+            n_precomputations = 1;
+        }
 
         let (load_counts, load_displacement) =
             calculate_precomputation_load(n_precomputations, size).unwrap();
@@ -847,11 +868,11 @@ where
         // Compute mandated local portion
         let local_load_count = load_counts[rank as usize];
         let local_load_displacement = load_displacement[rank as usize];
-        let equivalent_surface_order = &self.equivalent_surface_order[(local_load_displacement
-            as usize)
+
+        let iterator_r = &iterator[(local_load_displacement as usize)
             ..((local_load_displacement + local_load_count) as usize)];
-        let check_surface_order = &self.check_surface_order[(local_load_displacement as usize)
-            ..((local_load_displacement + local_load_count) as usize)];
+
+        // println!("Rank {:?} iterator {:?}", rank, iterator_r);
 
         let mut u_r = Vec::new();
         let mut st_r = Vec::new();
@@ -860,8 +881,7 @@ where
         let mut cutoff_rank_r = Vec::new();
         let mut directional_cutoff_ranks_r = Vec::new();
 
-        for (&equivalent_surface_order, &check_surface_order, level) in izip!(equivalent_surface_order, check_surface_order, 0..=total_depth)
-        {
+        for &((level, equivalent_surface_order), check_surface_order) in iterator_r.iter() {
             let transfer_vectors =
                 compute_transfer_vectors_at_level::<Scalar::Real>(level).unwrap();
 
@@ -1051,7 +1071,7 @@ where
             let c_u = Mutex::new(Vec::new());
             let c_vt = Mutex::new(Vec::new());
             let directional_cutoff_ranks =
-                Mutex::new(vec![0usize; self.source_to_target.transfer_vectors.len()]);
+                Mutex::new(vec![0i32; self.source_to_target.transfer_vectors.len()]);
 
             for _ in 0..NTRANSFER_VECTORS_KIFMM {
                 c_u.lock()
@@ -1105,7 +1125,7 @@ where
                 let vt_i_compressed = empty_array::<Scalar, 2>()
                     .simple_mult_into_resize(sigma_mat_i_compressed.r(), vt_i_compressed_.r());
 
[…]
