+++ kifmm/src/fmm/ghost_exchange.rs
@@ -187,8 +187,8 @@ where
         // TODO might be broken if level locals is incorrectly set
         // Insert received local data into target tree
         for (i, root) in expected_roots.iter().enumerate() {
-            let l = i * self.n_coeffs_equivalent_surface;
-            let r = l + self.n_coeffs_equivalent_surface;
+            let l = i * n_coeffs_equivalent_surface;
+            let r = l + n_coeffs_equivalent_surface;
             self.local_mut(root)
                 .unwrap()
                 .copy_from_slice(&receive_buffer[l..r]);
@@ -686,6 +686,8 @@ where
                 .all_to_all_varcount_into(&partition_send, &mut partition_receive);
         }
 
+        self.ghost_requested_queries_v = requested_queries.clone();
+
         // Store original ordering of received data temporarily
         let ghost_keys = requested_queries
             .into_iter()
@@ -750,115 +752,172 @@ where
         );
     }
 
+    // TODO the hard one
+    // Maybe need to re-think how I send back multipoles as I now need to keep a track of the buffer sizes which I don't
+    // do explicitl.y
     fn v_list_exchange_runtime(&mut self) {
-        // TODO
 
-        // // Available multipole data from received requests
-        // let mut available_multipoles = Vec::new();
+        // Available multipole data from received requests
+        let mut available_multipoles = Vec::new();
         // let mut available_multipoles_counts = Vec::new();
+        let mut available_multipoles_buffer_sizes_counts = Vec::new();
         // let mut available_multipoles_displacements = Vec::new();
 
-        // {
-        //     let mut counter = 0;
+        {
+            let mut counter = 0;
 
-        //     // Iterate over received data rank by rank
-        //     for (count, displacement) in self
-        //         .ghost_received_queries_counts_v
-        //         .iter()
-        //         .zip(self.ghost_received_queries_displacements_v.iter())
-        //     {
-        //         let l = *displacement as usize;
-        //         let r = l + (*count as usize);
+            // Iterate over received data rank by rank
+            for (count, displacement) in self
+                .ghost_received_queries_counts_v
+                .iter()
+                .zip(self.ghost_received_queries_displacements_v.iter())
+            {
+                let l = *displacement as usize;
+                let r = l + (*count as usize);
 
-        //         // Received queries from this rank
-        //         let received_queries_rank = &self.ghost_received_queries_v[l..r]; // received queries from this rank
+                // Received queries from this rank
+                let received_queries_rank = &self.ghost_received_queries_v[l..r]; // received queries from this rank
 
-        //         // Filter for available data corresponding to this request
-        //         let mut available_multipoles_rank = Vec::new();
+                // Filter for available data corresponding to this request
+                let mut available_multipoles_rank = Vec::new();
 
-        //         let mut counter_rank = 0i32;
+                // Associate each with expected buffer sizes for multipole data
+                let mut available_multipoles_rank_buffer_size = Vec::new();
 
-        //         // Only communicate back queries and associated data if multipole data is found
-        //         for &query in received_queries_rank.iter() {
-        //             let key = MortonKey::from_morton(query);
-        //             if let Some(multipole) = self.multipole(&key) {
-        //                 available_multipoles_rank.push(multipole);
-        //                 counter_rank += 1;
-        //             }
-        //         }
+                let mut counter_rank = 0i32;
 
-        //         // Update return buffers
-        //         let available_multipoles_rank = available_multipoles_rank.concat();
-        //         available_multipoles.extend(available_multipoles_rank);
-        //         available_multipoles_counts
[…]
