+++ kifmm/src/fmm/builder/multi_node.rs
@@ -422,6 +422,9 @@ where
                 ghost_requested_queries_v: Vec::default(),
                 ghost_requested_queries_key_to_index_v: HashMap::default(),
                 ghost_requested_queries_counts_v: Vec::default(),
+                ghost_requested_queries_v_buffer_sizes_counts: Vec::default(),
+                ghost_requested_queries_v_buffer_sizes_displacements: Vec::default(),
+                ghost_requested_queries_displacements_v: Vec::default(),
                 ghost_received_queries_displacements_v: Vec::default(),
                 ghost_received_queries_counts_v: Vec::default(),
                 ghost_received_queries_v: Vec::default(),
