+++ kifmm/src/fmm/ghost_exchange.rs
@@ -119,7 +119,6 @@ where
     }
 
     fn scatter_global_fmm_from_root(&mut self) {
-
         let rank = self.communicator.rank();
         let n_roots = self.tree.target_tree.trees.len();
 
@@ -554,13 +553,13 @@ where
         // filtering for available queries to be sent back to partners
         let mut received_queries = vec![0u64; total_receive_count];
 
-        // Available keys
+        // Filter for queries that exist
         let mut available_queries = Vec::new();
         let mut available_queries_counts = Vec::new();
         let mut available_queries_displacements = Vec::new();
 
         {
-            // Communicate queries
+            // First, communicate all queries, whether or not they exist
             let partition_send = Partition::new(
                 &self.tree.v_list_query.queries,
                 neighbourhood_send_counts,
@@ -576,7 +575,7 @@ where
             self.neighbourhood_communicator_v
                 .all_to_all_varcount_into(&partition_send, &mut partition_receive);
 
-            // Filter for locally available queries to send back
+            // Then, filter for queries that actually exist at this rank and send back this information
             let receive_counts = partition_receive.counts().iter().cloned().collect_vec();
             let receive_displacements = partition_receive.displs().iter().cloned().collect_vec();
 
@@ -654,7 +653,7 @@ where
                 .all_to_all_varcount_into(&partition_send, &mut partition_receive);
         }
 
-        // Create buffers to receive charge and coordinate data
+        // Create buffers to receive queries which have been requested from neighbours
         let total_receive_count_requested_queries =
             requested_queries_counts.iter().sum::<i32>() as usize;
         let mut requested_queries = vec![0u64; total_receive_count_requested_queries];
@@ -686,11 +685,10 @@ where
                 .all_to_all_varcount_into(&partition_send, &mut partition_receive);
         }
 
-        self.ghost_requested_queries_v = requested_queries.clone();
-
         // Store original ordering of received data temporarily
         let ghost_keys = requested_queries
-            .into_iter()
+            .iter()
+            .cloned()
             .map(MortonKey::<Scalar::Real>::from_morton)
             .collect_vec();
 
@@ -700,11 +698,36 @@ where
             ghost_requested_queries_key_to_index_v.insert(key, i);
         }
 
+        let mut ghost_requested_queries_v_buffer_sizes_counts = Vec::new();
+        for (idx, key) in ghost_keys.iter().enumerate() {
+            ghost_requested_queries_v_buffer_sizes_counts
+                .push(self.n_coeffs_equivalent_surface(key.level()));
+        }
+
+        let ghost_requested_queries_v_buffer_sizes_displacements =
+            ghost_requested_queries_v_buffer_sizes_counts
+                .iter()
+                .scan(0, |acc, &x| {
+                    let tmp = *acc;
+                    *acc += x;
+                    Some(tmp)
+                })
+                .collect_vec();
+
+        // Store the (available) requested queries, their counts, displacements and a mapping from their index
+        // in the request buffer to the morton key.
+        self.ghost_requested_queries_v_buffer_sizes_counts =
+            ghost_requested_queries_v_buffer_sizes_counts;
+        self.ghost_requested_queries_v_buffer_sizes_displacements =
+            ghost_requested_queries_v_buffer_sizes_displacements;
+        self.ghost_requested_queries_v = requested_queries;
+        self.ghost_requested_queries_displacements_v = requested_queries_displacements;
         self.ghost_requested_queries_counts_v = requested_queries_counts;
         self.ghost_requested_queries_key_to_index_v = ghost_requested_queries_key_to_index_v;
 
         // Create a set of all keys, ensure all sibling keys are included too
         // as this is required for efficient kernel application
[…]
