### 2025-05-02 ba73ed6
Add l2 error to grid search
