### 2025-05-07 #7124 @rdfriese
@jeremylt are you able to compile, I'm getting "a using two versions of `syn`" error in the rlst crate.

I'm running the latest rust version, so I wonder if its related to that
