### 2025-05-07 #7124 @jeremylt
Yeah, I'm getting something similar I think
```console
error[E0308]: mismatched types
  --> /home/jeremy/.cargo/git/checkouts/rlst-b28b4cb12e0eef38/dc3fb1e/proc-macro/src/tracing.rs:11:9
   |
11 |     id: String,
   |         ^^^^^^
   |         |
   |         expected fn pointer, found fn item
   |         arguments to this function are incorrect
   |
   = note: expected fn pointer `for<'a> fn(&'a syn::Meta) -> Result<_, _>`
                 found fn item `for<'a> fn(&'a syn::attr::Meta) -> Result<_, _> {<_ as FromMeta>::from_meta}`
help: the return type of this call is `for<'a> fn(&'a syn::attr::Meta) -> Result<_, darling::Error> {<_ as FromMeta>::from_meta}` due to the type of the argument passed
  --> /home/jeremy/.cargo/git/checkouts/rlst-b28b4cb12e0eef38/dc3fb1e/proc-macro/src/tracing.rs:11:9
   |
11 |     id: String,
   |         ^^^^^^ this argument influences the return type of `{{root}}`
note: function defined here
  --> /rustc/05f9846f893b09a1be1fc8560e33fc3c815cfecb/library/core/src/convert/mod.rs:104:14

error[E0308]: mismatched types
   --> /home/jeremy/.cargo/git/checkouts/rlst-b28b4cb12e0eef38/dc3fb1e/proc-macro/src/tracing.rs:11:9
    |
11  |     id: String,
    |         ^^^^^^
    |         |
    |         expected `syn::Meta`, found `syn::attr::Meta`
    |         arguments to this function are incorrect
    |
note: two different versions of crate `syn` are being used; two types coming from two different versions of the same crate are different types even if they look the same
   --> /home/jeremy/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/syn-1.0.109/src/attr.rs:336:1
    |
336 | / ast_enum_of_structs! {
337 | |     /// Content of a compile-time structured attribute.
338 | |     ///
339 | |     /// *This type is available only if Syn is built with the `"derive"` or `"full"`
...   |
370 | | }
    | |_^ this is the expected type `syn::Meta`
    |
   ::: /home/jeremy/.cargo/git/checkouts/rlst-b28b4cb12e0eef38/dc3fb1e/proc-macro/src/tracing.rs:3:5
    |
3   |   use darling::{ast::NestedMeta, Error, FromMeta};
    |       ------- one version of crate `syn` used here, as a dependency of crate `darling_core`
...
6   |   use syn::parse_macro_input;
    |       --- one version of crate `syn` used here, as a direct dependency of the current crate
    |
   ::: /home/jeremy/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/syn-2.0.101/src/attr.rs:451:1
    |
451 | / ast_enum! {
452 | |     /// Content of a compile-time structured attribute.
453 | |     ///
454 | |     /// ## Path
...   |
482 | | }
    | |_^ this is the found type `syn::attr::Meta`
    = help: you can use `cargo tree` to explore your dependency tree
    = note: this error originates in the macro `strip_attrs_pub` which comes from the expansion of the macro `ast_enum` (in Nightly builds, run with -Z macro-backtrace for more info)

For more information about this error, try `rustc --explain E0308`.
error: could not compile `rlst-proc-macro` (lib) due to 2 previous errors
warning: build failed, waiting for other jobs to finish...
```
