### 2025-05-08 #7124 @skailasa
Hi have bumped the relevant dependencies for the latest Rust versions, will update upon merge.
