### 2025-05-08 #7124 @skailasa
@jeremylt @rdfriese , I've bumped the RLST dependencies for the new version of Rust, you should be able to build now after cargo update.

Thank you
