### 2025-05-09 f435874
Add benchmarking scripts
