+++ scripts/src/bin/grid_search_single_node.rs
@@ -11,10 +11,10 @@ use kifmm::{
     traits::{
         fftw::Dft,
         fmm::{DataAccess, Evaluate},
-        general::single_node::{AsComplex, Epsilon, Hadamard8x8},
+        general::single_node::{AsComplex, Epsilon, GetCutoffRank, Hadamard8x8},
         tree::{SingleFmmTree, SingleTree},
     },
-    tree::helpers::points_fixture,
+    tree::helpers::{points_fixture, points_fixture_sphere},
     BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder,
 };
 use num::Float;
@@ -23,6 +23,7 @@ use rlst::{rlst_dynamic_array2, MatrixSvd, RawAccess, RawAccessMut, RlstScalar};
 
 fn grid_search_laplace_blas<T>(
     filename: String,
+    geometry: &str,
     n_points: usize,
     expansion_order_vec: &[usize],
     svd_threshold_vec: &[Option<T>],
@@ -65,9 +66,33 @@ fn grid_search_laplace_blas<T>(
     // Setup random sources and targets
     let n_sources = n_points;
     let n_targets = n_points;
-    let sources = points_fixture::<T::Real>(n_sources, None, None, Some(0));
-    let targets = points_fixture::<T::Real>(n_targets, None, None, Some(1));
+    let sources;
+    let targets;
+    if geometry == "uniform" {
+        sources = points_fixture::<T::Real>(n_sources, None, None, Some(0));
+        targets = points_fixture::<T::Real>(n_targets, None, None, Some(1));
+    } else if geometry == "sphere" {
+        sources = points_fixture_sphere::<T::Real>(n_sources);
+        targets = points_fixture_sphere::<T::Real>(n_targets);
+    }
+    // else if geometry == "spheroid" {
+    //     sources = points_fixture_oblate_spheroid::<T::Real>(
+    //         n_sources,
+    //         T::from(1.0).unwrap(),
+    //         T::from(0.5).unwrap(),
+    //     );
+    //     targets = points_fixture_oblate_spheroid::<T::Real>(
+    //         n_targets,
+    //         T::from(1.0).unwrap(),
+    //         T::from(0.5).unwrap(),
+    //     );
+    // }
+    else {
+        panic!("Unsupported geometry")
+    }
+
     let n_vecs = 1;
+
     let tmp = vec![T::one(); n_sources * n_vecs];
     let mut charges = rlst_dynamic_array2!(T, [n_sources, n_vecs]);
     charges.data_mut().copy_from_slice(&tmp);
@@ -125,9 +150,12 @@ fn grid_search_laplace_blas<T>(
             "min_rel_err".to_string(),
             "mean_rel_err".to_string(),
             "max_rel_err".to_string(),
+            "rel_l2_err".to_string(),
             "n_iter".to_string(),
             "n_oversamples".to_string(),
             "setup_time".to_string(),
+            "cutoff_rank".to_string(),
+            "rsvd".to_string(),
         ])
         .unwrap();
 
@@ -135,14 +163,6 @@ fn grid_search_laplace_blas<T>(
         "BLAS Pre-computation Time Elapsed {:?}",
         s.elapsed().as_secs()
     );
-
-    // Setup random sources and targets
-    let n_sources = 1000000;
-    let sources = points_fixture::<T::Real>(n_sources, None, None, Some(0));
-    let n_vecs = 1;
-    let tmp = vec![T::one(); n_sources * n_vecs];
-    let mut charges = rlst_dynamic_array2!(T, [n_sources, n_vecs]);
-    charges.data_mut().copy_from_slice(&tmp);
 
[…]
