+++ kifmm/benches/laplace_f32.rs
@@ -10,7 +10,7 @@ use kifmm::traits::tree::{SingleFmmTree, SingleTree};
 use kifmm::tree::helpers::points_fixture;
 use rlst::{rlst_dynamic_array2, RawAccess, RawAccessMut};
 
-fn laplace_potentials_f32(c: &mut Criterion) {
+fn laplace_potentials_f32_1e6(c: &mut Criterion) {
     let mut group = c.benchmark_group("F32 Potentials");
 
     group
@@ -22,51 +22,51 @@ fn laplace_potentials_f32(c: &mut Criterion) {
     let sources = points_fixture::<f32>(n_sources, None, None, Some(0));
     let targets = points_fixture::<f32>(n_targets, None, None, Some(1));
 
-    let nvecs = 1;
+    let nvecs = 10;
     let tmp = vec![1.0; n_sources * nvecs];
     let mut charges = rlst_dynamic_array2!(f32, [n_sources, nvecs]);
     charges.data_mut().copy_from_slice(&tmp);
 
-    // 3 Digits
+    // // 3 Digits
     {
-        // FFT based M2L for a vector of charges
-        // FMM parameters
-        let n_crit = None;
-        let depth = Some(4);
-        let e = 3;
-        let expansion_order = vec![e; depth.unwrap() as usize + 1];
-        let prune_empty = true;
-        let block_size = Some(32);
+        // // FFT based M2L for a vector of charges
+        // // FMM parameters
+        // let n_crit = None;
+        // let depth = Some(4);
+        // let e = 3;
+        // let expansion_order = vec![e; depth.unwrap() as usize + 1];
+        // let prune_empty = true;
+        // let block_size = Some(64);
 
-        let mut fmm_fft = SingleNodeBuilder::new(false)
-            .tree(sources.data(), targets.data(), n_crit, depth, prune_empty)
-            .unwrap()
-            .parameters(
-                charges.data(),
-                &expansion_order,
-                Laplace3dKernel::new(),
-                GreenKernelEvalType::Value,
-                FftFieldTranslation::new(block_size),
-            )
-            .unwrap()
-            .build()
-            .unwrap();
+        // let mut fmm_fft = SingleNodeBuilder::new(false)
+        //     .tree(sources.data(), targets.data(), n_crit, depth, prune_empty)
+        //     .unwrap()
+        //     .parameters(
+        //         charges.data(),
+        //         &expansion_order,
+        //         Laplace3dKernel::new(),
+        //         GreenKernelEvalType::Value,
+        //         FftFieldTranslation::new(block_size),
+        //     )
+        //     .unwrap()
+        //     .build()
+        //     .unwrap();
 
-        group.bench_function(format!("M2L=FFT digits=3"), |b| {
-            b.iter(|| fmm_fft.evaluate())
-        });
+        // group.bench_function(format!("N=1e6, M2L=FFT digits=3"), |b| {
+        //     b.iter(|| fmm_fft.evaluate())
+        // });
 
-        group.bench_function(format!("M2L=FFT digits=3, M2L "), |b| {
-            b.iter(|| {
-                for level in 2..=fmm_fft.tree().target_tree().depth() {
-                    fmm_fft.m2l(level).unwrap();
-                }
-            })
-        });
+        // group.bench_function(format!("N=1e6, M2L=FFT digits=3, M2L "), |b| {
+        //     b.iter(|| {
+        //         for level in 2..=fmm_fft.tree().target_tree().depth() {
+        //             fmm_fft.m2l(level).unwrap();
+        //         }
+        //     })
+        // });
 
-        group.bench_function(format!("M2L=FFT digits=3, P2P "), |b| {
[…]
