+++ .cargo/config.toml
@@ -5,7 +5,7 @@ rustflags = ["-C", "target-cpu=native"]
 rustflags = ["-C", "target-feature=+fma,+sse2,+avx,+avx2"]
 
 [target.aarch64]
-rustflags = ["-C", "target-feature=+neon"]
+rrustflags = ["-C", "target-feature=+neon,+fp-armv8,+fma"]
 
 [env]
 OPENBLAS_NUM_THREADS = "1"

[…]
