+++ kifmm/src/fmm/field_translation/source_to_target/blas/multi_node.rs
@@ -22,7 +22,7 @@ use crate::{
         field::SourceToTargetTranslation,
         fmm::{DataAccessMulti, HomogenousKernel, MetadataAccess},
         tree::{MultiFmmTree, MultiTree, SingleTree},
-        types::FmmError,
+        types::{FmmError, NumberOfFlops},
     },
     tree::constants::NTRANSFER_VECTORS_KIFMM,
     BlasFieldTranslationSaRcmp, DataAccess, KiFmm, MultiNodeFmmTree,
@@ -43,7 +43,7 @@ where
     KiFmm<Scalar, Kernel, BlasFieldTranslationSaRcmp<Scalar>>:
         DataAccess<Scalar = Scalar, Kernel = Kernel>,
 {
-    fn m2l(&self, level: u64) -> Result<(), FmmError> {
+    fn m2l(&self, level: u64) -> Result<NumberOfFlops, FmmError> {
         match self.fmm_eval_type {
             FmmEvalType::Vector => {
                 if let Some(targets) = self.tree().target_tree().keys(level) {
@@ -278,7 +278,7 @@ where
                     }
                 }
 
-                Ok(())
+                Ok(0)
             }
             FmmEvalType::Matrix(_) => Err(FmmError::Unimplemented(
                 "M2L unimplemented for matrix input with BLAS field translations".to_string(),
