+++ kifmm/src/fftw/dft.rs
@@ -194,9 +194,7 @@ impl Dft for f64 {
     type Plan = Plan<Self, Plan64>;
 
     fn count_flops(plan: &Self::Plan, add: &mut f64, mul: &mut f64, fmas: &mut f64) {
-        unsafe {
-            ffi::fftw_flops(plan.plan.0, add, mul, fmas)
-        }
+        unsafe { ffi::fftw_flops(plan.plan.0, add, mul, fmas) }
     }
 
     fn plan_forward(
@@ -386,9 +384,9 @@ impl Dft for c32 {
     type Plan = Plan<Self, Plan32>;
 
     fn count_flops(plan: &Self::Plan, add: &mut f64, mul: &mut f64, fmas: &mut f64) {
-       unsafe {
-        ffi::fftwf_flops(plan.plan.0, add, mul, fmas);
-       }
+        unsafe {
+            ffi::fftwf_flops(plan.plan.0, add, mul, fmas);
+        }
     }
 
     fn plan_forward(
@@ -569,9 +567,9 @@ impl Dft for c64 {
     type Plan = Plan<Self, Plan64>;
 
     fn count_flops(plan: &Self::Plan, add: &mut f64, mul: &mut f64, fmas: &mut f64) {
-       unsafe {
-        ffi::fftw_flops(plan.plan.0, add, mul, fmas);
-       }
+        unsafe {
+            ffi::fftw_flops(plan.plan.0, add, mul, fmas);
+        }
     }
 
     fn plan_forward(
