+++ kifmm/benches/laplace_f64_m1.rs
@@ -1,268 +1,0 @@
-use std::time::Duration;
-
-use criterion::measurement::{Measurement, WallTime};
-use criterion::{criterion_group, criterion_main, Criterion};
-use green_kernels::traits::Kernel;
-use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
-use kifmm::fftw::array::AlignedAllocable;
-use kifmm::fmm::types::FmmSvdMode;
-use kifmm::fmm::types::{BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
-use kifmm::linalg::rsvd::MatrixRsvd;
-use kifmm::traits::fftw::Dft;
-use kifmm::traits::field::{SourceToTargetTranslation, TargetTranslation};
-use kifmm::traits::fmm::{DataAccess, Evaluate};
-use kifmm::traits::general::single_node::{AsComplex, Epsilon, Hadamard8x8};
-use kifmm::traits::tree::{SingleFmmTree, SingleTree};
-use kifmm::tree::helpers::points_fixture;
-use kifmm::KiFmm;
-use num::Float;
-use rand_distr::uniform::SampleUniform;
-use rlst::{rlst_dynamic_array2, MatrixSvd, RawAccess, RawAccessMut, RlstScalar};
-
-fn benchmark_fft_m2l<
-    T: RlstScalar<Real = T>
-        + SampleUniform
-        + Float
-        + Epsilon
-        + MatrixSvd
-        + AsComplex
-        + Dft<InputType = T, OutputType = <T as AsComplex>::ComplexType>
-        + Default,
-    M: Measurement,
->(
-    group: &mut criterion::BenchmarkGroup<'_, M>,
-    digits: usize,
-    n_points: usize,
-    e: usize,
-    block_size: Option<usize>,
-    depth: Option<u64>,
-) where
-    <T as RlstScalar>::Real: Epsilon + RlstScalar + Default,
-    <T as Dft>::Plan: Sync,
-    <T as AsComplex>::ComplexType: AlignedAllocable,
-    <T as AsComplex>::ComplexType: Hadamard8x8<Scalar = <T as AsComplex>::ComplexType>,
-    T: AlignedAllocable,
-{
-    // FFT based M2L for a vector of charges
-    // FMM parameters
-    let n_crit = None;
-    let expansion_order = vec![e; depth.unwrap() as usize + 1];
-    let prune_empty = true;
-    let sources = points_fixture::<T>(n_points, None, None, Some(0));
-    let targets = points_fixture::<T>(n_points, None, None, Some(1));
-    let tmp = vec![T::one(); n_points];
-    let mut charges = rlst_dynamic_array2!(T, [n_points, 1]);
-    charges.data_mut().copy_from_slice(&tmp);
-
-    let mut fmm_fft = SingleNodeBuilder::new(false)
-        .tree(sources.data(), targets.data(), n_crit, depth, prune_empty)
-        .unwrap()
-        .parameters(
-            charges.data(),
-            &expansion_order,
-            Laplace3dKernel::new(),
-            GreenKernelEvalType::Value,
-            FftFieldTranslation::<T>::new(block_size),
-        )
-        .unwrap()
-        .build()
-        .unwrap();
-
-    group.bench_function(
-        format!("M2L=FFT digits={} n_points={}", digits, n_points),
-        |b| b.iter(|| fmm_fft.evaluate()),
-    );
-
-    group.bench_function(
-        format!("M2L=FFT digits={} n_points={},  M2L ", digits, n_points),
-        |b| {
-            b.iter(|| {
-                for level in 2..=fmm_fft.tree().target_tree().depth() {
-                    fmm_fft.m2l(level).unwrap();
[…]
