### 2025-05-13 9b6cbd3
Add AMD config to yaml
