+++ scripts/src/bin/count_flops.rs
@@ -1,0 +1,262 @@
+use std::fs;
+use std::path::PathBuf;
+
+use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+use kifmm::fftw::array::AlignedAllocable;
+use kifmm::fmm::types::FmmSvdMode;
+use kifmm::fmm::types::{BlasFieldTranslationSaRcmp, FftFieldTranslation, SingleNodeBuilder};
+use kifmm::linalg::rsvd::MatrixRsvd;
+use kifmm::traits::fftw::Dft;
+use kifmm::traits::fmm::Evaluate;
+use kifmm::traits::general::single_node::{AsComplex, Epsilon, Hadamard8x8};
+use kifmm::traits::types::NumberOfFlops;
+use kifmm::tree::helpers::points_fixture;
+use num::Float;
+use rand::distributions::uniform::SampleUniform;
+use rlst::{rlst_dynamic_array2, MatrixSvd, RawAccess, RawAccessMut, RlstScalar};
+use serde_yaml::Value;
+
+fn count_flops_fft_m2l<
+    T: RlstScalar<Real = T>
+        + SampleUniform
+        + Float
+        + Epsilon
+        + MatrixSvd
+        + AsComplex
+        + Dft<InputType = T, OutputType = <T as AsComplex>::ComplexType>
+        + Default,
+>(
+    n_points: usize,
+    e: usize,
+    block_size: Option<usize>,
+    depth: Option<u64>,
+) -> NumberOfFlops
+where
+    <T as RlstScalar>::Real: Epsilon + RlstScalar + Default,
+    <T as Dft>::Plan: Sync,
+    <T as AsComplex>::ComplexType: AlignedAllocable,
+    <T as AsComplex>::ComplexType: Hadamard8x8<Scalar = <T as AsComplex>::ComplexType>,
+    T: AlignedAllocable,
+{
+    // FFT based M2L for a vector of charges
+    // FMM parameters
+    let n_crit = None;
+    let expansion_order = vec![e; depth.unwrap() as usize + 1];
+    let prune_empty = true;
+    let sources = points_fixture::<T>(n_points, None, None, Some(0));
+    let targets = points_fixture::<T>(n_points, None, None, Some(1));
+    let tmp = vec![T::one(); n_points];
+    let mut charges = rlst_dynamic_array2!(T, [n_points, 1]);
+    charges.data_mut().copy_from_slice(&tmp);
+
+    let mut fmm_fft = SingleNodeBuilder::new(false)
+        .tree(sources.data(), targets.data(), n_crit, depth, prune_empty)
+        .unwrap()
+        .parameters(
+            charges.data(),
+            &expansion_order,
+            Laplace3dKernel::new(),
+            GreenKernelEvalType::Value,
+            FftFieldTranslation::<T>::new(block_size),
+        )
+        .unwrap()
+        .build()
+        .unwrap();
+
+    fmm_fft.evaluate().unwrap();
+    fmm_fft.nflops
+}
+
+fn count_flops_blas_m2l<T: RlstScalar<Real = T> + Epsilon + MatrixRsvd + Float + SampleUniform>(
+    n_points: usize,
+    n_vecs: usize,
+    e: usize,
+    surface_diff: Option<usize>,
+    depth: Option<u64>,
+    svd_mode: FmmSvdMode,
+    svd_threshold: Option<T>,
+) -> NumberOfFlops
+where
+    <T as RlstScalar>::Real: Epsilon,
+{
+    let sources = points_fixture::<T>(n_points, None, None, Some(0));
+    let targets = points_fixture::<T>(n_points, None, None, Some(1));
+
+    let tmp = vec![T::one(); n_points * n_vecs];
[…]
