+++ arch-conf.yaml
@@ -1,0 +1,262 @@
+arch:
+  APPLE_M1:
+    fp32:
+      blas:
+        n_points:
+          1000000:
+            digits:
+              3:
+                order: 3
+                depth: 5
+                svd_threshold: 0.001
+                surface_diff: null
+                n_oversamples: 5
+              4:
+                order: 3
+                depth: 5
+                svd_threshold: null
+                surface_diff: 1
+                n_oversamples: 10
+          8000000:
+            digits:
+              3:
+                order: 3
+                depth: 6
+                svd_threshold: 1e-7
+                surface_diff: null
+                n_oversamples: 10
+
+      fft:
+        n_points:
+          1000000:
+            digits:
+              3:
+                order: 3
+                depth: 5
+                block_size: 128
+              4:
+                order: 4
+                depth: 5
+                block_size: 64
+          8000000:
+            digits:
+              3:
+                order: 3
+                depth: 6
+                block_size: 64
+
+    fp64:
+      fft:
+        n_points:
+          1000000:
+            digits:
+              6:
+                order: 6
+                block_size: 128
+                depth: 5
+              8:
+                order: 8
+                block_size: 32
+                depth: 4
+              10:
+                order: 10
+                block_size: 64
+                depth: 4
+
+          8000000:
+            digits:
+              4:
+                order: 4
+                block_size: 128
+                depth: 6
+              6:
+                order: 6
+                block_size: 16
+                depth: 5
+              8:
+                order: 8
+                block_size: 64
+                depth: 5
+              10:
+                order: 10
+                block_size: 32
+                depth: 5
+
+      blas:
+        n_points:
+          1000000:
+            digits:
+              6:
+                order: 5
+                depth: 5
+                surface_diff: 1
+                n_oversamples: 5
+                svd_threshold: 0.001
+              8:
+                order: 8
+                depth: 5
+                surface_diff: 2
+                n_oversamples: 10
+                svd_threshold: 0.001
+              10:
+                order: 9
+                depth: 4
+                surface_diff: 2
+                n_oversamples: 10
+                svd_threshold: 1e-7
+
+          8000000:
+            digits:
+              4:
+                order: 3
+                depth: 6
+                surface_diff: 2
+                n_oversamples: 10
+                svd_threshold: 0.001
+              6:
+                order: 5
+                depth: 5
+                surface_diff: 1
+                n_oversamples: 20
+                svd_threshold: 0.001
+              8:
+                order: 7
+                depth: 5
+                surface_diff: 1
+                n_oversamples: 20
+                svd_threshold: 0.00001
+              10:
+                order: 9
+                depth: 5
+                surface_diff: 2
+                n_oversamples: 20
+                svd_threshold: 1e-7
+  AMD_3290X:
+    fp32:
+      blas:
+        n_points:
+          1000000:
+            digits:
+              3:
+                order: 3
+                depth: 4
+                surface_diff: null
+                n_oversamples: 20
+                svd_threshold: 0.001
+              4:
+                order: 3
+                depth: 4
+                surface_diff: 2
+                n_oversamples: 5
+                svd_threshold: 1e-7
+          8000000:
+            digits:
+              3:
+                order: 3
+                depth: 5
+                surface_diff: null
+                n_oversamples: 5
+                svd_threshold: null
+              4:
+                order: 3
+                depth: 5
+                surface_diff: 2
+                n_oversamples: 10
+                svd_threshold: 0.001
+
+      fft:
+        n_points:
+          1000000:
+            digits:
+              3:
+                order: 3
+                depth: 4
+                block_size: 64
+              4:
+                order: 4
+                depth: 4
+                block_size: 32
+          8000000:
+            digits:
+              3:
+                order: 3
[…]
