### 2025-05-13 b30ab69
Add count storage script
