+++ scripts/Cargo.toml
@@ -29,6 +29,11 @@ path = "src/bin/count_flops.rs"
 required-features = []
 
 [[bin]]
+name = "count_storage"
+path = "src/bin/count_storage.rs"
+required-features = []
+
+[[bin]]
 name = "grid_search_single_node"
 path = "src/bin/grid_search_single_node.rs"
 required-features = []
