### 2025-05-13 8d6ad10
Add displacement map time calculator
