+++ kifmm/src/bindings.rs
@@ -326,7 +326,7 @@ pub mod constructors {
     /// - `n_charges`: The length of the charges buffer.
     /// - `prune_empty`: A boolean flag indicating whether to prune empty leaf nodes, and their ancestors.
     /// - `n_crit`: Threshold for tree refinement, if set to 0 ignored. Otherwise will refine until threshold is
-    ///    reached based on a uniform particle distribution.
+    ///  reached based on a uniform particle distribution.
     /// - `depth`: The maximum depth of the tree, max supported depth is 16.
     /// - `singular_value_threshold`: Threshold for singular values used in compressing M2L matrices.
     /// - `surface_diff`: Set to 0 to disable, otherwise uses surface_diff+equivalent_surface_expansion_order = check_surface_expansion_order
@@ -427,7 +427,7 @@ pub mod constructors {
     /// - `n_charges`: The length of the charges buffer.
     /// - `prune_empty`: A boolean flag indicating whether to prune empty leaf nodes, and their ancestors.
     /// - `n_crit`: Threshold for tree refinement, if set to 0 ignored. Otherwise will refine until threshold is
-    ///    reached based on a uniform particle distribution.
+    ///  reached based on a uniform particle distribution.
     /// - `depth`: The maximum depth of the tree, max supported depth is 16.
     /// - `singular_value_threshold`: Threshold for singular values used in compressing M2L matrices.
     /// - `surface_diff`: Set to 0 to disable, otherwise uses surface_diff+equivalent_surface_expansion_order = check_surface_expansion_order
@@ -528,7 +528,7 @@ pub mod constructors {
     /// - `n_charges`: The length of the charges buffer.
     /// - `prune_empty`: A boolean flag indicating whether to prune empty leaf nodes, and their ancestors.
     /// - `n_crit`: Threshold for tree refinement, if set to 0 ignored. Otherwise will refine until threshold is
-    ///    reached based on a uniform particle distribution.
+    ///  reached based on a uniform particle distribution.
     /// - `depth`: The maximum depth of the tree, max supported depth is 16.
     /// - `singular_value_threshold`: Threshold for singular values used in compressing M2L matrices.
     /// - `surface_diff`: Set to 0 to disable, otherwise uses surface_diff+equivalent_surface_expansion_order = check_surface_expansion_order
@@ -645,7 +645,7 @@ pub mod constructors {
     /// - `n_charges`: The length of the charges buffer.
     /// - `prune_empty`: A boolean flag indicating whether to prune empty leaf nodes, and their ancestors.
     /// - `n_crit`: Threshold for tree refinement, if set to 0 ignored. Otherwise will refine until threshold is
-    ///    reached based on a uniform particle distribution.
+    ///  reached based on a uniform particle distribution.
     /// - `depth`: The maximum depth of the tree, max supported depth is 16.
     /// - `singular_value_threshold`: Threshold for singular values used in compressing M2L matrices.
     /// - `surface_diff`: Set to 0 to disable, otherwise uses surface_diff+equivalent_surface_expansion_order = check_surface_expansion_order
@@ -761,7 +761,7 @@ pub mod constructors {
     /// - `n_charges`: The length of the charges buffer.
     /// - `prune_empty`: A boolean flag indicating whether to prune empty leaf nodes, and their ancestors.
     /// - `n_crit`: Threshold for tree refinement, if set to 0 ignored. Otherwise will refine until threshold is
-    ///    reached based on a uniform particle distribution.
+    ///  reached based on a uniform particle distribution.
     /// - `depth`: The maximum depth of the tree, max supported depth is 16.
     /// - `block_size`: Parameter size controls cache utilisation in field translation, set to 0 to use default.
     ///
[…]
