### 2025-05-14 bce2dab
BUILD: Fix Python Build (#144)
* Make setuptools dependency explicit for mayavi compatibility
