### 2025-05-14 81ece30
Break up setup timing
