### 2025-05-15 bempp/kifmm#147 @jeremylt
Untested Code In Docs
This issue is for the  JOSS review https://github.com/openjournals/joss-reviews/issues/7124

I see code snippets in the docs but I don't see a CI job for doctests (unless I missed it?). I recommend using documentation tests so you verify that any API changes don't result in bad documentation examples.

https://doc.rust-lang.org/rustdoc/write-documentation/documentation-tests.html

In that vein, I'd also use more idiomatic error handling in those examples (with `?` over `.unwrap()`) in those snippets. The linked Rust documentation talks about how to do that with doctests.
