### 2025-05-15 bempp/kifmm#148 @jeremylt
Undocumented Code
This issue is for the  JOSS review https://github.com/openjournals/joss-reviews/issues/7124

It looks like not every function has docstrings. Is there a convention recorded somewhere about which function's do not receive docstrings?
