### 2025-05-16 0c28115
Start working on GPU kernel evals
