+++ kifmm/src/fmm/field_translation/metadata/single_node.rs
@@ -8,7 +8,7 @@ use green_kernels::{
     helmholtz_3d::Helmholtz3dKernel, laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait,
     types::GreenKernelEvalType,
 };
-use itertools::Itertools;
+use itertools::{izip, Itertools};
 use num::{Float, Zero};
 use rayon::prelude::*;
 use rlst::{
@@ -36,12 +36,14 @@ use crate::{
     traits::{
         fftw::{Dft, DftType},
         field::{
-            FieldTranslation as FieldTranslationTrait, SourceToTargetTranslationMetadata,
-            SourceTranslationMetadata, TargetTranslationMetadata,
+            FieldTranslation as FieldTranslationTrait, LeafTranslationMetadata,
+            SourceToTargetTranslationMetadata, SourceTranslationMetadata,
+            TargetTranslationMetadata,
         },
         fmm::{DataAccess, HomogenousKernel, Metadata, MetadataAccess},
         general::single_node::{AsComplex, Epsilon},
         tree::{Domain as DomainTrait, FmmTreeNode, SingleFmmTree, SingleTree},
+        types::FmmError,
     },
     tree::{
         constants::{
@@ -463,6 +465,109 @@ where
         self.uc2e_inv_2 = uc2e_inv_2;
         self.source = source;
         self.source_vec = source_vec;
+    }
+}
+
+impl<Scalar, Kernel, FieldTranslation> LeafTranslationMetadata
+    for KiFmm<Scalar, Kernel, FieldTranslation>
+where
+    Scalar: RlstScalar,
+    Kernel: KernelTrait<T = Scalar> + HomogenousKernel,
+    FieldTranslation: FieldTranslationTrait,
+{
+    fn leaf(&mut self) -> Result<(), FmmError> {
+        let Some(leaves) = self.tree.target_tree().all_leaves() else {
+            return Err(FmmError::Failed(
+                "Leaf setup failed, no leaves found in target tree".to_string(),
+            ));
+        };
+
+        // As a precomputation, set buffers to collect potentials for each leaf
+        // and their associated source/target coordinates
+        let mut all_source_coordinates_reordered_counts = Vec::new();
+        let mut all_charges_reordered_counts = Vec::new();
+        let mut all_target_coordinates_reordered =
+            vec![Scalar::default(); self.tree.target_tree().n_coordinates_tot().unwrap()];
+        let mut all_sources_coordinates_reordered = Vec::new();
+        let mut all_charges_reordered = Vec::new();
+
+        let all_target_coordinates = self.tree.target_tree().all_coordinates().unwrap();
+        let all_source_coordinates = self.tree.source_tree().all_coordinates().unwrap();
+
+        // Re-allocate the target coordinate data based on Morton ordering of leaves
+        let target_coordinates_displacements = self
+            .charge_index_pointer_targets
+            .iter()
+            .scan(0usize, |acc, &pointer| {
+                let tmp = pointer.1 - pointer.0; // Add current count
+                *acc += tmp;
+                Some(tmp)
+            })
+            .collect_vec();
+
+        for (leaf, index_pointer, &displacement) in izip!(
+            leaves,
+            &self.charge_index_pointer_targets,
+            target_coordinates_displacements.iter()
+        ) {
+            let target_coordinates_row_major =
+                &all_target_coordinates[index_pointer.0 * self.dim..index_pointer.1 * self.dim];
+            let count = (index_pointer.1 - index_pointer.0) * self.dim;
+
+            if count > 0 {
+                let target_coordinates_reordered = unsafe {
+                    let ptr = all_target_coordinates_reordered
+                        .as_mut_ptr()
+                        .add(displacement)
[…]
