+++ kifmm/src/p2p_kernel.rs
@@ -1,0 +1,10 @@
+//! Kernels for AMD,NVIDIA and CPU architectures
+pub mod cpu;
+
+#[cfg(feature = "cuda")]
+pub mod cuda;
+
+#[cfg(feature = "amd")]
+pub mod hip;
+
+
