+++ kifmm/src/fmm/data_access/single_node.rs
@@ -41,6 +41,7 @@ where
     }
 
     fn n_coeffs_equivalent_surface(&self, level: u64) -> usize {
+        println!("HERE level {:?} ncoeffs {:?}", level, &self.n_coeffs_equivalent_surface);
         self.n_coeffs_equivalent_surface[self.expansion_index(level)]
     }
 
