+++ kifmm/examples/mpi_test_fmm.rs
@@ -2,29 +2,161 @@
 
 #[cfg(feature = "mpi")]
 fn main() {
-    // use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
-    // use kifmm::{
-    //     fmm::types::MultiNodeBuilder,
-    //     traits::{
-    //         fmm::{DataAccessMulti, EvaluateMulti},
-    //         tree::{MultiFmmTree, MultiTree},
-    //     },
-    //     tree::{helpers::points_fixture, types::SortKind},
-    //     BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, SingleNodeBuilder,
-    // };
+    use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
+    use itertools::izip;
+    use kifmm::{
+        fmm::types::MultiNodeBuilder,
+        traits::{
+            fmm::{DataAccessMulti, EvaluateMulti},
+            tree::{MultiFmmTree, MultiTree},
+        },
+        tree::{helpers::points_fixture, types::SortKind},
+        BlasFieldTranslationSaRcmp, Evaluate, FftFieldTranslation, SingleNodeBuilder,
+    };
 
-    // use rand::{rngs::StdRng, Rng, SeedableRng};
-    // use rayon::ThreadPoolBuilder;
+    use rand::{rngs::StdRng, Rng, SeedableRng};
+    use rayon::ThreadPoolBuilder;
 
-    // use mpi::{
-    //     datatype::PartitionMut,
-    //     traits::{Communicator, Root},
-    // };
-    // use rlst::{rlst_dynamic_array1, RawAccess, RawAccessMut};
+    use mpi::{
+        datatype::PartitionMut,
+        traits::{Communicator, Root},
+    };
+    use rlst::{rlst_dynamic_array1, RawAccess, RawAccessMut};
 
-    // let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
-    // let world = universe.world();
-    // let comm = world.duplicate();
+    fn test_fmm(
+        name: String,
+        expansion_order: &[usize],
+        world: &mpi::topology::SimpleCommunicator,
+    ) {
+        let comm = world.duplicate();
+
+        // Tree parameters
+        let prune_empty = true;
+        let n_points = 10000;
+        let local_depth = 3;
+        let global_depth = 2;
+        let sort_kind = SortKind::Samplesort { n_samples: 100 };
+
+        // Fmm Parameters
+
+        let kernel = Laplace3dKernel::<f32>::new();
+        // let source_to_target =
+        //     BlasFieldTranslationSaRcmp::<f32>::new(None, None, kifmm::FmmSvdMode::Deterministic);
+        let source_to_target = FftFieldTranslation::<f32>::new(None);
+
+        // Generate some random test data local to each process
+        let points = points_fixture::<f32>(n_points, None, None, None);
+        let charges = vec![1f32; n_points];
+
+        let mut fmm = MultiNodeBuilder::new(false)
+            .tree(
+                &comm,
+                points.data(),
+                points.data(),
+                local_depth,
+                global_depth,
+                prune_empty,
+                sort_kind,
+            )
+            .unwrap()
+            .parameters(
+                &charges,
+                &expansion_order,
+                kernel,
+                GreenKernelEvalType::Value,
+                source_to_target,
+            )
+            .unwrap()
+            .build()
+            .unwrap();
+
+        // Perform upward and downward passes
+        fmm.evaluate().unwrap();
+
+        // Gather all coordinates for the test
[…]
