+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -762,6 +762,7 @@ where
     Scalar: RlstScalar + Default + MatrixRsvd + Equivalence + Float,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
 {
+    // TODO: The displacements by level must also be parallelised
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
 
@@ -1804,6 +1805,7 @@ where
         + Float,
     <Scalar as RlstScalar>::Real: RlstScalar + Default + Equivalence + Float,
 {
+    // TODO: the displacements by level must also be parallelised
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
         let start_level =
@@ -1862,238 +1864,264 @@ where
     }
 
     fn source_to_target(&mut self) {
-
         // TODO
 
         // // Compute the field translation operators
 
-        // // Pick a point in the middle of the domain
-        // let two = Scalar::real(2.0);
-        // let midway = self
-        //     .tree
-        //     .domain()
-        //     .side_length()
-        //     .iter()
-        //     .map(|d| *d / two)
-        //     .collect_vec();
+        // Pick a point in the middle of the domain
+        let two = Scalar::real(2.0);
+        let midway = self
+            .tree
+            .domain()
+            .side_length()
+            .iter()
+            .map(|d| *d / two)
+            .collect_vec();
 
-        // let point = midway
-        //     .iter()
-        //     .zip(self.tree.domain().origin())
-        //     .map(|(m, o)| *m + *o)
-        //     .collect_vec();
+        let point = midway
+            .iter()
+            .zip(self.tree.domain().origin())
+            .map(|(m, o)| *m + *o)
+            .collect_vec();
 
-        // let point = [point[0], point[1], point[2]];
+        let point = [point[0], point[1], point[2]];
 
-        // // Encode point in centre of domain and compute halo of parent, and their resp. children
-        // let key = MortonKey::from_point(&point, self.tree.source_tree().domain(), 3);
-        // let siblings = key.siblings();
-        // let parent = key.parent();
-        // let halo = parent.neighbors();
-        // let halo_children = halo.iter().map(|h| h.children()).collect_vec();
+        // Encode point in centre of domain and compute halo of parent, and their resp. children
+        let key = MortonKey::from_point(&point, self.tree.source_tree().domain(), 3);
+        let siblings = key.siblings();
+        let parent = key.parent();
+        let halo = parent.neighbors();
+        let halo_children = halo.iter().map(|h| h.children()).collect_vec();
 
-        // let equivalent_surface_order = self.equivalent_surface_order;
+        let iterator = if self.variable_expansion_order {
+            self.equivalent_surface_order
+                .iter()
+                .skip(2)
+                .cloned()
+                .collect_vec()
+        } else {
+            self.equivalent_surface_order.clone()
+        };
 
-        // // The child boxes in the halo of the sibling set
-        // let mut sources = vec![];
-        // // The sibling set
-        // let mut targets = vec![];
-        // // The transfer vectors corresponding to source->target translations
-        // let mut transfer_vectors = vec![];
-        // // Green's function evaluations for each source, target pair interaction
-        // let mut kernel_data_vec = vec![];
+        for &equivalent_surface_order in &iterator {
+            // The child boxes in the halo of the sibling set
+            let mut sources = vec![];
+            // The sibling set
+            let mut targets = vec![];
+            // The transfer vectors corresponding to source->target translations
+            let mut transfer_vectors = vec![];
+            // Green's function evaluations for each source, target pair interaction
+            let mut kernel_data_vec = vec![];
 
[…]
