+++ todo.md
@@ -1,0 +1,3 @@
+1. Go through todos and understand what's missing.
+2. source to target setup for fft-m2l, no need to parallelise
+3. helmholtz afterwards should be a copy and paste job

[…]
