+++ kifmm/examples/mpi_test_fmm.rs
@@ -40,9 +40,9 @@ fn main() {
         // Fmm Parameters
 
         let kernel = Laplace3dKernel::<f32>::new();
-        // let source_to_target =
-        //     BlasFieldTranslationSaRcmp::<f32>::new(None, None, kifmm::FmmSvdMode::Deterministic);
-        let source_to_target = FftFieldTranslation::<f32>::new(None);
+        let source_to_target =
+            BlasFieldTranslationSaRcmp::<f32>::new(None, None, kifmm::FmmSvdMode::Deterministic);
+        // let source_to_target = FftFieldTranslation::<f32>::new(None);
 
         // Generate some random test data local to each process
         let points = points_fixture::<f32>(n_points, None, None, None);
@@ -144,19 +144,19 @@ fn main() {
     let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Funneled).unwrap();
     let world = universe.world();
 
-    // let expansion_order = [4];
-    // test_fmm(
-    //     "single expansion order all levels".to_string(),
-    //     &expansion_order,
-    //     &world,
-    // );
-
-    let expansion_order = [4, 4, 5, 4, 5, 4];
+    let expansion_order = [4];
     test_fmm(
-        "variable expansion order per level".to_string(),
+        "single expansion order all levels".to_string(),
         &expansion_order,
         &world,
     );
+
+    // let expansion_order = [4, 4, 5, 4, 5, 4];
+    // test_fmm(
+    //     "variable expansion order per level".to_string(),
+    //     &expansion_order,
+    //     &world,
+    // );
 
     // // Tree parameters
     // let prune_empty = true;
