+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -756,13 +756,13 @@ where
     }
 }
 
+
 impl<Scalar> SourceToTargetTranslationMetadata
     for KiFmmMulti<Scalar, Laplace3dKernel<Scalar>, BlasFieldTranslationSaRcmp<Scalar>>
 where
     Scalar: RlstScalar + Default + MatrixRsvd + Equivalence + Float,
     <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
 {
-    // TODO: The displacements by level must also be parallelised
     fn displacements(&mut self, start_level: Option<u64>) {
         let mut displacements = Vec::new();
 
@@ -829,970 +829,1323 @@ where
     }
 
     fn source_to_target(&mut self) {
-        let size = self.communicator.size();
-        let rank = self.communicator.rank();
-        let total_depth = self.tree.source_tree().total_depth();
+        // Compute unique M2L interactions at level 3, shallowest level which contains them all
+        // Compute interaction matrices between source and unique targets, defined by unique transfer vectors
+
+        let equivalent_surface_order = self.equivalent_surface_order[0];
+        let check_surface_order = self.check_surface_order[0];
+        let n_coeffs_equivalent_surface = self.n_coeffs_equivalent_surface[0]; //cols
+        let n_coeffs_check_surface = self.n_coeffs_check_surface[0]; // rows
+
+        let mut se2tc_fat = rlst_dynamic_array2!(
+            Scalar,
+            [
+                n_coeffs_check_surface,
+                n_coeffs_equivalent_surface * NTRANSFER_VECTORS_KIFMM
+            ]
+        );
+        let mut se2tc_thin = rlst_dynamic_array2!(
+            Scalar,
+            [
+                n_coeffs_check_surface * NTRANSFER_VECTORS_KIFMM,
+                n_coeffs_equivalent_surface
+            ]
+        );
         let alpha = Scalar::real(ALPHA_INNER);
 
-        // Distribute SVD by level
-        // Number of pre-computations
-        // Distribute the SVD by level
-        let iterator = if self.variable_expansion_order {
-            (2..=total_depth)
-                .zip(self.equivalent_surface_order.iter().skip(2).cloned())
-                .zip(self.check_surface_order.iter().skip(2).cloned())
-                .collect_vec()
-        } else {
-            (2..=total_depth)
-                .zip(vec![
-                    *self.equivalent_surface_order.last().unwrap();
-                    (total_depth - 1) as usize
-                ])
-                .zip(vec![
-                    *self.check_surface_order.last().unwrap();
-                    (total_depth - 1) as usize
-                ])
-                .collect_vec()
-        };
+        for (i, t) in self.source_to_target.transfer_vectors.iter().enumerate() {
+            let source_equivalent_surface =
+                t.source
+                    .surface_grid(equivalent_surface_order, self.tree.domain(), alpha);
 
-        let n_precomputations;
-        if self.variable_expansion_order {
-            n_precomputations = iterator.len() as i32;
-        } else {
-            n_precomputations = 1;
+            let target_check_surface =
+                t.target
+                    .surface_grid(check_surface_order, self.tree.domain(), alpha);
+
+            let mut tmp_gram = rlst_dynamic_array2!(
+                Scalar,
+                [n_coeffs_check_surface, n_coeffs_equivalent_surface]
+            );
+
+            self.kernel.assemble_st(
+                GreenKernelEvalType::Value,
+                &target_check_surface,
+                &source_equivalent_surface,
+                tmp_gram.data_mut(),
+            );
+
+            let mut block = se2tc_fat.r_mut().into_subview(
+                [0, i * n_coeffs_equivalent_surface],
[…]
