### 2025-05-22 135baa7
Add parallelisation of setup with blas rcmp
