+++ kifmm/src/fmm/builder/multi_node.rs
@@ -306,32 +306,36 @@ where
             let total_depth = self.tree.as_ref().unwrap().source_tree.total_depth;
             let variable_expansion_order = self.variable_expansion_order.unwrap();
 
-            let global_equivalent_surface_order =
-                equivalent_surface_order[0..=(global_depth as usize)].to_vec();
-            let global_check_surface_order =
-                check_surface_order[0..=(global_depth as usize)].to_vec();
-            let global_n_coeffs_equivalent_surface =
-                n_coeffs_equivalent_surface[0..=(global_depth as usize)].to_vec();
-            let global_n_coeffs_check_surface =
-                n_coeffs_check_surface[0..=(global_depth as usize)].to_vec();
+            // let global_equivalent_surface_order =
+            //     equivalent_surface_order[0..=(global_depth as usize)].to_vec();
+            // let global_check_surface_order =
+            //     check_surface_order[0..=(global_depth as usize)].to_vec();
+            // let global_n_coeffs_equivalent_surface =
+            //     n_coeffs_equivalent_surface[0..=(global_depth as usize)].to_vec();
+            // let global_n_coeffs_check_surface =
+            //     n_coeffs_check_surface[0..=(global_depth as usize)].to_vec();
 
-            let local_equivalent_surface_order =
-                equivalent_surface_order[(global_depth as usize)..=(total_depth as usize)].to_vec();
-            let local_check_surface_order =
-                check_surface_order[(global_depth as usize)..=(total_depth as usize)].to_vec();
-            let local_n_coeffs_equivalent_surface = n_coeffs_equivalent_surface
-                [(global_depth as usize)..=(total_depth as usize)]
-                .to_vec();
-            let local_n_coeffs_check_surface =
-                n_coeffs_check_surface[(global_depth as usize)..=(total_depth as usize)].to_vec();
+            // let local_equivalent_surface_order =
+            //     equivalent_surface_order[(global_depth as usize)..=(total_depth as usize)].to_vec();
+            // let local_check_surface_order =
+            //     check_surface_order[(global_depth as usize)..=(total_depth as usize)].to_vec();
+            // let local_n_coeffs_equivalent_surface = n_coeffs_equivalent_surface
+            //     [(global_depth as usize)..=(total_depth as usize)]
+            //     .to_vec();
+            // let local_n_coeffs_check_surface =
+            //     n_coeffs_check_surface[(global_depth as usize)..=(total_depth as usize)].to_vec();
 
             let global_fmm: KiFmm<Scalar, Kernel, FieldTranslation> = KiFmm {
                 isa: self.isa.unwrap(),
-                equivalent_surface_order: global_equivalent_surface_order,
-                check_surface_order: global_check_surface_order,
+                // equivalent_surface_order: global_equivalent_surface_order,
+                // check_surface_order: global_check_surface_order,
+                equivalent_surface_order: equivalent_surface_order.to_vec(),
+                check_surface_order: check_surface_order.to_vec(),
                 variable_expansion_order,
-                n_coeffs_equivalent_surface: global_n_coeffs_equivalent_surface,
-                n_coeffs_check_surface: global_n_coeffs_check_surface,
+                // n_coeffs_equivalent_surface: global_n_coeffs_equivalent_surface,
+                // n_coeffs_check_surface: global_n_coeffs_check_surface,
+                n_coeffs_equivalent_surface: n_coeffs_equivalent_surface.to_vec(),
+                n_coeffs_check_surface: n_coeffs_check_surface.to_vec(),
                 fmm_eval_type,
                 kernel_eval_type,
                 kernel: kernel.clone(),
