+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -1,3 +1,5 @@
+use bytemuck::cast_slice;
+use bytemuck::Pod;
 use green_kernels::laplace_3d::Laplace3dKernel;
 use green_kernels::traits::Kernel as KernelTrait;
 use green_kernels::types::GreenKernelEvalType;
@@ -6,9 +8,11 @@ use itertools::Itertools;
 use mpi::datatype::Partition;
 use mpi::datatype::PartitionMut;
 use mpi::datatype::Partitioned;
+use mpi::topology::SimpleCommunicator;
 use mpi::traits::{Communicator, CommunicatorCollectives, Equivalence};
 use mpi::{Count, Rank};
 use num::{Float, Zero};
+use pulp::Scalar;
 use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 use rlst::{
     empty_array, rlst_array_from_slice2, rlst_dynamic_array2, rlst_dynamic_array3, Array,
@@ -19,6 +23,13 @@ use std::collections::{HashMap, HashSet};
 use std::sync::{Mutex, RwLock};
 
 use crate::fmm::field_translation::source_to_target::transfer_vector::compute_transfer_vectors_at_level;
+use crate::fmm::helpers::multi_node::deserialise_nested_vec;
+use crate::fmm::helpers::multi_node::deserialise_vec;
+use crate::fmm::helpers::multi_node::deserialise_vec_blasmetadata_sarcmp;
+use crate::fmm::helpers::multi_node::serialise_blas_metadata_sarcmp;
+use crate::fmm::helpers::multi_node::serialise_nested_vec;
+use crate::fmm::helpers::multi_node::serialise_vec;
+use crate::fmm::helpers::multi_node::serialise_vec_blasmetadata_sarcmp;
 use crate::fmm::helpers::multi_node::{
     coordinate_index_pointer_multi_node, leaf_expansion_pointers_multi_node,
     leaf_surfaces_multi_node, level_expansion_pointers_multi_node, level_index_pointer_multi_node,
@@ -205,11 +216,11 @@ where
         for (mat_1, mat_2) in izip!(local_uc2e_inv_1.iter(), local_uc2e_inv_2.iter()) {
             let mat_1_size = mat_1.shape()[0] * mat_1.shape()[1];
             local_msg_uc2e_inv_1[curr1..(curr1 + mat_1_size)].copy_from_slice(mat_1.data());
-            curr1 = mat_1_size;
+            curr1 += mat_1_size;
 
             let mat_2_size = mat_1.shape()[0] * mat_1.shape()[1];
             local_msg_uc2e_inv_2[curr2..(curr2 + mat_2_size)].copy_from_slice(mat_2.data());
-            curr2 = mat_1_size;
+            curr2 += mat_1_size;
         }
 
         // Setup buffers to receive data
@@ -533,11 +544,11 @@ where
         for (mat_1, mat_2) in izip!(local_dc2e_inv_1.iter(), local_dc2e_inv_2.iter()) {
             let mat_1_size = mat_1.shape()[0] * mat_1.shape()[1];
             local_msg_dc2e_inv_1[curr1..(curr1 + mat_1_size)].copy_from_slice(mat_1.data());
-            curr1 = mat_1_size;
+            curr1 += mat_1_size;
 
             let mat_2_size = mat_1.shape()[0] * mat_1.shape()[1];
             local_msg_dc2e_inv_2[curr2..(curr2 + mat_2_size)].copy_from_slice(mat_2.data());
-            curr2 = mat_1_size;
+            curr2 += mat_1_size;
         }
 
         // Setup buffers to receive data
@@ -756,7 +767,6 @@ where
     }
 }
 
-
 impl<Scalar> SourceToTargetTranslationMetadata
[…]

[…]
