### 2025-05-22 5321260
Clean up source/traget metadata parallelisation
