+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -1,18 +1,13 @@
-use bytemuck::cast_slice;
-use bytemuck::Pod;
-use green_kernels::laplace_3d::Laplace3dKernel;
-use green_kernels::traits::Kernel as KernelTrait;
-use green_kernels::types::GreenKernelEvalType;
-use itertools::izip;
-use itertools::Itertools;
-use mpi::datatype::Partition;
-use mpi::datatype::PartitionMut;
-use mpi::datatype::Partitioned;
-use mpi::topology::SimpleCommunicator;
-use mpi::traits::{Communicator, CommunicatorCollectives, Equivalence};
-use mpi::{Count, Rank};
+use green_kernels::{
+    laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait, types::GreenKernelEvalType,
+};
+use itertools::{izip, Itertools};
+use mpi::{
+    datatype::{Partition, PartitionMut, Partitioned},
+    traits::{Communicator, CommunicatorCollectives, Equivalence},
+    Count, Rank,
+};
 use num::{Float, Zero};
-use pulp::Scalar;
 use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 use rlst::{
     empty_array, rlst_array_from_slice2, rlst_dynamic_array2, rlst_dynamic_array3, Array,
@@ -22,52 +17,47 @@ use rlst::{
 use std::collections::{HashMap, HashSet};
 use std::sync::{Mutex, RwLock};
 
-use crate::fmm::field_translation::source_to_target::transfer_vector::compute_transfer_vectors_at_level;
-use crate::fmm::helpers::multi_node::all_gather_v_serialised;
-use crate::fmm::helpers::multi_node::deserialise_nested_vec;
-use crate::fmm::helpers::multi_node::deserialise_vec;
-use crate::fmm::helpers::multi_node::deserialise_vec_blas_metadata_sarcmp;
-use crate::fmm::helpers::multi_node::deserialise_vec_fft_metadata;
-use crate::fmm::helpers::multi_node::serialise_blas_metadata_sarcmp;
-use crate::fmm::helpers::multi_node::serialise_nested_vec;
-use crate::fmm::helpers::multi_node::serialise_vec;
-use crate::fmm::helpers::multi_node::serialise_vec_blas_metadata_sarcmp;
-use crate::fmm::helpers::multi_node::serialise_vec_fft_metadata;
-use crate::fmm::helpers::multi_node::{
-    coordinate_index_pointer_multi_node, leaf_expansion_pointers_multi_node,
-    leaf_surfaces_multi_node, level_expansion_pointers_multi_node, level_index_pointer_multi_node,
-    potential_pointers_multi_node,
-};
-use crate::fmm::helpers::single_node::ncoeffs_kifmm;
-use crate::fmm::helpers::single_node::{flip3, homogenous_kernel_scale, optionally_time};
-use crate::fmm::types::{BlasMetadataSaRcmp, FftMetadata};
-use crate::linalg::pinv::pinv;
-use crate::traits::fftw::{Dft, DftType};
-use crate::traits::fmm::{DataAccess, DataAccessMulti, Metadata, MetadataAccess};
-use crate::traits::general::{
-    multi_node::GhostExchange,
-    single_node::{AsComplex, Epsilon},
-};
-use crate::traits::tree::{Domain, FmmTreeNode, MultiFmmTree, MultiTree};
-use crate::traits::types::{CommunicationTime, CommunicationType};
-use crate::tree::constants::{NHALO, NSIBLINGS, NSIBLINGS_SQUARED, NTRANSFER_VECTORS_KIFMM};
[…]
