### 2025-05-22 c1969c6
Start adding setup code for helmholtz
