+++ kifmm/src/fmm/field_translation/metadata/multi_node.rs
@@ -1,5 +1,5 @@
 use green_kernels::{
-    laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait, types::GreenKernelEvalType,
+    helmholtz_3d::Helmholtz3dKernel, laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait, types::GreenKernelEvalType
 };
 use itertools::{izip, Itertools};
 use mpi::{
@@ -96,6 +96,224 @@ pub(crate) fn calculate_precomputation_load(
         None
     }
 }
+
+
+impl<Scalar, FieldTranslation> SourceTranslationMetadata
+    for KiFmmMulti<Scalar, Helmholtz3dKernel<Scalar>, FieldTranslation>
+where
+    Scalar: RlstScalar<Complex = Scalar> + Default + Epsilon + MatrixSvd + Equivalence + Float,
+    <Scalar as RlstScalar>::Real: Default + Sync + Send + Default + Equivalence + Float,
+    FieldTranslation: FieldTranslationTrait + Send + Sync + Default,
+    Self: DataAccessMulti + MetadataAccess,
+{
+    fn source(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+        let size = self.communicator.size();
+        let rank = self.communicator.rank();
+        let total_depth = self.tree.source_tree().total_depth();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain();
+
+        // Parallelise matrix inversion by level
+        let mut uc2e_inv_1_r = Vec::new();
+        let mut uc2e_inv_2_r = Vec::new();
+
+        // Calculate inverse upward check to equivalent matrices on each level
+        let iterator = if self.equivalent_surface_order.len() > 1 {
+            (0..=total_depth)
+                .zip(self.equivalent_surface_order.iter().cloned())
+                .zip(self.check_surface_order.iter().cloned())
+                .collect_vec()
+        } else {
+            let eq_value = *self.equivalent_surface_order.last().unwrap();
+            let ch_value = *self.check_surface_order.last().unwrap();
+            let eq_vec = vec![eq_value; total_depth as usize];
+            let ch_vec = vec![ch_value; total_depth as usize];
+
+            (0..total_depth)
+                .zip(eq_vec.into_iter())
+                .zip(ch_vec.into_iter())
+                .collect_vec()
+        };
+
+        // Distribute pseudo-inverse calculation by level
+        // Number of pre-computations
+        let n_precomputations = iterator.len() as i32;
+
+        let (load_counts, load_displacement) =
+            calculate_precomputation_load(n_precomputations, size).unwrap();
+
+        // Compute mandated local portion
+        let local_load_count = load_counts[rank as usize];
+        let local_load_displacement = load_displacement[rank as usize];
+
+        let iterator_r = &iterator[(local_load_displacement as usize)
+            ..((local_load_displacement + local_load_count) as usize)];
+
+        for &((level, equivalent_surface_order), check_surface_order) in iterator_r.iter() {
+            let curr = root.first_child_at_level(level).unwrap();
+            let upward_equivalent_surface =
+                curr.surface_grid(equivalent_surface_order, domain, alpha_inner);
+            let upward_check_surface = curr.surface_grid(check_surface_order, domain, alpha_outer);
+
+            let n_cols = ncoeffs_kifmm(equivalent_surface_order);
+            let n_rows = ncoeffs_kifmm(check_surface_order);
+
[…]
