+++ kifmm/src/tree/morton.rs
@@ -574,6 +574,15 @@ where
         Self::new(&self.anchor, 1 + self.morton)
     }
 
+    /// Return first child at a given level greater than or equal to the level of a morton key
+    pub fn first_child_at_level(&self, level: u64) -> Option<Self> {
+        if level >= self.level() {
+            Some(Self::new(&self.anchor, level - self.level() + self.morton))
+        } else {
+            None
+        }
+    }
+
     /// Return the first child of a Morton Key on the deepest level.
     pub fn finest_first_child(&self) -> Self {
         Self::new(&self.anchor, DEEPEST_LEVEL - self.level() + self.morton)
