### 2025-05-22 #7124 @diehlpk
@skailasa please have a look at the two issues above
