### 2025-05-22 bempp/kifmm#147 @diehlpk
@skailasa  please have a look
