### 2025-05-22 bempp/kifmm#148 @diehlpk
@skailasa please have a look
