+++ kifmm/src/fmm/field_translation/target/mod.rs
@@ -1,3 +1,4 @@
+//! # Local to local expansion field translation translations
 #[cfg(feature = "mpi")]
 mod multi_node;
 mod single_node;
