+++ kifmm/src/fmm/field_translation/metadata/multi_node/target/laplace.rs
@@ -1,0 +1,248 @@
+use green_kernels::{
+    laplace_3d::Laplace3dKernel, traits::Kernel as KernelTrait, types::GreenKernelEvalType,
+};
+use itertools::Itertools;
+use mpi::traits::{Communicator, Equivalence};
+use num::Float;
+use rlst::{
+    empty_array, rlst_dynamic_array2, MatrixSvd, MultIntoResize, RawAccess, RawAccessMut,
+    RlstScalar, Shape,
+};
+
+use crate::{
+    fmm::{
+        helpers::{
+            multi_node::{
+                all_gather_v_serialised, calculate_precomputation_load, deserialise_nested_array,
+                serialise_nested_array,
+            },
+            single_node::{homogenous_kernel_scale, ncoeffs_kifmm},
+        },
+        types::KiFmmMulti,
+    },
+    linalg::pinv::pinv,
+    traits::{
+        field::{FieldTranslation as FieldTranslationTrait, TargetTranslationMetadata},
+        fmm::{DataAccessMulti, MetadataAccess},
+        general::single_node::Epsilon,
+        tree::{FmmTreeNode, MultiFmmTree, MultiTree},
+    },
+    tree::{
+        constants::{ALPHA_INNER, ALPHA_OUTER},
+        types::MortonKey,
+    },
+};
+
+impl<Scalar, FieldTranslation> TargetTranslationMetadata
+    for KiFmmMulti<Scalar, Laplace3dKernel<Scalar>, FieldTranslation>
+where
+    Scalar: RlstScalar + Default + Epsilon + MatrixSvd + Equivalence + Float,
+    <Scalar as RlstScalar>::Real: Default + Equivalence + Float,
+    FieldTranslation: FieldTranslationTrait + Send + Sync,
+    Self: MetadataAccess,
+{
+    fn target(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+        let size = self.communicator.size();
+        let rank = self.communicator.rank();
+        let total_depth = self.tree.source_tree().total_depth();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain();
+
+        // Parallelise the matrix inversion over MPI ranks
+
+        // Local buffers
+        let mut dc2e_inv_1_r = Vec::new();
+        let mut dc2e_inv_2_r = Vec::new();
+
+        let iterator = if self.variable_expansion_order {
+            self.equivalent_surface_order
+                .iter()
+                .cloned()
+                .zip(self.check_surface_order.iter().cloned())
+                .collect_vec()
+        } else {
+            vec![*self.equivalent_surface_order.last().unwrap(); (total_depth - 1) as usize]
+                .iter()
+                .cloned()
+                .zip(vec![
+                    *self.check_surface_order.last().unwrap();
+                    (total_depth - 1) as usize
+                ])
+                .collect_vec()
+        };
+
+        // Distribute pseudo-inverse calculation, if have variable expansion order by level
+        // Number of pre-computations
+        let n_precomputations = if self.variable_expansion_order {
+            self.equivalent_surface_order.len() as i32
+        } else {
+            1
+        };
+
+        let (load_counts, load_displacement) =
+            calculate_precomputation_load(n_precomputations, size).unwrap();
+
+        // Compute mandated local portion
+        let local_load_count = load_counts[rank as usize];
+        let local_load_displacement = load_displacement[rank as usize];
+
+        let iterator_r = &iterator[(local_load_displacement as usize)
[…]
