+++ kifmm/src/fmm/field_translation/metadata/multi_node/target/helmholtz.rs
@@ -1,0 +1,248 @@
+use green_kernels::{
+    helmholtz_3d::Helmholtz3dKernel, traits::Kernel as KernelTrait, types::GreenKernelEvalType,
+};
+use itertools::Itertools;
+use mpi::traits::{Communicator, Equivalence};
+use num::Float;
+use rlst::{
+    empty_array, rlst_dynamic_array2, MatrixSvd, MultIntoResize, RawAccess, RawAccessMut,
+    RlstScalar, Shape,
+};
+
+use crate::{
+    fmm::{
+        helpers::{
+            multi_node::{
+                all_gather_v_serialised, calculate_precomputation_load, deserialise_nested_array,
+                serialise_nested_array,
+            },
+            single_node::ncoeffs_kifmm,
+        },
+        types::KiFmmMulti,
+    },
+    linalg::pinv::pinv,
+    traits::{
+        field::{FieldTranslation as FieldTranslationTrait, TargetTranslationMetadata},
+        fmm::{DataAccessMulti, MetadataAccess},
+        general::single_node::Epsilon,
+        tree::{FmmTreeNode, MultiFmmTree, MultiTree},
+    },
+    tree::{
+        constants::{ALPHA_INNER, ALPHA_OUTER},
+        types::MortonKey,
+    },
+};
+
+impl<Scalar, FieldTranslation> TargetTranslationMetadata
+    for KiFmmMulti<Scalar, Helmholtz3dKernel<Scalar>, FieldTranslation>
+where
+    Scalar: RlstScalar<Complex = Scalar> + Default + Epsilon + MatrixSvd + Equivalence + Float,
+    <Scalar as RlstScalar>::Real: Default + Sync + Send + Default + Equivalence + Float,
+    FieldTranslation: FieldTranslationTrait + Send + Sync + Default,
+    Self: DataAccessMulti + MetadataAccess,
+{
+    fn target(&mut self) {
+        let root = MortonKey::<Scalar::Real>::root();
+        let size = self.communicator.size();
+        let rank = self.communicator.rank();
+        let total_depth = self.tree.source_tree().total_depth();
+
+        // Cast surface parameters
+        let alpha_outer = Scalar::from(ALPHA_OUTER).unwrap().re();
+        let alpha_inner = Scalar::from(ALPHA_INNER).unwrap().re();
+        let domain = self.tree.domain();
+
+        // Parallelise matrix inversion by level
+        let mut dc2e_inv_1_r = Vec::new();
+        let mut dc2e_inv_2_r = Vec::new();
+
+        // Calculate inverse upward check to equivalent matrices on each level
+        let iterator = if self.equivalent_surface_order.len() > 1 {
+            (0..=total_depth)
+                .zip(self.equivalent_surface_order.iter().cloned())
+                .zip(self.check_surface_order.iter().cloned())
+                .collect_vec()
+        } else {
+            let eq_value = *self.equivalent_surface_order.last().unwrap();
+            let ch_value = *self.check_surface_order.last().unwrap();
+            let eq_vec = vec![eq_value; (total_depth + 1) as usize];
+            let ch_vec = vec![ch_value; (total_depth + 1) as usize];
+
+            (0..=total_depth).zip(eq_vec).zip(ch_vec).collect_vec()
+        };
+
+        // Distribute pseudo-inverse calculation by level
+        // Number of pre-computations
+        let n_precomputations = iterator.len() as i32;
+
+        let (load_counts, load_displacement) =
+            calculate_precomputation_load(n_precomputations, size).unwrap();
+
+        // Compute mandated local portion
+        let local_load_count = load_counts[rank as usize];
+        let local_load_displacement = load_displacement[rank as usize];
+
[…]
