### 2025-05-23 bempp/kifmm#147 @skailasa
Hi @jeremylt , the docs are built as a part of the CI with this script

https://github.com/bempp/kifmm/blob/main/.github/workflows/docs.yml

The build is currently failing, for unrelated reasons, and I can debug this- but its not because of API changes, have to verify with a team member who helps with CI.
