### 2025-05-23 d052806
More clippy, add test for new function
