### 2025-05-23 d75e86d
Add tests for serialisation
