+++ kifmm/src/fmm/helpers/multi_node.rs
@@ -500,3 +500,630 @@ pub(crate) fn calculate_precomputation_load(
         None
     }
 }
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use num::Complex;
+    use rand::distributions::{Distribution, Uniform};
+    use rlst::{dense::tools::RandScalar, DefaultIteratorMut};
+
+    fn test_array_real<T: RlstScalar + PartialOrd + RandScalar>()
+    where
+        T: rand::distributions::uniform::SampleUniform,
+    {
+        let m = 5;
+        let n = 4;
+        let mut expected = rlst_dynamic_array2!(T, [m, n]);
+        let mut rng = rand::thread_rng();
+        let between = Uniform::try_from(T::from(0.).unwrap()..T::from(1.0).unwrap()).unwrap();
+        expected
+            .iter_mut()
+            .for_each(|e| *e = between.sample(&mut rng));
+
+        let serialised = serialise_array(&expected);
+        let found = deserialise_array::<T>(&serialised).0;
+
+        (found.data().iter())
+            .zip(expected.data().iter())
+            .for_each(|(&e, &f)| {
+                assert!(RlstScalar::abs(e - f) <= T::from(1e-6).unwrap().re());
+            });
+    }
+
+    fn test_array_complex<T: RlstScalar<Real = T> + RandScalar + PartialOrd>()
+    where
+        T: rand::distributions::uniform::SampleUniform,
+        Complex<T>: RlstScalar,
+    {
+        let m = 5;
+        let n = 4;
+        let mut expected = rlst_dynamic_array2!(Complex<T>, [m, n]);
+
+        let mut rng = rand::thread_rng();
+        let between = Uniform::try_from(T::from(0.).unwrap()..T::from(1.0).unwrap()).unwrap();
+
+        expected.iter_mut().for_each(|e| {
+            *e = Complex {
+                re: between.sample(&mut rng),
+                im: between.sample(&mut rng),
+            }
+        });
+
+        let serialised = serialise_array(&expected);
+        let found = deserialise_array::<Complex<T>>(&serialised).0;
+
+        (found.data().iter())
+            .zip(expected.data().iter())
+            .for_each(|(&e, &f)| {
+                let diff = RlstScalar::powf(
+                    T::from(e.re()).unwrap() - T::from(f.re()).unwrap(),
+                    T::from(2.0).unwrap(),
+                ) + RlstScalar::powf(
+                    T::from(e.im()).unwrap() - T::from(f.im()).unwrap(),
+                    T::from(2.0).unwrap(),
+                );
+                let err = RlstScalar::powf(diff, T::from(0.5).unwrap().re());
+                println!("HERE {:?} {:?}", e, f);
+                assert!(err <= T::from(1e-6).unwrap().re());
+            });
+    }
+
+    fn test_array_real_empty<T: RlstScalar + PartialOrd + RandScalar>()
+    where
+        T: rand::distributions::uniform::SampleUniform,
+    {
+        let expected = rlst_dynamic_array2!(T, [0, 0]);
+        let serialised = serialise_array(&expected);
+        let found = deserialise_array::<T>(&serialised).0;
+        assert!(found.shape()[0] == 0);
+        assert!(found.shape()[1] == 0);
+        assert!(found.data().len() == 0);
+    }
+
+    fn test_array_complex_empty<T: RlstScalar + PartialOrd + RandScalar>()
+    where
+        T: rand::distributions::uniform::SampleUniform,
+        Complex<T>: RlstScalar,
+    {
+        let expected = rlst_dynamic_array2!(Complex<T>, [0, 0]);
+        let serialised = serialise_array(&expected);
[…]
