### 2025-05-26 c263aac
Re-add all mpi tests
