+++ kifmm/examples/mpi_test_fmm.rs
@@ -2,82 +2,73 @@
 
 #[cfg(feature = "mpi")]
 fn main() {
-    use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
-    use itertools::izip;
+    use green_kernels::{
+        helmholtz_3d::Helmholtz3dKernel, laplace_3d::Laplace3dKernel, traits::Kernel,
+        types::GreenKernelEvalType,
+    };
+    use itertools::{izip, Itertools};
     use kifmm::{
         fmm::types::MultiNodeBuilder,
         traits::{
-            fmm::{DataAccessMulti, EvaluateMulti},
+            fmm::EvaluateMulti,
             tree::{MultiFmmTree, MultiTree},
         },
         tree::{helpers::points_fixture, types::SortKind},
-        FftFieldTranslation,
+        FftFieldTranslation, MultiNodeFmmTree,
     };
 
     use mpi::{
         datatype::PartitionMut,
-        traits::{Communicator, Root},
+        topology::SimpleCommunicator,
+        traits::{Communicator, Equivalence, Root},
     };
-    use rlst::RawAccess;
+    use num::Float;
+    use rlst::{RawAccess, RlstScalar};
 
-    fn test_fmm(
+    fn _test_multi_node_helmholtz_fmm_helper<
+        T: RlstScalar<Complex = T> + Float + Default + Equivalence,
+    >(
         name: String,
-        expansion_order: &[usize],
-        world: &mpi::topology::SimpleCommunicator,
-    ) {
-        let comm = world.duplicate();
-
-        // Tree parameters
-        let prune_empty = true;
-        let n_points = 10000;
-        let local_depth = 3;
-        let global_depth = 2;
-        let sort_kind = SortKind::Samplesort { n_samples: 100 };
-
-        // Fmm Parameters
-
-        let kernel = Laplace3dKernel::<f32>::new();
-        // let source_to_target =
-        //     BlasFieldTranslationSaRcmp::<f32>::new(None, None, kifmm::FmmSvdMode::Deterministic);
-        let source_to_target = FftFieldTranslation::<f32>::new(None);
-
-        // Generate some random test data local to each process
-        let points = points_fixture::<f32>(n_points, None, None, None);
-        let charges = vec![1f32; n_points];
-
-        let mut fmm = MultiNodeBuilder::new(false)
-            .tree(
-                &comm,
-                points.data(),
-                points.data(),
-                local_depth,
-                global_depth,
-                prune_empty,
-                sort_kind,
-            )
-            .unwrap()
-            .parameters(
-                &charges,
-                &expansion_order,
-                kernel,
-                GreenKernelEvalType::Value,
-                source_to_target,
-            )
-            .unwrap()
-            .build()
-            .unwrap();
-
-        // Perform upward and downward passes
+        mut fmm: Box<
+            dyn EvaluateMulti<
+                Scalar = T,
+                Kernel = Helmholtz3dKernel<T>,
+                Tree = MultiNodeFmmTree<T::Real, SimpleCommunicator>,
+            >,
+        >,
+        eval_type: GreenKernelEvalType,
+        threshold: T::Real,
+    ) where
+        <T as RlstScalar>::Real: Equivalence,
+    {
+        // Run the FMM
         fmm.evaluate().unwrap();
 
-        // Gather all coordinates for the test
-        let root_process = fmm.communicator().process_at_rank(0);
-        let n_coords = fmm.tree().source_tree().coordinates.len() as i32;
-        let mut all_coordinates = vec![0f32; 3 * n_points * world.size() as usize];
+        // TODO add test for matrix input
+        let _eval_size = match eval_type {
[…]
