+++ kifmm/examples/mpi_test_trees.rs
@@ -2,207 +2,207 @@
 
 #[cfg(feature = "mpi")]
 fn main() {
-    // use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
-    // use kifmm::{
-    //     fmm::types::MultiNodeBuilder,
-    //     traits::{
-    //         fmm::{DataAccess, DataAccessMulti},
-    //         general::multi_node::GhostExchange,
-    //         tree::{MultiFmmTree, MultiTree},
-    //     },
-    //     tree::{
-    //         helpers::points_fixture,
-    //         types::{MortonKey, SortKind},
-    //     },
-    //     FftFieldTranslation, SingleNodeBuilder,
-    // };
+    use green_kernels::{laplace_3d::Laplace3dKernel, types::GreenKernelEvalType};
+    use kifmm::{
+        fmm::types::MultiNodeBuilder,
+        traits::{
+            fmm::{DataAccess, DataAccessMulti},
+            general::multi_node::GhostExchange,
+            tree::{MultiFmmTree, MultiTree},
+        },
+        tree::{
+            helpers::points_fixture,
+            types::{MortonKey, SortKind},
+        },
+        FftFieldTranslation, SingleNodeBuilder,
+    };
 
-    // use rayon::ThreadPoolBuilder;
+    use rayon::ThreadPoolBuilder;
 
-    // use mpi::{
-    //     datatype::PartitionMut,
-    //     traits::{Communicator, Root},
-    // };
-    // use rlst::RawAccess;
+    use mpi::{
+        datatype::PartitionMut,
+        traits::{Communicator, Root},
+    };
+    use rlst::RawAccess;
 
-    // let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
-    // let world = universe.world();
-    // let comm = world.duplicate();
+    let (universe, _threading) = mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
+    let world = universe.world();
+    let comm = world.duplicate();
 
-    // // Tree parameters
-    // let prune_empty = false;
-    // let n_points = 100000;
-    // let local_depth = 3;
-    // let global_depth = 2;
-    // let sort_kind = SortKind::Samplesort { n_samples: 100 };
+    // Tree parameters
+    let prune_empty = false;
+    let n_points = 100000;
+    let local_depth = 3;
+    let global_depth = 2;
+    let sort_kind = SortKind::Samplesort { n_samples: 100 };
 
-    // // Fmm Parameters
-    // let expansion_order = 4;
-    // let kernel = Laplace3dKernel::<f32>::new();
-    // let source_to_target = FftFieldTranslation::<f32>::new(None);
+    // Fmm Parameters
+    let expansion_order = 4;
+    let kernel = Laplace3dKernel::<f32>::new();
+    let source_to_target = FftFieldTranslation::<f32>::new(None);
 
-    // // Generate some random test data local to each process
-    // let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
-    // let charges = vec![1f32; n_points];
+    // Generate some random test data local to each process
+    let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
+    let charges = vec![1f32; n_points];
 
-    // ThreadPoolBuilder::new()
-    //     .num_threads(1)
-    //     .build_global()
-    //     .unwrap();
+    ThreadPoolBuilder::new()
+        .num_threads(1)
+        .build_global()
+        .unwrap();
 
-    // let mut fmm = MultiNodeBuilder::new(false)
-    //     .tree(
-    //         &comm,
-    //         points.data(),
-    //         points.data(),
[…]
