### 2025-05-26 e34736f
Start working on helmholtz tests
