+++ kifmm/examples/mpi_test_upward_pass.rs
@@ -2,7 +2,9 @@
 
 #[cfg(feature = "mpi")]
 fn main() {
+    use green_kernels::helmholtz_3d::Helmholtz3dKernel;
     use green_kernels::{laplace_3d::Laplace3dKernel, traits::Kernel, types::GreenKernelEvalType};
+    use itertools::izip;
     use kifmm::traits::general::multi_node::GhostExchange;
     use kifmm::{
         fmm::types::MultiNodeBuilder,
@@ -13,59 +15,35 @@ fn main() {
         tree::{constants::ALPHA_INNER, helpers::points_fixture, types::SortKind},
         Evaluate, FftFieldTranslation,
     };
-    use kifmm::{BlasFieldTranslationSaRcmp, FmmSvdMode};
+    use kifmm::{BlasFieldTranslationSaRcmp, FmmSvdMode, MultiNodeFmmTree};
+    use mpi::collective::SystemOperation;
+    use mpi::topology::SimpleCommunicator;
+    use mpi::traits::{CommunicatorCollectives, Equivalence};
     use mpi::{
         datatype::PartitionMut,
         traits::{Communicator, Root},
     };
-    use rlst::{RawAccess, RlstScalar};
+    use mpi_sys::RSMPI_SUM;
+    use num::{Float, One};
+    use num_complex::Complex32;
+    use rlst::{c32, RawAccess, RlstScalar};
 
-    fn test_upward_pass(
+    fn test_multi_node_helmholtz_upward_pass_helper<T: RlstScalar<Complex = T> + Float + Default>(
         name: String,
-        expansion_order: &[usize],
-        world: &mpi::topology::SimpleCommunicator,
-    ) {
-        let comm = world.duplicate();
-
-        // Tree parameters
-        let prune_empty = true;
-        let n_points = 10000;
-        let local_depth = 3;
-        let global_depth = 2;
-        let sort_kind = SortKind::Samplesort { n_samples: 100 };
-
-        // Fmm Parameters
-
-        let kernel = Laplace3dKernel::<f32>::new();
-        let source_to_target =
-            BlasFieldTranslationSaRcmp::<f32>::new(None, None, FmmSvdMode::Deterministic);
-        // let source_to_target = FftFieldTranslation::<f32>::new(None);
-
-        // Generate some random test data local to each process
-        let points = points_fixture::<f32>(n_points, None, None, None);
-        let charges = vec![1f32; n_points];
-
-        let mut fmm = MultiNodeBuilder::new(false)
-            .tree(
-                &comm,
-                points.data(),
-                points.data(),
-                local_depth,
-                global_depth,
-                prune_empty,
-                sort_kind,
-            )
-            .unwrap()
-            .parameters(
-                &charges,
-                &expansion_order,
-                kernel,
-                GreenKernelEvalType::Value,
-                source_to_target,
-            )
-            .unwrap()
-            .build()
-            .unwrap();
+        mut fmm: Box<
+            dyn EvaluateMulti<
+                Scalar = T,
+                Kernel = Helmholtz3dKernel<T>,
+                Tree = MultiNodeFmmTree<T::Real, SimpleCommunicator>,
+            >,
+        >,
+        eval_type: GreenKernelEvalType,
+        threshold: T::Real,
+    ) where
+        <T as RlstScalar>::Real: Equivalence,
+    {
+        // Test the global part of the upward pass
+        let alpha_inner = T::from(ALPHA_INNER).unwrap().re();
 
         // Perform partial upward pass on each rank
         fmm.evaluate_leaf_sources().unwrap();
@@ -74,21 +52,21 @@ fn main() {
         // Test at roots of local trees for result of partial upward passes
[…]
