### 2025-05-26 b52009a
Remove float dependence on potentiall complex scalar, to get upward pass test working
