+++ kifmm/examples/mpi_test_upward_pass.rs
@@ -24,11 +24,11 @@ fn main() {
         traits::{Communicator, Root},
     };
     use mpi_sys::RSMPI_SUM;
-    use num::{Float, One};
     use num_complex::Complex32;
+    use num::{Float, One};
     use rlst::{c32, RawAccess, RlstScalar};
 
-    fn test_multi_node_helmholtz_upward_pass_helper<T: RlstScalar<Complex = T> + Float + Default>(
+    fn test_multi_node_helmholtz_upward_pass_helper<T>(
         name: String,
         mut fmm: Box<
             dyn EvaluateMulti<
@@ -40,7 +40,8 @@ fn main() {
         eval_type: GreenKernelEvalType,
         threshold: T::Real,
     ) where
-        <T as RlstScalar>::Real: Equivalence,
+        T: RlstScalar<Complex = T> + Equivalence,
+        <T as RlstScalar>::Real: Equivalence + Float,
     {
         // Test the global part of the upward pass
         let alpha_inner = T::from(ALPHA_INNER).unwrap().re();
@@ -106,7 +107,9 @@ fn main() {
             T::real(0.0) // or handle division-by-zero error
         };
 
-        if fmm.rank() == 0 {
+        // if fmm.rank() == 0 {
+        // }
+
             println!(
                 "Local Upward Pass rank {:?} l2 err {:?} \n expected {:?} found {:?}",
                 fmm.rank(),
@@ -114,8 +117,6 @@ fn main() {
                 expected,
                 found
             );
-        }
-
         assert!(l2_error <= threshold);
         println!("...test_upward_pass_global_tree {} passed", name);
 
@@ -297,92 +298,92 @@ fn main() {
 
     }
 
-    // Test Laplace FMM
-    // N.B global tree refined to depth 3 to ensure that the global upward pass is also being run
-    {
-        let (universe, _threading) =
-            mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
-        let world = universe.world();
-        let comm = world.duplicate();
+    // // Test Laplace FMM
+    // // N.B global tree refined to depth 3 to ensure that the global upward pass is also being run
+    // {
+    //     let (universe, _threading) =
+    //         mpi::initialize_with_threading(mpi::Threading::Single).unwrap();
+    //     let world = universe.world();
+    //     let comm = world.duplicate();
 
-        let n_points = 10000;
-        let charges = vec![1f32; n_points];
-        let eval_type = GreenKernelEvalType::Value;
-        let source_to_target = FftFieldTranslation::new(None);
-        let sources = points_fixture(n_points, None, None, None);
-        let local_depth = 3;
-        let global_depth = 3;
-        let prune_empty = true;
+    //     let n_points = 10000;
+    //     let charges = vec![1f32; n_points];
+    //     let eval_type = GreenKernelEvalType::Value;
+    //     let source_to_target = FftFieldTranslation::new(None);
+    //     let sources = points_fixture(n_points, None, None, None);
+    //     let local_depth = 3;
+    //     let global_depth = 3;
+    //     let prune_empty = true;
 
-        // Single expansion order
-        {
-            let expansion_order = [5];
+    //     // Single expansion order
+    //     {
+    //         let expansion_order = [5];
 
-            let fmm = MultiNodeBuilder::new(false)
-                .tree(
-                    &comm.duplicate(),
-                    sources.data(),
-                    sources.data(),
-                    local_depth,
-                    global_depth,
-                    prune_empty,
-                    SortKind::Samplesort { n_samples: 10 },
-                )
-                .unwrap()
-                .parameters(
-                    &charges,
-                    &expansion_order,
-                    Laplace3dKernel::new(),
-                    eval_type,
[…]
