+++ kifmm/src/fmm/field_translation/metadata/multi_node/source/laplace.rs
@@ -13,8 +13,8 @@ use crate::{
     fmm::{
         helpers::{
             multi_node::{
-                all_gather_v_serialised, calculate_precomputation_load, deserialise_nested_array,
-                serialise_nested_array,
+                all_gather_v_serialised, calculate_precomputation_load,
+                deserialise_nested_array_2x2, serialise_nested_array_2x2,
             },
             single_node::ncoeffs_kifmm,
         },
@@ -124,8 +124,8 @@ where
         }
 
         // Need to gather results at all ranks
-        let uc2e_inv_1_r_serialised = serialise_nested_array(&uc2e_inv_1_r);
-        let uc2e_inv_2_r_serialised = serialise_nested_array(&uc2e_inv_2_r);
+        let uc2e_inv_1_r_serialised = serialise_nested_array_2x2(&uc2e_inv_1_r);
+        let uc2e_inv_2_r_serialised = serialise_nested_array_2x2(&uc2e_inv_2_r);
 
         let uc2e_inv_1_serialised =
             all_gather_v_serialised(&uc2e_inv_1_r_serialised, &self.communicator);
@@ -133,16 +133,16 @@ where
             all_gather_v_serialised(&uc2e_inv_2_r_serialised, &self.communicator);
 
         // Reconstruct metadata
-        let (mut uc2e_inv_1, mut rest) = deserialise_nested_array(&uc2e_inv_1_serialised);
+        let (mut uc2e_inv_1, mut rest) = deserialise_nested_array_2x2(&uc2e_inv_1_serialised);
         while !rest.is_empty() {
-            let (mut t1, t2) = deserialise_nested_array::<Scalar>(rest);
+            let (mut t1, t2) = deserialise_nested_array_2x2::<Scalar>(rest);
             uc2e_inv_1.append(&mut t1);
             rest = t2;
         }
 
-        let (mut uc2e_inv_2, mut rest) = deserialise_nested_array(&uc2e_inv_2_serialised);
+        let (mut uc2e_inv_2, mut rest) = deserialise_nested_array_2x2(&uc2e_inv_2_serialised);
         while !rest.is_empty() {
-            let (mut t1, t2) = deserialise_nested_array::<Scalar>(rest);
+            let (mut t1, t2) = deserialise_nested_array_2x2::<Scalar>(rest);
             uc2e_inv_2.append(&mut t1);
             rest = t2;
         }

[…]
