### 2025-05-28 9eb8c97
Found IA bug in helmholtz
