+++ kifmm/src/fmm/field_translation/metadata/multi_node/source_to_target/helmholtz.rs
@@ -1,6 +1,6 @@
 use std::{
     collections::{HashMap, HashSet},
-    sync::RwLock,
+    sync::{Mutex, RwLock},
 };
 
 use green_kernels::{
@@ -37,7 +37,7 @@ use crate::{
         tree::{Domain, FmmTreeNode, MultiFmmTree, MultiTree},
     },
     tree::{
-        constants::{ALPHA_INNER, NHALO, NSIBLINGS, NSIBLINGS_SQUARED},
+        constants::{ALPHA_INNER, NHALO, NSIBLINGS, NSIBLINGS_SQUARED, NTRANSFER_VECTORS_KIFMM},
         helpers::find_corners,
         types::MortonKey,
     },
@@ -775,6 +775,9 @@ where
         let start_level =
             start_level.unwrap_or_else(|| std::cmp::max(2, self.tree.source_tree().global_depth()));
 
+        let tmp = HashSet::new();
+        let target_tree_keys_set = self.tree.target_tree().all_keys_set().unwrap_or(&tmp);
+
         for level in start_level..=self.tree.source_tree().total_depth() {
             let sources = self.tree.source_tree().keys(level).unwrap();
             let n_sources = sources.len();
@@ -783,9 +786,6 @@ where
 
             let result = vec![vec![sentinel; n_sources]; 316];
             let result = result.into_iter().map(RwLock::new).collect_vec();
-
-            let tmp = HashSet::new();
-            let target_tree_keys_set = self.tree.target_tree().all_keys_set().unwrap_or(&tmp);
 
             sources
                 .into_par_iter()
@@ -840,245 +840,409 @@ where
     }
 
     fn source_to_target(&mut self) {
-        let size = self.communicator.size();
-        let rank = self.communicator.rank();
+        // Compute interaction matrices between source and unique targets, defined by unique transfer vectors
         let alpha = Scalar::real(ALPHA_INNER);
-        let total_depth = self.tree.source_tree().total_depth();
+        let depth = self.tree.source_tree().total_depth();
+
+        let mut result = BlasFieldTranslationIa::<Scalar>::default();
 
         let iterator = if self.variable_expansion_order {
-            (2..=total_depth)
+            (2..=depth)
                 .zip(self.equivalent_surface_order.iter().skip(2).cloned())
                 .zip(self.check_surface_order.iter().skip(2).cloned())
                 .collect_vec()
         } else {
-            (2..=total_depth)
+            (2..=depth)
                 .zip(vec![
                     *self.equivalent_surface_order.last().unwrap();
-                    (total_depth - 1) as usize
+                    (depth - 1) as usize
                 ])
                 .zip(vec![
                     *self.check_surface_order.last().unwrap();
-                    (total_depth - 1) as usize
+                    (depth - 1) as usize
                 ])
                 .collect_vec()
         };
 
-        // First need to enumerate tasks
-        let mut tasks = Vec::new();
-        for &((level, equivalent_surface_order), check_surface_order) in iterator.iter() {
+        for ((level, equivalent_surface_order), check_surface_order) in iterator {
             let transfer_vectors =
                 compute_transfer_vectors_at_level::<Scalar::Real>(level).unwrap();
-            for &t in transfer_vectors.iter() {
-                tasks.push((level, equivalent_surface_order, check_surface_order, t));
+
+            let mut level_result = BlasMetadataIa::default();
+            let level_u = Mutex::new(Vec::new());
+            let level_vt = Mutex::new(Vec::new());
+            let level_cutoff_rank = Mutex::new(vec![0usize; NTRANSFER_VECTORS_KIFMM]);
+
+            let domain = self.tree.source_tree().domain();
+
[…]
