+++ kifmm/examples/mpi_test_fmm.rs
@@ -8,13 +8,11 @@ fn main() {
     };
     use itertools::{izip, Itertools};
     use kifmm::{
-        fmm::types::MultiNodeBuilder,
-        traits::{
+        fmm::types::MultiNodeBuilder, linalg::rsvd::MatrixRsvd, traits::{
             fmm::EvaluateMulti,
+            general::single_node::Epsilon,
             tree::{MultiFmmTree, MultiTree},
-        },
-        tree::{helpers::points_fixture, types::SortKind},
-        BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, FftFieldTranslation, MultiNodeFmmTree,
+        }, tree::{helpers::points_fixture, types::SortKind}, BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, DataAccess, DataAccessMulti, Evaluate, FftFieldTranslation, MultiNodeFmmTree, SingleNodeBuilder
     };
 
     use mpi::{
@@ -22,10 +20,12 @@ fn main() {
         topology::SimpleCommunicator,
         traits::{Communicator, Equivalence, Root},
     };
-    use num::{Float, One};
+    use num::{Float, One, Zero};
     use rlst::{c32, RawAccess, RlstScalar};
 
-    fn test_multi_node_helmholtz_fmm_helper<T: RlstScalar<Complex = T> + Default + Equivalence>(
+    fn test_multi_node_helmholtz_fmm_helper<
+        T: RlstScalar<Complex = T> + Epsilon + MatrixRsvd + Default + Equivalence,
+    >(
         name: String,
         mut fmm: Box<
             dyn EvaluateMulti<
@@ -37,7 +37,7 @@ fn main() {
         eval_type: GreenKernelEvalType,
         threshold: T::Real,
     ) where
-        <T as RlstScalar>::Real: Equivalence + Float,
+        <T as RlstScalar>::Real: Equivalence + Float + Epsilon,
     {
         // Run the FMM
         fmm.evaluate().unwrap();
@@ -107,16 +107,48 @@ fn main() {
                 &mut expected,
             );
 
+            let single_fmm = SingleNodeBuilder::new(false)
+                .tree(&all_coordinates, &all_coordinates, None, Some(6), true)
+                .unwrap()
+                .parameters(
+                    &all_charges,
+                    &[5],
+                    Helmholtz3dKernel::<T>::new(T::real(0.0)),
+                    GreenKernelEvalType::Value,
+                    BlasFieldTranslationIa::new(None, None, kifmm::FmmSvdMode::Deterministic),
+                )
+                .unwrap()
+                .build()
+                .unwrap();
+
+            // Test metadata
+
             let found = fmm.potentials().unwrap();
 
             let mut num = T::real(0.0);
             let mut den = T::real(0.0);
-            for (&expected, &found) in izip!(&expected, found) {
-                num += RlstScalar::powf(RlstScalar::abs(expected - found), T::real(2.0));
-                den += RlstScalar::powf(RlstScalar::abs(expected), T::real(2.0));
+
+            for (expected, &found) in izip!(expected, found) {
+                // squared error in complex difference
+                let diff_re = expected.re() - found.re();
+                let diff_im = expected.im() - found.im();
+                // println!("expected {:?} found {:?}", expected, found);
+                num += RlstScalar::powf(diff_re, T::real(2.0))
+                    + RlstScalar::powf(diff_im, T::real(2.0));
+
+                // squared magnitude of expected
+                den += RlstScalar::powf(expected.re(), T::real(2.0))
+                    + RlstScalar::powf(expected.im(), T::real(2.0));
             }
 
-            let l2_error = RlstScalar::powf(num / den, T::real(0.5));
[…]
