+++ kifmm/src/fmm/field_translation/metadata/single_node/source_to_target/helmholtz.rs
@@ -51,6 +51,8 @@ where
         let mut displacements = Vec::new();
         let start_level = start_level.unwrap_or(2).max(2);
 
+        println!("HERE {:?}", self.tree.target_tree().all_keys_set().unwrap().len());
+
         for level in start_level..=self.tree.source_tree().depth() {
             let sources = self.tree.source_tree().keys(level).unwrap();
             let n_sources = sources.len();
@@ -278,6 +280,8 @@ where
             result.metadata.push(level_result);
             result.transfer_vectors.push(transfer_vectors);
         }
+
+        // println!("BAR {:?}", result.metadata.last().unwrap().u[0].data());
 
         self.source_to_target = result;
     }

[…]
