### 2025-05-28 bempp/kifmm#147 @jeremylt
I understand that the docs are built as part of CI. But I don't see where the code snippits inside of the documentation are being tested, like the example code here: https://bempp.github.io/kifmm/kifmm/index.html
