### 2025-05-28 bempp/kifmm#148 @jeremylt
Not documenting internal functions is mildly hostile to new contributors, but its a somewhat common practice. Is this policy documented somewhere in the contributing guidelines?  As far as I can tell, everything marked `pub` has at least a single line of documentation.
