+++ kifmm/examples/mpi_test_fmm.rs
@@ -8,11 +8,15 @@ fn main() {
     };
     use itertools::{izip, Itertools};
     use kifmm::{
-        fmm::types::MultiNodeBuilder, linalg::rsvd::MatrixRsvd, traits::{
+        fmm::types::MultiNodeBuilder,
+        linalg::rsvd::MatrixRsvd,
+        traits::{
             fmm::EvaluateMulti,
             general::single_node::Epsilon,
             tree::{MultiFmmTree, MultiTree},
-        }, tree::{helpers::points_fixture, types::SortKind}, BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, DataAccess, DataAccessMulti, Evaluate, FftFieldTranslation, MultiNodeFmmTree, SingleNodeBuilder
+        },
+        tree::{helpers::points_fixture, types::SortKind},
+        BlasFieldTranslationIa, BlasFieldTranslationSaRcmp, FftFieldTranslation, MultiNodeFmmTree,
     };
 
     use mpi::{
@@ -20,7 +24,7 @@ fn main() {
         topology::SimpleCommunicator,
         traits::{Communicator, Equivalence, Root},
     };
-    use num::{Float, One, Zero};
+    use num::{Float, One};
     use rlst::{c32, RawAccess, RlstScalar};
 
     fn test_multi_node_helmholtz_fmm_helper<
@@ -102,24 +106,10 @@ fn main() {
             fmm.kernel().evaluate_st(
                 GreenKernelEvalType::Value,
                 &all_coordinates,
-                &target_coordinates_rank,
+                target_coordinates_rank,
                 &all_charges,
                 &mut expected,
             );
-
-            let single_fmm = SingleNodeBuilder::new(false)
-                .tree(&all_coordinates, &all_coordinates, None, Some(6), true)
-                .unwrap()
-                .parameters(
-                    &all_charges,
-                    &[5],
-                    Helmholtz3dKernel::<T>::new(T::real(0.0)),
-                    GreenKernelEvalType::Value,
-                    BlasFieldTranslationIa::new(None, None, kifmm::FmmSvdMode::Deterministic),
-                )
-                .unwrap()
-                .build()
-                .unwrap();
 
             // Test metadata
 
@@ -132,7 +122,6 @@ fn main() {
                 // squared error in complex difference
                 let diff_re = expected.re() - found.re();
                 let diff_im = expected.im() - found.im();
-                // println!("expected {:?} found {:?}", expected, found);
                 num += RlstScalar::powf(diff_re, T::real(2.0))
                     + RlstScalar::powf(diff_im, T::real(2.0));
 
@@ -148,9 +137,6 @@ fn main() {
                 T::real(0.0) // or handle division-by-zero error
             };
 
-            // let l2_error = RlstScalar::powf(num / den, T::real(0.5));
-
-            println!("L2 {:?}", l2_error);
             assert!(l2_error <= threshold);
             println!("...test_helmholtz_fmm_{} passed", name);
         } else {
@@ -241,7 +227,7 @@ fn main() {
             fmm.kernel().evaluate_st(
                 GreenKernelEvalType::Value,
                 &all_coordinates,
-                &target_coordinates_rank,
+                target_coordinates_rank,
                 &all_charges,
                 &mut expected,
             );
@@ -257,7 +243,6 @@ fn main() {
 
             let l2_error = RlstScalar::powf(num / den, T::real(0.5));
 
-            // println!("L2 {:?}", l2_error);
             assert!(l2_error <= threshold);
             println!("...test_laplace_fmm_{} passed", name);
         } else {
@@ -271,186 +256,186 @@ fn main() {
         }
     }
 
[…]
