+++ kifmm/src/sorting/samplesort.rs
@@ -8,7 +8,7 @@ use mpi::{
     traits::{Communicator, CommunicatorCollectives, Equivalence},
     Count,
 };
-use rand::{rngs::StdRng, thread_rng, Rng, SeedableRng};
+use rand::{thread_rng, Rng};
 
 /// A sample sort implementation, the number of samples must be at most
 /// the size of the local array, and greater than 0.
@@ -46,9 +46,7 @@ where
         // 1. Collect k samples from each process onto all other processes
         let mut received_samples = vec![T::default(); n_samples * (size as usize)];
 
-        // TODO remove seeded range when I merge
-        let mut rng = StdRng::seed_from_u64(42);
-        // let mut rng = thread_rng();
+        let mut rng = thread_rng();
         let sample_idxs: Vec<usize> = (0..n_samples)
             .map(|_| rng.gen_range(0..array.len()))
             .collect();
