+++ kifmm/benches/helmholtz.rs
@@ -1,356 +1,0 @@
-use std::{fs, path::PathBuf, time::Duration};
-
-use criterion::{
-    criterion_group, criterion_main,
-    measurement::{Measurement, WallTime},
-    Criterion,
-};
-
-use pulp::{c32, c64};
-use rand_distr::uniform::SampleUniform;
-use rlst::{rlst_dynamic_array2, MatrixSvd, RawAccess, RawAccessMut, RlstScalar};
-use serde_yaml::Value;
-
-use green_kernels::{helmholtz_3d::Helmholtz3dKernel, types::GreenKernelEvalType};
-
-use kifmm::{
-    fftw::array::AlignedAllocable,
-    fmm::types::{FftFieldTranslation, FmmSvdMode, SingleNodeBuilder},
-    linalg::rsvd::MatrixRsvd,
-    traits::{
-        fftw::Dft,
-        field::{SourceToTargetTranslation, TargetTranslation},
-        fmm::{DataAccess, Evaluate},
-        general::single_node::{AsComplex, Epsilon, Hadamard8x8},
-        tree::{SingleFmmTree, SingleTree},
-    },
-    tree::helpers::points_fixture,
-    BlasFieldTranslationIa,
-};
-
-fn benchmark_fft_m2l<
-    T: RlstScalar<Complex = T>
-        + Epsilon
-        + MatrixSvd
-        + AsComplex
-        + Dft<InputType = T, OutputType = <T as AsComplex>::ComplexType>
-        + Default,
-    M: Measurement,
->(
-    group: &mut criterion::BenchmarkGroup<'_, M>,
-    digits: usize,
-    n_points: usize,
-    wavenumber: T::Real,
-    e: usize,
-    block_size: Option<usize>,
-    depth: Option<u64>,
-) where
-    <T as RlstScalar>::Real: Epsilon + RlstScalar + Default + SampleUniform,
-    <T as Dft>::Plan: Sync,
-    <T as AsComplex>::ComplexType: AlignedAllocable,
-    <T as AsComplex>::ComplexType: Hadamard8x8<Scalar = <T as AsComplex>::ComplexType>,
-    T: AlignedAllocable,
-{
-    // FFT based M2L for a vector of charges
-    // FMM parameters
-    let n_crit = None;
-    let expansion_order = vec![e; depth.unwrap() as usize + 1];
-    let prune_empty = true;
-    let sources = points_fixture::<T::Real>(n_points, None, None, Some(0));
-    let targets = points_fixture::<T::Real>(n_points, None, None, Some(1));
-    let tmp = vec![T::one(); n_points];
-    let mut charges = rlst_dynamic_array2!(T, [n_points, 1]);
-    charges.data_mut().copy_from_slice(&tmp);
-
-    let mut fmm_fft = SingleNodeBuilder::new(false)
-        .tree(sources.data(), targets.data(), n_crit, depth, prune_empty)
-        .unwrap()
-        .parameters(
-            charges.data(),
-            &expansion_order,
-            Helmholtz3dKernel::new(wavenumber),
-            GreenKernelEvalType::Value,
-            FftFieldTranslation::<T>::new(block_size),
-        )
-        .unwrap()
-        .build()
-        .unwrap();
-
-    // group.bench_function(
-    //     format!("M2L=FFT digits={} n_points={}", digits, n_points),
-    //     |b| b.iter(|| fmm_fft.evaluate()),
-    // );
-
-    // group.bench_function(
-    //     format!("M2L=FFT digits={} n_points={},  M2L ", digits, n_points),
-    //     |b| {
-    //         b.iter(|| {
[…]
