+++ kifmm/Cargo.toml
@@ -37,11 +37,11 @@ bytemuck = "1.16.0"
 superslice = {version = "1.0.0", optional = true}
 rand_distr = "0.4.3"
 rlst = { git = "https://github.com/linalg-rs/rlst.git", default-features = false }
-serde_yaml = "0.9.34"
 
 [dev-dependencies]
 approx = "0.5"
 criterion = { version = "0.5.*", features = ["html_reports"]}
+serde_yaml = "0.9.34"
 
 [build-dependencies]
 cbindgen = "0.26.0"
