### 2025-05-28 bempp/kifmm#148 @skailasa
Hi @jeremylt , understood, this is largely due to limited capacity on my end to support full documentation. I can add something to the contributing guidelines about this as suggested. Would the be sufficient?
