### 2025-05-28 bempp/kifmm#148 @jeremylt
Yeah, as long as the guidelines/expectations are clear
