### 2025-05-28 bempp/kifmm#147 @jeremylt
So the code is checked for compilation errors but not correctness if I understand?
