### 2025-05-28 bempp/kifmm#147 @jeremylt
The demonstration snippit is not tested for correctness though. That's what this link shows how to do
https://doc.rust-lang.org/rustdoc/write-documentation/documentation-tests.html
