### 2025-05-28 bempp/kifmm#147 @jeremylt
I really don't understand how making sure that the example works as expected (with hidden lines if you like) somehow is going to obfuscate the example? Testing for correctness ensures that the example actually shows what you think it does through and codebase changes.
