### 2025-05-28 bempp/kifmm#147 @jeremylt
Ultimately its your choice and y'all can do whatever you like. I just treat all code that isn't itself tested as wrong until proven correct.
