### 2025-05-28 bempp/kifmm#147 @jeremylt
As far as I can tell

1) Those exact lines in the documentation are tested to make sure they compile and run without error but are not checked for correctness

2) Similar lines of code are tested for correctness in the test suite

You really do not seem to want to test the lines of example code in the docs for correctness. That is something I personally try to avoid as I've been bitten by that sort of thing in my work, but it is ultimately your decision. I won't reject the paper over this; I just think its a (mild) risk that doesn't need to exist.
