### 2025-05-28 #7124 @jeremylt
Overall looking good.

Question:  Matthew Scroggs has quite a bit of git history in the repo. Should he also be an author?

https://github.com/bempp/kifmm/blob/main/paper/paper.md
