### 2025-05-29 bempp/kifmm#148 @skailasa
I've added a note about this to the readme/contributing section.
