### 2025-05-30 a4d44db
Add better timing capabilities
