### 2025-06-02 659bb0d
DOC: Update readme to reflect new API (#159)
