### 2025-06-11 #7124 @rdfriese
Overall looks good to me!

I created an issue at [https://github.com/linalg-rs/rlst/issues/127](https://github.com/linalg-rs/rlst/issues/127), about having the option for cargo to build  openblas incase the user doesn't have(or want) a system install. I recognize that is a different repo, but I think I saw @skailasa as a contributor so it may be something to quickly add and then it could be used in kifmm (and green-kernels).

I do not believe we can simply add the openblas-src to kifmm dependency list because rlst would not be able to link to it in this case. (they are being built in parallel, and I dont think the dependencies get updated LD_LIBRARY_PATHS pointing to their peers output directories)
