+++ paper/paper.md
@@ -71,7 +71,7 @@ Our principle contributions with `kifmm-rs` can therefore be summarised as:
 - _Optimisations for ARM and x86 targets_. ARM targets are increasingly common in high-performance computing systems, with portability enabled by Rust's LLVM based compiler.
 - _Competitive shared and distributed memory performance_. With state of the art M2L performance [@Kailasa2024], as well as a communication-optimal distributed memory implementation inspired by [@Ibeid2016].
 - _A C API_, using Rust's C ABI compatibility allowing for the construction of bindings into other languages, with full Python bindings for non-specialist users.
-- _A moderate frequency Helmholtz FMM_. Helmholtz FMMs are often presented for the low-frequency case [@wang2021exafmm,@Malhotra2015], due to the challenging data sizes involved. We present an extension of the kiFMM to the Helmholtz problem which has proven so far to be effective in single precision for relatively high wave numbers of up to $k \sim 100$.
+- _A moderate frequency Helmholtz FMM_. Helmholtz FMMs are often presented for the low-frequency case [@wang2021exafmm,@Malhotra2015], due to the challenging data sizes involved. We present an extension of the kiFMM to the Helmholtz problem which has proven so far to be effective in single precision for relatively high wave numbers of up to $kD \sim 100$, where $k$ is the wave number and $D$ is the diameter of the problem domain.
 
 `kifmm-rs` is currently a core library used as a part of the Bempp boundary element project [@bempp_rs]. A full API description is available as a part of published [documentation](https://bempp.github.io/kifmm/kifmm/index.html). [Python](https://github.com/bempp/kifmm/tree/main/kifmm/python/examples), [Rust](https://github.com/bempp/kifmm/tree/main/kifmm/examples) and [C](https://github.com/bempp/kifmm/tree/main/kifmm/c) examples can be found in the repository.
 
