+++ paper/paper.bib
@@ -7,7 +7,8 @@
   number={2},
   pages={591--626},
   year={2004},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2003.11.021}
 }
 
 
@@ -35,7 +36,8 @@ year = {2009}
   number={3},
   pages={808--830},
   year={2015},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.4208/cicp.020215.150515sw}
 }
 
 
@@ -46,7 +48,8 @@ year = {2009}
   volume={6},
   number={61},
   pages={3145},
-  year={2021}
+  year={2021},
+  doi={10.21105/joss.03145}
 }
 
 @article{fong2009black,
@@ -57,7 +60,8 @@ year = {2009}
   number={23},
   pages={8712--8725},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2009.08.031}
 }
 
 @article{bramas2020tbfmm,
@@ -67,7 +71,8 @@ year = {2009}
   volume={5},
   number={56},
   pages={2444},
-  year={2020}
+  year={2020},
+  doi={10.21105/joss.02444}
 }
 
 @inproceedings{blanchard2015scalfmm,
@@ -103,7 +108,8 @@ year = {2016}
   number={3},
   pages={1613--1619},
   year={2008},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2007.06.029}
 }
 
 @article{fu1998fast,
@@ -114,7 +120,8 @@ year = {2016}
   number={7},
   pages={1215--1229},
   year={1998},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/(SICI)1097-0207(19980815)42:7%3C1215::AID-NME406%3E3.0.CO;2-5}
 }
 
 @article{rokhlin1993diagonal,
@@ -125,7 +132,8 @@ year = {2016}
   number={1},
   pages={82--93},
   year={1993},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1006/acha.1993.1006}
 }
 
 @article{Greengard1987,
@@ -136,33 +144,18 @@ year = {2016}
   number={2},
   pages={325--348},
   year={1987},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/0021-9991(87)90140-9}
 }
 
 @article{Kailasa2024,
   title={M2L Translation Operators for Kernel Independent Fast Multipole Methods on Modern Architectures},
   author={Kailasa, Srinath and Betcke, Timo and Kazdadi, Sarah El},
   journal={arXiv preprint arXiv:2408.07436},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2408.07436}
 }
 
-@article{engquist2010fast,
-  title={Fast directional algorithms for the Helmholtz kernel},
-  author={Engquist, Bj{\"o}rn and Ying, Lexing},
-  journal={Journal of Computational and Applied Mathematics},
-  volume={234},
-  number={6},
-  pages={1851--1859},
-  year={2010},
-  publisher={Elsevier}
-}
-
-@book{steinbach2007numerical,
-  title={Numerical approximation methods for elliptic boundary value problems: finite and boundary elements},
-  author={Steinbach, Olaf},
-  year={2007},
-  publisher={Springer Science \& Business Media}
-}
 
 @misc{bempp_rs,
   author       = {Betcke, T. and Scroggs, M.},
