+++ kifmm/Cargo.toml
@@ -5,7 +5,7 @@ mpi = ["dep:mpi", "dep:superslice", "dep:memoffset", "dep:mpi-sys", "rlst/mpi"]
 
 [package]
 name = "kifmm"
-version = "2.0.1"
+version = "2.1.0-dev"
 edition = "2021"
 authors = ["Srinath Kailasa <srinathkailasa@gmail.com>"]
 description = "Kernel-independent fast multipole method"
