+++ kifmm/pyproject.toml
@@ -10,7 +10,7 @@ build-backend = "maturin"
 
 [project]
 name = "kifmm_py"
-version="2.0.1"
+version="2.1.0-dev"
 requires-python = "==3.10.*"
 classifiers = [
     "Programming Language :: Rust",
