+++ fftw-sys/Cargo.toml
@@ -11,7 +11,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-kifmm-fftw-src = { path = "../fftw-src" } 
+kifmm-fftw-src = { path = "../fftw-src" }
 libc = "0.2.153"
 num-complex = "0.4.5"
 
