### 2025-06-19 #7124 @skailasa
Hi @diehlpk, I've archived a release on Zenodo (DOI: 10.5281/zenodo.15696916). For the references, I've added as many DOIs as I could find, the software referenced often has not been given a DOI.
