### 2025-06-19 #7124 @diehlpk
@skailasa please check the invalid DOI.
