### 2025-06-20 #7124 @diehlpk
@openjournals/dev could you please have a look at the invalid DOI?
