### 2025-06-20 #7124 @diehlpk
> > [@skailasa](https://github.com/skailasa) please check the invalid DOI.
>
> Hi, that's the DOI I found using crossref (10.1002/(SICI)1097-0207(19980815)42:7%3C1215::AID-NME406%3E3.0.CO;2-5) Though the percentage signs seem to have been removed by editorialbot.

@skailasa I pinged the developers to have a look into the issue with the DOI.
