### 2025-06-23 52c4fba
ENH: Distributed Helmholtz Variant (#141)
- Distributed Helmholtz support
- Variable expansion order support for Distributed Laplace and Helmholtz
- Adds proper configuration for single node (laplace) benchmarks
- Removes redundant trait constraints on distributed trait impls
- Updates and adds to distributed FMM tests for both Laplace and Helmholtz


Closes #139 Closes #140
