+++ kifmm/src/bindings.rs
@@ -13,7 +13,7 @@ use crate::{
 pub mod types {
     use std::ffi::c_void;
 
-    use crate::traits::types::{CommunicationTime, FmmOperatorTime, MetadataTime};
+    use crate::traits::types::{CommunicationType, FmmOperatorType, MetadataType, OperatorTime};
 
     /// Static FMM type
     #[repr(C)]
@@ -136,20 +136,41 @@ pub mod types {
     }
 
     #[repr(C)]
+    #[derive(Debug, Clone, Copy)]
+    pub struct FmmOperatorEntry {
+        pub op_type: FmmOperatorType,
+        pub time: OperatorTime,
+    }
+
+    #[repr(C)]
     pub struct FmmOperatorTimes {
-        pub times: *mut FmmOperatorTime,
+        pub times: *mut FmmOperatorEntry,
         pub length: usize,
+    }
+
+    #[repr(C)]
+    #[derive(Debug, Clone, Copy)]
+    pub struct MetadataEntry {
+        pub metadata_type: MetadataType,
+        pub time: OperatorTime,
     }
 
     #[repr(C)]
     pub struct MetadataTimes {
-        pub times: *mut MetadataTime,
+        pub times: *mut MetadataEntry,
         pub length: usize,
     }
 
     #[repr(C)]
+    #[derive(Debug, Clone, Copy)]
+    pub struct CommunicationEntry {
+        pub comm_type: CommunicationType,
+        pub time: OperatorTime,
+    }
+
+    #[repr(C)]
     pub struct CommunicationTimes {
-        pub times: *mut CommunicationTime,
+        pub times: *mut CommunicationEntry,
         pub length: usize,
     }
 }
@@ -1560,7 +1581,7 @@ pub mod api {
     use itertools::Itertools;
 
     use crate::{
-        bindings::MetadataTimes,
+        bindings::{CommunicationEntry, FmmOperatorEntry, MetadataEntry, MetadataTimes},
         fmm::types::FmmEvalType,
         traits::{
             fmm::{ChargeHandler, DataAccess},
@@ -1602,24 +1623,46 @@ pub mod api {
                     let fmm =
                         pointer as *mut KiFmm<f32, Laplace3dKernel<f32>, FftFieldTranslation<f32>>;
 
-                    let length = unsafe { (*fmm).communication_times.len() };
-                    let times = unsafe { (*fmm).communication_times.as_mut_ptr() };
+                    let entries: Vec<CommunicationEntry> = unsafe {
+                        (*fmm)
+                            .communication_times
+                            .iter()
+                            .map(|(&comm_type, &time)| CommunicationEntry { comm_type, time })
+                            .collect()
+                    };
 
-                    let times = CommunicationTimes { times, length };
+                    // Leak the Vec to pass to C safely
+                    let length = entries.len();
+                    let mut boxed_entries = entries.into_boxed_slice();
+                    let ptr = boxed_entries.as_mut_ptr();
+                    std::mem::forget(boxed_entries); // Prevent Rust from freeing it
 
-                    Box::into_raw(Box::new(times))
+                    let boxed = Box::new(CommunicationTimes { times: ptr, length });
+
+                    Box::into_raw(boxed)
                 }
 
                 FmmTranslationCType::Blas => {
                     let fmm = pointer
                         as *mut KiFmm<f32, Laplace3dKernel<f32>, BlasFieldTranslationSaRcmp<f32>>;
 
-                    let length = unsafe { (*fmm).communication_times.len() };
-                    let times = unsafe { (*fmm).communication_times.as_mut_ptr() };
+                    let entries: Vec<CommunicationEntry> = unsafe {
+                        (*fmm)
+                            .communication_times
+                            .iter()
[…]
