+++ scripts/src/bin/fmm_m2l_fft_mpi_f32.rs
@@ -51,205 +51,205 @@ struct Args {
 }
 
 fn main() {
-    let (universe, _threading) =
-        mpi::initialize_with_threading(mpi::Threading::Serialized).unwrap();
-    let world = universe.world();
-    let comm = world.duplicate();
+    // let (universe, _threading) =
+    //     mpi::initialize_with_threading(mpi::Threading::Serialized).unwrap();
+    // let world = universe.world();
+    // let comm = world.duplicate();
 
-    // Tree parameters
-    let args = Args::parse();
-    let expansion_order = args.expansion_order;
-    let prune_empty = args.prune_empty;
-    let n_points = args.n_points;
-    let local_depth = args.local_depth;
-    let global_depth = args.global_depth;
-    let block_size = args.block_size;
-    let n_threads = args.n_threads;
-    let n_samples = args.n_samples;
-    let id = args.id;
+    // // Tree parameters
+    // let args = Args::parse();
+    // let expansion_order = args.expansion_order;
+    // let prune_empty = args.prune_empty;
+    // let n_points = args.n_points;
+    // let local_depth = args.local_depth;
+    // let global_depth = args.global_depth;
+    // let block_size = args.block_size;
+    // let n_threads = args.n_threads;
+    // let n_samples = args.n_samples;
+    // let id = args.id;
 
-    assert!(n_samples > 0 && n_samples < n_points);
+    // assert!(n_samples > 0 && n_samples < n_points);
 
-    let sort_kind = SortKind::Samplesort { n_samples };
+    // let sort_kind = SortKind::Samplesort { n_samples };
 
-    // Fmm Parameters
-    let kernel = Laplace3dKernel::<f32>::new();
+    // // Fmm Parameters
+    // let kernel = Laplace3dKernel::<f32>::new();
 
-    ThreadPoolBuilder::new()
-        .num_threads(n_threads)
-        .build_global()
-        .unwrap();
+    // ThreadPoolBuilder::new()
+    //     .num_threads(n_threads)
+    //     .build_global()
+    //     .unwrap();
 
-    let source_to_target = FftFieldTranslation::<f32>::new(Some(block_size));
+    // let source_to_target = FftFieldTranslation::<f32>::new(Some(block_size));
 
-    // Generate some random test data local to each process
-    let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
-    let charges = vec![1f32; n_points];
+    // // Generate some random test data local to each process
+    // let points = points_fixture::<f32>(n_points, None, None, Some(world.rank() as u64));
+    // let charges = vec![1f32; n_points];
 
-    let mut multi_fmm = MultiNodeBuilder::new(true)
-        .tree(
-            &comm,
-            points.data(),
-            points.data(),
-            local_depth,
-            global_depth,
-            prune_empty,
-            sort_kind.clone(),
-        )
-        .unwrap()
-        .parameters(
-            &charges,
-            expansion_order,
-            kernel.clone(),
-            green_kernels::types::GreenKernelEvalType::Value,
-            source_to_target,
-        )
-        .unwrap()
-        .build()
-        .unwrap();
+    // let mut multi_fmm = MultiNodeBuilder::new(true)
+    //     .tree(
+    //         &comm,
+    //         points.data(),
+    //         points.data(),
+    //         local_depth,
+    //         global_depth,
+    //         prune_empty,
+    //         sort_kind.clone(),
+    //     )
+    //     .unwrap()
+    //     .parameters(
+    //         &charges,
+    //         expansion_order,
+    //         kernel.clone(),
+    //         green_kernels::types::GreenKernelEvalType::Value,
+    //         source_to_target,
[…]
