+++ kifmm/src/fmm/data_access/mod.rs
@@ -1,3 +1,4 @@
+//! Functionality for accessing data from within built FMM objects
 #[cfg(feature = "mpi")]
 mod multi_node;
 mod single_node;
