### 2025-06-23 #7124 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
