### 2025-06-23 c4a31d1
small changes in paper.md
