+++ paper/paper.md
@@ -43,7 +43,7 @@ where $\symbfup{\phi}$ is an $M$ dimensional vector, $\mathbf{K}$ is a dense $M 
 å    \label{eq:sec:summary:laplace_kernel}
 \end{equation}
 
-Since their initial introduction FMMs have been developed for a wide variety of kernel functions such as the Helmholtz kernel which arises from the time-independent wave equation [@rokhlin1993diagonal],
+Since their initial introduction, FMMs have been developed for a wide variety of kernel functions such as the Helmholtz kernel that arises from the time-independent wave equation [@rokhlin1993diagonal],
 
 \begin{equation}
     K(x, y) = \begin{cases}
@@ -53,22 +53,22 @@ Since their initial introduction FMMs have been developed for a wide variety of 
   \label{eq:sec:summary:helmholtz_kernel}
 \end{equation}
 
-As well as the Stokes kernel describing fluid flow [@tornberg2008fast], and the Navier kernel used in elastostatics [@fu1998fast] among others. The major application of FMMs is found in the accelerated application of dense operator matrices that arise from the integral formulation of partial differential equations, and as a result FMM algorithms are a crucial component across simulations in many domains, from medical imaging to geophysics.
+as well as the Stokes kernel describing fluid flow [@tornberg2008fast], and the Navier kernel used in elastostatics [@fu1998fast], among others. The major application of FMMs is found in the accelerated application of dense operator matrices that arise from the integral formulation of partial differential equations, and as a result FMM algorithms are a crucial component across simulations in many domains, from medical imaging to geophysics.
 
-FMMs accelerate the evaluation of \eqref{eq:sec:summary:matvec} by decomposing the problem domain using a hierarchical tree, a quadtree in 2D and an octree in 3D, the algorithm consisting of a series of operations collectively referred to as _field translations_. Evaluations of the off-diagonal blocks of the matrix $\mathbf{K}$ in \eqref{eq:sec:summary:matvec} correspond to the `multipole to local' or M2L field translation which summarise the long-range interactions of a local cluster of target points, and the evaluation of the diagonal blocks correspond to unavoidable direct evaluations of the kernel function for evaluating near-range interactions. M2L is the most challenging optimisation bottleneck in FMMs due to its memory-bound nature.
+FMMs accelerate the evaluation of \eqref{eq:sec:summary:matvec} by decomposing the problem domain using a hierarchical tree, a quadtree in 2D, and an octree in 3D, the algorithm consisting of a series of operations collectively referred to as _field translations_. Evaluations of the off-diagonal blocks of the matrix $\mathbf{K}$ in \eqref{eq:sec:summary:matvec} correspond to the `multipole to local' or M2L field translation that summarise the long-range interactions of a local cluster of target points, and the evaluation of the diagonal blocks correspond to unavoidable direct evaluations of the kernel function for evaluating near-range interactions. M2L is the most challenging optimisation bottleneck in FMMs due to its memory-bound nature.
 
-Many parts of an implementation are common across FMM algorithms, such as the tree setup and the kernel function implementation. As a result, our framework is presented as a set of modular, re-usable, sub-components of each of the key algorithmic sub-components: (i) the tree in shared and distributed memory, (ii) the field translation operations and (iii) the kernel evaluation. We use our framework to develop an implementation of the so-called kernel independent Fast Multiple Method (kiFMM) [@Ying2004], compatible with a wide variety of elliptic PDE kernels with an implementation for the Laplace and Helmholtz kernels provided currently. The kiFMM has favourable implementation properties due to its formulation in terms of high operational-intensity BLAS operations.
[…]
