### 2025-06-23 #7124 @danielskatz
👋 @skailasa - I've suggested a number of small changes in https://github.com/bempp/kifmm/pull/166 - please merge this, or let me know what you disagree with, then we can continue to acceptance and publication
