### 2025-06-24 86cf610
PAPER: minor changes for JOSS publication (#166)
* small changes in paper.md

* minor changes in paper.bib
