+++ paper/paper.bib
@@ -42,7 +42,7 @@ year = {2009}
 
 
 @article{wang2021exafmm,
-  title={ExaFMM: a high-performance fast multipole method library with C++ and Python interfaces},
+  title={ExaFMM: a high-performance fast multipole method library with {C}++ and {P}ython interfaces},
   author={Wang, Tingyu and Yokota, Rio and Barba, Lorena A},
   journal={Journal of Open Source Software},
   volume={6},
@@ -65,7 +65,7 @@ year = {2009}
 }
 
 @article{bramas2020tbfmm,
-  title={TBFMM: A C++ generic and parallel fast multipole method library},
+  title={TBFMM: A {C}++ generic and parallel fast multipole method library},
   author={Bramas, B{\'e}renger},
   journal={Journal of Open Source Software},
   volume={5},
@@ -101,7 +101,7 @@ year = {2016}
 }
 
 @article{tornberg2008fast,
-  title={A fast multipole method for the three-dimensional Stokes equations},
+  title={A fast multipole method for the three-dimensional {S}tokes equations},
   author={Tornberg, Anna-Karin and Greengard, Leslie},
   journal={Journal of Computational Physics},
   volume={227},
@@ -125,7 +125,7 @@ year = {2016}
 }
 
 @article{rokhlin1993diagonal,
-  title={Diagonal forms of translation operators for the Helmholtz equation in three dimensions},
+  title={Diagonal forms of translation operators for the {H}elmholtz equation in three dimensions},
   author={Rokhlin, Vladimir},
   journal={Applied and computational harmonic analysis},
   volume={1},
@@ -159,7 +159,7 @@ year = {2016}
 
 @misc{bempp_rs,
   author       = {Betcke, T. and Scroggs, M.},
-  title        = {Bempp-rs: Boundary element methods in Rust},
+  title        = {Bempp-rs: Boundary element methods in {R}ust},
   year         = 2024,
   url          = {https://github.com/bempp/bempp-rs},
   note         = {Accessed: 2024-06-18},
