### 2025-06-24 #7124 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
